% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{addLineBreakToCaption}
\alias{addLineBreakToCaption}
\title{addLineBreakToCaption}
\usage{
addLineBreakToCaption(
  captions,
  maxLines = reEnv$maxLinesPerLegendCaption,
  width = reEnv$maxWidthPerLegendCaption
)
}
\arguments{
\item{captions}{Array of character strings to render}

\item{maxLines}{Maximum number of lines directly setting the maximum number of line breaks allowed.}

\item{width}{Maximum number of characters per line desired.
Due to `maxLines`, the returned width can be wider than `width`.}
}
\value{
A character vector of wrapped strings with line breaks at sensible places.
}
\description{
Add line breaks to get prettier captions
}
\examples{
# Use cat to display result of line break character
cat(addLineBreakToCaption("this-is-a-long-sentence-with-dashes", maxLines = 2, width = 25))

cat(addLineBreakToCaption("this is a sentence with spaces", maxLines = 2, width = 25))

cat(addLineBreakToCaption("this_is_a_long_sentence_without_preferential_splits", maxLines = 2, width = 25))

cat(addLineBreakToCaption("this too short to split", maxLines = 3, width = 40))

cat(addLineBreakToCaption("this forces the sentence to use one line", maxLines = 1, width = 5))

}
