% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addMarkdownToc}
\alias{addMarkdownToc}
\title{addMarkdownToc}
\usage{
addMarkdownToc(fileName, tocTitle = "# Table of Contents")
}
\arguments{
\item{fileName}{name of .md file to update}

\item{tocTitle}{Markdown text for table of content title}
}
\description{
Add table of content to a markdown file
}
\keyword{internal}
