% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-comparison-time-profile.R
\name{addOutputToComparisonTimeProfile}
\alias{addOutputToComparisonTimeProfile}
\title{addOutputToComparisonTimeProfile}
\usage{
addOutputToComparisonTimeProfile(
  outputMapping,
  simulationDuration,
  axesProperties,
  plotObject,
  configurationPlan,
  logFolder
)
}
\arguments{
\item{outputMapping}{list of mapping elements from `OutputMappings` field in configuration plan}

\item{simulationDuration}{Duration of simulation in X axis unit}

\item{axesProperties}{list of axes properties obtained from `getAxesProperties`}

\item{plotObject}{ggplot object}

\item{configurationPlan}{A `ConfigurationPlan` object}

\item{logFolder}{folder where the logs are saved}
}
\value{
A ggplot object
}
\description{
Add plot layers for an output mapping from comparison time profile plot
}
\keyword{internal}
