% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addRmdFigureChunk}
\alias{addRmdFigureChunk}
\title{addRmdFigureChunk}
\usage{
addRmdFigureChunk(fileName, figureFile, figureCaption = "")
}
\arguments{
\item{fileName}{name of Rmd file}

\item{figureFile}{figure to include}

\item{figureCaption}{caption of figure}
}
\value{
R markdown text with figure and save the Rmd document
}
\description{
Add a figure in a Rmd document
}
