% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addRmdTableChunk}
\alias{addRmdTableChunk}
\title{addRmdTableChunk}
\usage{
addRmdTableChunk(fileName, tableFile, tableCaption = "")
}
\arguments{
\item{fileName}{name of Rmd file}

\item{tableFile}{table to include as a csv file}

\item{tableCaption}{caption of table}
}
\value{
R markdown text with figure and save the Rmd document
}
\description{
Add a table in a Rmd document
}
