% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addRmdTextChunk}
\alias{addRmdTextChunk}
\title{addRmdTextChunk}
\usage{
addRmdTextChunk(fileName, text)
}
\arguments{
\item{fileName}{name of Rmd file}

\item{text}{text to include in the document}
}
\description{
Add markdown text to a Rmd document
}
