% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addTableAndFigureNumbersToMarkdown}
\alias{addTableAndFigureNumbersToMarkdown}
\title{addTableAndFigureNumbersToMarkdown}
\usage{
addTableAndFigureNumbersToMarkdown(fileName)
}
\arguments{
\item{fileName}{name of .md file to update}
}
\description{
Reference tables and figures in a report
}
\keyword{internal}
