% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addTextChunk}
\alias{addTextChunk}
\title{addTextChunk}
\usage{
addTextChunk(fileName, text)
}
\arguments{
\item{fileName}{name of .md file}

\item{text}{text to include in the document}
}
\description{
Add a text chunk to a .md document
}
\examples{
\dontrun{
resetReport("report.md")
addTextChunk(fileName = "report.md", text = "new text")
}

}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{getFigurePathsFromReport}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
