% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{addXParametersForDemogrpahyPlot}
\alias{addXParametersForDemogrpahyPlot}
\title{addXParametersForDemogrpahyPlot}
\usage{
addXParametersForDemogrpahyPlot(workflow, parameters)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}

\item{parameters}{list of demography parameters to be used as x-parameters}
}
\description{
Append x parameters for range plots of demography plot task.
The method update directly the input workflow
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- PopulationWorkflow$new(worflowType, workflowFolder, simulationSets)

# Get the list of parameters in x-axis for range plots
addXParametersForDemogrpahyPlot(
  workflow = myWorkflow,
  parameters = StandardPath$GestationalAge
)
}

}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemogrpahyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemogrpahyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemogrpahyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setXParametersForDemogrpahyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemogrpahyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
