% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{addXParametersForPkParametersPlot}
\alias{addXParametersForPkParametersPlot}
\title{addXParametersForPkParametersPlot}
\usage{
addXParametersForPkParametersPlot(workflow, parameters)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}

\item{parameters}{list of parameters to be used as x-parameters}
}
\description{
Apppend x-parameters for range plots of PK parameters plot task.
The method update directly the input workflow
}
