% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{addYParametersForDemogrpahyPlot}
\alias{addYParametersForDemogrpahyPlot}
\title{addYParametersForDemogrpahyPlot}
\usage{
addYParametersForDemogrpahyPlot(workflow, parameters)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}

\item{parameters}{list of demography parameters to be used as x-parameters}
}
\description{
Append y parameters for range plots of demography plot task.
The method update directly the input workflow
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- PopulationWorkflow$new(worflowType, workflowFolder, simulationSets)

# Add parameters in y-axis for range plots and histograms
addYParametersForDemogrpahyPlot(
workflow = myWorkflow, 
parameters = StandardPath$GestationalAge
)

}

}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemogrpahyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getXParametersForDemogrpahyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemogrpahyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setXParametersForDemogrpahyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemogrpahyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
