% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{addYParametersForPkParametersPlot}
\alias{addYParametersForPkParametersPlot}
\title{addYParametersForPkParametersPlot}
\usage{
addYParametersForPkParametersPlot(workflow, parameters)
}
\description{
addYParametersForPkParametersPlot
}
