% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{adjustTitlePage}
\alias{adjustTitlePage}
\title{adjustTitlePage}
\usage{
adjustTitlePage(fileName, qualificationVersionInfo = NULL)
}
\arguments{
\item{fileName}{name of .md file to update}

\item{qualificationVersionInfo}{A `QualificationVersionInfo`object defining Qualification Version Information to be displayed on title page}
}
\description{
Adust Qualification Version Information to be displayed on title page
}
