% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\docType{data}
\name{AggregationConfiguration}
\alias{AggregationConfiguration}
\title{AggregationConfiguration}
\format{An object of class \code{list} of length 2.}
\usage{
AggregationConfiguration
}
\description{
Aggregation default properties  (which functions and their captions)
}
\keyword{datasets}
