% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{alignXTicks}
\alias{alignXTicks}
\title{alignXTicks}
\usage{
alignXTicks(plotObject)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}
}
\value{
A \code{ggplot} object
}
\description{
Use the \code{plotConfiguration} of a \code{plotObject}
to check and perform vertical alignment of x-axis tick labels
}
\keyword{internal}
