% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\docType{data}
\name{AllAvailableTasks}
\alias{AllAvailableTasks}
\title{AllAvailableTasks}
\format{
An object of class \code{list} of length 9.
}
\usage{
AllAvailableTasks
}
\description{
Names of all existing tasks that can be performed by `MeanModelWorkflow` or `PopulationWorkflow` objects
}
\examples{
AllAvailableTasks$simulate
AllAvailableTasks$plotTimeProfilesAndResiduals
AllAvailableTasks$plotDemography
AllAvailableTasks$plotAbsorption
AllAvailableTasks$plotMassBalance
}
\keyword{datasets}
