% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\docType{data}
\name{AllPKParameterNames}
\alias{AllPKParameterNames}
\title{AllPKParameterNames}
\format{An object of class \code{character} of length 13.}
\usage{
AllPKParameterNames
}
\description{
array of standard PK parameters from OSP Suite
}
\keyword{datasets}
