% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\docType{data}
\name{AllPKParameters}
\alias{AllPKParameters}
\title{AllPKParameters}
\format{An object of class \code{list} of length 13.}
\usage{
AllPKParameters
}
\description{
list of usual PK parameters from OSP Suite
}
\keyword{datasets}
