% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{analyzeCoreSensitivity}
\alias{analyzeCoreSensitivity}
\title{analyzeCoreSensitivity}
\usage{
analyzeCoreSensitivity(
  simulation,
  variableParameterPaths = NULL,
  variationRange = 0.1,
  numberOfCores = NULL,
  debugLogFileName = file.path(getwd(), defaultFileNames$logDebugFile()),
  infoLogFileName = file.path(getwd(), defaultFileNames$logInfoFile()),
  errorLogFileName = file.path(getwd(), defaultFileNames$logErrorFile()),
  nodeName = NULL,
  showProgress = FALSE
)
}
\arguments{
\item{simulation}{simulation class object}

\item{variableParameterPaths}{paths of parameters to be analyzed}

\item{variationRange}{variation range for sensitivity analysis}

\item{numberOfCores}{Number of cores to use on local node.  This parameter
should be should be set to 1 when parallelizing over many nodes.}

\item{debugLogFileName}{path to debug log file}

\item{infoLogFileName}{path to info log file}

\item{errorLogFileName}{path to error log file}

\item{nodeName}{identifier for node used in parallel computation of sensitivity}
}
\value{
sensitivity analysis results
}
\description{
Run a sensitivity analysis for a single individual,
varying only the set of parameters variableParameterPaths
}
