% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{anchor}
\alias{anchor}
\title{anchor}
\usage{
anchor(name)
}
\arguments{
\item{name}{Name/identifier of the anchor tag}
}
\value{
A character string
}
\description{
Create an anchor tag for markdown document
}
\examples{
anchor("section-1")

}
