% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{asisCodeIds}
\alias{asisCodeIds}
\title{asisCodeIds}
\format{
An object of class \code{character} of length 18.
}
\usage{
asisCodeIds
}
\description{
Some of the identifiers above need to be treated
as is by [getIdentifierInfo()]
}
\keyword{datasets}
