% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{asTimeAfterDose}
\alias{asTimeAfterDose}
\title{asTimeAfterDose}
\usage{
asTimeAfterDose(data, timeOffset, maxTime = NULL)
}
\arguments{
\item{data}{A data.frame that includes the variable \code{Time}}

\item{timeOffset}{Value shifting time}

\item{maxTime}{Optional maximum time value before shift}
}
\value{
A data.frame
}
\description{
Transform time data as time after dose data
}
\keyword{internal}
