% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{autoAxesTicksFromLimits}
\alias{autoAxesTicksFromLimits}
\title{autoAxesTicksFromLimits}
\usage{
autoAxesTicksFromLimits(limits)
}
\arguments{
\item{limits}{Min and max values of axis range}
}
\value{
Numeric values of ticks to display
}
\description{
Defines auto axis ticks from limits for log scale plots.
For wide range, log tick labels display every factor of 10
For small range, log tick labels displayed at 1,2,5 of every factor of 10
}
\keyword{internal}
