% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{buildDDIDataFrameFunctions}
\alias{buildDDIDataFrameFunctions}
\title{Names of functions for extracting data for each DDI plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
buildDDIDataFrameFunctions
}
\description{
Names of functions for extracting data for each DDI plot type
}
\keyword{internal}
