% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\docType{data}
\name{buildGOFDataFrameFunctions}
\alias{buildGOFDataFrameFunctions}
\title{Names of functions for extracting data for each GOF plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
buildGOFDataFrameFunctions
}
\description{
Names of functions for extracting data for each GOF plot type
}
\keyword{datasets}
