% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{buildQualificationDDIDataframe}
\alias{buildQualificationDDIDataframe}
\title{buildQualificationDDIDataframe}
\usage{
buildQualificationDDIDataframe(dataframe, metadata, pkParameter, plotType)
}
\arguments{
\item{dataframe}{data.frame}

\item{metadata}{meta data on `data`}

\item{pkParameter}{for which DDI ratios are to be evaluated}

\item{plotType}{for which DDI ratios are to be evaluated.  `plotType` is either `predictedVsObserved` or `residualsVsObserved`.}
}
\value{
dataframe for plotting goodness of fit of residuals vs time type
}
\description{
Build dataframe for DDI
}
\keyword{internal}
