% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{buildQualificationDDIPredictedVsObserved}
\alias{buildQualificationDDIPredictedVsObserved}
\title{buildQualificationDDIPredictedVsObserved}
\usage{
buildQualificationDDIPredictedVsObserved(dataframe, metadata)
}
\arguments{
\item{dataframe}{data.frame}

\item{metadata}{meta data on `data`}
}
\value{
dataframe for plotting goodness of fit of predictedVsObserved type
}
\description{
Build dataframe for observation vs prediction
}
