% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{buildQualificationDDIResidualsVsObserved}
\alias{buildQualificationDDIResidualsVsObserved}
\title{buildQualificationDDIResidualsVsObserved}
\usage{
buildQualificationDDIResidualsVsObserved(dataframe, metadata)
}
\arguments{
\item{dataframe}{data.frame}

\item{metadata}{meta data on `data`}
}
\value{
dataframe for plotting goodness of fit of residuals vs time type
}
\description{
Build dataframe for residuals vs observed
}
\keyword{internal}
