% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateArithmeticErrorRange}
\alias{calculateArithmeticErrorRange}
\title{calculateArithmeticErrorRange}
\usage{
calculateArithmeticErrorRange(values, errorValues)
}
\arguments{
\item{values}{Numeric values of the arithmetic mean}

\item{errorValues}{Numeric values of the arithmetic error}
}
\value{
A named list, with \code{ymin} and \code{ymax}, of the range calculated from the arithmetic mean and errors.
}
\description{
Calculate the range from the arithmetic mean and error.
}
