% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateGeometricErrorRange}
\alias{calculateGeometricErrorRange}
\title{calculateGeometricErrorRange}
\usage{
calculateGeometricErrorRange(values, errorValues)
}
\arguments{
\item{values}{Numeric values of the geometric mean}

\item{errorValues}{Numeric values of the geometric error}
}
\value{
A named list, with \code{ymin} and \code{ymax}, of the range calculated from the geometric mean and errors.
}
\description{
Calculate the range from the geometric mean and error.
}
