% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateGMFE}
\alias{calculateGMFE}
\title{calculateGMFE}
\usage{
calculateGMFE(x, y)
}
\arguments{
\item{x}{x values to compare}

\item{y}{y values to compare}
}
\value{
GMFE
}
\description{
Calculate Geometric Mean Fold Error between \code{x} and \code{y}.
Strictly positive pairs of values are kept in the calculation
}
\examples{
# GMFE
calculateGMFE(c(1,2, 2.1, 3.3), c(1.3, 1.9, 3.0))
}
