% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{calculatePKAnalysesRatio}
\alias{calculatePKAnalysesRatio}
\title{calculatePKAnalysesRatio}
\usage{
calculatePKAnalysesRatio(structureSets, settings)
}
\description{
Calculate and save statistics of ratios of PK parameters for each simulation set but reference
If simulation sets use the same population, calculate individual ratios and
get summary statistics from these individual ratios.
If simulation sets use different populations, perform Monte Carlo Sampling,
calculate individual ratios for each repetition of the Monte Carlo Sampling,
get summary statistics from these individual ratios for each repetition, and
get the median of the summary statistics as best approximation of the ratio summary statistics.
}
\keyword{internal}
