% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{calculatePKParameters}
\alias{calculatePKParameters}
\title{calculatePKParameters}
\usage{
calculatePKParameters(structureSet, settings = NULL)
}
\arguments{
\item{structureSet}{\code{SimulationStructure} R6 class object contain paths of files to be used}

\item{settings}{list of options to be passed on the function}
}
\value{
pkAnalysis object
}
\description{
Calculate PK parameters from simulated time profiles
}
\keyword{internal}
