% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-pk-parameters-task.R
\name{CalculatePKParametersTask}
\alias{CalculatePKParametersTask}
\title{CalculatePKParametersTask}
\description{
R6 class for CalculatePKParametersTask settings
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Tasks]{ospsuite.reportingengine::Tasks}} -> \code{CalculatePKParametersTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulationFilePath}}{simulation folder}

\item{\code{simulationResultFilePaths}}{simulation files}

\item{\code{pkParametersToEvaluate}}{list of PK parameters to evaluate}

\item{\code{userDefinedPKFunctions}}{list of user defined functions to calculate PK parameters}

\item{\code{pkParameterResultsFilePath}}{files where PK parameters are saved}

\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CalculatePKParametersTask$new()}}
\item \href{#method-clone}{\code{CalculatePKParametersTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="activate">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-activate}{\code{ospsuite.reportingengine::Tasks$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-inactivate}{\code{ospsuite.reportingengine::Tasks$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="print">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-print}{\code{ospsuite.reportingengine::Tasks$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-validateInput}{\code{ospsuite.reportingengine::Tasks$validateInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `CalculatePKParametersTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters inherited from R6 class `Task` object}

\item{\code{simulationFilePath}}{simulation folder}

\item{\code{simulationResultFilePaths}}{simulation files}

\item{\code{pkParametersToEvaluate}}{list of PK parameters to evaluate}

\item{\code{userDefinedPKFunctions}}{list of user defined functions to calculate PK parameters}

\item{\code{pkParameterResultsFilePath}}{files where PK parameters are saved}

\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `CalculatePKParametersTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
