% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{checkAllCoresSuccessful}
\alias{checkAllCoresSuccessful}
\title{checkAllCoresSuccessful}
\usage{
checkAllCoresSuccessful(coreResults)
}
\arguments{
\item{coreResults}{list of results returned by each core after an mpi.remote.exec command is complete}
}
\description{
Check that all cores executed an mpi.remote.exec command successfully.
}
\keyword{internal}
