% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{checkIsSamePopulation}
\alias{checkIsSamePopulation}
\title{checkIsSamePopulation}
\usage{
checkIsSamePopulation(simulationSet, referenceSet)
}
\arguments{
\item{simulationSet}{A `PopulationSimulationSet` object}

\item{referenceSet}{A `PopulationSimulationSet` object}
}
\value{
A logical
}
\description{
Check if 2 simulation sets use the same population.
Same population is identified as
1- same population file and 2- same study design file (if defined)
}
\keyword{internal}
