% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{checkLibraryLoaded}
\alias{checkLibraryLoaded}
\title{checkLibraryLoaded}
\usage{
checkLibraryLoaded(libraryName)
}
\arguments{
\item{libraryName}{the library to be checked if loaded successfully}
}
\description{
Check that a library has been successfully loaded on the core.
}
