% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{checkMetaDataIsConsistent}
\alias{checkMetaDataIsConsistent}
\title{checkMetaDataIsConsistent}
\usage{
checkMetaDataIsConsistent(metaData)
}
\arguments{
\item{metaData}{A data.frame summarizing meta data of multiple \code{SimulationSet} and \code{Output} objects}
}
\description{
Check consistency of metadata units, dimensions and residuals scale
}
\keyword{internal}
