% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{checkNotNull}
\alias{checkNotNull}
\title{checkNotNull}
\usage{
checkNotNull(object)
}
\arguments{
\item{object}{to be checked if not null}
}
\description{
Check that an object is not null.
}
