% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{checkPKParameterEndTime}
\alias{checkPKParameterEndTime}
\title{checkPKParameterEndTime}
\usage{
checkPKParameterEndTime(pkParameterName, endTime)
}
\arguments{
\item{pkParameterName}{PK Parameter name. See \code{ospsuite::allPKParameterNames()} for more details}

\item{endTime}{Qualification \code{EndTime} field}
}
\description{
Check consistency of PK Parameter end time
}
\keyword{internal}
