% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{checkPopulationLoaded}
\alias{checkPopulationLoaded}
\title{checkPopulationLoaded}
\usage{
checkPopulationLoaded(population)
}
\arguments{
\item{population}{the population object to be checked if loaded successfully}
}
\description{
Check that the population has been loaded on the core.
}
