% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{checkSamePopulationIds}
\alias{checkSamePopulationIds}
\title{checkSamePopulationIds}
\usage{
checkSamePopulationIds(setIds, referenceSetIds, setName, referenceSetName)
}
\arguments{
\item{setIds}{A vector of IndividualIds for a simulation set}

\item{referenceSetIds}{A vector of IndividualIds for the reference simulation set}

\item{setName}{Name of simulation set for warning message}

\item{referenceSetName}{Name of the reference simulation set for warning message}
}
\description{
Check if PKAnalses with same population actually use the same IndividualIds
}
\keyword{internal}
