% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{checkSimulationLoaded}
\alias{checkSimulationLoaded}
\title{checkSimulationLoaded}
\usage{
checkSimulationLoaded(simulation)
}
\arguments{
\item{simulation}{the simulation object to be checked if loaded successfully}

\item{simulationFilePath}{the path to the simulation object to be checked if loaded successfully}
}
\description{
Check that the simulation has been loaded on the core.
}
