% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{concatenateDataDisplayName}
\alias{concatenateDataDisplayName}
\title{concatenateDataDisplayName}
\usage{
concatenateDataDisplayName(inputs, sep = " - ")
}
\arguments{
\item{inputs}{Vector of inputs to concatenate}

\item{sep}{Separator for display names}
}
\value{
Character of concatenated inputs
}
\description{
Concatenate inputs for dataDisplayName
}
\keyword{internal}
