% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{concatenateDataSelection}
\alias{concatenateDataSelection}
\title{concatenateDataSelection}
\usage{
concatenateDataSelection(inputs, sep = ") & (")
}
\arguments{
\item{inputs}{Vector of inputs to concatenate}

\item{sep}{Separator character corresponding to logical &}
}
\value{
Character of concatenated inputs
}
\description{
Concatenate inputs for dataSelection
}
