% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationFontsFields}
\alias{ConfigurationFontsFields}
\title{ConfigurationFontsFields}
\format{
An object of class \code{character} of length 5.
}
\usage{
ConfigurationFontsFields
}
\description{
Enum defining possible fonts fields defined in configuration plan
Note that some fields might not be used/converted by the current `opsuite.reportingengine` or `tlf` version (e.g. FontFamilyName)
}
\keyword{datasets}
