% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationLinetypes}
\alias{ConfigurationLinetypes}
\title{ConfigurationLinetypes}
\format{
An object of class \code{list} of length 3.
}
\usage{
ConfigurationLinetypes
}
\description{
List defining linetype following tlf nomenclature from a LineStyle field of configuration plan
}
\keyword{internal}
