% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration-plan.R
\docType{class}
\name{ConfigurationPlan}
\alias{ConfigurationPlan}
\title{ConfigurationPlan}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConfigurationPlan
}
\description{
R6 class for ConfigurationPlan
}
\section{Fields}{

\describe{
\item{\code{reInputPath}}{Input path to run the Configuration Plan}

\item{\code{reOutputPath}}{Output path where Configuration Plan results will be stored}

\item{\code{simulationMapping}}{R6 class mapping simulations to needed for plots}

\item{\code{observedDataSets}}{R6 class storing the observed data}

\item{\code{plots}}{R6 class defining the plots to be performed in Configuration Plan}

\item{\code{inputs}}{Inputs to be moved as is in Configuration Plan output}

\item{\code{sections}}{Description of output structure tree as a data.frame}

\item{\code{intro}}{Introduction to be moved as is in Configuration Plan output}
}}

\keyword{datasets}
