% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationPlots}
\alias{ConfigurationPlots}
\title{ConfigurationPlots}
\format{
An object of class \code{list} of length 5.
}
\usage{
ConfigurationPlots
}
\description{
Enum defining possible plots defined in configuration plan
}
\keyword{datasets}
