% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationPlotSettings}
\alias{ConfigurationPlotSettings}
\title{ConfigurationPlotSettings}
\format{
An object of class \code{list} of length 3.
}
\usage{
ConfigurationPlotSettings
}
\description{
Enum defining possible plot settings fields defined in configuration plan
}
\keyword{datasets}
