% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationScales}
\alias{ConfigurationScales}
\title{ConfigurationScales}
\format{
An object of class \code{list} of length 2.
}
\usage{
ConfigurationScales
}
\description{
List defining scale following tlf nomenclature from a Scaling field of configuration plan
}
\keyword{internal}
