% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationShapes}
\alias{ConfigurationShapes}
\title{ConfigurationShapes}
\format{
An object of class \code{list} of length 14.
}
\usage{
ConfigurationShapes
}
\description{
List mapping configuration plan symbols to tlf shapes.
The current mapping leads to open shapes
}
\keyword{internal}
