% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{copyReport}
\alias{copyReport}
\title{copyReport}
\usage{
copyReport(from, to, copyWordReport = TRUE, keep = FALSE)
}
\arguments{
\item{from}{path of initial .md file to copy}

\item{to}{path of destination .md file to be copied}

\item{copyWordReport}{logical defining if .docx report is also copied}

\item{keep}{logical defining if initial .md file and figures are kept}
}
\description{
Copy markdown report and its figures (using their paths)
}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{getAnchorName}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
