% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createFolder}
\alias{createFolder}
\title{createFolder}
\usage{
createFolder(folderName, printConsole = FALSE)
}
\arguments{
\item{folderName}{Path of folder to be created}
}
\description{
Create a folder with path folderName and log in logDebug
}
