% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{createQualificationReport}
\alias{createQualificationReport}
\title{createQualificationReport}
\usage{
createQualificationReport(
  qualificationRunnerFolder,
  pkSimPortableFolder = NULL,
  workingDirectory = getwd(),
  createWordReport = TRUE,
  watermark = ""
)
}
\arguments{
\item{qualificationRunnerFolder}{Folder where QualificationRunner.exe is located}

\item{pkSimPortableFolder}{Folder where PK-Sim is located.
If not specified, installation path will be read from the registry (available only in case of full **non-portable** installation).
This option is **MANDATORY** for the portable version of PK-Sim.}

\item{workingDirectory}{Folder where the qualification inputs are located
The `workingDirectory` is assumed with the following structure
- `input`: input qualification plan directory
  - `qualification_plan.json`: qualification plan file
- `re_input`: directory generated from qualification runner used by workflow
  - `report-configuration-plan.json` configuration plan file defining the report configuration 
- `re_output`: directory generated from workflow
  - `report.md`: final markdown report
  - `report.docx`: final word report if `createWordReport` is `TRUE`}

\item{createWordReport}{Logical defining if a `docx` version of the report should also be created.
Note that `pandoc` installation is required for this feature
[https://github.com/Open-Systems-Pharmacology/OSPSuite.ReportingEngine/wiki/Installing-pandoc]}

\item{watermark}{Character string that will appear in all generated plots
Default is no watermark. `Label` objects from `tlf` package can be used to specifiy watermark font.}
}
\description{
Run a qualification workflow to create a qualification report.
}
