% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{createSectionOutput}
\alias{createSectionOutput}
\title{createSectionOutput}
\usage{
createSectionOutput(configurationPlan)
}
\arguments{
\item{configurationPlan}{A \code{ConfigurationPlan} object}
}
\value{
Names of created appendices
}
\description{
Create the Sections output and their markdown content
}
\keyword{internal}
