% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{data}
\name{DataSelectionKeys}
\alias{DataSelectionKeys}
\title{DataSelectionKeys}
\format{
An object of class \code{list} of length 2.
}
\usage{
DataSelectionKeys
}
\description{
List of available short keys for observed data selection
}
\examples{
DataSelectionKeys$NONE
DataSelectionKeys$ALL
}
\keyword{datasets}
