% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-source.R
\name{DataSource}
\alias{DataSource}
\title{DataSource}
\description{
R6 class managing observed data in simulation sets
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataFile}}{file name of the observed dataset}

\item{\code{metaDataFile}}{file name of the observed data dictionary}

\item{\code{caption}}{displayed legend of the data in the report}

\item{\code{descriptor}}{displayed data source descriptor in the report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSource-new}{\code{DataSource$new()}}
\item \href{#method-DataSource-getCaption}{\code{DataSource$getCaption()}}
\item \href{#method-DataSource-getRelativeDataPath}{\code{DataSource$getRelativeDataPath()}}
\item \href{#method-DataSource-clone}{\code{DataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-new}{}}}
\subsection{Method \code{new()}}{
Create a new `DataSource` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$new(dataFile, metaDataFile, caption = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataFile}}{file name of the observed dataset}

\item{\code{metaDataFile}}{file name of the observed data dictionary}

\item{\code{caption}}{displayed legend of the data in the report}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DataSource` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-getCaption"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-getCaption}{}}}
\subsection{Method \code{getCaption()}}{
Get the report caption that informs about the data source
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$getCaption(workflowFolder = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workflowFolder}}{Path of workflow folder}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-getRelativeDataPath"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-getRelativeDataPath}{}}}
\subsection{Method \code{getRelativeDataPath()}}{
Get the relative data path compared to workflow folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$getRelativeDataPath(workflowFolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workflowFolder}}{Path of workflow folder}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
