% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{dataSourceNamesFromIdentifier}
\alias{dataSourceNamesFromIdentifier}
\title{dataSourceNamesFromIdentifier}
\usage{
dataSourceNamesFromIdentifier(dataSourceId)
}
\arguments{
\item{dataSourceId}{Content translated from cell corresponding to dataSource in SimulationSets sheet.}
}
\value{
names of Output objects in SimulationSets sheet
}
\description{
Get the names of dataSource from simulation set identifier
}
\keyword{internal}
