% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{DataSourcesCodeIdentifiers}
\alias{DataSourcesCodeIdentifiers}
\title{DataSourceCodeIdentifiers}
\format{
An object of class \code{list} of length 4.
}
\usage{
DataSourcesCodeIdentifiers
}
\description{
Data Sources Code Identifiers
}
\keyword{internal}
