% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\docType{data}
\name{ddiPKRatioColumnName}
\alias{ddiPKRatioColumnName}
\title{Column names to check in observed data based on Configuration Plan PK Parameter}
\format{
An object of class \code{list} of length 2.
}
\usage{
ddiPKRatioColumnName
}
\description{
Column names to check in observed data based on Configuration Plan PK Parameter
}
\keyword{internal}
