% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{ddiPlotAxesSettings}
\alias{ddiPlotAxesSettings}
\title{Names of fields in configuration plane containing axes settings data for each DDI plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
ddiPlotAxesSettings
}
\description{
Names of fields in configuration plane containing axes settings data for each DDI plot type
}
\keyword{internal}
