% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{ddiPlotTypeSpecifications}
\alias{ddiPlotTypeSpecifications}
\title{Specifications of each DDI plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
ddiPlotTypeSpecifications
}
\description{
Specifications of each DDI plot type
}
\keyword{internal}
