% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{ddiSubplotTypes}
\alias{ddiSubplotTypes}
\title{Allowed DDI subplot types}
\format{
An object of class \code{list} of length 3.
}
\usage{
ddiSubplotTypes
}
\description{
Allowed DDI subplot types
}
\keyword{internal}
