% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{defaultMassBalanceGroupings}
\alias{defaultMassBalanceGroupings}
\title{defaultMassBalanceGrouping}
\usage{
defaultMassBalanceGroupings(simulationResults, compoundNames)
}
\arguments{
\item{simulationResults}{A `SimulationResults` object}

\item{compoundNames}{Names of simulation molecules}
}
\description{
Get mass balance default inclusion/exclusion defined as groupings list
In the Matlab version, the grouping was as followed:
1) Saliva is excluded from the mass balance calculation
2) Lumen is a parent compartment that regroups many sub-compartments but Feces
3) All the rest of the sub-compartments are used normally
}
\keyword{internal}
