% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\docType{data}
\name{defaultMaxSimulationsPerCore}
\alias{defaultMaxSimulationsPerCore}
\title{defaultMaxSimulationsPerCore}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
defaultMaxSimulationsPerCore
}
\description{
Default value for a scale factor used in a parallel simulation.  The product of this scale factor and the number of allowable cores (allowedCores) sets the maximum number of simulations that may be run on one core.
}
\keyword{datasets}
