% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\docType{data}
\name{defaultQuantileVec}
\alias{defaultQuantileVec}
\title{defaultQuantileVec}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
defaultQuantileVec
}
\description{
default quantiles for population sensitivity analysis
}
\keyword{datasets}
