% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\docType{data}
\name{defaultSensitivityAnalysisNumberOfCores}
\alias{defaultSensitivityAnalysisNumberOfCores}
\title{defaultSensitivityAnalysisNumberOfCores}
\format{An object of class \code{numeric} of length 1.}
\usage{
defaultSensitivityAnalysisNumberOfCores
}
\description{
default numberOfCores for sensitivity analysis
}
\keyword{datasets}
