% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\docType{data}
\name{defaultSimulationNumberOfCores}
\alias{defaultSimulationNumberOfCores}
\title{defaultSimulationNumberOfCores}
\format{An object of class \code{numeric} of length 1.}
\usage{
defaultSimulationNumberOfCores
}
\description{
default numberOfCores for simulation
}
\keyword{datasets}
