% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\docType{data}
\name{defaultVariationRange}
\alias{defaultVariationRange}
\title{defaultVariationRange}
\format{An object of class \code{numeric} of length 1.}
\usage{
defaultVariationRange
}
\description{
default parameter variation range for sensitivity analysis
}
\keyword{datasets}
