% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\docType{data}
\name{DemographyDefaultParameters}
\alias{DemographyDefaultParameters}
\title{DemographyDefaultParameters}
\format{
An object of class \code{list} of length 5.
}
\usage{
DemographyDefaultParameters
}
\description{
Demography Default Parameters
}
\keyword{internal}
