% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{displayConfigurationPlanPlotInfo}
\alias{displayConfigurationPlanPlotInfo}
\title{displayConfigurationPlanPlotInfo}
\usage{
displayConfigurationPlanPlotInfo(configurationPlanField)
}
\arguments{
\item{configurationPlanField}{A list extracted from a configuration plan field}
}
\value{
Array of name properties and their value
}
\description{
Display information of a configuration plan \code{Plot} field
}
\examples{

# Example of list with properties
plotField <- list(
  PlotNumber = 5,
  Simulation = "Name of Simulation",
  Project = "Name of Project",
  Options = list(width = 10, height = 10, units = "cm")
)

# Log messages are usually displayed through `cat()`, `error()` or `warning()`
cat(displayConfigurationPlanPlotInfo(plotField))

}
