% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{displayRatioMeasureAnalyticalSolution}
\alias{displayRatioMeasureAnalyticalSolution}
\title{displayRatioMeasureAnalyticalSolution}
\usage{
displayRatioMeasureAnalyticalSolution(referenceData, comparisonData)
}
\arguments{
\item{referenceData}{A data.frame of PK Parameter values for reference Population}

\item{comparisonData}{A data.frame of PK Parameter values for Population to compare}
}
\value{
A character
}
\description{
Display Ratio Measure Analytical Solution message
}
\keyword{internal}
