% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mass-balance.R
\name{envList2PathStringsVector}
\alias{envList2PathStringsVector}
\title{pathStringsVector}
\usage{
envList2PathStringsVector(envList)
}
\description{
Input is a list envList of environments that is output by the ospsuite::getAllXXXMatching() functions

Output is a list of the paths in each environment in the input list
}
