% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{evalDataFilter}
\alias{evalDataFilter}
\title{evalDataFilter}
\usage{
evalDataFilter(data, filterExpression)
}
\arguments{
\item{data}{data.frame containing observed data}

\item{filterExpression}{expression
character string filter to be applied}
}
\value{
vector of logicals corresponding to the evaluation of the filter
}
\description{
Evaluate a data filter by converting the variable names of the data.frame
into names of variables to be evaluated in the filter expression.
}
