% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\docType{data}
\name{ExportPlotConfiguration}
\alias{ExportPlotConfiguration}
\title{ExportPlotConfiguration}
\format{
An object of class \code{ExportPlotConfiguration} (inherits from \code{R6}) of length 6.
}
\usage{
ExportPlotConfiguration
}
\description{
Properties of exported plots
}
\keyword{datasets}
