% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileExtension}
\alias{fileExtension}
\title{fileExtension}
\usage{
fileExtension(file)
}
\arguments{
\item{file}{File name or path}
}
\value{
The file extension as a character string
}
\description{
Return the extension of a file
}
\keyword{internal}
