% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{formatAxisProperties}
\alias{formatAxisProperties}
\title{formatAxisProperties}
\usage{
formatAxisProperties(axisField)
}
\arguments{
\item{axisField}{List of configuration plan field properties}
}
\value{
A list of properties compatible with \code{tlf} package nomenclature
}
\description{
Format configuration plan field values to a list compatible with \code{tlf} package nomenclature
}
\keyword{internal}
