% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pop-pk-parameters.R
\name{formatPKParametersTable}
\alias{formatPKParametersTable}
\title{formatPKParametersTable}
\usage{
formatPKParametersTable(structureSet, pkParametersTable, populationTable)
}
\arguments{
\item{pkParametersTable}{A data.frame of PK parameters}

\item{populationTable}{A data.frame of population parameters}

\item{structureSets}{A `SimulationStructure` objects}
}
\value{
A data.frame and its metaData including the values of PK parameters across Population Simulation sets
}
\description{
Format data.frame of PK and Population Parameters from a simulation set
}
\keyword{internal}
