% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateResultFileNames}
\alias{generateResultFileNames}
\title{generateResultFileNames}
\usage{
generateResultFileNames(
  numberOfCores,
  folderName,
  fileName,
  separator = "-",
  extension = ".csv"
)
}
\value{
A list of filenames to be output by each core
}
\description{
#Generate a listcontaining names of CSV result files that will be output by each corein parallel computation
}
