% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geomean}
\alias{geomean}
\title{geomean}
\usage{
geomean(x, na.rm = TRUE)
}
\arguments{
\item{x}{values}

\item{na.rm}{logical defining removal of `NA` values}
}
\value{
Geometric mean
}
\description{
Calculate the geometric mean
}
