% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geomeanDividedBySD}
\alias{geomeanDividedBySD}
\title{geomeanDividedBySD}
\usage{
geomeanDividedBySD(x, na.rm = TRUE)
}
\arguments{
\item{x}{values}

\item{na.rm}{logical defining removal of `NA` values}
}
\value{
Geometric mean / geometric SD
}
\description{
Calculate the geometric mean / geometric SD
}
