% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geomeanMultipliedBySD}
\alias{geomeanMultipliedBySD}
\title{geomeanMultipliedBySD}
\usage{
geomeanMultipliedBySD(x, na.rm = TRUE)
}
\arguments{
\item{x}{values}

\item{na.rm}{logical defining removal of `NA` values}
}
\value{
Geometric mean * geometric SD
}
\description{
Calculate the geometric mean * geometric SD
}
