% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getAggregateFromStat}
\alias{getAggregateFromStat}
\title{getAggregateFromStat}
\usage{
getAggregateFromStat(statisticId, time, outputValues)
}
\arguments{
\item{statisticId}{Statistic Id as defined in \code{ConfigurationPlan} used for data aggregation}

\item{time}{Time values on which output values are aggregated}

\item{outputValues}{Output values to be aggregated}
}
\value{
A data.frame of aggregated data to display
}
\description{
Get summary statistics of output values
}
\keyword{internal}
