% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllowedCores}
\alias{getAllowedCores}
\title{getAllowedCores}
\usage{
getAllowedCores()
}
\value{
Allowed number of CPU cores for computation
}
\description{
Relevant only when reporting engine is executed on a Linux Kubernetes cluster.
}
\keyword{internal}
