% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getAnchorName}
\alias{getAnchorName}
\title{getAnchorName}
\usage{
getAnchorName(tag)
}
\arguments{
\item{tag}{Character string}
}
\value{
A character Name/identifier of the anchor tag
}
\description{
Get the name/identifier an anchor tag
}
\examples{
getAnchorName('<a id="section-1"></a>')

# Works also for tag name instead of id 
getAnchorName('<a name="section-1"></a>')

}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
