% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{getApplicationTimesForPath}
\alias{getApplicationTimesForPath}
\title{getApplicationTimesForPath}
\usage{
getApplicationTimesForPath(simulation, path)
}
\arguments{
\item{simulation}{A \code{Simulation} object}

\item{path}{A simulation path}
}
\value{
Time values of applications
}
\description{
getApplicationTimesForPath
}
\keyword{internal}
