% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{getApplicationTimesForSimulation}
\alias{getApplicationTimesForSimulation}
\title{getApplicationTimesForSimulation}
\usage{
getApplicationTimesForSimulation(simulation, outputs)
}
\arguments{
\item{simulation}{A `Simulation` object}

\item{output}{A list of `Output` objects}
}
\value{
Time values of applications
}
\description{
getApplicationTimesForSimulation
}
\keyword{internal}
