% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getAxesPropertiesForTimeProfiles}
\alias{getAxesPropertiesForTimeProfiles}
\title{getAxesPropertiesForTimeProfiles}
\usage{
getAxesPropertiesForTimeProfiles(configurationPlanAxes)
}
\arguments{
\item{configurationPlanAxes}{list of axes properties defined in `Axes` field of configuration plan}
}
\value{
A list of properties for identified `x` and `y` axes that follows `tlf` package nomenclature
}
\description{
Identify and convert properties from `Axes` field of configuration plan
}
