% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getAxesPropertiesFromName}
\alias{getAxesPropertiesFromName}
\title{getAxesPropertiesFromName}
\usage{
getAxesPropertiesFromName(configurationPlan, plotName)
}
\arguments{
\item{configurationPlan}{A `ConfigurationPlan` object}

\item{plotName}{Field name of the plot in the configuration plan `AxesSettings`}
}
\value{
A list of properties for axes identified for `x`, `y` and `y2` axes.
The identified properties are directly compatible with `tlf` package nomenclature
}
\description{
Get axes properties from global field `AxesSettings` of configuration plan if defined,
otherwise use default values.
}
\keyword{internal}
