% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{getAxesSettings}
\alias{getAxesSettings}
\title{getAxesSettings}
\usage{
getAxesSettings(axesSettingsFromConfigurationPlot)
}
\arguments{
\item{axesSettingsFromConfigurationPlot}{is a field from the `configurationPlan$plots` list}
}
\value{
`axesSettings`, a list of settings for each of the X and Y axis.  Each list contains the unit, dimensions, and scaling type for each axes and option to plot grid lines.
}
\description{
Read axes settings for plots.
}
\keyword{internal}
