% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getCaptionFromStat}
\alias{getCaptionFromStat}
\title{getCaptionFromStat}
\usage{
getCaptionFromStat(statisticId, outputName)
}
\arguments{
\item{statisticId}{Statistic Id as defined in \code{ConfigurationPlan} used for data aggregation}

\item{outputName}{Display name of output values}
}
\value{
A character value used as caption for the time profile plot
}
\description{
Get display legend caption for time profile plot
}
\keyword{internal}
