% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getColorFromOutputGroup}
\alias{getColorFromOutputGroup}
\title{getColorFromOutputGroup}
\usage{
getColorFromOutputGroup(
  group,
  data,
  dataMapping,
  legendVariable = "legend",
  colorVariable = "color"
)
}
\arguments{
\item{group}{A data.frame mapping properties to output groups}

\item{data}{A data.frame}

\item{dataMapping}{A `DataMapping` object}

\item{legendVariable}{Name of legend variable in`group`}

\item{colorVariable}{Name of color variable in`group`}
}
\value{
A sorted array of color values
}
\description{
Get the appropriate colors from an output group
}
\keyword{internal}
