% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getColorGroupForPKParameterPlot}
\alias{getColorGroupForPKParameterPlot}
\title{getColorGroupForPKParameterPlot}
\usage{
getColorGroupForPKParameterPlot(
  output,
  referenceSetName = NULL,
  simulationSetNames
)
}
\arguments{
\item{output}{An `Output` object}

\item{referenceSetName}{Display name of reference simulation set}

\item{simulationSetNames}{Display names of simulation sets}
}
\value{
A data.frame mapping colors to names
}
\description{
Map colors to population names
}
\keyword{internal}
