% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getComparisonHistogramResults}
\alias{getComparisonHistogramResults}
\title{getComparisonHistogramResults}
\usage{
getComparisonHistogramResults(
  demographyPaths,
  simulationSetNames,
  data,
  metaData,
  observedData,
  settings = NULL,
  simulationSetDescriptor = "",
  demographyResults = list()
)
}
\arguments{
\item{demographyPaths}{Names of demography variables to be displayed}

\item{simulationSetNames}{Names of simulation sets}

\item{data}{A data.frame of simulated demography values across the simulationSets}

\item{metaData}{A list of meta data indicating the display properties of the data}

\item{observedData}{A data.frame of observed demography values across the simulationSets}

\item{settings}{A list of plot settings}

\item{simulationSetDescriptor}{Character describing the population sets within the report}

\item{demographyResults}{A list of `TaskResults` objects}
}
\value{
A list of `TaskResults` objects
}
\description{
Get Comparison Histogram results for Pediatric workflows
}
\keyword{internal}
