% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\name{getComparisonTimeProfileOutputsDataframe}
\alias{getComparisonTimeProfileOutputsDataframe}
\title{getComparisonTimeProfileOutputsDataframe}
\usage{
getComparisonTimeProfileOutputsDataframe(configurationPlan)
}
\arguments{
\item{configurationPlan}{The configuration plan of a Qualification workflow read from json file.}
}
\value{
A dataframe containing data for generating time profile plots
}
\description{
Get a dataframe relating project, simulation and output path for each comparison time profile plot
}
