% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getCompoundNameFromPath}
\alias{getCompoundNameFromPath}
\title{getCompoundNameFromPath}
\usage{
getCompoundNameFromPath(path)
}
\arguments{
\item{path}{A quantitiy path from the configuration plan
For instance, "S2|Organism|PeripheralVenousBlood|Midazolam|Plasma (Peripheral Venous Blood)"
or "Midazolam 600mg SD|ObservedData|Peripheral Venous Blood|Plasma|Rifampin|Conc"}
}
\value{
A string corresponding to the compound name of a configuration plan quantity path
}
\description{
Get the compound name from a configuration plan quantity path
}
\examples{
\dontrun{
getCompoundNameFromPath("S2|Organism|PeripheralVenousBlood|Midazolam|Plasma (Peripheral Venous Blood)")
# > "Midazolam"
getCompoundNameFromPath("Midazolam 600mg SD|ObservedData|Peripheral Venous Blood|Plasma|Rifampin|Conc")
# > "Rifampin"
}
}
\keyword{internal}
