% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mass-balance.R
\name{getCompoundsList}
\alias{getCompoundsList}
\title{getCompoundsList}
\usage{
getCompoundsList(simulation)
}
\description{
Function to return a list of the compounds in a simulation.

Input is a simulation ("simulation") loaded from a PKML file using ospsuite::loadSimulation()

Output is a vector ("uniqueCompounds") of strings of unique compounds
}
