% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getCurvePropertiesForTimeProfiles}
\alias{getCurvePropertiesForTimeProfiles}
\title{getCurvePropertiesForTimeProfiles}
\usage{
getCurvePropertiesForTimeProfiles(
  configurationPlanCurve,
  simulation,
  simulationResults,
  axesProperties,
  configurationPlan,
  logFolder
)
}
\arguments{
\item{configurationPlanCurve}{list of properties defined in `Curves` field of configuration plan}

\item{simulation}{A `Simulation` object from `ospsuite` package
that includes required information to identify and convert the data requested from `configurationPlanCurve` properties}

\item{simulationResults}{A `SimulationResults` object from `ospsuite` package
that includes the data requested from `configurationPlanCurve` properties}

\item{axesProperties}{list of axes properties obtained from `getAxesForTimeProfiles`}

\item{configurationPlan}{A `ConfigurationPlan` object that includes methods to find observed data}

\item{logFolder}{folder where the logs are saved}
}
\value{
A list of data and properties to be plotted and that follows `tlf` package nomenclature
}
\description{
Identify and convert properties from `Curves` field of configuration plan
}
