% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getDataSourcesContent}
\alias{getDataSourcesContent}
\title{getDataSourcesContent}
\usage{
getDataSourcesContent(excelFile, dataSourcesTable, simulationSources)
}
\arguments{
\item{excelFile}{name of the Excel file from which the R script is created}

\item{dataSourcesTable}{Data.frame read from the Excel sheet "DataSources"}

\item{simulationSources}{Names of DataSource objects used by simulation sets}
}
\value{
Character vector defining the \code{DataSource} objects
}
\description{
Creates a character vector to be written in a workflow .R script defining \code{DataSource} objects.
}
\keyword{internal}
