% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\name{getDDIOutputsDataframe}
\alias{getDDIOutputsDataframe}
\title{getDDIOutputsDataframe}
\usage{
getDDIOutputsDataframe(configurationPlan)
}
\arguments{
\item{configurationPlan}{The configuration plan of a Qualification workflow read from json file.}
}
\value{
A list containing data for generating DDI plots
}
\description{
Get a dataframe relating project, simulation, output, pk parameter, start time, end time for each DDI plot component
}
\keyword{internal}
