% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getDDISection}
\alias{getDDISection}
\title{getDDISection}
\usage{
getDDISection(dataframe, metadata, sectionID, idPrefix, captionSuffix = NULL)
}
\arguments{
\item{dataframe}{of DDI data for current DDI section}

\item{metadata}{for DDI plots}

\item{sectionID}{of report}

\item{idPrefix}{unique ID to index task results}

\item{captionSuffix}{to append to qualification plan title}
}
\value{
a `list` of DDI results for the current DDI section
}
\description{
Plot observation vs prediction for DDI qualification workflow
}
\keyword{internal}
