% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getDDITable}
\alias{getDDITable}
\title{getDDITable}
\usage{
getDDITable(dataframe)
}
\arguments{
\item{dataframe}{for generating DDI summary table}
}
\value{
Summary table for DDI plot
}
\description{
Summary table for DDI
}
\keyword{internal}
