% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getDefaultDemographyXParameters}
\alias{getDefaultDemographyXParameters}
\title{getDefaultDemographyXParameters}
\usage{
getDefaultDemographyXParameters(workflowType)
}
\arguments{
\item{workflowType}{Name of workflow type.
Use enum \code{PopulationWorkflowTypes} to get a list of available workflow types.}
}
\value{
names of default demography parameters
}
\description{
Get names of default demography parameters in x axis of demography plots.
}
\examples{

getDefaultDemographyXParameters(PopulationWorkflowTypes$pediatric)

}
