% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames-nomenclature.R
\name{getDefaultFolderName}
\alias{getDefaultFolderName}
\title{getDefaultFolderName}
\usage{
getDefaultFolderName(..., suffix = "", sep = "-")
}
\arguments{
\item{...}{names}

\item{suffix}{Suffix to be added at the end of the folder name.
Dafault `suffix` is the curent date}

\item{sep}{separation between names and suffix. Default is `-`}
}
\value{
default filename adding default suffix and file extension
}
\description{
Create a default folder name for reporting engine results and figures
}
\keyword{internal}
