% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{getDefaultMCRandomSeed}
\alias{getDefaultMCRandomSeed}
\title{getDefaultMCRandomSeed}
\usage{
getDefaultMCRandomSeed()
}
\value{
Random seed number
}
\description{
Get the default random seed when performing Monte Carlo sampling
}
\examples{
getDefaultMCRandomSeed()

}
