% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{getDefaultMCRepetitions}
\alias{getDefaultMCRepetitions}
\title{getDefaultMCRepetitions}
\usage{
getDefaultMCRepetitions()
}
\value{
Random seed number
}
\description{
Get the default number of repetitions when performing Monte Carlo sampling
}
\examples{
getDefaultMCRepetitions()

}
