% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getDefaultPkParametersXParameters}
\alias{getDefaultPkParametersXParameters}
\title{getDefaultPkParametersXParameters}
\usage{
getDefaultPkParametersXParameters(workflowType)
}
\description{
getDefaultPkParametersXParameters
}
