% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getDefaultPropertiesFromTheme}
\alias{getDefaultPropertiesFromTheme}
\title{getDefaultPropertyFromTheme}
\usage{
getDefaultPropertiesFromTheme(
  plotName,
  propertyType = "points",
  propertyNames = as.character(tlf::AestheticProperties)
)
}
\arguments{
\item{plotName}{Name of the plot in Theme (eg `"plotTimeProfile"`)}

\item{propertyType}{One of `"points"`, `"lines`, `"ribbons"` or `"errorbars"`}

\item{propertyName}{Name of the aesthetic property (eg `"color"`)}
}
\value{
Property value
}
\description{
Get default property value from current reEnv theme
}
\keyword{internal}
