% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{getDefaultRETheme}
\alias{getDefaultRETheme}
\title{getDefaultRETheme}
\usage{
getDefaultRETheme()
}
\value{
A \code{Theme} object from \code{tlf} package
}
\description{
Get default plot settings for RE package
}
