% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getDefaultTimeProfileAxesSettings}
\alias{getDefaultTimeProfileAxesSettings}
\title{getDefaultTimeProfileAxesSettings}
\usage{
getDefaultTimeProfileAxesSettings()
}
\value{
List of `x` and `y` axes settings
}
\description{
Get the default axes settings for mean and population time profiles
for keeping compatibility with configuration plan from Matlab version
}
\keyword{internal}
