% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getDefaultTotalSensitivityThreshold}
\alias{getDefaultTotalSensitivityThreshold}
\title{getDefaultTotalSensitivityThreshold}
\usage{
getDefaultTotalSensitivityThreshold(
  totalSensitivityThreshold,
  variableParameterPaths
)
}
\arguments{
\item{totalSensitivityThreshold}{cut-off used for plots of the most sensitive parameters}

\item{variableParameterPaths}{vector of paths of parameters to vary when performing sensitivity analysis}
}
\description{
return the default totalSensitivityThreshold to be used in a population sensitivity analysis plot
}
