% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getDemographyAcrossPopulations}
\alias{getDemographyAcrossPopulations}
\title{getDemographyAcrossPopulations}
\usage{
getDemographyAcrossPopulations(structureSets, demographyPaths)
}
\arguments{
\item{structureSets}{List of `SimulationStructure` objects}

\item{demographyPaths}{Paths of demography variables to display}
}
\value{
A list of `data` and its `metaData`
}
\description{
Get Demography data across populations of simulation sets
}
\keyword{internal}
