% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getDemographyAggregatedData}
\alias{getDemographyAggregatedData}
\title{getDemographyAggregatedData}
\usage{
getDemographyAggregatedData(
  data,
  xParameterName,
  yParameterName,
  bins = NULL,
  stairstep = TRUE
)
}
\description{
getDemographyAggregatedData
}
