% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getDictionaryVariable}
\alias{getDictionaryVariable}
\title{getDictionaryVariable}
\usage{
getDictionaryVariable(
  dictionary,
  variableID,
  idColumn = dictionaryParameters$ID,
  datasetColumn = dictionaryParameters$datasetColumn
)
}
\arguments{
\item{dictionary}{A data.frame from dictionary}

\item{variableID}{An identifier}

\item{idColumn}{The column name used for identification}

\item{datasetColumn}{The column name used mapping the id to variable}
}
\value{
A variable name from dictionary
}
\description{
Get the variable name from dictionary
}
\keyword{internal}
