% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getDisplaySensitivityParameters}
\alias{getDisplaySensitivityParameters}
\title{getDisplaySensitivityParameters}
\usage{
getDisplaySensitivityParameters(
  parameterNames,
  maxLinesPerParameter,
  maxWidthPerParameter,
  logFolder = getwd()
)
}
\arguments{
\item{parameterNames}{vector parameters displayed in the sensitivity plots}

\item{maxLinesPerParameter}{maxLinesPerParameter maximum number of lines allowed per displayed parameters}

\item{maxWidthPerParameter}{maximum number of characters allowed per lines of displayed parameters}

\item{logFolder}{folder where the logs are saved}
}
\value{
Displayed parameter names for sensitivity breaking lines when too long
}
\description{
Displayed parameter names for sensitivity breaking lines when too long
}
