% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{getElapsedTime}
\alias{getElapsedTime}
\title{getElapsedTime}
\usage{
getElapsedTime(tic, unit = "min")
}
\arguments{
\item{tic}{Numeric start time}

\item{unit}{display unit of elapsed time}
}
\value{
Character displaying elapsed time in `unit`
}
\description{
Get computer time as a character
}
\keyword{internal}
