% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFileExtension}
\alias{getFileExtension}
\title{getFileExtension}
\usage{
getFileExtension(filePath)
}
\arguments{
\item{filePath}{character string containing the name of the path or file to trim}
}
\value{
extension character string of the trimmed filed name
}
\description{
Get extension of a file
}
\examples{
\dontrun{
pathName <- "folder/subfolder/testFile.txt"
getFileExtension(pathName)
}
}
