% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getFileLocationFromType}
\alias{getFileLocationFromType}
\title{getFileLocationFromType}
\usage{
getFileLocationFromType(location, type, excelFile)
}
\arguments{
\item{location}{Path of the file if type is FILE or sheetname if type is SHEET}

\item{type}{Location type: either "SHEET" or "FILE"}

\item{excelFile}{name of the Excel file from which the R script is created}
}
\value{
Character of location to provide
}
\description{
Get the right path text and copy the sheet content as a file if type is SHEET
}
