% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getFirstLineElement}
\alias{getFirstLineElement}
\title{getFirstLineElement}
\usage{
getFirstLineElement(lineContent, split = " ")
}
\arguments{
\item{lineContent}{Character string}

\item{split}{character pattern to split between elements/words}
}
\value{
Character string
}
\description{
Get first element/word of a line
}
\keyword{internal}
