% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{getGOFAxesUnits}
\alias{getGOFAxesUnits}
\title{getGOFAxesUnits}
\usage{
getGOFAxesUnits(gofPlan, settings)
}
\arguments{
\item{gofPlan}{List providing the configuration of the goodness of fit results}

\item{settings}{settings for the task}
}
\value{
A list of units for goodness of fit results
}
\description{
Get list of units from axes properties and settings
}
