% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\name{getGOFOutputsDataframe}
\alias{getGOFOutputsDataframe}
\title{getGOFOutputsDataframe}
\usage{
getGOFOutputsDataframe(configurationPlan)
}
\arguments{
\item{configurationPlan}{The configuration plan of a Qualification workflow read from json file.}
}
\value{
A dataframe containing data for generating GOF plots
}
\description{
Get a dataframe relating project, simulation and output path for each GOF plot component
}
\keyword{internal}
