% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getGuestDeltaFromConfigurationPlan}
\alias{getGuestDeltaFromConfigurationPlan}
\title{getGuestDeltaFromConfigurationPlan}
\usage{
getGuestDeltaFromConfigurationPlan(ddiRatioPlan)
}
\arguments{
\item{ddiRatioPlan}{Properties defined in `DDIRatioPlots` field of confiuration plan}
}
\value{
A named list of Guest et al. values
}
\description{
Get Guest et al. delta values for a DDI Ratio plot from the configuration plan
}
\keyword{internal}
