% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getHeaderNames}
\alias{getHeaderNames}
\title{getHeaderNames}
\usage{
getHeaderNames(excelTable)
}
\arguments{
\item{excelTable}{table read from Excel sheet}
}
\value{
names of objects in header
}
\description{
Get the names of objects in header
}
\keyword{internal}
