% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getIdentifierInfo}
\alias{getIdentifierInfo}
\title{getIdentifierInfo}
\usage{
getIdentifierInfo(workflowTable, simulationIndex, codeId)
}
\arguments{
\item{workflowTable}{Data.frame read from one of the available Excel sheets}

\item{simulationIndex}{Column to read after removing "Code Identifier" and "Description"}

\item{codeId}{Line to read in the data.frame corresponding to a specific value of "Code Identifier"}
}
\value{
Information from a data.frame matching a certain \code{simulationIndex} column and \code{codeId} line
}
\description{
Get and format the information from a data.frame matching a certain \code{simulationIndex} column and \code{codeId} line
}
\keyword{internal}
