% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getIndividualSAResultsFileName}
\alias{getIndividualSAResultsFileName}
\title{getIndividualSAResultsFileName}
\usage{
getIndividualSAResultsFileName(individualId, resultsFileName)
}
\arguments{
\item{individualId}{id of individual}

\item{resultsFileName}{root name of population sensitivity analysis results CSV files}
}
\description{
Function to build name of individual SA results file
}
\keyword{internal}
