% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getIntroFromReportTitle}
\alias{getIntroFromReportTitle}
\title{getIntroFromReportTitle}
\usage{
getIntroFromReportTitle(reportTitle = NULL, intro = "temp-report-title.md")
}
\arguments{
\item{reportTitle}{Report title page
If \code{reportTitle} is an existing file, its content will be used as cover page.
If \code{reportTitle} is one character string, it is assumed as a title.
Thus the markdown title tag is internally added to \code{reportTitle}.
If \code{reportTitle} is multiple character strings, it is assumed as the cover page content.
and used \emph{as is} in the cover page.}

\item{intro}{temporary introduction file deleted when merged to final report
Parameter is named intro to stay consistent with Qualification Workflow nomenclature}
}
\value{
\code{intro}, path of temporary introduction file if created
}
\description{
Get introduction file used as cover page of report
}
\keyword{internal}
