% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getLegendAesOverride}
\alias{getLegendAesOverride}
\title{getLegendAesOverride}
\usage{
getLegendAesOverride(plotObject)
}
\arguments{
\item{plotObject}{A ggplot object}
}
\value{
A list of aesthetic values
}
\description{
In time profiles, legends are merged into one unique legend
The displayed legend is stored in the \code{plotObject} within the color guide field \code{override.aes}.
This function simply gets the list from that field for updating the current legend
}
\keyword{internal}
