% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getLegendScalingFactors}
\alias{getLegendScalingFactors}
\title{getLegendScalingFactors}
\usage{
getLegendScalingFactors(legendPosition = tlf::LegendPositions$outsideTop)
}
\arguments{
\item{legendPosition}{The name of the legend position as defined by `tlf` enum `LegendPositions`}
}
\value{
A list of scaling values for `width` and `height`
}
\description{
Get factors to scale plot dimensions accounting for the location of the legend.
Initial estimates based on Abdullah's tests
TODO: improve this directly from tlf
}
\keyword{internal}
