% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getLineBreakWidth}
\alias{getLineBreakWidth}
\title{getLineBreakWidth}
\usage{
getLineBreakWidth(element = "legend", plotConfiguration)
}
\arguments{
\item{element}{The name of element to which the line break should be added.
If applied to the legend, use \code{"legend"}.
If applied to a plot label use e.g. \code{"ylabel"} or \code{"title"}.
If applied to tick labels use \code{"yticks"} or \code{"yticklabels"}.}

\item{plotConfiguration}{A \code{PlotConfiguration} object from the \code{tlf} package}
}
\value{
An integer as max character width before using line breaks
}
\description{
Calculate the maximum number of characters before breaking lines.
This aims at preventing as much as possible legends shrinking the plot and legends not fully displayed
}
\keyword{internal}
