% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{getLogPaths}
\alias{getLogPaths}
\title{getLogPaths}
\usage{
getLogPaths(logLevel, logFolder)
}
\arguments{
\item{logLevel}{Name of log level. Use helper enum `LogLevels` for finding appropriate log level.}

\item{logFolder}{Folder in which the log files are saved}
}
\value{
An array of file paths
}
\description{
Get paths of log files from requested log level
}
\keyword{internal}
