% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{getMassBalanceDataMapping}
\alias{getMassBalanceDataMapping}
\title{getMassBalanceDataMapping}
\usage{
getMassBalanceDataMapping(plotType)
}
\arguments{
\item{plotType}{One of the 5 plot types displayed by mass balance task}
}
\value{
A `tlf` `DataMapping` object
}
\description{
Get mass balance data mapping
}
\keyword{internal}
