% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{getMassBalancePlotConfiguration}
\alias{getMassBalancePlotConfiguration}
\title{getMassBalancePlotConfiguration}
\usage{
getMassBalancePlotConfiguration(plotType, data, metaData, settings)
}
\arguments{
\item{plotType}{One of the 5 plot types displayed by mass balance task}

\item{data}{data.frame}

\item{metaData}{meta data on \code{data}}

\item{settings}{User-defined options}
}
\value{
A \code{tlf} \code{PlotConfiguration} object
}
\description{
Get mass balance time profile plot configuration
}
\keyword{internal}
