% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getMeanPKAnalysesFromOuptut}
\alias{getMeanPKAnalysesFromOuptut}
\title{getMeanPKAnalysesFromOuptut}
\usage{
getMeanPKAnalysesFromOuptut(data, output, molWeight = NULL)
}
\arguments{
\item{data}{A data.frame of PK Analyses}

\item{output}{An `Output` object defining `pkParameters`}

\item{molWeight}{Molecular weight for converting into PK Parameter `displayUnit`}
}
\value{
A data.frame with `Path`, `Parameter`, `Value` and `Unit` to display in final report
}
\description{
Get PK analyses from an `Output` object
}
\keyword{internal}
