% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getMeanPKAnalysesFromOutput}
\alias{getMeanPKAnalysesFromOutput}
\title{getMeanPKAnalysesFromOutput}
\usage{
getMeanPKAnalysesFromOutput(data, output, molWeight = NULL)
}
\arguments{
\item{data}{A data.frame of PK Analyses}

\item{output}{An \code{Output} object defining \code{pkParameters}}

\item{molWeight}{Molecular weight for converting into PK Parameter \code{displayUnit}}
}
\value{
A data.frame with \code{Path}, \code{Parameter}, \code{Value} and \code{Unit} to display in final report
}
\description{
Get PK analyses from an \code{Output} object
}
\keyword{internal}
