% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getMechanismName}
\alias{getMechanismName}
\title{getMechanismName}
\usage{
getMechanismName(mechanism)
}
\arguments{
\item{mechanism}{name of mechanism as read form DDI observed data file}
}
\value{
Display name of mechanism to be used in DDI report
}
\description{
Remove underscores from mechanism name read from DDI observed data file according to the dictionary defined in `reEnv$ddiRatioSubsetsDictionary`
}
\keyword{internal}
