% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getMolWeightForCompound}
\alias{getMolWeightForCompound}
\title{getMolWeightForCompound}
\usage{
getMolWeightForCompound(compoundName, simulation)
}
\arguments{
\item{compoundName}{Name of a compound e.g. "Midazolam"}

\item{simulation}{A `Simulation` object}
}
\value{
molecular weight in base unit or NA if not found
}
\description{
Get molecular weight in base unit for a requested compound in a simulation
}
