% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getMolWeightForObservedData}
\alias{getMolWeightForObservedData}
\title{get}
\usage{
getMolWeightForObservedData(pathArray, configurationPlan, simulation)
}
\arguments{
\item{pathArray}{Configuration plan path as an array for observed data}

\item{configurationPlan}{configuration plan object}

\item{simulation}{simulation object}
}
\value{
molecular weight in base unit or NA if not found
}
\description{
Translate value of `Scaling` property from configuration plan into tlf `scale`
}
