% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getMolWeightForPath}
\alias{getMolWeightForPath}
\title{getMolWeightForPath}
\usage{
getMolWeightForPath(pathName, simulation)
}
\arguments{
\item{pathName}{List of PK senstivities computed by runSensitivityAnalysis}

\item{simulation}{List of PlotConfiguration class objects for each plot}
}
\value{
Molecular weight value for the path
}
\description{
Get Molecular Weight for quantity path name
}
