% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getMonteCarloMedians}
\alias{getMonteCarloMedians}
\title{getMonteCarloMedians}
\usage{
getMonteCarloMedians(listOfData)
}
\arguments{
\item{listOfData}{A list of summary statistics matrices}
}
\value{
A data.frame displaying summary statistics by column
}
\description{
Get all median values from a list of Monte Carlo simulations
}
\keyword{internal}
