% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getObjectNameAsString}
\alias{getObjectNameAsString}
\title{getObjectNameAsString}
\usage{
getObjectNameAsString(object)
}
\arguments{
\item{object}{the name of which is to be returned}
}
\value{
the name of the \code{object} as a string
}
\description{
Return the name of an object as a string
}
\keyword{internal}
