% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getObjectNamesInGlobalEnv}
\alias{getObjectNamesInGlobalEnv}
\title{getObjectNamesInGlobalEnv}
\usage{
getObjectNamesInGlobalEnv(objectType)
}
\arguments{
\item{objectType}{Object type or class}
}
\value{
Array of Number of \code{Output} objects in the Global Environment
}
\description{
Get object names of certain type/class in the Global Environment
}
\keyword{internal}
