% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getObservedCurveProperties}
\alias{getObservedCurveProperties}
\title{getObservedCurveProperties}
\usage{
getObservedCurveProperties(
  configurationPlanCurve,
  simulation,
  axesProperties,
  configurationPlan
)
}
\arguments{
\item{configurationPlanCurve}{\code{Curves} fields of configuration plan}

\item{simulation}{A \code{Simulation} object from \code{ospsuite} package
that includes required information to identify and convert the data requested from \code{configurationPlanCurve} properties}

\item{axesProperties}{list of axes properties obtained from \code{getAxesProperties}}

\item{configurationPlan}{A \code{ConfigurationPlan} object that includes methods to find observed data}
}
\value{
A named list data and meta data parameters
}
\description{
Get Curve Properties and values of Observed data
}
\keyword{internal}
