% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getObservedDataFromConfigurationPlan}
\alias{getObservedDataFromConfigurationPlan}
\title{getObservedDataFromConfigurationPlan}
\usage{
getObservedDataFromConfigurationPlan(observedDataId, configurationPlan)
}
\arguments{
\item{observedDataId}{Identifier of observed data}

\item{configurationPlan}{A \code{ConfigurationPlan} object that includes methods to find the data}
}
\value{
list of including data and metaData to perform time profile plot
}
\description{
Get selected observed data from a \code{ConfigurationPlan} object
}
\keyword{internal}
