% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getObservedDataFromOutput}
\alias{getObservedDataFromOutput}
\title{getObservedDataFromOutput}
\usage{
getObservedDataFromOutput(
  output,
  data,
  dataMapping,
  simulationQuantity,
  molWeight,
  timeUnit,
  logFolder
)
}
\arguments{
\item{output}{An `Output` object}

\item{data}{A data.frame}

\item{dataMapping}{A list mapping the variable of data}

\item{simulationQuantity}{Dimension/quantity for unit conversion of dependent variable}

\item{molWeight}{Molar weight for unit conversion of dependent variable}

\item{timeUnit}{time unit for unit conversion of time}

\item{logFolder}{folder where the logs are saved}
}
\value{
list of data and lloq data.frames
}
\description{
Get selected observed data from an Output object
}
