% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getObservedDataIdFromPath}
\alias{getObservedDataIdFromPath}
\title{getObservedDataIdFromPath}
\usage{
getObservedDataIdFromPath(path)
}
\arguments{
\item{path}{A quantitiy path from the configuration plan
For instance, "S2|Organism|PeripheralVenousBlood|Midazolam|Plasma (Peripheral Venous Blood)"
or "Midazolam 600mg SD|ObservedData|Peripheral Venous Blood|Plasma|Rifampin|Conc"}
}
\value{
A string corresponding to the `id` of a configuration plan observed dataset
}
\description{
Get an observed dataset id from a configuration plan quantity path
}
\examples{
\dontrun{
getObservedDataIdFromPath("S2|Organism|PeripheralVenousBlood|Midazolam|Plasma (Peripheral Venous Blood)")
# > NULL
getObservedDataIdFromPath("Midazolam 600mg SD|ObservedData|Peripheral Venous Blood|Plasma|Rifampin|Conc")
# > "Midazolam 600mg SD"
}
}
