% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getObservedDemographyAcrossPopulations}
\alias{getObservedDemographyAcrossPopulations}
\title{getObservedDemographyAcrossPopulations}
\usage{
getObservedDemographyAcrossPopulations(
  structureSets,
  demographyPaths,
  metaData
)
}
\arguments{
\item{structureSets}{List of `SimulationStructure` objects}

\item{demographyPaths}{Paths of demography variables to display}

\item{metaData}{List of display names and units of demography variables}
}
\value{
A data.frame
}
\description{
Get Observed Demography data across populations of simulation sets
}
\keyword{internal}
