% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getObservedDemographyFromSimulationSet}
\alias{getObservedDemographyFromSimulationSet}
\title{getObservedDemographyFromSimulationSet}
\usage{
getObservedDemographyFromSimulationSet(structureSet, demographyPaths, metaData)
}
\arguments{
\item{structureSet}{A `SimulationStructure` object}

\item{demographyPaths}{Names of demography variables to display}

\item{metaData}{Meta data of demography variables to display}
}
\value{
A data.frame
}
\description{
getObservedDemographyFromSimulationSet
}
\keyword{internal}
