% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getOutpuNamesFromSimulationSetIdentifer}
\alias{getOutpuNamesFromSimulationSetIdentifer}
\title{getOutpuNamesFromSimulationSetIdentifer}
\usage{
getOutpuNamesFromSimulationSetIdentifer(outputsId)
}
\arguments{
\item{outputsId}{Content translated from cell corresponding to outputs in SimulationSets sheet.
The `outputsId` are expected as `"c(outputName1, outputName2, etc.)"`}
}
\value{
names of Output objects in SimulationSets sheet
}
\description{
Get the names of output name from simulation set identifier
}
\keyword{internal}
