% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getOutputGroups}
\alias{getOutputGroups}
\title{getOutputGroups}
\usage{
getOutputGroups(metaDataFrame)
}
\arguments{
\item{metaDataFrame}{A data.frame with variables `path`, `dimension`, `unit`, `group`, `color`, `fill`}
}
\value{
List of data.frames data.frame with variables `path`, `dimension`, `unit`, `group`, `color`, `fill`
}
\description{
Group paths with same group IDs and check their units and residuals are consistent
}
\keyword{internal}
