% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getOutputNames}
\alias{getOutputNames}
\title{getOutputNames}
\usage{
getOutputNames(excelFile, outputSheet, simulationSetName)
}
\arguments{
\item{excelFile}{name of the Excel file from which the R script is created}

\item{outputSheet}{name of sheet defining an `Output` object}

\item{simulationSetName}{name of simulation set}
}
\value{
Name of `Output` object
}
\description{
Get the names of `Output` objets
}
