% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getOutputPathsInSimulationSet}
\alias{getOutputPathsInSimulationSet}
\title{getOutputPathsInSimulationSet}
\usage{
getOutputPathsInSimulationSet(simulationSet)
}
\arguments{
\item{simulationSet}{SimulationSet object or derived class}
}
\value{
Path names of outputs in `simulationSet`
}
\description{
getOutputPathsInSimulationSet
}
