% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pop-pk-parameters.R
\name{getOutputsForPKPlot}
\alias{getOutputsForPKPlot}
\title{getOutputsForPKPlot}
\usage{
getOutputsForPKPlot(structureSets)
}
\arguments{
\item{structureSets}{List of \code{SimulationStructure} objects}
}
\value{
list of \code{Output} objects
}
\description{
Get the list of outputs and their PK parameters for population PK parameter plot
}
\keyword{internal}
