% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\name{getOutputsFromConfigurationPlan}
\alias{getOutputsFromConfigurationPlan}
\title{getOutputsFromConfigurationPlan}
\usage{
getOutputsFromConfigurationPlan(configurationPlan)
}
\arguments{
\item{configurationPlan}{The configuration plan of a Qualification workflow read from json file.}
}
\value{
A dataframe of project, simulation, output paths and (if applicable) pk parameters and start and end times of interval over which the pk parameter is evaluated
}
\description{
Get a list of outputs from simulation and from `ConfigurationPlan`
}
\keyword{internal}
