% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{getOutputsTimeProfile}
\alias{getOutputsTimeProfile}
\title{getOutputsTimeProfile}
\usage{
getOutputsTimeProfile(plot)
}
\arguments{
\item{plot}{is a descriptor of the plot that is read from the `ConfigurationPlan` that is generated by the configuration plan json}
}
\value{
A vector of output paths that are to be used in the time profile plot descriptor `plot`
}
\description{
Get a vector of output paths that are to be used in a time profile plot
}
