% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mass-balance.R
\name{getPathsForMoleculeAmount}
\alias{getPathsForMoleculeAmount}
\title{getPathsForMoleculeAmount}
\usage{
getPathsForMoleculeAmount(simulation, compound)
}
\description{
Input is a simulation ("simulation") loaded from a PKML file using ospsuite::loadSimulation() and the name of a compound ("compound") in that simulation

Output is a list ("pth") of environments pointing to amounts of the compound in the model containers
}
