% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getPkAnalysisResultsFileNames}
\alias{getPkAnalysisResultsFileNames}
\title{getPkAnalysisResultsFileNames}
\usage{
getPkAnalysisResultsFileNames(workflow)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}
}
\value{
Names of the the expected PK analysis result files
}
\description{
Get the expected PK analysis result files obtained from a workflow
}
