% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPKParameterGroupsInOutput}
\alias{getPKParameterGroupsInOutput}
\title{getPKParametersInOutput}
\usage{
getPKParameterGroupsInOutput(output)
}
\arguments{
\item{output}{Output object}
}
\value{
Names of pkParameters in `output`
}
\description{
getPKParametersInOutput
}
\keyword{internal}
