% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getPKParameterMeasure}
\alias{getPKParameterMeasure}
\title{getPKParameterMeasure}
\usage{
getPKParameterMeasure(data, dataMapping)
}
\arguments{
\item{data}{A data.frame of PK Parameter values across Population Simulation sets}

\item{dataMapping}{A `BoxWhiskerDataMapping` object}
}
\value{
A data.frame summarizing the PK Parameter distribution by population
}
\description{
Get table summarizing the PK Parameter distribution by population
}
\keyword{internal}
