% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKParameterRatioMeasureFromMCSampling}
\alias{getPKParameterRatioMeasureFromMCSampling}
\title{getPKParameterRatioMeasureFromMCSampling}
\usage{
getPKParameterRatioMeasureFromMCSampling(
  comparisonData,
  referenceData,
  dataMapping,
  settings = NULL
)
}
\arguments{
\item{comparisonData}{A data.frame of PK Parameter values for Population to compare}

\item{referenceData}{A data.frame of PK Parameter values for reference Population}

\item{dataMapping}{A `BoxWhiskerDataMapping` object}

\item{settings}{A list of task settings}
}
\value{
A data.frame
}
\description{
Get PK Parameter Ratio Measure From Monte Carlo Sampling
}
\keyword{internal}
