% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKParameterRatioPKTable}
\alias{getPKParameterRatioPKTable}
\title{getPKParameterRatioPKTable}
\usage{
getPKParameterRatioPKTable(
  data,
  dataMapping,
  structureSets,
  referenceSimulationSetName,
  settings = NULL
)
}
\arguments{
\item{data}{A data.frame of PK Parameter values for across Populations}

\item{dataMapping}{A `BoxWhiskerDataMapping` object}

\item{structureSets}{`SimulationStructure` R6 class object}

\item{referenceSimulationSetName}{Name of reference simulation set}

\item{settings}{A list of task settings}
}
\description{
Get Data for Measure Table and Plots of PK Ratios
}
\keyword{internal}
