% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKParameterRatioTable}
\alias{getPKParameterRatioTable}
\title{getPKParameterRatioTable}
\usage{
getPKParameterRatioTable(pkParameter, outputPath, structureSets)
}
\arguments{
\item{pkParameter}{PK Parameter name}

\item{outputPath}{Quantity Path name}

\item{structureSets}{`SimulationStructure` R6 class object}
}
\description{
Get Table of summary statistics of PK parameter ratios across simulation sets
}
\keyword{internal}
