% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pop-pk-parameters.R
\name{getPKParametersAcrossPopulations}
\alias{getPKParametersAcrossPopulations}
\title{getPKParametersAcrossPopulations}
\usage{
getPKParametersAcrossPopulations(structureSets)
}
\arguments{
\item{structureSets}{list of \code{SimulationStructures} objects}
}
\value{
list of data.frame and its metaData including the values of PK parameters across Population Simulation sets
}
\description{
Get the values of PK parameters across Population Simulation sets
}
\keyword{internal}
