% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getPKParametersContent}
\alias{getPKParametersContent}
\title{getPKParametersContent}
\usage{
getPKParametersContent(pkParametersTable)
}
\arguments{
\item{pkParametersTable}{Data.frame read from the Excel sheet "PK Parameters"}
}
\value{
Character vector updating the PKParameters objects
}
\description{
Creates a character vector to be written in a workflow .R script updating the PKParameters objects
}
