% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getPKParametersInfoContent}
\alias{getPKParametersInfoContent}
\title{getPKParametersInfoContent}
\usage{
getPKParametersInfoContent(excelFile, pkParametersSheet)
}
\arguments{
\item{excelFile}{name of the Excel file from which the R script is created}

\item{pkParametersSheet}{name of the Excel sheet that defines the \code{pkParameters} information}
}
\value{
Character vector defining \code{pkParameters} input of \code{Output} object
}
\description{
Creates a character vector to be written in a workflow .R script defining \code{pkParameters} input of \code{Output} object.
}
\keyword{internal}
