% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getPKRatioForMapping}
\alias{getPKRatioForMapping}
\title{getPKRatioForMapping}
\usage{
getPKRatioForMapping(
  pkRatioMapping,
  pkParameterNames,
  configurationPlan,
  settings
)
}
\arguments{
\item{pkRatioMapping}{List providing the PK ratio mapping from field \code{PKRatios} of configuration plan}

\item{pkParameterNames}{Names of PK Parameters as defined by users}

\item{configurationPlan}{A \code{ConfigurationPlan} object}

\item{settings}{settings for the task}
}
\value{
list with \code{data} and \code{metaData}
}
\description{
Get data of pk ratio from field \code{PKRatios} of configuration plan
}
\keyword{internal}
