% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\name{getPKRatioOutputsDataframe}
\alias{getPKRatioOutputsDataframe}
\title{getPKRatioOutputsDataframe}
\usage{
getPKRatioOutputsDataframe(configurationPlan)
}
\arguments{
\item{configurationPlan}{The configuration plan of a Qualification workflow read from json file.}
}
\value{
A list containing data for generating DDI plots
}
\description{
Get a dataframe relating project, simulation, output path and pk parameter for each PK ratio plot
}
\keyword{internal}
