% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKRatioSummaryForDifferentPopulations}
\alias{getPKRatioSummaryForDifferentPopulations}
\title{getPKRatioSummaryForDifferentPopulations}
\usage{
getPKRatioSummaryForDifferentPopulations(structureSet, referenceSet, settings)
}
\arguments{
\item{structureSet}{A \code{SimulationStructure} object of the population to compare}

\item{referenceSet}{A \code{SimulationStructure} object of the reference population}
}
\value{
A data.frame of the PK Parameter ratios summary statistics
}
\description{
Calculate and save summary statistics of ratios of PK parameters if populations are identical
}
\keyword{internal}
