% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKRatioSummaryFromAnalyticalSolution}
\alias{getPKRatioSummaryFromAnalyticalSolution}
\title{getPKRatioSummaryFromAnalyticalSolution}
\usage{
getPKRatioSummaryFromAnalyticalSolution(
  pkData,
  referencePKData,
  simulationSetName
)
}
\arguments{
\item{pkData}{A data.frame of PK Parameter values for the Population to compare}

\item{referencePKData}{A data.frame of PK Parameter values for reference Population}

\item{simulationSetName}{Name of simulation set}
}
\value{
A data.frame of the PK Parameter ratios summary statistics
}
\description{
Get PK Ratio Summary from Analytical Solution
}
\keyword{internal}
