% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKRatioSummaryFromSettings}
\alias{getPKRatioSummaryFromSettings}
\title{getPKRatioSummaryFromSettings}
\usage{
getPKRatioSummaryFromSettings(pkData, referencePKData, settings)
}
\arguments{
\item{pkData}{A matrix of PK Parameter values for Population to compare}

\item{referencePKData}{A matrix of PK Parameter values for reference Population}

\item{settings}{A list of task settings}
}
\value{
A data.frame of the PK Parameter ratios summary statistics
}
\description{
Get PK Parameter Ratio Measure From Monte Carlo Sampling
}
\keyword{internal}
