% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKRatioSummaryStatistics}
\alias{getPKRatioSummaryStatistics}
\title{getPKRatioSummaryStatistics}
\usage{
getPKRatioSummaryStatistics(pkData, referencePKData)
}
\arguments{
\item{pkData}{A matrix of PK Parameter values for Population to compare}

\item{referencePKData}{A matrix of PK Parameter values for reference Population}
}
\value{
A matrix of the PK Parameter ratios summary statistics
}
\description{
Calculate and save summary statistics of ratios of PK parameters
Note that this function computes on matrix objects to be faster
than on data.frame when Monte Carlo simulation is performed
}
\keyword{internal}
