% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getPKResultsDataFrame}
\alias{getPKResultsDataFrame}
\title{getPKResultsDataFrame}
\usage{
getPKResultsDataFrame(pkParameterResultsFilePath, pkParameterSelection)
}
\arguments{
\item{pkParameterResultsFilePath}{Path to PK parameter results CSV file}

\item{pkParameterSelection}{PK parameters used for selection of individuals for whom to calculate sensitivity}
}
\value{
pkResultsDataFrame, a dataframe storing the contents of the CSV file with path pkParameterResultsFilePath
}
\description{
Read PK parameter results into a dataframe and set QuantityPath,Parameter and Unit columns as factors
}
