% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getPopSensDfForPkAndOutput}
\alias{getPopSensDfForPkAndOutput}
\title{getPopSensDfForPkAndOutput}
\usage{
getPopSensDfForPkAndOutput(
  simulation,
  sensitivityResultsFolder,
  indexDf,
  output,
  pkParameter,
  totalSensitivityThreshold
)
}
\arguments{
\item{simulation}{loaded from the used in the simulationFile path in the simulation set}

\item{sensitivityResultsFolder}{location of sensitivity analysis results}

\item{indexDf}{dataframe with contents of population sensitivity results index file}

\item{output}{pathID of output for which to obtain the population sensitivity results}

\item{pkParameter}{name of PK parameter for which to obtain the population sensitivity results}

\item{totalSensitivityThreshold}{cut-off used for plots of the most sensitive parameters}
}
\value{
sortedFilteredIndividualsDfForPKParameter dataframe of population-wide sensitivity results for pkParameter and output
}
\description{
Retrieve dataframe of ranked and filtered population sensitivity results for a given PK parameter and model output pathID
}
