% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getPopulationPKAnalysesFromOuptut}
\alias{getPopulationPKAnalysesFromOuptut}
\title{getPopulationPKAnalysesFromOuptut}
\usage{
getPopulationPKAnalysesFromOuptut(
  data,
  metaData,
  output,
  pkParameter,
  molWeight = NULL
)
}
\arguments{
\item{data}{data.frame of the PK Analyses across Population Simulation sets}

\item{metaData}{metaData (dimension and unit) of the PK Analyses across Population Simulation sets}

\item{output}{An `Output ` object}

\item{pkParameter}{`pkParameter` from `Output ` object}

\item{molWeight}{Molecular weight of compound (if unit conversion needed)}
}
\value{
list of data.frame and its metaData including the values of PK parameters specified by `pkParameter` and `Output` objects
}
\description{
Get the values of PK parameters specified by an `Output` object from a data.frame
}
\keyword{internal}
