% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getPopulationResultsFromOutput}
\alias{getPopulationResultsFromOutput}
\title{getPopulationResultsFromOutput}
\usage{
getPopulationResultsFromOutput(
  simulationPathResults,
  output,
  simulationQuantity,
  molWeight,
  simulationSet,
  settings = NULL
)
}
\arguments{
\item{simulationPathResults}{list with simulated data included}

\item{output}{An `Output` object}

\item{simulationQuantity}{Dimension/quantity for unit conversion of dependent variable}

\item{molWeight}{Molar weight for unit conversion of dependent variable}

\item{simulationSet}{`SimulationSet` object}

\item{settings}{TaskSetting object}
}
\value{
list of data and metaData
}
\description{
Get simulated population data from an Output object
}
\keyword{internal}
