% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getPopulationSensitivityAnalysisResultsFileNames}
\alias{getPopulationSensitivityAnalysisResultsFileNames}
\title{getPopulationSensitivityAnalysisResultsFileNames}
\usage{
getPopulationSensitivityAnalysisResultsFileNames(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` object or derived class}
}
\value{
Names of the the expected sensitivity analysis result files
}
\description{
Get the expected sensitivity analysis result files obtained from a population workflow
}
