% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getQualificationDDIPlotData}
\alias{getQualificationDDIPlotData}
\title{getQualificationDDIPlotData}
\usage{
getQualificationDDIPlotData(configurationPlan)
}
\arguments{
\item{configurationPlan}{The configuration plan of a Qualification workflow read from json file.}
}
\value{
plotDDIdata, a list of lists of the form list(dataframe,metadata) specific to each DID plot
}
\description{
Build dataframes and metadata for each DDI plot
}
\keyword{internal}
