% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getQualificationDDIRatioMeasure}
\alias{getQualificationDDIRatioMeasure}
\title{getQualificationDDIRatioMeasure}
\usage{
getQualificationDDIRatioMeasure(summaryDataFrame, pkParameterName, delta)
}
\arguments{
\item{summaryDataFrame}{data.frame with DDI Ratios}

\item{pkParameterName}{Name of PK Parameter as defined by users}

\item{delta}{Delta value from Guest et al. formula}
}
\value{
A data.frame
}
\description{
Get qualification measure of DDI ratio from field `DDIRatioPlots` of configuration plan
}
\keyword{internal}
