% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{getQualificationGOFData}
\alias{getQualificationGOFData}
\title{getQualificationGOFData}
\usage{
getQualificationGOFData(gofPlan, configurationPlan, axesUnits)
}
\arguments{
\item{gofPlan}{List providing the mapping of observed and simulated data}

\item{configurationPlan}{A `ConfigurationPlan` object}

\item{axesUnits}{list of axes properties obtained from `getGOFAxesUnits`}
}
\value{
list with `data` and `metaData`
}
\description{
Get data of goodness of fit from field `GOFMergedPlots` of configuration plan
}
\keyword{internal}
