% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{getQualificationGOFGMFE}
\alias{getQualificationGOFGMFE}
\title{getQualificationGOFGMFE}
\usage{
getQualificationGOFGMFE(data)
}
\arguments{
\item{data}{data.frame with residuals values}
}
\value{
A data.frame
}
\description{
Get Geometric Mean Fold Error for GOF plots
}
\keyword{internal}
