% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{getQualificationGOFPlotData}
\alias{getQualificationGOFPlotData}
\title{getQualificationGOFPlotData}
\usage{
getQualificationGOFPlotData(configurationPlan)
}
\arguments{
\item{configurationPlan}{A `ConfigurationPlan` object}
}
\value{
plotGOFdata, a list of lists of the form list(dataframe,metadata) specific to each GOF plot
}
\description{
Build dataframes and metadata for each GOF plot
}
\keyword{internal}
