% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioData}
\alias{getQualificationPKRatioData}
\title{getQualificationPKRatioData}
\usage{
getQualificationPKRatioData(pkRatioPlan, configurationPlan, settings)
}
\arguments{
\item{pkRatioPlan}{List providing the PK ratio mapping from field \code{PKRatioPlots} of configuration plan}

\item{configurationPlan}{A \code{ConfigurationPlan} object}

\item{settings}{settings for the task}
}
\value{
list with \code{data} and \code{metaData}
}
\description{
Get data of pk ratio from field \code{PKRatioPlots} of configuration plan
}
\keyword{internal}
