% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioGMFE}
\alias{getQualificationPKRatioGMFE}
\title{getQualificationPKRatioGMFE}
\usage{
getQualificationPKRatioGMFE(pkParameterNames, data, settings)
}
\arguments{
\item{data}{data.frame with PK Ratios}

\item{settings}{settings for the task}

\item{pkParameterName}{Name of PK Parameter as defined by users}

\item{metaData}{metaData with units and dimension for labeling the table header}
}
\value{
A data.frame
}
\description{
Get plot of pk ratio from field `PKRatioPlots` of configuration plan
}
