% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioMeasure}
\alias{getQualificationPKRatioMeasure}
\title{getQualificationPKRatioMeasure}
\usage{
getQualificationPKRatioMeasure(pkParameterName, data, metaData, settings)
}
\arguments{
\item{pkParameterName}{Name of PK Parameter as defined by users}

\item{data}{data.frame with PK Ratios}

\item{metaData}{metaData with units and dimension for labeling the table header}

\item{settings}{settings for the task}
}
\value{
A data.frame
}
\description{
Get plot of pk ratio from field `PKRatioPlots` of configuration plan
}
