% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioPlot}
\alias{getQualificationPKRatioPlot}
\title{getQualificationPKRatioPlot}
\usage{
getQualificationPKRatioPlot(pkParameterName, data, metaData, axesProperties)
}
\arguments{
\item{pkParameterName}{Name of PK Parameter as defined by users}

\item{data}{data.frame with PK Ratios}

\item{metaData}{metaData with units and dimension for labeling the table header}

\item{axesProperties}{list of axes properties obtained from `getAxesProperties`}
}
\value{
A ggplot object
}
\description{
Get plot of pk ratio from field `PKRatioPlots` of configuration plan
}
\keyword{internal}
