% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioTable}
\alias{getQualificationPKRatioTable}
\title{getQualificationPKRatioTable}
\usage{
getQualificationPKRatioTable(data, metaData)
}
\arguments{
\item{data}{data.frame with PK Ratios}

\item{metaData}{metaData with units and dimension for labeling the table header}
}
\value{
A data.frame with correct numeric format
}
\description{
Get data of pk ratio from field `PKRatioPlots` of configuration plan
}
