% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getRatioMeasureAnalyticalSolution}
\alias{getRatioMeasureAnalyticalSolution}
\title{getRatioMeasureAnalyticalSolution}
\usage{
getRatioMeasureAnalyticalSolution(x, y)
}
\arguments{
\item{x}{Array of PK parameter numeric values}

\item{y}{Array of PK parameter numeric values}
}
\value{
A list including geoMean, geoSD and geoCV
}
\description{
Get Ratio Measure Analytical Solution
}
\keyword{internal}
