% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getReaderFunction}
\alias{getReaderFunction}
\title{getReaderFunction}
\usage{
getReaderFunction(fileName, nlines = 2)
}
\arguments{
\item{fileName}{Name of file to be read}

\item{nlines}{Number of lines to look at for checking consistency within file.
Note that, unlike \code{read.csv}, \code{count.fields} does not have options for handling encoding or escape characters.
Thus, using \code{nlines} as low as possible reduces the chances of inconsistent column widths caused by such characters.}
}
\value{
Reader function such as \code{read.csv}
}
\description{
Get most appropriate reader function by guessing file separator
File separator is guessed by checking number of fields/columns along lines
}
\keyword{internal}
