% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{getReferencePopulationName}
\alias{getReferencePopulationName}
\title{getReferencePopulationName}
\usage{
getReferencePopulationName(structureSets)
}
\arguments{
\item{structureSets}{An array of \code{SimulationStructure} objects}
}
\description{
Get Simulation Set Name from reference Population Simulation Set
}
\keyword{internal}
