% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-info.R
\name{getReportingEngineInfo}
\alias{getReportingEngineInfo}
\title{getReportingEngineInfo}
\usage{
getReportingEngineInfo()
}
\value{
list of user information on reporting engine
}
\description{
Get relevant information regarding the Reporting Engine
}
