% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{getRESettings}
\alias{getRESettings}
\title{getRESettings}
\usage{
getRESettings(settingName)
}
\arguments{
\item{settingName}{setting name as defined in enum `reSettingsNames`}
}
\description{
Get properties from name of the default/global settings stored in `reEnv`.
}
