% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getResiduals}
\alias{getResiduals}
\title{getResiduals}
\usage{
getResiduals(observedData, simulatedData, metaData = NULL, dataMapping = NULL)
}
\arguments{
\item{observedData}{data.frame of time profile observed data}

\item{simulatedData}{data.frame of time profile simulated data}

\item{metaData}{meta data on `data`}

\item{dataMapping}{XYDataMapping class object mapping what variable is to compare}
}
\value{
residualsData data.frame with Time, Observed, Simulated, Residuals
}
\description{
This function may be reshape to be more generic later on
Currently, the input variable data is a data.frame with "Time", "Concentration" and "Legend"
The function get the simulated data with the time the closest to the observed data times
}
