% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getResidualsPlotResultsInGroup}
\alias{getResidualsPlotResultsInGroup}
\title{getResidualsPlotResultsInGroup}
\usage{
getResidualsPlotResultsInGroup(
  data,
  metaData,
  outputId,
  structureSet = NULL,
  settings = NULL
)
}
\arguments{
\item{data}{A data.frame of residuals data}

\item{metaData}{metaData represented as a data.frame}

\item{outputId}{Output identifier to provide unique id name}

\item{structureSet}{A \code{SimulationStructure} object or \code{NULL} if performing residuals across simulations}

\item{settings}{Optional settings for the plots. In particular, includes reference data for population time profile.}
}
\value{
List of \code{plots}, their \code{captions} and \code{data} to export
}
\description{
Get plots and their captions for residuals
}
\keyword{internal}
