% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{getScheduleOfPlots}
\alias{getScheduleOfPlots}
\title{getScheduleOfPlots}
\usage{
getScheduleOfPlots(configurationPlan)
}
\arguments{
\item{configurationPlan}{is a `ConfigurationPlan` object that is generated by the configuration plan json}
}
\value{
A dataframe describing the the type of each plot, the order of each plot among its plot type, and the section id of each plot
}
\description{
Get a dataframe of plots from the `ConfigurationPlan` object `configurationPlan` that is generated by the configuration plan json
}
