% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{getScheduleOfPlotsOfType}
\alias{getScheduleOfPlotsOfType}
\title{getScheduleOfPlotsOfType}
\usage{
getScheduleOfPlotsOfType(pltList, plotType)
}
\arguments{
\item{pltList}{is a list of plots that is to be generated, all of the plot type `plotType`}

\item{plotType}{describes the type of the plots described in `pltList`}
}
\value{
A dataframe describing the the the order of each plot among its plot type `plotType`, and the section id of each plot
}
\description{
Get a dataframe of plots of the plot type `plotType` to be generated from the `ConfigurationPlan` object that is generated by the configuration plan json
}
