% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getScriptDocumentation}
\alias{getScriptDocumentation}
\title{getScriptDocumentation}
\usage{
getScriptDocumentation(excelFile, colSep = "\\t")
}
\arguments{
\item{excelFile}{name of the Excel file from which the R script is created}

\item{colSep}{Displayed column separator written in script}
}
\value{
Character vector translating the Documentation sheet
}
\description{
Get R script from Documentation sheet
}
\keyword{internal}
