% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getSelectedRows}
\alias{getSelectedRows}
\title{getSelectedRows}
\usage{
getSelectedRows(data, dataSelection)
}
\arguments{
\item{data}{A data.frame}

\item{dataSelection}{Character string or expression evaluated to select data
The enum helper `DataSelectionKeys` provides keys for selected all or none of the data}
}
\value{
A data.frame of selected data
}
\description{
Get selected rows from data and its selection
The function leverage `dplyr::filter` to select the rows
}
\examples{
data <- data.frame(
x = seq(0,9),
y = seq(10,19),
mdv = c(1,1, rep(0, 8)),
groups = rep(c("A", "B"), 5)
)

# Select all the rows
getSelectedRows(data, DataSelectionKeys$ALL)

# Select no row
getSelectedRows(data, DataSelectionKeys$NONE)

# Select rows from group A
getSelectedData(data, "groups \%in\% 'A'")

# Get rows of missing dependent variable (mdv)
getSelectedRows(data, "mdv == 0")

}
\seealso{
DataSelectionKeys
}
