% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getSensitivityAnalysisResultsFileNames}
\alias{getSensitivityAnalysisResultsFileNames}
\title{getSensitivityAnalysisResultsFileNames}
\usage{
getSensitivityAnalysisResultsFileNames(workflow)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}
}
\value{
Names of the the expected sensitivity analysis result files
}
\description{
Get the expected sensitivity analysis result files obtained from a workflow
}
