% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getSensitivityDataFrameForIndividuals}
\alias{getSensitivityDataFrameForIndividuals}
\title{getSensitivityDataFrameForIndividuals}
\usage{
getSensitivityDataFrameForIndividuals(
  indexDf,
  structureSet,
  pkParameter,
  output
)
}
\arguments{
\item{indexDf}{of population distribution of pkParameter, used to select individuals for sensitivity analysis}

\item{structureSet}{`SimulationStructure` R6 class object}

\item{pkParameter}{name of PK parameter for which to obtain the population sensitivity results}

\item{output}{pathID of output for which to obtain the population sensitivity results}
}
\value{
individualsDfForPKParameter dataframe of population-wide sensitivity results for pkParameter and output
}
\description{
retrieve the dataframe of individuals' sensitivity analysis results for given output and pkParameter
}
