% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getSensitivityVariableParameterPaths}
\alias{getSensitivityVariableParameterPaths}
\title{getSensitivityVariableParameterPaths}
\usage{
getSensitivityVariableParameterPaths(excelFile, sensitivityParametersSheet)
}
\arguments{
\item{excelFile}{name of the Excel file from which the R script is created}

\item{sensitivityParametersSheet}{name of sheet defining the parameters to vary}
}
\value{
Text of vector of paths to vary
}
\description{
Get the variable paths for sensitivity analysis settings
}
\keyword{internal}
