% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getSimulationDescriptor}
\alias{getSimulationDescriptor}
\title{getSimulationDescriptor}
\usage{
getSimulationDescriptor(workflow)
}
\arguments{
\item{workflow}{A `Workflow` object}
}
\value{
character describing simulation sets
}
\description{
Get workflow simulation set descriptor
}
