% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-user-helpers.R
\name{getSimulationParameterDisplayPaths}
\alias{getSimulationParameterDisplayPaths}
\title{getSimulationParameterDisplayPaths}
\usage{
getSimulationParameterDisplayPaths(parameterPaths, simulation, dictionary)
}
\arguments{
\item{parameterPaths}{Paths of a parameter in simulation}

\item{simulation}{A `Simulation` object from `ospsuite`}

\item{dictionary}{A data.frame mapping `parameterPaths` to `parameterDisplayPaths`}
}
\value{
An array of `parameterDisplayPaths` matching the `parameterPaths`
}
\description{
getSimulationParameterDisplayPaths
}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemographyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemographyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemographyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemographyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setXParametersForDemographyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemographyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
