% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getSimulationResultFileNames}
\alias{getSimulationResultFileNames}
\title{getSimulationResultFileNames}
\usage{
getSimulationResultFileNames(workflow)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}
}
\value{
Names of the the expected simulation result files
}
\description{
Get the expected simulation result files obtained from a workflow
}
