% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getSimulationSetType}
\alias{getSimulationSetType}
\title{getSimulationSetType}
\usage{
getSimulationSetType(workflowMode)
}
\arguments{
\item{workflowMode}{Either `PopulationWorkflow` or `MeanModelWorkflow`}
}
\value{
`PopulationSimulationSet` or `SimulationSet` as character
}
\description{
Output the SimulationSet object matching the appropriate workflow type
}
\keyword{internal}
