% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{getSimulationTimeRanges}
\alias{getSimulationTimeRanges}
\title{getSimulationTimeRanges}
\usage{
getSimulationTimeRanges(simulation, path, simulationSet)
}
\arguments{
\item{simulation}{A \code{Simulation} object}

\item{path}{Field \code{path} from \code{Output} object}

\item{simulationSet}{A \code{SimulationSet} or \code{PopulationSimulationSet} object}
}
\value{
Lists including \code{values} and \code{name} of time ranges.
Also includes logical field \code{keep} to define if a specific application range is kept in report.
}
\description{
Get time ranges for time profile plots according to applications and user defined settings
}
\keyword{internal}
