% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getSmartZoomLimits}
\alias{getSmartZoomLimits}
\title{getSmartZoomLimits}
\usage{
getSmartZoomLimits(dataVector, residualsVsObserved = FALSE)
}
\arguments{
\item{dataVector}{a numeric vector containing the x or y values of datapoints to be plotted}

\item{residualsVsObserved}{A logical values indicating if DDI Ratios are plotted as residuals vs observed}
}
\value{
a list containing the minimum and maximum axes limits
}
\description{
Get default axis limits for DDI Ratio plot
}
\keyword{internal}
