% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getSplitPositions}
\alias{getSplitPositions}
\title{getSplitPositions}
\usage{
getSplitPositions(possibleSplits, splitWidth, numberOfSplits)
}
\arguments{
\item{possibleSplits}{Positions where a space or a dash was flaged}

\item{splitWidth}{Maximum number of characters desired per lines}

\item{numberOfSplits}{Maximum number of line breaks to use}
}
\value{
Position where to insert a line break character
}
\description{
Algorithm that gets positions where splitting a character string for sensible line breaks
}
\keyword{internal}
