% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{getStatisticsFromType}
\alias{getStatisticsFromType}
\title{getStatisticsFromType}
\usage{
getStatisticsFromType(statisticsType)
}
\arguments{
\item{statisticsType}{Statistics summarizing time profile simulated data
as defined by helper enum `StatisticsType`}
}
\value{
A list including `y`, `ymin` and `ymax` summary statistics as well as their `caption`
}
\description{
Get statistics
}
\examples{
\dontrun{
getStatisticsFromType(StatisticsTypes$`Arithmetic mean`)
}

}
