% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getTaskInputs}
\alias{getTaskInputs}
\title{getTaskInputs}
\usage{
getTaskInputs(task)
}
\arguments{
\item{task}{`Task` object or derived class}
}
\value{
Names of the files required to perform the task
}
\description{
Get the names of the files required to perform the task
}
