% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\name{getTimeFromPlan}
\alias{getTimeFromPlan}
\title{getTimeFromPlan}
\usage{
getTimeFromPlan(configurationPlanField, fieldName)
}
\arguments{
\item{configurationPlanField}{The configuration plan field listing \code{endTime} value and unit}

\item{fieldName}{The name of the field to extract the time from (e.g., "EndTime")}
}
\value{
\code{endTime} value in base unit
}
\description{
Get endTime value in base unit from configuration plan field
}
\keyword{internal}
