% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getTimeProfilePlotConfiguration}
\alias{getTimeProfilePlotConfiguration}
\title{getTimeProfilePlotConfiguration}
\usage{
getTimeProfilePlotConfiguration(
  workflowType,
  group,
  data,
  metaData,
  observedData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{workflowType}{Workflow type, either `"mean"` or `"population"`}

\item{group}{A data.frame mapping properties to output groups}

\item{data}{A data.frame}

\item{metaData}{List of metaData defining dimensions and units in the data.frame}

\item{dataMapping}{List mapping x, y and color variables to `data`}

\item{plotConfiguration}{A user-defined `TimeProfilePlotConfiguration` object}
}
\value{
A `TimeProfilePlotConfiguration` object
}
\description{
Define a `TimeProfilePlotConfiguration` object
}
\keyword{internal}
