% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getTimeProfilePlotResults}
\alias{getTimeProfilePlotResults}
\title{getTimeProfilePlotResults}
\usage{
getTimeProfilePlotResults(
  workflowType,
  timeRange,
  simulatedData,
  observedData = NULL,
  metaDataFrame,
  timeProfileMapping,
  structureSet,
  settings = NULL
)
}
\arguments{
\item{workflowType}{\code{"mean"} or \code{"population"} workflow}

\item{timeRange}{array of time values defining range of simulated data.
Note that \code{timeRange} accounts for user defined \code{timeOffset}}

\item{simulatedData}{data.frame of simulated data}

\item{observedData}{data.frame of observed data}

\item{metaDataFrame}{metaData represented as a data.frame}

\item{timeProfileMapping}{list of variables to map in the time profile plots}

\item{structureSet}{A \code{SimulationStructure} object}

\item{settings}{Optional settings for the plots. In particular, includes reference data for population time profile.}
}
\value{
List of \code{plots} and their \code{captions}
}
\description{
Get plots and their captions for mean or population time profiles
}
\keyword{internal}
