% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getTimeTicksFromUnit}
\alias{getTimeTicksFromUnit}
\title{getTimeTicksFromUnit}
\usage{
getTimeTicksFromUnit(unit, timeValues = NULL)
}
\arguments{
\item{unit}{A time unit as defined in `ospsuite::ospUnits$Time`}

\item{timeValues}{Numeric values used by the data}
}
\value{
List of `ticks` and their `ticklabels`
}
\description{
Defines auto time ticks from time unit and time values
}
\keyword{internal}
