% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getTitleInfo}
\alias{getTitleInfo}
\title{getTitleInfo}
\usage{
getTitleInfo(fileContent, titlePattern = "#", titleLevels = 6)
}
\arguments{
\item{fileContent}{Content of a markdown or text file read as an array of character strings}

\item{titlePattern}{character pattern referencing titles in first element of line}

\item{titleLevels}{levels of titles in the report}
}
\value{
List of title information including \code{line}, \code{content}, \code{reference}, \code{count}, \code{level}
}
\description{
Get section titles information from report content
}
\keyword{internal}
