% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getUserDefPKParametersContent}
\alias{getUserDefPKParametersContent}
\title{getUserDefPKParametersContent}
\usage{
getUserDefPKParametersContent(userDefPKParametersTable)
}
\arguments{
\item{userDefPKParametersTable}{Data.frame read from the Excel sheet "PK Parameters"}
}
\value{
A list of script content, associated with its potential warnings and errors for updating the PKParameters objects
}
\description{
Creates a character vector to be written in a workflow .R script updating the PKParameters objects
}
\keyword{internal}
