% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getWorkflowContent}
\alias{getWorkflowContent}
\title{getWorkflowContent}
\usage{
getWorkflowContent(workflowTable)
}
\arguments{
\item{workflowTable}{Data.frame read from the Excel sheet "Workflow and Tasks"}
}
\value{
Character vector defining the `Workflow` object
}
\description{
Creates a character vector to be written in a workflow .R script defining `Workflow` object
}
