% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWorkflowParameterDisplayPaths}
\alias{getWorkflowParameterDisplayPaths}
\title{getWorkflowParameterDisplayPaths}
\usage{
getWorkflowParameterDisplayPaths(workflow)
}
\arguments{
\item{workflow}{Object of class `MeanModelWorkflow` or `PopulationWorkflow`}
}
\description{
Get mapping between parameters and their display paths in a workflow
to replace standard display of parameter paths.
}
