% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getXParametersForDemographyPlot}
\alias{getXParametersForDemographyPlot}
\title{getXParametersForDemographyPlot}
\usage{
getXParametersForDemographyPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of x parameters used for demography range plots
}
\description{
getXParametersForDemographyPlot
}
