% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getXParametersForPkParametersPlot}
\alias{getXParametersForPkParametersPlot}
\title{getXParametersForPkParametersPlot}
\usage{
getXParametersForPkParametersPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of x-parameters used for PK parameters range plots
}
\description{
getXParametersForPkParametersPlot
}
