% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{getYParametersForDemogrpahyPlot}
\alias{getYParametersForDemogrpahyPlot}
\title{getYParametersForDemogrpahyPlot}
\usage{
getYParametersForDemogrpahyPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of y parameters used for demography histogram and range plots
}
\description{
getYParametersForDemogrpahyPlot
}
