% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\docType{data}
\name{gofPlotAxesSettings}
\alias{gofPlotAxesSettings}
\title{Names of fields in configuration plane containing axes settings data for each GOF plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
gofPlotAxesSettings
}
\description{
Names of fields in configuration plane containing axes settings data for each GOF plot type
}
\keyword{internal}
