% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-settings.R
\name{GofTaskSettings}
\alias{GofTaskSettings}
\title{GofTaskSettings}
\description{
R6 class defining properties of time profiles and residuals plot task settings
}
\keyword{internal}
\section{Super class}{
\code{\link[ospsuite.reportingengine:TaskSettings]{ospsuite.reportingengine::TaskSettings}} -> \code{GofTaskSettings}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{referenceData}}{Data results obtained by TimeProfilesAndResiduals task corresponding to referencePopulation}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{includeReferenceData}}{logical defining if reference population should be included in
time profiles and residual plots when applicable}

\item{\code{outputSelections}}{is a subset of paths of all outputs belonging to a simulation set, to be used to generate a plot with only selected outputs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GofTaskSettings$new()}}
\item \href{#method-clone}{\code{GofTaskSettings$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a `GofTaskSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GofTaskSettings$new(
  taskName = AllAvailableTasks$plotTimeProfilesAndResiduals,
  outputSelections = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{taskName}}{name of the task using the settings}

\item{\code{outputSelections}}{subset of simulationSet outputs to be used in GoF plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `GofTaskSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GofTaskSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
