% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{hasAnchor}
\alias{hasAnchor}
\title{hasAnchor}
\usage{
hasAnchor(tag)
}
\arguments{
\item{tag}{Character string}
}
\value{
A logical
}
\description{
Check if a character string includes an anchor tag
}
\examples{
# Flags both anchors using id or name
hasAnchor('<a id="section-1"></a>')
hasAnchor('<a name="section-1"></a>')

hasAnchor("# section 1")

}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{getFigurePathsFromReport}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
