% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{initializeRmdFile}
\alias{initializeRmdFile}
\title{initializeRmdFile}
\usage{
initializeRmdFile(
  fileName,
  title = "",
  outputFormat = "html_document",
  numberSections = TRUE
)
}
\arguments{
\item{fileName}{name of Rmd file to initialize}

\item{title}{title of the document}

\item{outputFormat}{output format of the document}

\item{numberSections}{logical to get numbered sections}
}
\description{
Initialize a R markdown document by Writing its header in YAML
}
