% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{introToYamlHeader}
\alias{introToYamlHeader}
\title{introToYamlHeader}
\usage{
introToYamlHeader(introContent)
}
\arguments{
\item{introContent}{Character array of the intro content}
}
\value{
A character array of yaml content
}
\description{
Translate an markdown introduction file into yaml header
In order to include introduction before the table of content,
it needs to be included as cover page features through a yaml header.
A yaml header provides additional arguments to pandoc when translating the md report.
Cover page features can be created with each their own style in the reference doc
}
\keyword{internal}
