% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isDimension}
\alias{isDimension}
\title{Check if the provided values are included all available dimensions}
\usage{
isDimension(values)
}
\arguments{
\item{values}{Vector of dimensions}
}
\value{
TRUE if the values are included all available dimensions
}
\description{
Check if the provided values are included all available dimensions
}
