% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isFileExtension}
\alias{isFileExtension}
\title{Check if the provided path has required extension}
\usage{
isFileExtension(file, extension)
}
\arguments{
\item{extension}{extension of the file required after "."}

\item{path}{(character) file or path name to be checked}
}
\value{
TRUE if the path includes the extension
}
\description{
Check if the provided path has required extension
}
