% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isLoadedPackage}
\alias{isLoadedPackage}
\title{isLoadedPackage}
\usage{
isLoadedPackage(packageName)
}
\arguments{
\item{packageName}{Name of the package}
}
\description{
Check that a R package has been successfully loaded
}
