% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{isSamePopulationInSets}
\alias{isSamePopulationInSets}
\title{isSamePopulationInSets}
\usage{
isSamePopulationInSets(structureSets)
}
\arguments{
\item{structureSets}{A `SimulationStructure`}
}
\value{
list
}
\description{
Check wither simulation sets use same population as reference set
}
\keyword{internal}
