% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-info.R
\name{isValidatedSystem}
\alias{isValidatedSystem}
\title{isValidatedSystem}
\usage{
isValidatedSystem()
}
\value{
logical assessing if system is validated
}
\description{
Check if environment variable is defined and as value \code{'prodrun'}
}
