% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lastPathElement}
\alias{lastPathElement}
\title{lastPathElement}
\usage{
lastPathElement(path)
}
\arguments{
\item{path}{simulation path}
}
\value{
last path element as character string
}
\description{
lastPathElement
}
