% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadCalculatePKParametersTask}
\alias{loadCalculatePKParametersTask}
\title{loadCalculatePKParametersTask}
\usage{
loadCalculatePKParametersTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `calculatePKParameters` task}
}
\value{
A `CalculatePKParametersTask` object
}
\description{
Define `calculatePKParameters` task and its settings
}
