% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadCalculateSensitivityTask}
\alias{loadCalculateSensitivityTask}
\title{loadCalculateSensitivityTask}
\usage{
loadCalculateSensitivityTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `calculateSensitivity` task}
}
\value{
A `PopulationSensitivityAnalysisTask` for `PopulationWorkflow` objects.
A `SensitivityAnalysisTask` object otherwise
}
\description{
Define `calculateSensitivity` task and its settings
}
