% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{loadConfigurationPlan}
\alias{loadConfigurationPlan}
\title{loadConfigurationPlan}
\usage{
loadConfigurationPlan(jsonFile, workflowFolder)
}
\arguments{
\item{jsonFile}{name of the JSON snapshot file including the configuration plan}

\item{workflowFolder}{path of the output folder created or used by the Workflow.}
}
\value{
A `ConfigurationPlan` object including the content of json file
}
\description{
Load a `ConfigurationPlan` object from json file
}
