% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadGOFMergedTask}
\alias{loadGOFMergedTask}
\title{loadGOFMergedTask}
\usage{
loadGOFMergedTask(workflow, configurationPlan)
}
\arguments{
\item{workflow}{`QualificationWorkflow` object}

\item{configurationPlan}{A `ConfigurationPlan` object}
}
\value{
A `QualificationTask` object
}
\description{
loadGOFMergedTask
}
