% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{loadObservedDataFromSimulationSet}
\alias{loadObservedDataFromSimulationSet}
\title{loadObservedDataFromSimulationSet}
\usage{
loadObservedDataFromSimulationSet(simulationSet, logFolder)
}
\arguments{
\item{simulationSet}{A `SimulationSet` object}

\item{logFolder}{folder where the logs are saved}
}
\value{
list of data and dataMapping
}
\description{
Load observed data and its dataMapping from a simulationSet
}
