% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{loadPackageOnCores}
\alias{loadPackageOnCores}
\title{loadPackageOnCores}
\usage{
loadPackageOnCores(packageName)
}
\arguments{
\item{packageName}{Name of package to be loaded}
}
\description{
Load \code{packageName} namespace to core, load the library and check that it has been loaded successfully
}
\keyword{internal}
