% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{loadPKAnalysesFromCSV}
\alias{loadPKAnalysesFromCSV}
\title{loadPKAnalysesFromCSV}
\usage{
loadPKAnalysesFromCSV(filePath, simulation, to = "PKAnalyses")
}
\arguments{
\item{filePath}{Full path of the file containing the PK-Analyses to load}

\item{simulation}{A `Simulation` object}

\item{to}{Format of the loaded output`}
}
\value{
A `PKAnalyses` object if `to="PKAnalyses"`
A `data.frame` if `to="data.frame"`
A `tibble` if `to="tibble"`
}
\description{
Load PK analyses from a CSV file
Wrap the `ospsuite::importPKAnalysesFromCSV()` to provide more useful warning messages.
}
\keyword{internal}
