% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{loadPKAnalysesFromStructureSet}
\alias{loadPKAnalysesFromStructureSet}
\title{loadPKAnalysesFromStructureSet}
\usage{
loadPKAnalysesFromStructureSet(
  structureSet,
  to = "PKAnalyses",
  useCache = FALSE
)
}
\arguments{
\item{structureSet}{A \code{SimulationStructure} object}

\item{to}{Format of the loaded output`}
}
\value{
A \code{PKAnalyses} object if \code{to="PKAnalyses"}
A \code{data.frame} if \code{to="data.frame"}
A \code{tibble} if \code{to="tibble"}
}
\description{
loadPKAnalysesFromStructureSet
}
\keyword{internal}
