% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotAbsorptionTask}
\alias{loadPlotAbsorptionTask}
\title{loadPlotAbsorptionTask}
\usage{
loadPlotAbsorptionTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`MeanModelWorkflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `plotAbsorption` task}
}
\value{
A `PlotTask` object
}
\description{
Define `plotAbsorption` task and its settings
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- MeanModelWorkflow$new(workflowFolder, simulationSets)

# (Re)load a default plotAbsorption task for workflow
myWorkflow$plotAbsorption <- loadPlotAbsorptionTask(
  workflow = myWorkflow,
  active = TRUE
)

# Load a user-defined simulation task for workflow
myWorkflow$userDefinedTasks[["plotAbsorption"]] <- loadPlotAbsorptionTask(
  workflow = myWorkflow,
  active = TRUE
)
}

}
\seealso{
Other workflow tasks: 
\code{\link{GofPlotTask}},
\code{\link{PlotTask}},
\code{\link{PopulationPlotTask}},
\code{\link{PopulationSensitivityAnalysisTask}},
\code{\link{QualificationTask}},
\code{\link{SensitivityAnalysisTask}},
\code{\link{Task}},
\code{\link{addUserDefinedTask}()},
\code{\link{loadCalculatePKParametersTask}()},
\code{\link{loadCalculateSensitivityTask}()},
\code{\link{loadGOFMergedTask}()},
\code{\link{loadPlotDDIRatioTask}()},
\code{\link{loadPlotDemographyTask}()},
\code{\link{loadPlotMassBalanceTask}()},
\code{\link{loadPlotPKParametersTask}()},
\code{\link{loadPlotPKRatioTask}()},
\code{\link{loadPlotSensitivityTask}()},
\code{\link{loadPlotTimeProfilesAndResidualsTask}()},
\code{\link{loadQualificationComparisonTimeProfileTask}()},
\code{\link{loadQualificationTimeProfilesTask}()},
\code{\link{loadSimulateTask}()}
}
\concept{workflow tasks}
