% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotDDIRatioTask}
\alias{loadPlotDDIRatioTask}
\title{loadPlotDDIRatioTask}
\usage{
loadPlotDDIRatioTask(workflow, configurationPlan)
}
\arguments{
\item{workflow}{`QualificationWorkflow` object}

\item{configurationPlan}{A `ConfigurationPlan` object}
}
\value{
A `QualificationTask` object
}
\description{
Define `plotDDIRatio` task and its settings
}
