% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotDemographyTask}
\alias{loadPlotDemographyTask}
\title{loadPlotDemographyTask}
\usage{
loadPlotDemographyTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{\code{PopulationWorkflow} object or derived class}

\item{active}{logical defining if \code{Task} will be run by workflow.
Default value is \code{FALSE}}

\item{settings}{specific settings for \code{plotDemography} task}
}
\value{
A \code{PlotTask} object
}
\description{
Define \code{plotDemography} task and its settings
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- PopulationWorkflow$new(
  workflowType,
  workflowFolder,
  simulationSets
)

# (Re)load a default plotDemography task for workflow
myWorkflow$plotDemography <- loadPlotDemographyTask(
  workflow = myWorkflow,
  active = TRUE
)

# Load a user-defined simulation task for workflow
myWorkflow$userDefinedTasks[["plotDemography"]] <- loadPlotDemographyTask(
  workflow = myWorkflow,
  active = TRUE
)
}

}
\seealso{
Other workflow tasks: 
\code{\link{GofPlotTask}},
\code{\link{PlotTask}},
\code{\link{PopulationPlotTask}},
\code{\link{PopulationSensitivityAnalysisTask}},
\code{\link{QualificationTask}},
\code{\link{SensitivityAnalysisTask}},
\code{\link{Task}},
\code{\link{addUserDefinedTask}()},
\code{\link{loadCalculatePKParametersTask}()},
\code{\link{loadCalculateSensitivityTask}()},
\code{\link{loadGOFMergedTask}()},
\code{\link{loadPlotAbsorptionTask}()},
\code{\link{loadPlotDDIRatioTask}()},
\code{\link{loadPlotMassBalanceTask}()},
\code{\link{loadPlotPKParametersTask}()},
\code{\link{loadPlotPKRatioTask}()},
\code{\link{loadPlotSensitivityTask}()},
\code{\link{loadPlotTimeProfilesAndResidualsTask}()},
\code{\link{loadQualificationComparisonTimeProfileTask}()},
\code{\link{loadQualificationTimeProfilesTask}()},
\code{\link{loadSimulateTask}()}
}
\concept{workflow tasks}
