% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotMassBalanceTask}
\alias{loadPlotMassBalanceTask}
\title{loadPlotMassBalanceTask}
\usage{
loadPlotMassBalanceTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`MeanModelWorkflow` object}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `plotMassBalance` task}
}
\value{
A `PlotTask` object
}
\description{
Define `plotMassBalance` task and its settings
}
