% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotPKParametersTask}
\alias{loadPlotPKParametersTask}
\title{loadPlotPKParametersTask}
\usage{
loadPlotPKParametersTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `plotPKParameters` task}
}
\value{
A `PopulationSensitivityAnalysisTask` for `PopulationWorkflow` objects.
A `SensitivityAnalysisTask` object otherwise
}
\description{
Define `plotPKParameters` task and its settings
}
