% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotPKRatioTask}
\alias{loadPlotPKRatioTask}
\title{loadPlotPKRatioTask}
\usage{
loadPlotPKRatioTask(workflow, configurationPlan)
}
\arguments{
\item{workflow}{`QualificationWorkflow` object}

\item{configurationPlan}{A `ConfigurationPlan` object}
}
\value{
A `QualificationTask` object
}
\description{
Define `plotPKRatio` task and its settings
}
