% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotSensitivityTask}
\alias{loadPlotSensitivityTask}
\title{loadPlotSensitivityTask}
\usage{
loadPlotSensitivityTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `plotSensitivity` task}
}
\value{
A `PopulationSensitivityAnalysisTask` for `PopulationWorkflow` objects.
A `SensitivityAnalysisTask` object otherwise
}
\description{
Define `plotSensitivity` task and its settings
}
