% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{loadPopulationOnCores}
\alias{loadPopulationOnCores}
\title{loadPopulationOnCores}
\usage{
loadPopulationOnCores(populationFiles, logFolder)
}
\arguments{
\item{populationFiles}{population files to be loaded on cores}

\item{logFolder}{folder where the logs are saved}
}
\description{
Send population file names to cores, check that population is loaded on each core successfully
}
