% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadQualificationComparisonTimeProfileTask}
\alias{loadQualificationComparisonTimeProfileTask}
\title{loadQualificationComparisonTimeProfileTask}
\usage{
loadQualificationComparisonTimeProfileTask(workflow, configurationPlan)
}
\arguments{
\item{workflow}{`QualificationWorkflow` object}

\item{configurationPlan}{A `ConfigurationPlan` object}
}
\value{
A `QualificationTask` object
}
\description{
Define `plotComparisonTimeProfileTask` task and its settings
}
