% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{loadQualificationWorkflow}
\alias{loadQualificationWorkflow}
\title{loadQualificationWorkflow}
\usage{
loadQualificationWorkflow(workflowFolder, jsonFile)
}
\arguments{
\item{workflowFolder}{path of the output folder created or used by the Workflow.}

\item{jsonFile}{name of the JSON snapshot file including the configuration plan}
}
\value{
A `QualificationWorkflow` object
}
\description{
Load a `ConfigurationPlan` object from json file
}
