% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{loadRESettings}
\alias{loadRESettings}
\title{loadRESettings}
\usage{
loadRESettings(file)
}
\arguments{
\item{file}{`.RData` file containing the settings}
}
\description{
Load reporting engine global settings from a file
}
