% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadSimulateTask}
\alias{loadSimulateTask}
\title{loadSimulateTask}
\usage{
loadSimulateTask(workflow, active = TRUE, settings = NULL)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `TRUE`}

\item{settings}{specific settings for `simulate` task}
}
\value{
A `SimulationTask` object
}
\description{
Define `simulate` task and its settings
}
