% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{loadSimulationOnCores}
\alias{loadSimulationOnCores}
\title{loadSimulationOnCores}
\usage{
loadSimulationOnCores(structureSet, logFolder)
}
\arguments{
\item{structureSet}{containing simulationSet which contains path to simulation file and pathIDs to be loaded in simulation object as outputs}

\item{logFolder}{folder where the logs are saved}
}
\description{
Send structureSet to core, check its simulation has been loaded successfully
}
