% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{loadStudyDesign}
\alias{loadStudyDesign}
\title{loadStudyDesign}
\usage{
loadStudyDesign(studyDesignFile)
}
\arguments{
\item{studyDesignFile}{file name of study design table}
}
\value{
A `StudyDesign` object`
}
\description{
Load a StudyDesign object from a file containing a study design table
In this table, Line 1 is path,  Line 2 is unit, Line 3 is type and subsequent lines are values.
The cells for type can include "SOURCE" OR "TARGET". "SOURCE" type can also include subtype "MIN", "MAX" or "EQUALS".
In the current version, unit is expected to be intern units of simulation.
}
