% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadUserDefinedTask}
\alias{loadUserDefinedTask}
\title{loadUserDefinedTask}
\usage{
loadUserDefinedTask(
  workflow,
  taskFunction,
  taskName = "userDefinedTask",
  active = TRUE,
  settings = NULL
)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{taskFunction}{Function to be called by `Task` object}

\item{taskName}{character name of task to initialize reporting of task}

\item{active}{logical to set if task is run by workflow}

\item{settings}{list of input arguments that can be used by `settings` input argument of `taskFunction`}
}
\value{
Updated `Workflow` object
}
\description{
Upload a user-defined function as workflow task
}
