% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{logErrorMessage}
\alias{logErrorMessage}
\title{Log the error with a message}
\usage{
logErrorMessage(message, logFolderPath = getwd())
}
\arguments{
\item{message}{message to display and then log}

\item{logFolderPath}{path where logs are saved}
}
\description{
Log the error with a message
}
