% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{logErrorThenStop}
\alias{logErrorThenStop}
\title{Log the error with a message and then stop, displaying same message.}
\usage{
logErrorThenStop(message, logFolderPath = getwd())
}
\arguments{
\item{message}{message to display and then log}

\item{logFolderPath}{path where logs are saved}
}
\description{
Log the error with a message and then stop, displaying same message.
}
\keyword{internal}
