% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{Logging}
\alias{Logging}
\title{Logging}
\description{
R6 class managing logging
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{messages}}{Array of messages to log.
Each element of `messages` is a list
that includes `time`, `type` and `text` of message to log.}

\item{\code{folder}}{Folder where log files are saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Logging-info}{\code{Logging$info()}}
\item \href{#method-Logging-error}{\code{Logging$error()}}
\item \href{#method-Logging-debug}{\code{Logging$debug()}}
\item \href{#method-Logging-textToMessage}{\code{Logging$textToMessage()}}
\item \href{#method-Logging-addMessage}{\code{Logging$addMessage()}}
\item \href{#method-Logging-showMessages}{\code{Logging$showMessages()}}
\item \href{#method-Logging-printMessage}{\code{Logging$printMessage()}}
\item \href{#method-Logging-saveMessage}{\code{Logging$saveMessage()}}
\item \href{#method-Logging-writeAsJson}{\code{Logging$writeAsJson()}}
\item \href{#method-Logging-write}{\code{Logging$write()}}
\item \href{#method-Logging-reset}{\code{Logging$reset()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-info"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-info}{}}}
\subsection{Method \code{info()}}{
Add an `Info` message to log
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$info(text, display = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{content of the `Info` message}

\item{\code{display}}{Logical setting if message is displayed on console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-error"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-error}{}}}
\subsection{Method \code{error()}}{
Add an `Error` message to log
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$error(text, display = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{content of the `Error` message}

\item{\code{display}}{Logical setting if message is displayed on console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-debug"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-debug}{}}}
\subsection{Method \code{debug()}}{
Add a `Debug` message to log
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$debug(text, display = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{content of the `Debug` message}

\item{\code{display}}{Logical setting if message is displayed on console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-textToMessage"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-textToMessage}{}}}
\subsection{Method \code{textToMessage()}}{
Add meta data to `text` to create an informative message to log
with a flag for its `logType` (`"Info"`, `"Error"`, or `"Debug"`)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$textToMessage(text, logType)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{content of the message}

\item{\code{logType}}{type of message (`"Info"`, `"Error"`, or `"Debug"`)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-addMessage"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-addMessage}{}}}
\subsection{Method \code{addMessage()}}{
Add a message to the list of logged `messages`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$addMessage(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{content of the message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-showMessages"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-showMessages}{}}}
\subsection{Method \code{showMessages()}}{
Display logged `messages` as a data.frame
Can be filtered according to `logTypes` (`"Info"`, `"Error"`, and/or `"Debug"`)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$showMessages(logTypes = LogTypes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logTypes}}{types of message (`"Info"`, `"Error"`, or `"Debug"`)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-printMessage"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-printMessage}{}}}
\subsection{Method \code{printMessage()}}{
Print the log messages on R/RStudio console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$printMessage(message, display = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{Message to print}

\item{\code{display}}{Logical setting if message is displayed on console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-saveMessage"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-saveMessage}{}}}
\subsection{Method \code{saveMessage()}}{
Write a message in its log file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$saveMessage(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{Message to print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-writeAsJson"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-writeAsJson}{}}}
\subsection{Method \code{writeAsJson()}}{
Write the log messages in a json log file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$writeAsJson(fileName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{Name of json log file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-write"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-write}{}}}
\subsection{Method \code{write()}}{
Write the log messages in txt log files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$write()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logging-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Logging-reset}{}}}
\subsection{Method \code{reset()}}{
Reset/empty log messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logging$reset(folder = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder}}{Folder where logs are saved}
}
\if{html}{\out{</div>}}
}
}
}
