% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{logicalCodeIds}
\alias{logicalCodeIds}
\title{logicalCodeIds}
\format{
An object of class \code{character} of length 5.
}
\usage{
logicalCodeIds
}
\description{
Some of the identifiers above need to be treated
as logical by [getIdentifierInfo()] while usually input as `"Yes"`/`"No"`
}
\keyword{internal}
