% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\docType{data}
\name{LogLevels}
\alias{LogLevels}
\title{LogLevels}
\format{
An object of class \code{list} of length 4.
}
\usage{
LogLevels
}
\description{
Helper enum listing all available log levels
}
\examples{
LogLevels$Info
LogLevels$Warning
LogLevels$Error
LogLevels$Debug
}
\keyword{datasets}
