% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logMessage}
\alias{logMessage}
\title{logMessage}
\usage{
logMessage(message, logLevel = LogLevels$Info, logFolder = NULL)
}
\arguments{
\item{message}{message to report}

\item{logLevel}{Name of log level. Use helper enum `LogLevels` for finding appropriate log level.}

\item{logFolder}{Directory in which logs are saved.
If `NULL` the message is only reported on console}
}
\description{
Report a logging message
}
