% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logWorkflow}
\alias{logWorkflow}
\title{logWorkflow}
\usage{
logWorkflow(
  message,
  pathFolder = getwd(),
  logTypes = c(LogTypes$Info, LogTypes$Debug)
)
}
\arguments{
\item{message}{message to save in log files}

\item{pathFolder}{folder where the logs are saved
Default is `getwd()`}

\item{logTypes}{types of logs where message is saved
Use enum `LogTypes` to get all the logTypes.
Default is c("Info", "Debug")}
}
\value{

}
\description{
Save messages into log files
}
