% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{lookupPKParameterDisplayName}
\alias{lookupPKParameterDisplayName}
\title{lookupPKParameterDisplayName}
\usage{
lookupPKParameterDisplayName(output, pkParameter)
}
\arguments{
\item{output}{an Output object}

\item{pkParameter}{a string with the name of the pkParameter from the population sensitivity results index file}
}
\value{
the display name of the input pk parameter or the string pkParameter itself if no display name found
}
\description{
lookupPKParameterDisplayName
}
