% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{mapStudyDesignSources}
\alias{mapStudyDesignSources}
\title{mapStudyDesignSources}
\usage{
mapStudyDesignSources(data, population, simulation)
}
\arguments{
\item{data}{data.frame read from a study design file}

\item{population}{\code{Population} object}

\item{simulation}{\code{Simulation} object}
}
\value{
vector of expressions assigning target values
Must be the same length as target values
}
\description{
mapStudyDesignSources
}
\keyword{internal}
