% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{mapStudyDesignTargets}
\alias{mapStudyDesignTargets}
\title{mapStudyDesignTargets}
\usage{
mapStudyDesignTargets(data, population, simulation)
}
\arguments{
\item{data}{data.frame read from a study design file}

\item{population}{\code{Population} object}

\item{simulation}{\code{Simulation} object}
}
\value{
list of \code{StudyDesignTarget} objects
}
\description{
mapStudyDesignTargets
}
\keyword{internal}
