% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{massMoleConversion}
\alias{massMoleConversion}
\title{massMoleConversion}
\usage{
massMoleConversion(dimension)
}
\arguments{
\item{dimension}{is string that is from among the valid OSP dimensions}
}
\value{
If \code{dimension} is \code{Concentration (mass)} or \code{Mass}, then return \code{Concentration (molar)} or \code{Amount} respectively, otherwise return \code{dimension}.
}
\description{
Function to map \code{Concentration (mass)} or \code{Mass} dimensions to \code{Concentration (molar)} and \code{Amount} respectively.
}
\keyword{internal}
