% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{mcSampling}
\alias{mcSampling}
\title{mcSampling}
\usage{
mcSampling(
  dataSize,
  sampleSize,
  n = getDefaultMCRepetitions(),
  seed = getDefaultMCRandomSeed()
)
}
\arguments{
\item{dataSize}{Number of rows/unique individuals in \code{data}}

\item{sampleSize}{Number of sampled individuals in each Monte Carlo repetition}

\item{n}{Number of repetitions}

\item{seed}{Random Seed Number in order to get repeatable results}
}
\value{
A list of \code{n} elements that include \code{sampleSize} integers sampled from \code{1:dataSize}
}
\description{
Perform repeatable Monte Carlo random sampling
for a data.frame obtained from a \code{PKAnalyses} object
}
\keyword{internal}
