% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model.R
\docType{class}
\name{MeanModel}
\alias{MeanModel}
\title{MeanModel}
\description{
MeanModel R6 class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanModel$new()}}
\item \href{#method-simulateMeanModel}{\code{MeanModel$simulateMeanModel()}}
\item \href{#method-calculateMassBalance}{\code{MeanModel$calculateMassBalance()}}
\item \href{#method-calculateMeanModelPKParameters}{\code{MeanModel$calculateMeanModelPKParameters()}}
\item \href{#method-plotMeanModelTimeProfiles}{\code{MeanModel$plotMeanModelTimeProfiles()}}
\item \href{#method-plotMeanModelResiduals}{\code{MeanModel$plotMeanModelResiduals()}}
\item \href{#method-plotMeanModelMassBalance}{\code{MeanModel$plotMeanModelMassBalance()}}
\item \href{#method-plotMeanModelAbsorption}{\code{MeanModel$plotMeanModelAbsorption()}}
\item \href{#method-plotMeanModelPkParameters}{\code{MeanModel$plotMeanModelPkParameters()}}
\item \href{#method-calculateMeanModelSensitivity}{\code{MeanModel$calculateMeanModelSensitivity()}}
\item \href{#method-plotMeanModelSensitivity}{\code{MeanModel$plotMeanModelSensitivity()}}
\item \href{#method-clone}{\code{MeanModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$new(
  modelFilePath = NULL,
  simModel = NULL,
  modelDisplayName = NULL,
  simulationOutputsFolder = "."
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateMeanModel"></a>}}
\subsection{Method \code{simulateMeanModel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$simulateMeanModel(saveSimulation = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateMassBalance"></a>}}
\subsection{Method \code{calculateMassBalance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$calculateMassBalance(saveMassBalance = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateMeanModelPKParameters"></a>}}
\subsection{Method \code{calculateMeanModelPKParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$calculateMeanModelPKParameters(
  userDefinedPKFunctions = NULL,
  savePKAnalysis = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanModelTimeProfiles"></a>}}
\subsection{Method \code{plotMeanModelTimeProfiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$plotMeanModelTimeProfiles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanModelResiduals"></a>}}
\subsection{Method \code{plotMeanModelResiduals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$plotMeanModelResiduals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanModelMassBalance"></a>}}
\subsection{Method \code{plotMeanModelMassBalance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$plotMeanModelMassBalance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanModelAbsorption"></a>}}
\subsection{Method \code{plotMeanModelAbsorption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$plotMeanModelAbsorption()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanModelPkParameters"></a>}}
\subsection{Method \code{plotMeanModelPkParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$plotMeanModelPkParameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateMeanModelSensitivity"></a>}}
\subsection{Method \code{calculateMeanModelSensitivity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$calculateMeanModelSensitivity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanModelSensitivity"></a>}}
\subsection{Method \code{plotMeanModelSensitivity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$plotMeanModelSensitivity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
