% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-set.R
\name{MeanModelSet}
\alias{MeanModelSet}
\title{MeanModelSet}
\description{
R6 class representing Reporting Engine Mean Model Set
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{simulationName}}{display name of simulation}

\item{\code{pathID}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{pathName}}{display name for `pathID`}

\item{\code{pathUnit}}{display unit for `pathID`}

\item{\code{pkParameters}}{PK parameters function names to be calculated from the simulation (e.g. `C_max`).}

\item{\code{pkParametersNames}}{display names for `pkParameters`}

\item{\code{pkParameterUnits}}{display units for `pkParameters`}

\item{\code{dataFilter}}{filter to compare with observed data}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pkParametersNames}}{display names for `pkParameters`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanModelSet$new()}}
\item \href{#method-clone}{\code{MeanModelSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `MeanModelSet` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelSet$new(
  simulationFile,
  simulationName = NULL,
  pathID,
  pathName = NULL,
  pathUnit = NULL,
  pkParameters = AllPKParameters,
  pkParametersNames = NULL,
  pkParametersUnits = NULL,
  dataFilter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{simulationName}}{display name of simulation}

\item{\code{pathID}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{pathName}}{display name for `pathID`}

\item{\code{pathUnit}}{display unit for `pathID`}

\item{\code{pkParameters}}{PK parameters function names to be calculated from the simulation (e.g. `C_max`).
Default value is enum `AllPKParameters`.}

\item{\code{pkParametersNames}}{display names for `pkParameters`}

\item{\code{dataFilter}}{filter to compare with observed data}

\item{\code{pkParameterUnits}}{display units for `pkParameters`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelSet` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
