% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-workflow.R
\name{MeanModelWorkflow}
\alias{MeanModelWorkflow}
\title{MeanModelWorkflow}
\description{
R6 class for Reporting Engine Mean Model Workflow
}
\seealso{
Other workflows: 
\code{\link{PopulationWorkflow}},
\code{\link{QualificationWorkflow}}
}
\concept{workflows}
\section{Super class}{
\code{ospsuite.reportingengine::Workflow} -> \code{MeanModelWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulate}}{\code{SimulationTask} object for time profile simulations}

\item{\code{calculatePKParameters}}{\code{CalculatePKParametersTask} object for PK parameters calculation}

\item{\code{calculateSensitivity}}{\code{SensitivityAnalysisTask} object for sensitivity analysis}

\item{\code{plotTimeProfilesAndResiduals}}{\code{PlotTask} object for goodness of fit plots}

\item{\code{plotMassBalance}}{\code{PlotTask} object for mass balance plot}

\item{\code{plotAbsorption}}{\code{PlotTask} object for absorption plot}

\item{\code{plotPKParameters}}{\code{PlotTask} object for PK parameters plot}

\item{\code{plotSensitivity}}{\code{PlotTask} object for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeanModelWorkflow-new}{\code{MeanModelWorkflow$new()}}
\item \href{#method-MeanModelWorkflow-runWorkflow}{\code{MeanModelWorkflow$runWorkflow()}}
\item \href{#method-MeanModelWorkflow-clone}{\code{MeanModelWorkflow$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-activateTasks'><code>ospsuite.reportingengine::Workflow$activateTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getActiveTasks'><code>ospsuite.reportingengine::Workflow$getActiveTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getAllPlotTasks'><code>ospsuite.reportingengine::Workflow$getAllPlotTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getAllTasks'><code>ospsuite.reportingengine::Workflow$getAllTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getInactiveTasks'><code>ospsuite.reportingengine::Workflow$getInactiveTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getParameterDisplayPaths"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getParameterDisplayPaths'><code>ospsuite.reportingengine::Workflow$getParameterDisplayPaths()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getSimulationDescriptor"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getSimulationDescriptor'><code>ospsuite.reportingengine::Workflow$getSimulationDescriptor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getWatermark"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getWatermark'><code>ospsuite.reportingengine::Workflow$getWatermark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-inactivateTasks'><code>ospsuite.reportingengine::Workflow$inactivateTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="print"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-print'><code>ospsuite.reportingengine::Workflow$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="printReportingEngineInfo"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-printReportingEngineInfo'><code>ospsuite.reportingengine::Workflow$printReportingEngineInfo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setParameterDisplayPaths"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setParameterDisplayPaths'><code>ospsuite.reportingengine::Workflow$setParameterDisplayPaths()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setSimulationDescriptor"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setSimulationDescriptor'><code>ospsuite.reportingengine::Workflow$setSimulationDescriptor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setWatermark"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setWatermark'><code>ospsuite.reportingengine::Workflow$setWatermark()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeanModelWorkflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeanModelWorkflow-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MeanModelWorkflow} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object \code{Workflow}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MeanModelWorkflow} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeanModelWorkflow-runWorkflow"></a>}}
\if{latex}{\out{\hypertarget{method-MeanModelWorkflow-runWorkflow}{}}}
\subsection{Method \code{runWorkflow()}}{
Run mean model workflow tasks for all simulation sets if tasks are activated
The order of tasks is as follows:
\enumerate{
\item Run simulations
\item Perform PK and sensitivity analyses
\item Perform plot tasks\if{html}{\out{<br>}}
a. time profiles and residual plots\if{html}{\out{<br>}}
b. absorption plots\if{html}{\out{<br>}}
c. mass balance plots\if{html}{\out{<br>}}
d. PK and sensitivity analyses tables and plots
\item Render report
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in the workflow folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeanModelWorkflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeanModelWorkflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
