% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{mergeMarkdowndFiles}
\alias{mergeMarkdowndFiles}
\title{mergeMarkdowndFiles}
\usage{
mergeMarkdowndFiles(inputFiles, outputFile, logFolder = getwd())
}
\arguments{
\item{inputFiles}{name of .md files to merge}

\item{outputFile}{text to include in the document}

\item{logFolder}{folder where the logs are saved}
}
\description{
Merge all appendices reports into one final report
}
