% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newOutputColor}
\alias{newOutputColor}
\title{newOutputColor}
\usage{
newOutputColor()
}
\value{
A color from OSP Suite Color Map
}
\description{
Find a color for new \code{Output} objects
}
\keyword{internal}
