% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{numberSections}
\alias{numberSections}
\title{numberSections}
\usage{
numberSections(fileName, logFolder, tocPattern = "#", tocLevels = 3)
}
\arguments{
\item{fileName}{name of .md file to update}

\item{logFolder}{folder where the logs are saved}

\item{tocPattern}{character pattern referencing sections in first element of line}

\item{tocLevels}{levels of sections in the report}
}
\value{
Table of content referencing sections following a markdown format
}
\description{
Reference sections of a report
}
