% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{numberTablesAndFigures}
\alias{numberTablesAndFigures}
\title{numberTablesAndFigures}
\usage{
numberTablesAndFigures(
  fileName,
  logFolder = getwd(),
  figurePattern = "Figure:",
  tablePattern = "Table:"
)
}
\arguments{
\item{fileName}{name of .md file to update}

\item{logFolder}{folder where the logs are saved}

\item{figurePattern}{character pattern referencing figures in first element of line}

\item{tablePattern}{character pattern referencing tables in first element of line}
}
\description{
Reference tables and figures in a report
}
\keyword{internal}
