% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{OptionalSettings}
\alias{OptionalSettings}
\title{OptionalSettings}
\format{
An object of class \code{list} of length 12.
}
\usage{
OptionalSettings
}
\description{
Workflow optional settings
}
\keyword{internal}
