% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{parseObservationsDataFrame}
\alias{parseObservationsDataFrame}
\title{parseObservationsDataFrame}
\usage{
parseObservationsDataFrame(observationsDataFrame)
}
\arguments{
\item{observationsDataFrame}{from CSV}
}
\value{
A named list with `time` and `output` fields.  Each field contains a list that is output by `separateVariableFromUnit` with fields that store the variable name and the unit.
}
\description{
Function to read the variable names and units in the first two columns of an observations data frame used for qualification. First column stores timepoints.  Second column stores measurements of a quantity corresponding to timepoints in first column.
}
