% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseVariableFromObject}
\alias{parseVariableFromObject}
\title{parseVariableFromObject}
\usage{
parseVariableFromObject(objectName, variableName, keepIfNull = FALSE)
}
\arguments{
\item{objectName}{Name of the object whose field is updated}

\item{variableName}{Name of the variable and field of `objectName`}

\item{keepIfNull}{logical `variableName <- objectName$variableName \%||\% variableName`}
}
\value{
An expression to `eval()`
}
\description{
Create an expression of type `variableName <- objectName$variableName`
}
\keyword{internal}
