% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-of-work.R
\name{performSensitivityAnalysis}
\alias{performSensitivityAnalysis}
\title{performSensitivityAnalysis}
\usage{
performSensitivityAnalysis(simFileName, simFileFolder, paramRange = NULL,
  pkParameterName = "AUC", totalSensitivityThreshold = 1)
}
\arguments{
\item{pkParameterName}{name of parameter to be analyzed}

\item{totalSensitivityThreshold}{numeric value between 0 and 1.
Close to 0, only the most sensitive output paths are returned.
Close to 1, almost all the output paths are returned.}

\item{simulation}{simulation class object}
}
\value{
sensitivityResults
}
\description{
Run a sensitivity analysis from a simulation
}
