% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-outputs.R
\docType{data}
\name{pkDictionaryQualificationOSP}
\alias{pkDictionaryQualificationOSP}
\title{Dictionary for `ospsuite` PK parameters defined by Configuration Plan}
\format{
An object of class \code{list} of length 43.
}
\usage{
pkDictionaryQualificationOSP
}
\description{
Dictionary for `ospsuite` PK parameters defined by Configuration Plan
}
\keyword{internal}
