% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk-parameter-info.R
\name{PkParameterInfo}
\alias{PkParameterInfo}
\title{PkParameterInfo}
\description{
R6 class representing workflow selected PK parameter.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pkParameter}}{PK Parameter name for the simulation (e.g. `AUC_inf`)}

\item{\code{displayName}}{display name for `pkParameter`}

\item{\code{displayUnit}}{display unit for `pkParameter`}

\item{\code{group}}{Grouping identifier}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PkParameterInfo$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `PkParameterInfo` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PkParameterInfo$new(
  pkParameter,
  displayName = NULL,
  displayUnit = NULL,
  group = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pkParameter}}{PK Parameter name for the simulation (e.g. `AUC_inf`)
Tips: use ospsuite::allPKParameterNames() to get a list of available PK parameters}

\item{\code{displayName}}{display name for `pkParameter`}

\item{\code{displayUnit}}{display unit for `pkParameter`}

\item{\code{group}}{Grouping identifier. Default group associate `pkParameter`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PkParameterInfo` object
}
}
}
