% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{pkParameterTableAsRelativeChange}
\alias{pkParameterTableAsRelativeChange}
\title{pkParameterTableAsRelativeChange}
\usage{
pkParameterTableAsRelativeChange(pkParametersTable, referenceSimulationSetName)
}
\arguments{
\item{pkParametersTable}{PK Parameter Measure summarizing key statistics for each population}

\item{referenceSimulationSetName}{Reference}
}
\description{
Get Measure Table for Parallel Population Workflow
}
\keyword{internal}
