% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-absorption.R
\name{plotAbsorptionTimeProfile}
\alias{plotAbsorptionTimeProfile}
\title{plotAbsorptionTimeProfile}
\usage{
plotAbsorptionTimeProfile(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on \code{data}}

\item{dataMapping}{\code{XYGDataMapping} R6 class object from \code{tlf} library}

\item{plotConfiguration}{\code{PlotConfiguration} R6 class object from \code{tlf} library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot absorption time profile
}
