% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{plotBoxWhiskerLog}
\alias{plotBoxWhiskerLog}
\title{plotBoxWhiskerLog}
\usage{
plotBoxWhiskerLog(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame of the ratios}

\item{plotConfiguration}{PlotConfiguration R6 class object}
}
\value{
ggplot object
}
\description{
Plot box-whiskers of PK parameter values in log scale
}
\keyword{internal}
