% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{plotDDIXLabel}
\alias{plotDDIXLabel}
\title{Labels for DDI plot X-axis}
\format{
An object of class \code{list} of length 2.
}
\usage{
plotDDIXLabel
}
\description{
Labels for DDI plot X-axis
}
\keyword{datasets}
