% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{plotDDIYLabel}
\alias{plotDDIYLabel}
\title{Labels for DDI plot Y-axis}
\format{
An object of class \code{list} of length 2.
}
\usage{
plotDDIYLabel
}
\description{
Labels for DDI plot Y-axis
}
\keyword{datasets}
