% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-of-work.R
\name{plotDemography}
\alias{plotDemography}
\title{plotDemography}
\usage{
plotDemography(simulation, population,
  parameterNames = c(StandardPath$Age, StandardPath$Weight,
  StandardPath$Height), plotConfiguration = NULL)
}
\arguments{
\item{simulation}{simulation class object}

\item{population}{population class object}

\item{plotConfiguration}{HistogramPlotConfiguration class object}

\item{demographyParameters}{Paths of the demography parameters to plot}
}
\value{
demographyPlots list of ggplot objects
}
\description{
Plot histograms of demography parameters from a simulated population
}
