% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{plotDemographyHistogram}
\alias{plotDemographyHistogram}
\title{plotDemographyHistogram}
\usage{
plotDemographyHistogram(
  data,
  bins = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame of the ratios}

\item{plotConfiguration}{PlotConfiguration R6 class object}
}
\value{
ggplot object
}
\description{
Plot histograms of ratios as is
}
