% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{plotDemographyParameters}
\alias{plotDemographyParameters}
\title{plotDemographyParameters}
\usage{
plotDemographyParameters(
  structureSets,
  logFolder = getwd(),
  settings = NULL,
  workflowType = PopulationWorkflowTypes$parallelComparison,
  xParameters = getDefaultXParametersForWorkflowType(workflowType),
  yParameters = NULL
)
}
\arguments{
\item{structureSets}{`SimulationStructure` R6 class object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{list of settings for the output table/plot}

\item{workflowType}{workflowType Type of population workflow.
Use enum `PopulationWorkflowTypes` to get list of workflow types.}

\item{xParameters}{list of parameters to be plotted along x axis}

\item{yParameters}{list of parameters to be plotted along y axis}
}
\value{
list of plots and tables with summary of demography parameters
}
\description{
Plot demography parameters box plots and tables
}
