% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\docType{data}
\name{plotGOFFunctions}
\alias{plotGOFFunctions}
\title{Names of functions for plotting GOF plots for each GOF plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
plotGOFFunctions
}
\description{
Names of functions for plotting GOF plots for each GOF plot type
}
\keyword{datasets}
