% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plotGoodnessOfFit}
\alias{plotGoodnessOfFit}
\title{plotGoodnessOfFit}
\usage{
plotGoodnessOfFit(
  populationSimulation,
  population,
  observedData = NULL,
  quantity = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{populationSimulation}{class object from runSimulation(simulation, population)}

\item{population}{population class object}

\item{observedData}{list of datasets corresponding to observations}

\item{quantity}{Paths of the quantity parameters to plot}

\item{plotConfiguration}{List of PlotConfiguration class objects for each plot}
}
\value{
gofPlots list of ggplot objects
}
\description{
Plot goodness of fit diagnostics including time profiles, vpc,
observations vs predictions, residuals plots (residuals vs time, vs predictions, qq-plots and histogram)
}
