% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{plotMassBalanceTimeProfile}
\alias{plotMassBalanceTimeProfile}
\title{plotMassBalanceTimeProfile}
\usage{
plotMassBalanceTimeProfile(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{dataMapping}{`XYGDataMapping` R6 class object from `tlf` library}

\item{plotCOnfiguration}{`PlotConfiguration` R6 class object from `tlf` library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot mass balance time profile
}
