% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-absorption.R
\name{plotMeanAbsorption}
\alias{plotMeanAbsorption}
\title{plotMeanAbsorption}
\usage{
plotMeanAbsorption(structureSet, settings = NULL)
}
\arguments{
\item{structureSet}{\code{SimulationStructure} R6 class object}

\item{settings}{Options to be defined
(e.g. plotConfigurations list of \code{PlotConfiguration} objects)}
}
\value{
list of \code{ggplot} objects
}
\description{
Plot absorption diagnostics time profile
}
\keyword{internal}
