% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{plotMeanGoodnessOfFit}
\alias{plotMeanGoodnessOfFit}
\title{plotMeanGoodnessOfFit}
\usage{
plotMeanGoodnessOfFit(structureSet, plotConfigurations = NULL)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{plotConfigurations}{List of `PlotConfiguration` R6 class objects for each goodness of fit plot}
}
\value{
list of `residuals` objects
}
\description{
Plot goodness of fit diagnostics including time profiles,
observations vs predictions, residuals plots (residuals vs time, vs predictions, qq-plots and histogram)
}
