% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{plotMeanMassBalance}
\alias{plotMeanMassBalance}
\title{plotMeanMassBalance}
\usage{
plotMeanMassBalance(structureSet, plotConfigurations = NULL)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{plotConfigurations}{List of `PlotConfiguration` R6 class objects for each goodness of fit plot}
}
\value{
list of `ggplot` objects
}
\description{
Plot mass balance diagnostics time profiles, cumulative time profiles,
normalized time profiles and cumulative normalized time profiles
}
