% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{plotMeanPKParameters}
\alias{plotMeanPKParameters}
\title{plotMeanPKParameters}
\usage{
plotMeanPKParameters(structureSet, plotConfigurations)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{plotConfigurations}{list of configurations for the output table/plot}
}
\value{
data.frame with calculated PK parameters for the simulation set
}
\description{
Plot PK parameters table
}
