% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{plotMeanResVsPred}
\alias{plotMeanResVsPred}
\title{plotMeanResVsPred}
\usage{
plotMeanResVsPred(data, metaData = NULL, plotConfiguration = NULL)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{plotCOnfiguration}{`PlotConfiguration` R6 class object from `tlf` library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot observation vs prediction for mean model workflow
}
