% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotMeanResVsTime}
\alias{plotMeanResVsTime}
\title{plotMeanResVsTime}
\usage{
plotMeanResVsTime(
  data,
  metaData = NULL,
  caption = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{caption}{legend caption for data.frame}

\item{plotConfiguration}{`PlotConfiguration` R6 class object from `tlf` library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot Residual vs time for mean model workflow
}
