% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{plotMeanSensitivity}
\alias{plotMeanSensitivity}
\title{plotMeanSensitivity}
\usage{
plotMeanSensitivity(structureSet, logFolder = getwd(), settings)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{list of settings for the output table/plot}
}
\value{
list of plots and tables
}
\description{
Plot sensitivity analysis results for mean models
}
