% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{plotMeanTimeProfile}
\alias{plotMeanTimeProfile}
\title{plotMeanTimeProfile}
\usage{
plotMeanTimeProfile(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{dataMapping}{`TimeProfileDataMapping` R6 class object from `tlf` library}

\item{plotCOnfiguration}{`TimeProfilePlotConfiguration` R6 class object from `tlf` library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot time profile for mean model workflow
}
