% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotMeanTimeProfileLog}
\alias{plotMeanTimeProfileLog}
\title{plotMeanTimeProfileLog}
\usage{
plotMeanTimeProfileLog(
  simulatedData,
  observedData = NULL,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{simulatedData}{data.frame of simulated data}

\item{observedData}{data.frame of observed data}

\item{metaData}{meta data on \code{data}}

\item{dataMapping}{A list to be integrated to \code{tlf} DataMapping objects}

\item{plotConfiguration}{\code{PlotConfiguration} class object from \code{tlf} library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot time profile for mean model workflow
}
