% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotPopTimeProfile}
\alias{plotPopTimeProfile}
\title{plotPopTimeProfile}
\usage{
plotPopTimeProfile(
  simulatedData,
  observedData = NULL,
  dataMapping = NULL,
  metaData = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{simulatedData}{data.frame of simulated data}

\item{observedData}{data.frame of observed data}

\item{dataMapping}{A list to be integrated to \code{tlf} DataMapping objects}

\item{metaData}{meta data on \code{data}}

\item{plotConfiguration}{\code{TimeProfilePlotConfiguration} R6 class object from \code{tlf} library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot time profile for population model workflow
}
