% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{plotPopulationPKParameters}
\alias{plotPopulationPKParameters}
\title{plotPopulationPKParameters}
\usage{
plotPopulationPKParameters(
  structureSets,
  logFolder = getwd(),
  settings = NULL,
  workflowType = PopulationWorkflowTypes$parallelComparison
)
}
\arguments{
\item{structureSets}{`SimulationStructure` R6 class object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{list of settings for the output table/plot}

\item{workflowType}{workflowType Type of population workflow.
Use enum `PopulationWorkflowTypes` to get list of workflow types.}
}
\value{
data.frame with calculated PK parameters for the simulation set
}
\description{
Plot PK parameters box plots and tables
}
