% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{plotPopulationSensitivity}
\alias{plotPopulationSensitivity}
\title{plotPopulationSensitivity}
\usage{
plotPopulationSensitivity(
  structureSets,
  settings,
  workflowType = PopulationWorkflowTypes$parallelComparison,
  xParameters = NULL,
  yParameters = NULL
)
}
\arguments{
\item{structureSets}{list of \code{SimulationStructure} objects}

\item{settings}{list of settings for the population sensitivity plot}

\item{workflowType}{Element from \code{PopulationWorkflowTypes}}

\item{xParameters}{selected parameters to be plotted in x axis}

\item{yParameters}{selected parameters to be plotted in y axis}
}
\value{
a structured list of plots for each possible combination of pathID output-pkParameter that is found in sensitivity results index file
}
\description{
Retrieve list of plots of population sensitivity analyses across all populations
}
\keyword{internal}
