% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{plotQualificationDDIs}
\alias{plotQualificationDDIs}
\title{plotQualificationDDIs}
\usage{
plotQualificationDDIs(configurationPlan, settings)
}
\arguments{
\item{configurationPlan}{A \code{ConfigurationPlan} object}

\item{settings}{A \code{TaskSettings} object}
}
\value{
list of qualification DDI ggplot objects
}
\description{
Plot observation vs prediction for DDI qualification workflow
}
\keyword{internal}
