% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification-gof.R
\name{plotQualificationGOFPredictedVsObserved}
\alias{plotQualificationGOFPredictedVsObserved}
\title{plotQualificationGOFPredictedVsObserved}
\usage{
plotQualificationGOFPredictedVsObserved(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
ggplot object of time profile for qualification workflow
}
\description{
Plot observation vs prediction for qualification workflow
}
