% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{plotQualificationGOFResidualsOverTime}
\alias{plotQualificationGOFResidualsOverTime}
\title{plotQualificationGOFResidualsOverTime}
\usage{
plotQualificationGOFResidualsOverTime(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
ggplot object of residuals over time profile for qualification workflow
}
\description{
Plot observation vs prediction for qualification workflow
}
