% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{plotQualificationMeanTimeProfile}
\alias{plotQualificationMeanTimeProfile}
\title{plotQualificationMeanTimeProfile}
\usage{
plotQualificationMeanTimeProfile(
  configurationPlanCurves,
  simulation,
  simulationResults,
  axesProperties,
  configurationPlan,
  plotConfiguration
)
}
\arguments{
\item{configurationPlanCurves}{\code{Curves} fields of configuration plan}

\item{simulation}{A \code{Simulation} object from \code{ospsuite} package
that includes required information to identify and convert the data requested from \code{configurationPlanCurve} properties}

\item{simulationResults}{A \code{SimulationResults} object from \code{ospsuite} package
that includes the data requested from \code{configurationPlanCurve} properties}

\item{axesProperties}{list of axes properties obtained from \code{getAxesProperties}}

\item{configurationPlan}{A \code{ConfigurationPlan} object that includes methods to find observed data}

\item{plotConfiguration}{A \code{TimeProfilePlotConfiguration} object}
}
\value{
Mean time profile plot as a \code{ggplot} object
}
\description{
Plot mean time profile for qualification workflow
}
\keyword{internal}
