% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{plotQualificationPKRatio}
\alias{plotQualificationPKRatio}
\title{plotQualificationPKRatio}
\usage{
plotQualificationPKRatio(configurationPlan, settings)
}
\arguments{
\item{configurationPlan}{A \code{ConfigurationPlan} object}

\item{settings}{settings for the task}
}
\value{
list with \code{plots} and \code{tables}
}
\description{
Plot PK Ratio plots for qualification workflow
}
\keyword{internal}
