% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{plotQualificationPopulationTimeProfile}
\alias{plotQualificationPopulationTimeProfile}
\title{plotQualificationPopulationTimeProfile}
\usage{
plotQualificationPopulationTimeProfile(
  simulationAnalysis,
  observedDataCollection,
  simulation,
  simulationResults,
  axesProperties,
  configurationPlan,
  plotObject
)
}
\arguments{
\item{simulationAnalysis}{Field \code{Analysis} from \code{ConfigurationPlan} population time profile plot}

\item{observedDataCollection}{Field \code{ObservedDataCollection} from \code{ConfigurationPlan} population time profile plot}

\item{simulation}{A \code{Simulation} object from \code{ospsuite} package
that includes required information to identify and convert the data requested from \code{configurationPlanCurve} properties}

\item{simulationResults}{A \code{SimulationResults} object from \code{ospsuite} package
that includes the data requested from \code{configurationPlanCurve} properties}

\item{axesProperties}{list of axes properties obtained from \code{getAxesProperties}}

\item{configurationPlan}{A \code{ConfigurationPlan} object that includes methods to find observed data}

\item{plotObject}{A \code{ggplot} object}
}
\value{
Population time profile plot as a \code{ggplot} object
}
\description{
Plot population time profile for qualification workflow
}
\keyword{internal}
