% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotResidualsHistogram}
\alias{plotResidualsHistogram}
\title{plotResidualsHistogram}
\usage{
plotResidualsHistogram(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  bins = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{dataMapping}{`HistogramDataMapping` R6 class object from `tlf` library}

\item{plotConfiguration}{`PlotConfiguration` R6 class object from `tlf` library}

\item{bins}{number of bins defined in the histogram}
}
\value{
ggplot object of log residuals histogram
}
\description{
Plot histogram of residuals
}
