% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotResidualsQQPlot}
\alias{plotResidualsQQPlot}
\title{plotResidualsQQPlot}
\usage{
plotResidualsQQPlot(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{dataMapping}{`HistogramDataMapping` R6 class object from `tlf` library}

\item{plotConfiguration}{`PlotConfiguration` R6 class object from `tlf` library}
}
\value{
ggplot object of log residuals qq-plot
}
\description{
Plot quantile-quantile plot for residuals
}
