% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-of-work.R, R/utils-plots.R
\name{plotSensitivity}
\alias{plotSensitivity}
\title{plotSensitivity}
\usage{
plotSensitivity(pkSensitivities = NULL, plotConfiguration = NULL)

plotSensitivity(pkSensitivities = NULL, plotConfiguration = NULL)
}
\arguments{
\item{pkSensitivities}{List of PK senstivities computed by runSensitivityAnalysis}

\item{plotConfiguration}{List of PlotConfiguration class objects for each plot}

\item{pkSensitivities}{List of PK senstivities computed by runSensitivityAnalysis}

\item{plotConfiguration}{List of PlotConfiguration class objects for each plot}
}
\value{
sensitivityPlot ggplot object

sensitivityPlot ggplot object
}
\description{
Plot sensitivity analysis results

Plot sensitivity analysis results
}
