% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-task.R
\name{PlotTask}
\alias{PlotTask}
\title{PlotTask}
\description{
R6 class for PlotTask settings
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Task]{ospsuite.reportingengine::Task}} -> \code{PlotTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{section title in the report corresponding to the task}

\item{\code{fileName}}{name of report appendix file associated to task}

\item{\code{getTaskResults}}{function called by task that computes and format figure results}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlotTask$new()}}
\item \href{#method-saveResults}{\code{PlotTask$saveResults()}}
\item \href{#method-runTask}{\code{PlotTask$runTask()}}
\item \href{#method-clone}{\code{PlotTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate">}\href{../../ospsuite.reportingengine/html/Task.html#method-activate}{\code{ospsuite.reportingengine::Task$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Task.html#method-inactivate}{\code{ospsuite.reportingengine::Task$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print">}\href{../../ospsuite.reportingengine/html/Task.html#method-print}{\code{ospsuite.reportingengine::Task$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Task.html#method-validateInput}{\code{ospsuite.reportingengine::Task$validateInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `PlotTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$new(reportTitle = NULL, fileName = NULL, getTaskResults = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reportTitle}}{title to be printed in the report}

\item{\code{fileName}}{name of report appendix file associated to task}

\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{...}}{input parameters inherited from `Task` R6 class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PlotTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveResults"></a>}}
\subsection{Method \code{saveResults()}}{
Save results from task run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$saveResults(set, taskResults)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set}}{R6 class `SimulationStructure`}

\item{\code{taskResults}}{list of results from task run.
Results contains at least 2 fields: `plots` and `tables`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runTask"></a>}}
\subsection{Method \code{runTask()}}{
Run task and save its output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$runTask(structureSets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{structureSets}}{list of `SimulationStructure` R6 class}

\item{\code{self$fileName}}{name of report file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
