% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{plotTornado}
\alias{plotTornado}
\title{plotTornado}
\usage{
plotTornado(data, settings = NULL)
}
\arguments{
\item{data}{data.frame}

\item{settings}{`SensitivityPlotSettings` object}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot sensitivity results in a tornado plot
}
