% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-sensitivity-analysis-settings.R
\name{PopulationSensitivityAnalysisSettings}
\alias{PopulationSensitivityAnalysisSettings}
\title{PopulationSensitivityAnalysisSettings}
\description{
R6 class for Population Sensitivity Analysis Settings
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{variationRange}}{variation range for sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{quantileVec}}{vector of quantiles to be calculated}

\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}

\item{\code{pkParameterSelection}}{list of selected PK parameters for sensitivity analysis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationSensitivityAnalysisSettings$new()}}
\item \href{#method-updateVariationRange}{\code{PopulationSensitivityAnalysisSettings$updateVariationRange()}}
\item \href{#method-updateNumberOfCores}{\code{PopulationSensitivityAnalysisSettings$updateNumberOfCores()}}
\item \href{#method-updateQuantileVec}{\code{PopulationSensitivityAnalysisSettings$updateQuantileVec()}}
\item \href{#method-updateVariableParameterPaths}{\code{PopulationSensitivityAnalysisSettings$updateVariableParameterPaths()}}
\item \href{#method-updatePKParameterSelection}{\code{PopulationSensitivityAnalysisSettings$updatePKParameterSelection()}}
\item \href{#method-updateShowProgress}{\code{PopulationSensitivityAnalysisSettings$updateShowProgress()}}
\item \href{#method-clone}{\code{PopulationSensitivityAnalysisSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `PopulationSensitivityAnalysisSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$new(
  variationRange = NULL,
  numberOfCores = NULL,
  quantileVec = NULL,
  variableParameterPaths = NULL,
  pkParameterSelection = NULL,
  showProgress = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variationRange}}{variation range for sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{quantileVec}}{vector of quantiles to be calculated}

\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}

\item{\code{pkParameterSelection}}{list of selected PK parameters for sensitivity analysis}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationSensitivityAnalysisSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateVariationRange"></a>}}
\subsection{Method \code{updateVariationRange()}}{
Update variation range of sensitivity analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$updateVariationRange(variationRange)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variationRange}}{variation range for sensitivity analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateNumberOfCores"></a>}}
\subsection{Method \code{updateNumberOfCores()}}{
Update number of cores to be used during the sensitivity analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$updateNumberOfCores(numberOfCores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{number of cores for parallel computation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateQuantileVec"></a>}}
\subsection{Method \code{updateQuantileVec()}}{
Update vector of quantiles to be calculated for sensitivity analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$updateQuantileVec(quantileVec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantileVec}}{vector of quantiles to be calculated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateVariableParameterPaths"></a>}}
\subsection{Method \code{updateVariableParameterPaths()}}{
Update vector of paths of parameters to vary during sensitivity analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$updateVariableParameterPaths(
  variableParameterPaths
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePKParameterSelection"></a>}}
\subsection{Method \code{updatePKParameterSelection()}}{
Update list of selected PK parameters for sensitivity analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$updatePKParameterSelection(
  pkParameterSelection
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pkParameterSelection}}{list of selected PK parameters for sensitivity analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateShowProgress"></a>}}
\subsection{Method \code{updateShowProgress()}}{
Update `showProgress`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$updateShowProgress(showProgress)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{showProgress}}{is a logical input.  TRUE shows progress of sensitivity analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
