% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-sensitivity-analysis-task.R
\name{PopulationSensitivityAnalysisTask}
\alias{PopulationSensitivityAnalysisTask}
\title{PopulationSensitivityAnalysisTask}
\description{
R6 class for PopulationSensitivityAnalysisTask settings
}
\seealso{
Other workflow tasks: 
\code{\link{GofPlotTask}},
\code{\link{PlotTask}},
\code{\link{PopulationPlotTask}},
\code{\link{QualificationTask}},
\code{\link{SensitivityAnalysisTask}},
\code{\link{Task}},
\code{\link{addUserDefinedTask}()},
\code{\link{loadCalculatePKParametersTask}()},
\code{\link{loadCalculateSensitivityTask}()},
\code{\link{loadGOFMergedTask}()},
\code{\link{loadPlotAbsorptionTask}()},
\code{\link{loadPlotDDIRatioTask}()},
\code{\link{loadPlotDemographyTask}()},
\code{\link{loadPlotMassBalanceTask}()},
\code{\link{loadPlotPKParametersTask}()},
\code{\link{loadPlotPKRatioTask}()},
\code{\link{loadPlotSensitivityTask}()},
\code{\link{loadPlotTimeProfilesAndResidualsTask}()},
\code{\link{loadQualificationComparisonTimeProfileTask}()},
\code{\link{loadQualificationTimeProfilesTask}()},
\code{\link{loadSimulateTask}()}
}
\concept{workflow tasks}
\section{Super classes}{
\code{ospsuite.reportingengine::Task} -> \code{ospsuite.reportingengine::SensitivityAnalysisTask} -> \code{PopulationSensitivityAnalysisTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{settings}}{A \code{SensitivityAnalysisSettings} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PopulationSensitivityAnalysisTask-new}{\code{PopulationSensitivityAnalysisTask$new()}}
\item \href{#method-PopulationSensitivityAnalysisTask-saveResults}{\code{PopulationSensitivityAnalysisTask$saveResults()}}
\item \href{#method-PopulationSensitivityAnalysisTask-clone}{\code{PopulationSensitivityAnalysisTask$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-activate'><code>ospsuite.reportingengine::Task$activate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getAbsolutePath"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getAbsolutePath'><code>ospsuite.reportingengine::Task$getAbsolutePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getInputs"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getInputs'><code>ospsuite.reportingengine::Task$getInputs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getRelativePath"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getRelativePath'><code>ospsuite.reportingengine::Task$getRelativePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-inactivate'><code>ospsuite.reportingengine::Task$inactivate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-print'><code>ospsuite.reportingengine::Task$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-validateInput'><code>ospsuite.reportingengine::Task$validateInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateStructureSetInput"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-validateStructureSetInput'><code>ospsuite.reportingengine::Task$validateStructureSetInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="SensitivityAnalysisTask" data-id="runTask"><a href='../../ospsuite.reportingengine/html/SensitivityAnalysisTask.html#method-SensitivityAnalysisTask-runTask'><code>ospsuite.reportingengine::SensitivityAnalysisTask$runTask()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationSensitivityAnalysisTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationSensitivityAnalysisTask-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{PopulationSensitivityAnalysisTask} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisTask$new(
  getTaskResults = NULL,
  settings = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{settings}}{\code{SensitivityAnalysisSettings} object}

\item{\code{...}}{parameters inherited from R6 class \code{SensitivityAnalysisTask} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PopulationSensitivityAnalysisTask} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationSensitivityAnalysisTask-saveResults"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationSensitivityAnalysisTask-saveResults}{}}}
\subsection{Method \code{saveResults()}}{
Save the task results related to a \code{structureSet}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisTask$saveResults(structureSet, taskResults)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{structureSet}}{A \code{SimulationStructure} object defining the properties of a simulation set}

\item{\code{taskResults}}{list of results from task run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationSensitivityAnalysisTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationSensitivityAnalysisTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
