% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-simulation-set.R
\name{PopulationSimulationSet}
\alias{PopulationSimulationSet}
\title{PopulationSimulationSet}
\description{
R6 class representing Reporting Engine Population Model Set
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:SimulationSet]{ospsuite.reportingengine::SimulationSet}} -> \code{PopulationSimulationSet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{populationFile}}{name of csv file to be used for the population}

\item{\code{populationName}}{display name of population}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationSimulationSet$new()}}
\item \href{#method-copyInputFiles}{\code{PopulationSimulationSet$copyInputFiles()}}
\item \href{#method-clone}{\code{PopulationSimulationSet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `PopulationSimulationSet` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSet$new(
  simulationSetName = NULL,
  simulationFile,
  simulationName = NULL,
  pathID = NULL,
  pathName = NULL,
  pathUnit = NULL,
  pkParameters = NULL,
  pkParametersNames = NULL,
  pkParametersUnits = NULL,
  dataFilter = NULL,
  observedDataFile = NULL,
  observedMetaDataFile = NULL,
  populationFile,
  populationName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{simulationName}}{display name of simulation}

\item{\code{pathID}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{pathName}}{display name for `pathID`}

\item{\code{pathUnit}}{display unit for `pathID`}

\item{\code{pkParameters}}{PK parameters function names to be calculated from the simulation (e.g. `C_max`).
Default value is obtained from enum(ospsuite::allPKParameterNames()).}

\item{\code{pkParametersNames}}{display names for `pkParameters`}

\item{\code{pkParametersUnits}}{display units for `pkParameters`}

\item{\code{dataFilter}}{characters or expression to filter the observed data}

\item{\code{observedDataFile}}{name of csv file to be used for observed data}

\item{\code{observedMetaDataFile}}{name of csv file to be used as dictionary of the observed data}

\item{\code{populationFile}}{name of csv file to be used for the population}

\item{\code{populationName}}{display name of population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelSet` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyInputFiles"></a>}}
\subsection{Method \code{copyInputFiles()}}{
Copy input files into a simulation set specific filder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSet$copyInputFiles(inputFilesFolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputFilesFolder}}{where input are located}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
