% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-simulation-settings.R
\name{PopulationSimulationSettings}
\alias{PopulationSimulationSettings}
\title{PopulationSimulationSettings}
\description{
R6 class for Population Simulation Settings
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numberOfCores}}{number of cores for parallel computation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationSimulationSettings$new()}}
\item \href{#method-updateNumberOfCores}{\code{PopulationSimulationSettings$updateNumberOfCores()}}
\item \href{#method-updateShowProgress}{\code{PopulationSimulationSettings$updateShowProgress()}}
\item \href{#method-clone}{\code{PopulationSimulationSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `PopulationSimulationSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSettings$new(
  numberOfCores = defaultSimulationNumberOfCores,
  showProgress = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{showProgress}}{option to print progress of simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationSimulationSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateNumberOfCores"></a>}}
\subsection{Method \code{updateNumberOfCores()}}{
Update the `numberOfCores`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSettings$updateNumberOfCores(numberOfCores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{is the number of cores to use for simulation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateShowProgress"></a>}}
\subsection{Method \code{updateShowProgress()}}{
Update `showProgress`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSettings$updateShowProgress(showProgress)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{showProgress}}{is a logical input.  TRUE shows progress of simulation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
