% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{class}
\name{PopulationWorkflow}
\alias{PopulationWorkflow}
\title{PopulationWorkflow}
\usage{
PopulationWorkflow
}
\description{
Population Workflow Task for Reporting Engine
}
\section{Fields}{

\describe{
\item{\code{populationSimulation}}{Population simulation task}

\item{\code{pkParametersCalculation}}{PK parameters calculation task}

\item{\code{sensitivityAnalysis}}{Sensitivity analysis task}

\item{\code{demographyPlot}}{Plot demography task}

\item{\code{timeProfilePlot}}{Plot time profile task}

\item{\code{pkParametersPlot}}{Plot PK parameters task}

\item{\code{sensitivityPlot}}{Plot sensitiviy task}
}}

\section{Methods}{

\describe{
\item{new()}{Initilialize reporting engine population workflow}
\item{runPopulationSimulation()}{Simulate population(s)}
\item{runPKParametersCalculation(Calculate population PK parameters)}{}
\item{runSensitivityAnalysis(Calculate population sensitivity)}{}
\item{plotDemography()}{}
\item{plotSensitivity()}{}
\item{runWorkflow()}{}
}
}

\keyword{datasets}
