% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R
\docType{data}
\name{PopulationWorkflowTypes}
\alias{PopulationWorkflowTypes}
\title{PopulationWorkflowTypes}
\format{
An object of class \code{list} of length 3.
}
\usage{
PopulationWorkflowTypes
}
\description{
List of population workflow available types
}
\examples{
PopulationWorkflowTypes$pediatric
PopulationWorkflowTypes$parallelComparison
PopulationWorkflowTypes$ratioComparison
}
\keyword{datasets}
