% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{prettyCaption}
\alias{prettyCaption}
\title{prettyCaption}
\usage{
prettyCaption(
  captions,
  maxLines = reEnv$maxLinesPerLegendCaption,
  width = reEnv$maxWidthPerLegendCaption
)
}
\arguments{
\item{captions}{Array of character strings to render}

\item{maxLines}{Maximum number of lines directly setting the maximum number of line breaks allowed.}

\item{width}{Maximum number of characters per line desired.
Due to `maxLines`, the returned width can be wider than `width`.}
}
\value{
A character vector of wrapped strings with line breaks at sensible places.
}
\description{
Get prettied captions with line breaks to prevent cropping of long captions
}
\examples{

cat(prettyCaption("this-is-a-long-sentence-with-dashes", maxLines = 2, width = 25))

cat(prettyCaption("this is a sentence with spaces", maxLines = 2, width = 25))

cat(prettyCaption("this_is_a_long_sentence_without_preferential_splits", maxLines = 2, width = 25))

cat(prettyCaption("this too short to split", maxLines = 3, width = 40))

cat(prettyCaption("this forces the sentence to use one line", maxLines = 1, width = 5))
}
