% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{qualificationCatch}
\alias{qualificationCatch}
\title{qualificationCatch}
\usage{
qualificationCatch(expr, configurationPlanField = NULL)
}
\arguments{
\item{expr}{Evaluated code chunks}

\item{configurationPlanField}{A list extracted from a configuration plan field}
}
\description{
Add configuration plan plot information when catching a warning/error
}
\keyword{internal}
