% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-workflow.R
\name{QualificationWorkflow}
\alias{QualificationWorkflow}
\title{QualificationWorkflow}
\description{
R6 class for Reporting Engine Qualification Workflow
}
\seealso{
Other workflows: 
\code{\link{MeanModelWorkflow}},
\code{\link{PopulationWorkflow}}
}
\concept{workflows}
\section{Super class}{
\code{ospsuite.reportingengine::Workflow} -> \code{QualificationWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configurationPlan}}{`ConfigurationPlan` object}

\item{\code{simulate}}{`SimulationTask` object for time profile simulations}

\item{\code{calculatePKParameters}}{`CalculatePKParametersTask` object for PK parameters calculation}

\item{\code{plotTimeProfiles}}{`PlotTask` object for time profile plots}

\item{\code{plotComparisonTimeProfile}}{`PlotTask` object for comparison of time profiles plots}

\item{\code{plotGOFMerged}}{`PlotTask` object for goodness of fit plots}

\item{\code{plotPKRatio}}{`PlotTask` object for PK ratio plot}

\item{\code{plotDDIRatio}}{`PlotTask` object for DDI ratio plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QualificationWorkflow-new}{\code{QualificationWorkflow$new()}}
\item \href{#method-QualificationWorkflow-runWorkflow}{\code{QualificationWorkflow$runWorkflow()}}
\item \href{#method-QualificationWorkflow-updateConfigurationPlan}{\code{QualificationWorkflow$updateConfigurationPlan()}}
\item \href{#method-QualificationWorkflow-clone}{\code{QualificationWorkflow$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-activateTasks'><code>ospsuite.reportingengine::Workflow$activateTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getActiveTasks'><code>ospsuite.reportingengine::Workflow$getActiveTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getAllPlotTasks'><code>ospsuite.reportingengine::Workflow$getAllPlotTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getAllTasks'><code>ospsuite.reportingengine::Workflow$getAllTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getInactiveTasks'><code>ospsuite.reportingengine::Workflow$getInactiveTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getParameterDisplayPaths"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getParameterDisplayPaths'><code>ospsuite.reportingengine::Workflow$getParameterDisplayPaths()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getSimulationDescriptor"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getSimulationDescriptor'><code>ospsuite.reportingengine::Workflow$getSimulationDescriptor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getWatermark"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getWatermark'><code>ospsuite.reportingengine::Workflow$getWatermark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-inactivateTasks'><code>ospsuite.reportingengine::Workflow$inactivateTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="print"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-print'><code>ospsuite.reportingengine::Workflow$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="printReportingEngineInfo"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-printReportingEngineInfo'><code>ospsuite.reportingengine::Workflow$printReportingEngineInfo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setParameterDisplayPaths"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setParameterDisplayPaths'><code>ospsuite.reportingengine::Workflow$setParameterDisplayPaths()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setSimulationDescriptor"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setSimulationDescriptor'><code>ospsuite.reportingengine::Workflow$setSimulationDescriptor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setWatermark"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setWatermark'><code>ospsuite.reportingengine::Workflow$setWatermark()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationWorkflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationWorkflow-new}{}}}
\subsection{Method \code{new()}}{
Create a new `QualificationWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$new(configurationPlan, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configurationPlan}}{A `ConfigurationPlan` object}

\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `QualificationWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationWorkflow-runWorkflow"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationWorkflow-runWorkflow}{}}}
\subsection{Method \code{runWorkflow()}}{
Run qualification workflow tasks for all simulation sets if tasks are activated
The order of tasks is as follows:
# 1) Run simulations
# 2) Perform PK analyses
# 3) Perform plot tasks
## 3.a) time profiles and residual plots
## 3.b) comparison time profiles plots
## 3.c) PK ratio tables and plots
## 3.d) DDI ratio tables and plots
# 4) Render report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in the workflow folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationWorkflow-updateConfigurationPlan"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationWorkflow-updateConfigurationPlan}{}}}
\subsection{Method \code{updateConfigurationPlan()}}{
Update the content of the workflow `configurationPlan`.
Caution, updating the `configurationPlan` using this method won't update the workflow simulations and their results.
Use the method only to bypass reloading a full workflow if only plot aesthetics or section content is changed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$updateConfigurationPlan(configurationPlanFile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configurationPlanFile}}{path to the json file corresponding to the Configuration Plan of a Qualification workflow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationWorkflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationWorkflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
