% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{ratioBoxplot}
\alias{ratioBoxplot}
\title{ratioBoxplot}
\usage{
ratioBoxplot(data, plotConfiguration = NULL)
}
\arguments{
\item{data}{data.frame of the ratios}

\item{plotConfiguration}{PlotConfiguration R6 class object}
}
\value{
ggplot object
}
\description{
Plot box-whiskers of ratios as is
}
