% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{rbindPKParametersTables}
\alias{rbindPKParametersTables}
\title{rbindPKParametersTables}
\usage{
rbindPKParametersTables(pkParametersTableAcrossPopulations, pkParametersTable)
}
\arguments{
\item{pkParametersTableAcrossPopulations}{A data.frame of PK and Population Parameters across simulation sets}

\item{pkParametersTable}{A data.frame of PK and Population Parameters for a simulation set}
}
\value{
A data.frame of PK and Population Parameters across simulation sets
}
\description{
Concatenate data.frames of PK and Population Parameters across simulation sets
}
\keyword{internal}
