% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{readObservedDataFile}
\alias{readObservedDataFile}
\title{readObservedDataFile}
\usage{
readObservedDataFile(fileName, header = TRUE, encoding = "UTF-8")
}
\arguments{
\item{fileName}{name of file to be read}

\item{header}{logical indicating if data has a header}

\item{encoding}{encoding of the file}
}
\value{
data.frame containing observed data
}
\description{
Read observed data file with Nonmem format.
Can read csv files as well as
}
