% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\docType{data}
\name{reEnv}
\alias{reEnv}
\title{reEnv}
\format{
An object of class \code{environment} of length 38.
}
\usage{
reEnv
}
\description{
Environment that holds various global variables and settings for the RE library.
It is not exported and should not be directly manipulated by other packages.
}
\section{Fields}{

\describe{
\item{\code{packageName}}{Name of the package.
This will be used to retrieve information on the package at run time.}

\item{\code{log}}{`Logging` object defining how to record log messages}

\item{\code{defaultLogPrint}}{List of logical values indicating whether to print log messages on console}

\item{\code{formatNumericsDigits}}{Number of digits to display for numeric values in reports}

\item{\code{formatNumericsScientific}}{Logical value indicating whether to use scientific notation for numeric values in reports}

\item{\code{defaultBins}}{Default number of bins to use in histograms}

\item{\code{binUsingQuantiles}}{Logical value indicating whether splitting method for bins
`TRUE` uses same number of points in bins through quantiles
`FALSE` uses same interval ranges}

\item{\code{defaultStairstep}}{Logical value indicating whether to use `stairstep` lines in range plots}

\item{\code{defaultSimulationSetDescriptor}}{Default descriptor for simulation sets to include in reports}

\item{\code{maximalParametersPerSensitivityPlot}}{Maximum number of parameters to include in a sensitivity plot}

\item{\code{maxWidthPerParameter}}{Maximum number of characters allowed before adding line breaks
in sensitivity plot parameter captions.}

\item{\code{maxLinesPerParameter}}{Maximum number of lines allowed when line breaks need to be included
in sensitivity plot parameter captions.}

\item{\code{maxWidthPerLegendCaption}}{Maximum number of characters allowed before adding line breaks in legend captions.}

\item{\code{maxLinesPerLegendCaption}}{Maximum number of lines allowed when line breaks need to be included in legend captions.}

\item{\code{autoAxisLimitMargin}}{For DDI Ratio plots,
value of the margin to add to the limits of the axis when left undefined}

\item{\code{blankLinesBetweenArtifacts}}{Default number of line breaks to include in reports after inclusion of artifact}

\item{\code{fontScaleFactor}}{Arbitrary scale factor for font size when exporting plot as png to render prettier captions}

\item{\code{workflowWatermarkMessage}}{Default watermark message to print when system is not validated}

\item{\code{residualsHistogramLabel}}{Default ylabel in goodness of fit residuals histogram plots}

\item{\code{demographyHistogramLabel}}{Default ylabel in demography histogram plots}

\item{\code{residualsQQLabel}}{Default ylabel in goodness of fit QQ-plots}

\item{\code{referenceColor}}{Color of reference simulationSet in plots for Population Workflows}

\item{\code{referenceFill}}{Fill color of reference simulationSet in plots for Population Workflows}

\item{\code{defaultErrorbarCapSize}}{Default size of the error bar caps in the plots}

\item{\code{defaultSimulationNumberOfCores}}{default numberOfCores for simulation}

\item{\code{defaultSensitivityAnalysisNumberOfCores}}{default numberOfCores for sensitivity analysis}

\item{\code{defaultVariationRange}}{default parameter variation range for sensitivity analysis}

\item{\code{defaultQuantileVec}}{default quantiles for population sensitivity analysis}

\item{\code{defaultMaxSimulationsPerCore}}{Default value for a scale factor used in a parallel simulation.
The product of this scale factor and the number of allowable cores (`allowedCores`)
sets the maximum number of simulations that may be run on one core.}

\item{\code{defaultMCRandomSeed}}{Default Random Seed for Monte Carlo sampling when calculating PK Ratios in population workflows}

\item{\code{defaultMCRepetitions}}{Default number of repetitions for Monte Carlo sampling when calculating PK Ratios in population workflows}

\item{\code{defaultPKParametersHeader}}{Default header for PK parameters summary tables in reports}

\item{\code{pkRatioDictionary}}{Dictionary of column names for PK Ratio tables}

\item{\code{ddiRatioListColumnMappings}}{Dictionary of column names for DDI Ratio tables}

\item{\code{ddiRatioSubsetsDictionary}}{Dictionary of sub-sections names for DDI Ratios}

\item{\code{theme}}{A `Theme` object from the `{tlf}` package defining default properties of the figures}

\item{\code{defaultPlotFormat}}{List of default format when exporting plots}

\item{\code{defaultTimeProfileStatistics}}{A list of functions and labels defining the summary statistics and captions for population time profiles}
}}

\keyword{internal}
