% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{removeCommentsFromWorkflowContent}
\alias{removeCommentsFromWorkflowContent}
\title{removeCommentsFromWorkflowContent}
\usage{
removeCommentsFromWorkflowContent(workflowContent, commentPattern = "#")
}
\arguments{
\item{workflowContent}{Character vector with content of the workflow file}

\item{commentPattern}{Character starting a comment}
}
\value{
Workflow content without its comments
}
\description{
Remove comments from workflow file content
}
\keyword{internal}
