% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeForbiddenLetters}
\alias{removeForbiddenLetters}
\title{removeForbiddenLetters}
\usage{
removeForbiddenLetters(
  text,
  forbiddenLetters = "[[:punct:][:blank:]]",
  replacement = "_"
)
}
\arguments{
\item{text}{character string to be evaluated}

\item{forbiddenLetters}{characters to be removed if in the \code{text}.
Default value of \code{forbiddenLetters} is \code{"[[:punct:]]"}
meaning that all pointuation characters are forbidden.}

\item{replacement}{character replacing the \code{forbiddenLetters}.
Default value of \code{forbiddenLetters} is "_".}
}
\value{
\code{text} character string with forbidden letters replaced
}
\description{
Trim path and extension of a file
}
\examples{
\dontrun{
removeForbiddenLetters(text)
}
}
