% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeMissingValues}
\alias{removeMissingValues}
\title{removeMissingValues}
\usage{
removeMissingValues(data, dataMapping = NULL)
}
\arguments{
\item{data}{data.frame}

\item{dataMapping}{name of variable on which the missing values are checked}
}
\value{
filtered data.frame
}
\description{
removeMissingValues
}
\keyword{internal}
