% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeNegativeValues}
\alias{removeNegativeValues}
\title{removeNegativeValues}
\usage{
removeNegativeValues(data, dataMapping = NULL)
}
\arguments{
\item{data}{data.frame}

\item{dataMapping}{name of variable on which the missing values are checked}
}
\value{
filtered data.frame
}
\description{
removeNegativeValues
}
\keyword{internal}
