% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{removePathStrings}
\alias{removePathStrings}
\title{removePathStrings}
\usage{
removePathStrings(pth)
}
\arguments{
\item{pth}{string from which to remove}
}
\description{
remove occurences of any of the substrings "Applications-","Neighborhoods-","Organism-","ProtocolSchemaItem-" from a path
}
