% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{removeTempPopFiles}
\alias{removeTempPopFiles}
\title{removeTempPopFiles}
\usage{
removeTempPopFiles(folderName, fileNamesVec)
}
\description{
Function to remove temporary population data files previously created when splitting main population data file for parallelization.

Input is vector of strings containting names of temporary population data files, as output by function splitPopDataFile

No output.
}
