% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{renderReport}
\alias{renderReport}
\title{renderReport}
\usage{
renderReport(
  fileName,
  createWordReport = FALSE,
  numberSections = TRUE,
  intro = NULL,
  wordConversionTemplate = NULL
)
}
\arguments{
\item{fileName}{name of .md file to render}

\item{createWordReport}{option for creating Markdown-Report only but not a Word-Report}

\item{numberSections}{logical defining if sections are numbered}

\item{intro}{name of .md file that include introduction (before toc)}

\item{wordConversionTemplate}{optional docx template for rendering a tuned Word-Report document}
}
\description{
Render report with number sections and table of content
}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
