% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{renderRmdFile}
\alias{renderRmdFile}
\title{renderRmdFile}
\usage{
renderRmdFile(fileName)
}
\arguments{
\item{fileName}{name of Rmd file}
}
\description{
Translate R markdown document into markdown and html document
}
