% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{renderSaParameterDisplayName}
\alias{renderSaParameterDisplayName}
\title{renderSaParameterDisplayName}
\usage{
renderSaParameterDisplayName(paths, depth = 3)
}
\arguments{
\item{paths}{vector of paths to render}

\item{depth}{levels in parameter path}
}
\value{
Updated paths
}
\description{
Render display names for sensitivity plots keeping elements in `depth` of parameter paths,
and removing following keys `Organism`, `Applications`, `ProtocolSchemaItem`, and `Neighborhoods`
}
