% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{repeatableSampling}
\alias{repeatableSampling}
\title{repeatableSampling}
\usage{
repeatableSampling(
  x,
  size,
  n = getDefaultMCRepetitions(),
  seed = getDefaultMCRandomSeed()
)
}
\arguments{
\item{x}{Array to be sampled}

\item{size}{Number of samples}

\item{n}{Number of repetitions}

\item{seed}{Random Seed}
}
\value{
A list of `n` arrays of length `size` sampled from `x`
}
\description{
Perform repeatable random sampling
}
\keyword{internal}
