% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replaceInfWithNA}
\alias{replaceInfWithNA}
\title{replaceInfWithNA}
\usage{
replaceInfWithNA(data)
}
\arguments{
\item{data}{numeric vector}
}
\value{
numeric vector
}
\description{
replaceInfWithNA
}
\keyword{internal}
