% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-info.R
\docType{class}
\name{ReportingEngineInfo}
\alias{ReportingEngineInfo}
\title{ReportingEngineInfo}
\usage{
ReportingEngineInfo
}
\description{
R6 class providing relevant information regarding the Reporting Engine
}
\section{Fields}{

\describe{
\item{\code{Date}}{(public)}

\item{\code{computerName}}{(private) name of computer}

\item{\code{userName}}{(private) name of user}

\item{\code{login}}{(private)}

\item{\code{tlfVersion}}{(private) version of tlf package}

\item{\code{ospsuiteVersion}}{(private) version of osp suite package}

\item{\code{Rversion}}{(private) version of R}

\item{\code{isValidatedSystem}}{(private) logical for computers that are evaluatued validated for Reporting Engine}
}}

\section{Methods}{

\describe{
\item{new()}{Initialize ReportingEngineInfo}
\item{print()}{Print Reporting Engine Inforation}
}
}

\keyword{datasets}
