% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{resetLogs}
\alias{resetLogs}
\title{resetLogs}
\usage{
resetLogs(folder)
}
\arguments{
\item{folder}{Folder where logs are saved}
}
\description{
Reset/empty messages of global logging system
}
\keyword{internal}
