% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{resetReport}
\alias{resetReport}
\title{resetReport}
\usage{
resetReport(fileName, logFolder = getwd())
}
\arguments{
\item{fileName}{name of .md file to reset}

\item{logFolder}{folder where the logs are saved}
}
\description{
Initialize a report, warning if a previous version already exists
}
\examples{
\dontrun{
resetReport("report.md")
}
}
