% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{resetRESettingsToDefault}
\alias{resetRESettingsToDefault}
\title{resetRESettingsToDefault}
\usage{
resetRESettingsToDefault()
}
\description{
Reset the global settings stored in `reEnv` to default values defined by the package.
}
