% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\docType{data}
\name{residualsVsObservedFlag}
\alias{residualsVsObservedFlag}
\title{DDI plot type identifier}
\format{
An object of class \code{list} of length 2.
}
\usage{
residualsVsObservedFlag
}
\description{
DDI plot type identifier
}
\keyword{datasets}
