% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runMeanModelSensitivity}
\alias{runMeanModelSensitivity}
\title{runMeanModelSensitivity}
\usage{
runMeanModelSensitivity(structureSet, logFolder = getwd(), settings = NULL)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{list of settings for the output table/plot}
}
\value{
data.frame with calculated PK parameters for the simulation set
}
\description{
Run mean model sensitivity analysis
}
