% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runParallelPopulationSimulation}
\alias{runParallelPopulationSimulation}
\title{runParallelPopulationSimulation}
\usage{
runParallelPopulationSimulation(
  numberOfCores,
  inputFolderName,
  simulationFileName,
  populationFileName,
  resultsFolderName,
  resultsFileName
)
}
\value{
Simulation results for population
}
\description{
Spawn cores, divide population among cores, run population simulation on cores, save results as CSV.
}
