% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runParallelSensitivityAnalysis}
\alias{runParallelSensitivityAnalysis}
\title{runParallelSensitivityAnalysis}
\usage{
runParallelSensitivityAnalysis(
  simFilePath,
  variableParameterPaths,
  individualParameters,
  variationRange,
  numberOfCores,
  logFolder = getwd()
)
}
\arguments{
\item{simFilePath}{path to simulation file}

\item{variableParameterPaths}{paths to parameters to vary in sensitivity analysis}

\item{individualParameters}{is an object storing an individual's parameters, obtained
from a population object's getParameterValuesForIndividual() function.}

\item{variationRange}{variation range for sensitivity analysis}

\item{numberOfCores}{is the number of cores over which to parallelize the sensitivity analysis}

\item{logFolder}{folder where the logs are saved}
}
\value{
SA results for population
}
\description{
Spawn cores, divide parameters among cores, run sensitivity analysis on cores
for a single individual, save results as CSV.
}
