% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runPopulationSensitivityAnalysis}
\alias{runPopulationSensitivityAnalysis}
\title{runPopulationSensitivityAnalysis}
\usage{
runPopulationSensitivityAnalysis(structureSet, settings)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{settings}{list of settings for the population sensitivity analysis}
}
\description{
runPopulationSensitivityAnalysis
}
\keyword{internal}
