% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runSensitivity}
\alias{runSensitivity}
\title{runSensitivity}
\usage{
runSensitivity(
  structureSet,
  settings,
  individualId = NULL,
  logFolder = getwd(),
  resultsFileName = NULL
)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{settings}{list of settings for the sensitivity analysis}

\item{individualId}{ID of individual in population data file for whom to perform sensitivity analysis}

\item{logFolder}{folder where the logs are saved}

\item{resultsFileName}{root name of population sensitivity analysis results CSV files}
}
\value{
SA results for individual or population
}
\description{
Determine whether to run SA for individual or population.  If for individual,  pass simulation to individualSensitivityAnalysis.
If SA is for population, loop thru population file, extract parameters for each individual, and pass them to individualSensitivityAnalysis.
}
\keyword{internal}
