% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-of-work.R
\name{runSensitivityAnalysis}
\alias{runSensitivityAnalysis}
\title{runSensitivityAnalysis}
\usage{
runSensitivityAnalysis()
}
\arguments{
\item{simulation}{simulation class object}
}
\value{
sensitivityResults
}
\description{
Run a sensitivity analysis from a simulation
}
