% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveFigure}
\alias{saveFigure}
\title{saveFigure}
\usage{
saveFigure(plotObject, fileName, logFolder, simulationSetName = NULL)
}
\arguments{
\item{plotObject}{A `ggplot` object}

\item{fileName}{Name of the file in which `plotObject` is saved}

\item{logFolder}{folder where the logs are saved}

\item{simulationSetName}{Name of the simulation set for `PlotTask` results}
}
\value{

}
\description{
Save figure and catches
}
\keyword{internal}
