% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{saveLogs}
\alias{saveLogs}
\title{saveLogs}
\usage{
saveLogs(folder = NULL)
}
\arguments{
\item{folder}{Directory into which logs will be saved}
}
\description{
Save workflow logs to their respective files
}
