% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{saveLogsToJson}
\alias{saveLogsToJson}
\title{saveLogsToJson}
\usage{
saveLogsToJson(jsonFile)
}
\arguments{
\item{jsonFile}{Path of json file saving log messages}
}
\description{
Save workflow logs to a json file
}
