% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{saveRESettings}
\alias{saveRESettings}
\title{saveRESettings
Save the current reporting engine global settings in a `.RData` file}
\usage{
saveRESettings(file)
}
\arguments{
\item{file}{`.RData` file containing the settings}
}
\description{
saveRESettings
Save the current reporting engine global settings in a `.RData` file
}
