% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-results.R
\name{saveTaskResults}
\alias{saveTaskResults}
\title{saveTaskResults}
\usage{
saveTaskResults(
  id = NULL,
  sectionId = NULL,
  plot = NULL,
  plotCaption = NULL,
  includePlot = NULL,
  table = NULL,
  tableCaption = NULL,
  includeTable = NULL,
  textChunk = NULL,
  includeTextChunk = NULL,
  taskResults = NULL
)
}
\arguments{
\item{id}{unique identifer of results, helps in deriving final plot name}

\item{sectionId}{unique identifer of section results, used only by qualification workflows}

\item{plot}{`ggplot` object corresponding to the figure to be saved}

\item{plotCaption}{text included into the report explaining the figure}

\item{includePlot}{logical indicating if the plot should be included in final report}

\item{table}{data.frame corresponding to the table to be saved}

\item{tableCaption}{text included into the report explaining the table}

\item{includeTable}{logical indicating if the table should be included in final report}

\item{textChunk}{text included into the report explaining the table}

\item{includeTextChunk}{logical indicating if the text chunk should be included in final report}
}
\value{
A `TaskResults` object
}
\description{
Save task results within a `TaskResults` object
}
