% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{sectionsAsDataFrame}
\alias{sectionsAsDataFrame}
\title{sectionsAsDataFrame}
\usage{
sectionsAsDataFrame(
  sectionsIn,
  sectionsOut = data.frame(),
  parentFolder = "images",
  sectionLevel = 1
)
}
\arguments{
\item{sectionsIn}{list including Id and Title of section}

\item{sectionsOut}{data.frame including id, path, title}

\item{parentFolder}{For subsections only, path of parent section}

\item{sectionLevel}{Section level defining the level of markdown title}
}
\value{
A data.frame including information about every section and subsection
}
\description{
Recursively parse Sections field of configuration plan
to create a data.frame easier to handle by the wrkflow
}
\keyword{internal}
