% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis-settings.R
\name{SensitivityAnalysisSettings}
\alias{SensitivityAnalysisSettings}
\title{SensitivityAnalysisSettings}
\description{
R6 class for Population Sensitivity Analysis Settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{variationRange}}{variationRange of the sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{quantileVec}}{vector of quantiles to be calculated}

\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}

\item{\code{showProgress}}{is a logical input.  TRUE shows progress of sensitivity analysis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensitivityAnalysisSettings$new()}}
\item \href{#method-clone}{\code{SensitivityAnalysisSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a `SensitivityAnalysisSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityAnalysisSettings$new(
  variationRange = NULL,
  numberOfCores = NULL,
  quantileVec = NULL,
  variableParameterPaths = NULL,
  showProgress = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variationRange}}{variation range for sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{quantileVec}}{vector of quantiles to be calculated}

\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}

\item{\code{showProgress}}{sensitivity analysis progress printed to console if TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SensitivityAnalysisSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityAnalysisSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
