% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis-task.R
\name{SensitivityAnalysisTask}
\alias{SensitivityAnalysisTask}
\title{SensitivityAnalysisTask}
\description{
R6 class for SensitivityAnalysisTask settings
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Tasks]{ospsuite.reportingengine::Tasks}} -> \code{SensitivityAnalysisTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inputFolderName}}{input folder}

\item{\code{simulationFileName}}{name of simulation files}

\item{\code{populationFileName}}{name of population files}

\item{\code{resultsFolderName}}{name of folder where results are saved}

\item{\code{resultsFileName}}{name of file where results are saved}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{inputFolderName}}{input folder}

\item{\code{simulationFileName}}{name of simulation files}

\item{\code{populationFileName}}{name of population files}

\item{\code{resultsFolderName}}{name of folder where results are saved}

\item{\code{resultsFileName}}{name of file where results are saved}

\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensitivityAnalysisTask$new()}}
\item \href{#method-clone}{\code{SensitivityAnalysisTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="activate">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-activate}{\code{ospsuite.reportingengine::Tasks$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-inactivate}{\code{ospsuite.reportingengine::Tasks$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="print">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-print}{\code{ospsuite.reportingengine::Tasks$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-validateInput}{\code{ospsuite.reportingengine::Tasks$validateInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `SensitivityAnalysisTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityAnalysisTask$new(
  numberOfCores = 1,
  quantileVec = c(0.05, 0.5, 0.95),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{...}}{parameters inherited from R6 class `Task` object}

\item{\code{inputFolderName}}{input folder}

\item{\code{simulationFileName}}{name of simulation files}

\item{\code{populationFileName}}{name of population files}

\item{\code{resultsFolderName}}{name of folder where results are saved}

\item{\code{resultsFileName}}{name of file where results are saved}

\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SensitivityAnalysisTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityAnalysisTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
