% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-plot-settings.R
\name{SensitivityPlotSettings}
\alias{SensitivityPlotSettings}
\title{SensitivityPlotSettings}
\description{
R6 class for sensitivity analysis plot settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{totalSensitivityThreshold}}{cut-off used for plots of the most sensitive parameters}

\item{\code{maximalParametersPerSensitivityPlot}}{is the maximal number of parameters to display in a sensitivity plot}

\item{\code{plotConfiguration}}{`PlotConfiguration` R6 class object from `tlf` library}

\item{\code{xAxisFontSize}}{for sensitivity plot.
This Value will overwrite values defined by plot configuration object}

\item{\code{yAxisFontSize}}{for sensitivity plot.
This Value will overwrite values defined by plot configuration object}

\item{\code{xLabel}}{for sensitivity plot.
This Value will overwrite values defined by plot configuration object}

\item{\code{yLabel}}{for sensitivity plot.
This Value will overwrite values defined by plot configuration object}

\item{\code{colorPalette}}{for sensitivity plot.
This Value will overwrite values defined by plot configuration object}

\item{\code{maxLinesPerParameter}}{maximum number of lines allowed per displayed parameters}

\item{\code{maxWidthPerParameter}}{maximum number of characters allowed per lines of displayed parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SensitivityPlotSettings-new}{\code{SensitivityPlotSettings$new()}}
\item \href{#method-SensitivityPlotSettings-clone}{\code{SensitivityPlotSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SensitivityPlotSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-SensitivityPlotSettings-new}{}}}
\subsection{Method \code{new()}}{
Create a `SensitivityPlotSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityPlotSettings$new(
  totalSensitivityThreshold = NULL,
  variableParameterPaths = NULL,
  maximalParametersPerSensitivityPlot = NULL,
  plotConfiguration = NULL,
  xAxisFontSize = 6,
  yAxisFontSize = 6,
  maxLinesPerParameter = NULL,
  maxWidthPerParameter = NULL,
  xLabel = "Sensitivity",
  yLabel = NULL,
  colorPalette = "Spectral"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalSensitivityThreshold}}{cut-off used for plots of the most sensitive parameters}

\item{\code{variableParameterPaths}}{paths that were varied in the sensitivity analysis.  If supplied totalSensitivityThreshold = 1, else 0.9.}

\item{\code{maximalParametersPerSensitivityPlot}}{maximalParametersPerSensitivityPlot is the maximal number of parameters to display in a sensitivity plot}

\item{\code{plotConfiguration}}{`PlotConfiguration` object from `tlf` library}

\item{\code{xAxisFontSize}}{Font size of x-axis labels for sensitivity plot}

\item{\code{yAxisFontSize}}{Font size of y-axis labels for sensitivity plot}

\item{\code{maxLinesPerParameter}}{maxLinesPerParameter maximum number of lines allowed per displayed parameters}

\item{\code{maxWidthPerParameter}}{maximum number of characters allowed per lines of displayed parameters}

\item{\code{xLabel}}{Label of x-axis for sensitivity plot}

\item{\code{yLabel}}{Label of y-axis for sensitivity plot}

\item{\code{colorPalette}}{Name of a color palette to be used by `ggplot2::scale_fill_brewer()` for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SensitivityPlotSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SensitivityPlotSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SensitivityPlotSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityPlotSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
