% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{separateVariableFromUnit}
\alias{separateVariableFromUnit}
\title{separateVariableFromUnit}
\usage{
separateVariableFromUnit(variableUnitString)
}
\arguments{
\item{variableUnitString}{is a string that has the format 'Variable [unit]' or 'Variable'}
}
\value{
A named list, with fields 'name' and 'unit'.  If variableUnitString lacks a '[unit]' substring, then the unit field is returned as an empty string.
}
\description{
Split a string containing a variable and unit description into a list of two strings that separates the variable and the unit.
}
\keyword{internal}
