% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{setApplicationRangeInclusion}
\alias{setApplicationRangeInclusion}
\title{setApplicationRangeInclusion}
\usage{
setApplicationRangeInclusion(workflow, applicationRange, keep)
}
\arguments{
\item{workflow}{`Workflow` object}

\item{applicationRange}{name of range (as defined in enum `ApplicationRanges`)}

\item{keep}{logical defining if the application is kept in the report}
}
\description{
For multiple applications only, 
set whether range (as defined in enum `ApplicationRanges`) is kept in the report
}
