% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{setDefaultAutoAxisLimitMargin}
\alias{setDefaultAutoAxisLimitMargin}
\title{setDefaultAutoAxisLimitMargin}
\usage{
setDefaultAutoAxisLimitMargin(margin)
}
\arguments{
\item{margin}{numeric value between 0 and 1
defining the margin of automated axis limits}
}
\description{
setDefaultAutoAxisLimitMargin
}
\examples{
setDefaultAutoAxisLimitMargin(0.1)
}
