% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{setDefaultBins}
\alias{setDefaultBins}
\title{setDefaultBins}
\usage{
setDefaultBins(bins)
}
\arguments{
\item{bins}{number or edges of bins}
}
\description{
setDefaultBins
}
\examples{
setDefaultBins(10)
}
