% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{setDefaultMCRandomSeed}
\alias{setDefaultMCRandomSeed}
\title{setDefaultMCRandomSeed}
\usage{
setDefaultMCRandomSeed(seed)
}
\arguments{
\item{seed}{Random seed number as an integer}
}
\description{
Set the default random seed when performing Monte Carlo sampling
}
\examples{
setDefaultMCRandomSeed(123456)

}
