% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{setDefaultMCRepetitions}
\alias{setDefaultMCRepetitions}
\title{setDefaultMCRepetitions}
\usage{
setDefaultMCRepetitions(n)
}
\arguments{
\item{n}{Number of repetitions}
}
\description{
Set the default number of repetitions when performing Monte Carlo sampling
}
\examples{
setDefaultMCRepetitions(1e4)

}
