% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{setDefaultNumericFormat}
\alias{setDefaultNumericFormat}
\title{setDefaultNumericFormat}
\usage{
setDefaultNumericFormat(digits = NULL, nsmall = NULL, scientific = NULL)
}
\arguments{
\item{digits}{Number of significant digits}

\item{nsmall}{Number of decimal digits
(Note: `nsmall` needs to be consistent with `digits` for being correctly used)}

\item{scientific}{Logical defining if numeric format uses a scientific expression}
}
\description{
Set default format for numeric values output in reports
}
