% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{setDefaultREErrorbarCapSize}
\alias{setDefaultREErrorbarCapSize}
\title{setDefaultREErrorbarCapSize}
\usage{
setDefaultREErrorbarCapSize(size)
}
\arguments{
\item{size}{Size of error bar caps}
}
\description{
Set the default size of error bar caps
}
\examples{
setDefaultREErrorbarCapSize(5)

}
