% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{setDefaultStairstep}
\alias{setDefaultStairstep}
\title{setDefaultStairstep}
\usage{
setDefaultStairstep(stairstep)
}
\arguments{
\item{stairstep}{logical defining if stairstep should be plotted by default when aggregation is performed}
}
\description{
setDefaultStairstep
}
\examples{
setDefaultStairstep(TRUE)
}
