% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{setDefaultTheme}
\alias{setDefaultTheme}
\title{setDefaultTheme}
\usage{
setDefaultTheme(theme = NULL)
}
\arguments{
\item{theme}{`Theme` object from `tlf` package
If `NULL`, the current theme is re-initialized to the reporting engine default}
}
\description{
Set the default plot settings for a workflow
}
