% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{setDefaultThemeFromJson}
\alias{setDefaultThemeFromJson}
\title{setDefaultThemeFromJson}
\usage{
setDefaultThemeFromJson(jsonFile)
}
\arguments{
\item{jsonFile}{path to json file where theme properties are stored}
}
\description{
setDefaultThemeFromJson
}
