% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{setDefaultTimeProfileStatistics}
\alias{setDefaultTimeProfileStatistics}
\title{setDefaultTimeProfileStatistics}
\usage{
setDefaultTimeProfileStatistics(
  statisticsType = NULL,
  y = NULL,
  ymin = NULL,
  ymax = NULL,
  yCaption = NULL,
  rangeCaption = NULL
)
}
\arguments{
\item{statisticsType}{Name of statistics type as defined in enum \code{StatisticsTypes}}

\item{y}{Function or function name for middle values statistics}

\item{ymin}{Function or function name for min values statistics}

\item{ymax}{Function or function name for max values statistics}

\item{yCaption}{Legend caption for middle values statistics}

\item{rangeCaption}{Legend caption for range values statistics}
}
\description{
Set default statistics used in population time profiles and residuals plots
}
\examples{
\dontrun{
# Set the default statistics as geometric mean
setDefaultTimeProfileStatistics(statisticsType = StatisticsTypes$`Geometric mean`)

# Set the default legend caption displayed for range
setDefaultTimeProfileStatistics(rangeCaption = "90\% population range")
}

}
