% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{setLogFolder}
\alias{setLogFolder}
\title{setLogFolder}
\usage{
setLogFolder(folder)
}
\arguments{
\item{folder}{Folder where logs are saved}
}
\description{
Set folder where logs are saved
}
\keyword{internal}
