% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setNumberOfLocalCores}
\alias{setNumberOfLocalCores}
\title{setNumberOfLocalCores}
\usage{
setNumberOfLocalCores(settings, task)
}
\value{
Settings for simulation or sensitivity analysis updated with available number of local cores
}
\description{
setNumberOfLocalCores
}
