% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{setPlotFormat}
\alias{setPlotFormat}
\title{setPlotFormat}
\usage{
setPlotFormat(format, width = NULL, height = NULL, units = NULL)
}
\arguments{
\item{format}{file format of the exported plots}

\item{width}{plot width in `unit`}

\item{height}{plot height in `unit`}

\item{units}{units of `width` and `height`}
}
\value{
demographyPlots list of ggplot objects
}
\description{
Set plot format
}
