% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{setSimulationDescriptor}
\alias{setSimulationDescriptor}
\title{setSimulationDescriptor}
\usage{
setSimulationDescriptor(workflow, text)
}
\arguments{
\item{workflow}{A `Workflow` object}

\item{text}{Character describing simulation sets}
}
\description{
Set workflow simulation set descriptor
}
