% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{setWatermarkConfiguration}
\alias{setWatermarkConfiguration}
\title{setWatermarkConfiguration}
\usage{
setWatermarkConfiguration(watermark = NULL)
}
\arguments{
\item{watermark}{character or \code{Label} class object from `tlf` package}
}
\description{
Set default watermark configuration for current theme
}
