% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setWorkflowParameterDisplayPaths}
\alias{setWorkflowParameterDisplayPaths}
\title{setWorkflowParameterDisplayPaths}
\usage{
setWorkflowParameterDisplayPaths(parameterDisplayPaths, workflow)
}
\arguments{
\item{parameterDisplayPaths}{data.frame mapping Parameters with their display paths
Variables of the data.frame should include `parameter` and `displayPath`.}

\item{workflow}{Object of class `MeanModelWorkflow` or `PopulationWorkflow`}
}
\description{
Set mapping between parameters and their display paths in a workflow
to replace standard display of parameter paths.
}
