% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{setXParametersForDemographyPlot}
\alias{setXParametersForDemographyPlot}
\title{setXParametersForDemographyPlot}
\usage{
setXParametersForDemographyPlot(workflow, parameters)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}

\item{parameters}{list of demography parameters to be used as x-parameters}
}
\description{
Set x parameters for range plots of demography plot task.
The method update directly the input workflow
}
