% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{setYParametersForDemogrpahyPlot}
\alias{setYParametersForDemogrpahyPlot}
\title{setYParametersForDemogrpahyPlot}
\usage{
setYParametersForDemogrpahyPlot(workflow, parameters)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}

\item{parameters}{list of demography parameters to be used as y-parameters}
}
\description{
Set y-parameters for histograms and range plots of demography plot task.
The method update directly the input workflow
}
