% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{setYParametersForPkParametersPlot}
\alias{setYParametersForPkParametersPlot}
\title{setYParametersForPkParametersPlot}
\usage{
setYParametersForPkParametersPlot(workflow, parameters)
}
\description{
setYParametersForPkParametersPlot
}
