% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{showLogMessages}
\alias{showLogMessages}
\title{showLogMessages}
\usage{
showLogMessages(logTypes = LogTypes)
}
\arguments{
\item{logTypes}{Select specific logs in \code{LogTypes} that will be displayed in the data.frame.}
}
\description{
Display log messages as a data.frame
}
