% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModel}
\alias{simulateModel}
\title{simulateModel}
\usage{
simulateModel(
  simFilePath,
  popDataFilePath = NULL,
  resultsFilePath,
  debugLogFileName = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logDebugFile()),
  infoLogFileName = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logInfoFile()),
  errorLogFileName = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logErrorFile()),
  nodeName = NULL,
  showProgress = FALSE
)
}
\arguments{
\item{simFilePath}{path to pkml model file}

\item{popDataFilePath}{path to the population data file}

\item{resultsFilePath}{path to simulation results CSV files}
}
\value{
Simulation results for individual or population
}
\description{
Simulate model, either for an individual or for a given population.  Calculate and save PK parameters as an option.
}
