% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModelForPopulation}
\alias{simulateModelForPopulation}
\title{simulateModelForPopulation}
\usage{
simulateModelForPopulation(structureSets, settings = NULL)
}
\arguments{
\item{structureSets}{List of `SimulationStructure` objects contain paths of files to be used}

\item{settings}{list of simulation settings}
}
\value{
Simulation results for individual or population
}
\description{
Simulate model for a population.  Run parallel workflow if numberOfCores > 1 AND population size is >1.
}
\keyword{internal}
