% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModelOnCore}
\alias{simulateModelOnCore}
\title{simulateModelOnCore}
\usage{
simulateModelOnCore(
  simulation,
  population,
  debugLogFileName = file.path(getwd(), defaultFileNames$logDebugFile()),
  nodeName = NULL,
  showProgress = FALSE
)
}
\arguments{
\item{simulation}{A \code{Simulation} object}

\item{population}{A \code{Population} object}

\item{debugLogFileName}{path to file where core debug logs are saved}

\item{nodeName}{node name for parallel simulations}

\item{showProgress}{option to print progress of simulation to console}
}
\description{
Simulate model, either for an individual or for a given population.
}
