% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModelParallel}
\alias{simulateModelParallel}
\title{simulateModelParallel}
\usage{
simulateModelParallel(structureSets, settings = NULL)
}
\arguments{
\item{structureSets, }{a list of \code{SimulationStructure} R6 class objects contain paths of files to be used}

\item{settings}{list of options to be passed to the function}
}
\value{
List of simulation results for each simulation set
}
\description{
Simulate models within a list of structure sets in parallel for an individual.
}
\keyword{internal}
