% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulation.R
\name{simulatePopulation}
\alias{simulatePopulation}
\title{simulatePopulation}
\usage{
simulatePopulation(simFileName, simFileFolder, popDataFileName,
  popDataFileFolder, resultFileName, resultFileFolder)
}
\value{
Simulation results for population
}
\description{
Simulate model for a given population
}
