% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{SimulationCodeIdentifiers}
\alias{SimulationCodeIdentifiers}
\title{SimulationCodeIdentifiers}
\format{
An object of class \code{list} of length 14.
}
\usage{
SimulationCodeIdentifiers
}
\description{
Simulation Sets Code Identifiers
}
\keyword{internal}
