% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-set.R
\name{SimulationSet}
\alias{SimulationSet}
\title{SimulationSet}
\description{
R6 class representing Reporting Engine Mean Model Set
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{outputs}}{list of `Output` R6 class objects}

\item{\code{dataSource}}{A `DataSource` object}

\item{\code{dataSelection}}{character or expression used to select a subset of observed data}

\item{\code{timeUnit}}{display unit for time variable}

\item{\code{applicationRanges}}{named list of logicals defining which Application ranges are included in}

\item{\code{minimumSimulationEndTime}}{is the minimum length of time for which a simulation must be run}

\item{\code{timeOffset}}{shift of display time in time profile plots}

\item{\code{massBalanceSettings}}{List of mass balance settings
reported time profiles and residual plots when applicable}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulationSet-new}{\code{SimulationSet$new()}}
\item \href{#method-SimulationSet-clone}{\code{SimulationSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationSet-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SimulationSet` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationSet$new(
  simulationSetName,
  simulationFile,
  outputs = NULL,
  dataSource = NULL,
  dataSelection = DataSelectionKeys$ALL,
  timeUnit = "h",
  applicationRanges = ApplicationRanges,
  minimumSimulationEndTime = NULL,
  timeOffset = 0,
  massBalanceFile = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{outputs}}{list of `Output` R6 class objects}

\item{\code{dataSource}}{A `DataSource` object}

\item{\code{dataSelection}}{characters or expression to select subset the observed data
By default, all the data is selected.
When using a character array, selections are concatenated with the `&` sign}

\item{\code{timeUnit}}{display unit for time variable. Default is "h"}

\item{\code{applicationRanges}}{names of application ranges to include in the report. Names are available in enum `ApplicationRanges`.}

\item{\code{minimumSimulationEndTime}}{is the minimum length of time for which a simulation must be run}

\item{\code{timeOffset}}{shift of display time in time profile plots}

\item{\code{massBalanceFile}}{List of mass balance settings provided either as a json file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SimulationSet` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulationSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimulationSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
