% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-structure.R
\name{SimulationStructure}
\alias{SimulationStructure}
\title{SimulationStructure}
\description{
R6 class representing Reporting Engine generic Workflow
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reportingEngineInfo}}{R6 class object with relevant information about reporting engine}

\item{\code{simulationSets}}{list of `MeanModelSet` R6 class objects}

\item{\code{observedData}}{list of observed `data` and `metaData`}

\item{\code{reportFolder}}{path where report and logs are saved}

\item{\code{resultsFolder}}{path where results are saved}

\item{\code{figuresFolder}}{path where figure are saved}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{reportingEngineInfo}}{R6 class object with relevant information about reporting engine}

\item{\code{simulationSets}}{list of `MeanModelSet` R6 class objects}

\item{\code{observedData}}{list of observed `data` and `metaData`}

\item{\code{reportFolder}}{path where report and logs are saved}

\item{\code{resultsFolder}}{path where results are saved}

\item{\code{figuresFolder}}{path where figure are saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimulationStructure$new()}}
\item \href{#method-clone}{\code{SimulationStructure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationStructure$new(
  simulationSet,
  workflowResultsFolder,
  workflowFiguresFolder = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationStructure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
