% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{sortAndFilterIndividualsDF}
\alias{sortAndFilterIndividualsDF}
\title{sortAndFilterIndividualsDF}
\usage{
sortAndFilterIndividualsDF(individualsDfForPKParameter, rankFilter)
}
\arguments{
\item{individualsDfForPKParameter}{dataframe of results of sensitivity analysis possibly multiple individuals}

\item{rankFilter}{results for only the 'rankFilter' most sensitive parameters will be returned}
}
\value{
sortedFilteredIndividualsDfForPKParameter dataframe of population-wide sensitivity results for pkParameter and output
}
\description{
sort dataframe individualsDfForPKParameter of individuals' sensitivity analysis results and filter out all except the rankFilter most sensitive parameters
}
