% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{splitPopDataFile}
\alias{splitPopDataFile}
\title{splitPopDataFile}
\usage{
splitPopDataFile(fileName, folderName, numberOfSlaves,
  numberOfCommentLines)
}
\description{
Function to split a population data file for parallel processing

Input is a CSV population data file

Output is vector of strings of the name of new CSV files
}
