% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{StandardExcelSheetNames}
\alias{StandardExcelSheetNames}
\title{StandardExcelSheetNames}
\format{
An object of class \code{list} of length 7.
}
\usage{
StandardExcelSheetNames
}
\description{
Enum defining the standard names for the Excel sheets of template
}
\keyword{internal}
