% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\docType{data}
\name{StandardPlotTasks}
\alias{StandardPlotTasks}
\title{StandardPlotTasks}
\format{
An object of class \code{list} of length 3.
}
\usage{
StandardPlotTasks
}
\description{
Names of plot tasks performed by both `MeanModelWorkflow` and `PopulationWorkflow` objects
}
\examples{

# Lists all available standard plot task names available in both mean and population workflows
StandardPlotTasks

}
\seealso{
Other enum helpers: 
\code{\link{AllAvailableTasks}},
\code{\link{ApplicationRanges}},
\code{\link{DataSelectionKeys}},
\code{\link{PopulationWorkflowTypes}},
\code{\link{ResidualScales}},
\code{\link{StandardSimulationTasks}},
\code{\link{StatisticsTypes}},
\code{\link{tlfSettingsNames}}
}
\concept{enum helpers}
\keyword{datasets}
