% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\docType{data}
\name{StandardSimulationTasks}
\alias{StandardSimulationTasks}
\title{StandardSimulationTasks}
\format{An object of class \code{list} of length 3.}
\usage{
StandardSimulationTasks
}
\description{
Names of simulation tasks performed by both `MeanModelWorkflow` and `PopulationWorkflow` objects
}
\keyword{datasets}
