% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{startQualificationRunner}
\alias{startQualificationRunner}
\title{startQualificationRunner}
\usage{
startQualificationRunner(
  qualificationRunnerFolder,
  qualificationPlanFile,
  outputFolder,
  pkSimPortableFolder = NULL,
  moBiPortableFolder = NULL,
  configurationPlanName = NULL,
  overwrite = TRUE,
  logFile = NULL,
  logLevel = NULL,
  displayVersion = FALSE
)
}
\arguments{
\item{qualificationRunnerFolder}{Folder where QualificationRunner.exe is located}

\item{qualificationPlanFile}{full path of the input qualification plan}

\item{outputFolder}{Name of output folder created by the qualification runner}

\item{pkSimPortableFolder}{Folder where PK-Sim is located.
If not specified, installation path will be read from the registry (available only in case of full \strong{non-portable} installation).
This option is \strong{MANDATORY} for the portable version of PK-Sim.}

\item{moBiPortableFolder}{Folder where MoBi is located.
If not specified, installation path will be read from the registry (available only in case of full \strong{non-portable} installation).
This option is \strong{MANDATORY} for the portable version of MoBi.
Only required for qualification plans including MoBi projects.}

\item{configurationPlanName}{Name of the configuration plan to be generated. Default is \code{"report-configuration-plan"}}

\item{overwrite}{Logical defining if the contents of the output folder will be deleted, even if it not empty. Default is false.}

\item{logFile}{Full path of log file where log output will be written. A log file will not be created if this value is not provided.}

\item{logLevel}{Log verbosity (Debug, Information, Warning, Error). Default is Information.}

\item{displayVersion}{Logical defining if version information is displayed}
}
\description{
Starts the qualification runner and creates inputs for the reporting engine
}
\seealso{
Other qualification workflow: 
\code{\link{ConfigurationPlan}},
\code{\link{QualificationVersionInfo}},
\code{\link{adjustTitlePage}()},
\code{\link{loadConfigurationPlan}()},
\code{\link{loadQualificationWorkflow}()}
}
\concept{qualification workflow}
