% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\docType{data}
\name{StatisticsTypes}
\alias{StatisticsTypes}
\title{StatisticsTypes}
\format{
An object of class \code{list} of length 5.
}
\usage{
StatisticsTypes
}
\description{
List of available statistic types summarizing data for time profile plots
\itemize{
\item \code{"2.5th-97.5th Percentiles"} summarizes data using median, 2.5th and 97.5th percentiles
\item \code{"5th-95th Percentiles"} summarizes data using median, 5th and 95th percentiles
\item \code{"10th-90th Percentiles"} summarizes data using median, 10th and 90th percentiles
\item \code{"Geometric mean"} summarizes data using geometric mean and mean */ geometric standard deviation
\item \code{"Arithmetic mean"} summarizes data using arithmetic mean and mean +/- standard deviation
}
}
\seealso{
Other enum helpers: 
\code{\link{AllAvailableTasks}},
\code{\link{ApplicationRanges}},
\code{\link{DataSelectionKeys}},
\code{\link{PopulationWorkflowTypes}},
\code{\link{ResidualScales}},
\code{\link{StandardPlotTasks}},
\code{\link{StandardSimulationTasks}},
\code{\link{reSettingsNames}}
}
\concept{enum helpers}
\keyword{datasets}
