% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{summaryStatisticsToDataFrame}
\alias{summaryStatisticsToDataFrame}
\title{summaryStatisticsToDataFrame}
\usage{
summaryStatisticsToDataFrame(data, monteCarlo = FALSE)
}
\arguments{
\item{data}{A matrix of summary statistics}
}
\value{
A data.frame displaying summary statistics by column
}
\description{
Translate matrix of summary statistics to data.frame with appropriate names
}
\keyword{internal}
