% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-settings.R
\name{TaskSettings}
\alias{TaskSettings}
\title{TaskSettings}
\description{
R6 class defining properties of plot task settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{bins}}{number or edges of bins when aggregation is used}

\item{\code{stairstep}}{logical defining if plot uses stairstep when aggregation is used}

\item{\code{digits}}{number significant digits used in reporting of numbers}

\item{\code{nsmall}}{number decimal digits a accordingly with digits used in reporting of numbers}

\item{\code{scientific}}{logical defining scientific notation is used in reporting of numbers}

\item{\code{scales}}{named list of logicals defining if log/linear plots are included}

\item{\code{plotConfigurations}}{named list of `PlotConfiguration` objects
defining aesthetic properties of reported plots}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TaskSettings$new()}}
\item \href{#method-clone}{\code{TaskSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a `TaskSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskSettings$new(taskName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{taskName}}{name of the task using the settings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TaskSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
