% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{tlfLinetype}
\alias{tlfLinetype}
\title{tlfLinetype}
\usage{
tlfLinetype(configurationLinetype)
}
\arguments{
\item{configurationLinetype}{Value of the configuration plan plot property}
}
\value{
\code{linetype} values that follows \code{tlf} package nomenclature
}
\description{
Translate value of \code{LineStyle} property from configuration plan into tlf \code{linetype}
}
\keyword{internal}
