% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{tlfScale}
\alias{tlfScale}
\title{tlfScale}
\usage{
tlfScale(configurationScale)
}
\arguments{
\item{configurationScale}{Value of the configuration plan plot property}
}
\value{
`scale` values that follows `tlf` package nomenclature
}
\description{
Translate value of `Scaling` property from configuration plan into tlf `scale`
}
