% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\docType{data}
\name{tlfSettingsNames}
\alias{tlfSettingsNames}
\title{reSettingsNames}
\format{
An object of class \code{list} of length 33.
}
\usage{
tlfSettingsNames
}
\description{
Names of the default/global settings stored in `reEnv`.
Can be used with `getRESettings()`
}
\seealso{
Other enum helpers: 
\code{\link{AllAvailableTasks}},
\code{\link{ApplicationRanges}},
\code{\link{DataSelectionKeys}},
\code{\link{PopulationWorkflowTypes}},
\code{\link{ResidualScales}},
\code{\link{StandardPlotTasks}},
\code{\link{StandardSimulationTasks}},
\code{\link{StatisticsTypes}}
}
\concept{enum helpers}
\keyword{datasets}
