% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{tlfShape}
\alias{tlfShape}
\title{tlfShape}
\usage{
tlfShape(configurationShape)
}
\arguments{
\item{configurationShape}{Value of the configuration plan plot property}
}
\value{
`shape` values that follows `tlf` package nomenclature
}
\description{
Translate value of `Symbol` property from configuration plan into tlf `shape`
}
\keyword{internal}
