% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{translateDataSelection}
\alias{translateDataSelection}
\title{translateDataSelection}
\usage{
translateDataSelection(dataSelection)
}
\arguments{
\item{dataSelection}{characters or expression to select subset the observed data}
}
\value{
characters or expression to select subset the observed data
}
\description{
Translate `dataSelection` input by user into characters/expression understood by `getSelectedData`
}
