% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{translateForWord}
\alias{translateForWord}
\title{translateForWord}
\usage{
translateForWord(mdText)
}
\arguments{
\item{mdText}{Character strings of markdown text}
}
\value{
Updated md text
}
\description{
Translate markdown text into a version that Pandoc can convert to Word with appropriate formatting
}
\keyword{internal}
