% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{trimFile}
\alias{trimFile}
\title{trimFile}
\usage{
trimFile(fileName)
}
\arguments{
\item{fileName}{path of .md file to trim}
}
\description{
Trim all the duplicated blank lines in a markdown report file as well as pre-title blank space
}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{getFigurePathsFromReport}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
