% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimFileName}
\alias{trimFileName}
\title{trimFileName}
\usage{
trimFileName(path, extension = NULL, sep = "/")
}
\arguments{
\item{path}{character string containing the name of the path or file to trim}

\item{extension}{character string containing the extension file}

\item{sep}{character string separating path elements. "/" is default value.}
}
\value{
fileName character string of the trimmed filed name
}
\description{
Trim path and extension of a file
}
\examples{
\dontrun{
pathName <- "folder/subfolder/testFile.txt"
trimFileName(pathName, extension = "txt")
}
}
