% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{updateArtifactNumbers}
\alias{updateArtifactNumbers}
\title{updateArtifactNumbers}
\usage{
updateArtifactNumbers(
  fileContent,
  pattern,
  replacement,
  anchorId,
  captionBelow = FALSE
)
}
\arguments{
\item{fileContent}{Content of a markdown or text file read as an array of character strings}

\item{pattern}{character pattern referencing figures in first element of line}

\item{replacement}{character replacing pattern in updated caption name}

\item{anchorId}{character pattern referencing anchor tags}

\item{captionBelow}{logical defining if caption is below artifact}
}
\value{
Array of character strings
}
\description{
Update artifact (figure or table) captions and references in report
}
\keyword{internal}
