% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{updateAxesMargin}
\alias{updateAxesMargin}
\title{updateAxesMargin}
\usage{
updateAxesMargin(axesProperties, sideMarginsEnabled = TRUE)
}
\arguments{
\item{axesProperties}{List of axes properties}

\item{sideMarginsEnabled}{Logical defining if side margins are enabled.}
}
\value{
List of axes properties
}
\description{
Update axes properties based on configuration plan settings for side margins
}
\keyword{internal}
