% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{updateFigureNumbers}
\alias{updateFigureNumbers}
\title{updateFigureNumbers}
\usage{
updateFigureNumbers(fileContent)
}
\arguments{
\item{fileContent}{Content of a markdown or text file read as an array of character strings}
}
\value{
Array of character strings
}
\description{
Update figure captions and references in report
}
\keyword{internal}
