% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{updateIndividualParametersOnCores}
\alias{updateIndividualParametersOnCores}
\title{updateIndividualParametersOnCores}
\usage{
updateIndividualParametersOnCores(individualParameters)
}
\arguments{
\item{individualParameters}{parameters to update}
}
\description{
Update individual parameters on core
}
\keyword{internal}
