% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{updatePlotAxes}
\alias{updatePlotAxes}
\title{updatePlotAxes}
\usage{
updatePlotAxes(plotObject, axesProperties)
}
\arguments{
\item{plotObject}{A ggplot object}

\item{axesProperties}{list of axes properties obtained from \code{getAxesForTimeProfiles}}
}
\value{
A ggplot object
}
\description{
Update the axes, grid and legend properties of a plot object based on the identified axes properties
}
\keyword{internal}
