% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{updatePlotConfigurationTimeTicks}
\alias{updatePlotConfigurationTimeTicks}
\title{updatePlotConfigurationTimeTicks}
\usage{
updatePlotConfigurationTimeTicks(
  data,
  metaData,
  dataMapping,
  plotConfiguration
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on `data`}

\item{dataMapping}{`XYGDataMapping` R6 class object from `tlf` library}

\item{plotConfiguration}{`PlotConfiguration` R6 class object from `tlf` library}
}
\value{
A `PlotConfiguration` object
}
\description{
Update time ticks based on selected time unit in `PlotConfiguration` objects
}
\keyword{internal}
