% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{updatePlotDimensions}
\alias{updatePlotDimensions}
\title{updatePlotDimensions}
\usage{
updatePlotDimensions(plotObject)
}
\arguments{
\item{plotObject}{A `ggplot` object}
}
\value{
A `ggplot` object
}
\description{
Update plot dimensions based on size and position of legend
}
\keyword{internal}
