% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{updateQualificationTimeProfilePlotConfiguration}
\alias{updateQualificationTimeProfilePlotConfiguration}
\title{updateQualificationTimeProfilePlotConfiguration}
\usage{
updateQualificationTimeProfilePlotConfiguration(
  simulatedMetaData = NULL,
  observedMetaData = NULL,
  requestedAxes = "Y",
  axesProperties = NULL,
  plotConfiguration
)
}
\arguments{
\item{simulatedMetaData}{List of meta data on simulated data obtained from \code{getSimulatedCurveProperties}}

\item{observedMetaData}{List of meta data on observed data obtained from \code{getObservedCurveProperties}}

\item{requestedAxes}{Array of requested axes included in \code{"Y"}, \code{"Y2"} and \code{"Y3"}}

\item{axesProperties}{list of axes properties obtained from \code{getAxesProperties}}

\item{plotConfiguration}{A \code{PlotConfiguration} object}
}
\value{
Updated \code{TimeProfilePlotConfiguration} object
}
\description{
Update TimeProfilePlotConfiguration properties based on meta data and requested axes
}
\keyword{internal}
