% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{updateTableNumbers}
\alias{updateTableNumbers}
\title{updateTableNumbers}
\usage{
updateTableNumbers(fileContent)
}
\arguments{
\item{fileContent}{Content of a markdown or text file read as an array of character strings}
}
\value{
Array of character strings
}
\description{
Update table captions and references in report
}
\keyword{internal}
