% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{updateTargetValues}
\alias{updateTargetValues}
\title{updateTargetValues}
\usage{
updateTargetValues(values, targetValues, sourceExpressions, data)
}
\arguments{
\item{values}{initial vector to update}

\item{targetValues}{vector of values to be assigned using `sourceExpressions`}

\item{sourceExpressions}{study design expressions to be evaluated}

\item{data}{population data as data.frame}
}
\description{
updateTargetValues
}
