% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{updateVpcPlotColor}
\alias{updateVpcPlotColor}
\title{updateVpcPlotColor}
\usage{
updateVpcPlotColor(plotObject, output, referenceSimulationSetName = NULL)
}
\arguments{
\item{plotObject}{A ggplot object}

\item{output}{An `Output` object}

\item{referenceSimulationSetName}{Display name of reference simulation set}
}
\value{
A ggplot object
}
\description{
Update colors of lines and ranges of VPC plots
}
\keyword{internal}
