% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-defined-pk-function.R
\docType{class}
\name{UserDefinedPKFunction}
\alias{UserDefinedPKFunction}
\title{UserDefinedPKFunction}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserDefinedPKFunction
}
\description{
UserDefinedPKFunction R6 class
}
\keyword{datasets}
