% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{validateHasRunOnAllCores}
\alias{validateHasRunOnAllCores}
\title{validateHasRunOnAllCores}
\usage{
validateHasRunOnAllCores(coreResults, inputName, inputType, runType = "load")
}
\arguments{
\item{coreResults}{list of logical results returned by each core after an mpi.remote.exec command is complete}

\item{inputName}{Name of the input to be loaded}

\item{inputType}{Type of input to be loaded}

\item{runType}{Type of run executed on `{Rmpi}` cores}
}
\description{
Validate if all cores executed an mpi.remote.exec command successfully.
}
\keyword{internal}
