% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{validateNoDuplicate}
\alias{validateNoDuplicate}
\title{validateNoDuplicate}
\usage{
validateNoDuplicate(
  values,
  variableName = NULL,
  na.rm = TRUE,
  nullAllowed = FALSE
)
}
\arguments{
\item{values}{An array to validate}

\item{variableName}{Name of variable that can be used to display a useful message}

\item{na.rm}{logical indicating if \code{NA} values should be removed before the check}

\item{nullAllowed}{logical indicating if \code{NULL} values should be allowed}
}
\description{
Leverage \code{ospsuite.utils::validateHasOnlyDistinctValues()} to
validate that a vector has only distinct values and display a useful message.
}
\keyword{internal}
