% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{validateObservedMetaDataFile}
\alias{validateObservedMetaDataFile}
\title{Check the consistency between observed data and its dictionary.
Units for `dv `and `time` need to be defined at least once in either
the observed dataset, its dictionary or outputs
In case of multiple definitions, warnings will thrown and the following priorities will be applied:
1. Use units from outputs
2. Use units from observed dataset
3. Use units from dictionary}
\usage{
validateObservedMetaDataFile(observedMetaDataFile, observedDataFile, outputs)
}
\arguments{
\item{observedMetaDataFile}{Path of meta data file on observed dataset (also called dictionary)}

\item{observedDataFile}{Path of observed dataset}

\item{outputs}{list or array of `Output` objects}
}
\description{
Check the consistency between observed data and its dictionary.
Units for `dv `and `time` need to be defined at least once in either
the observed dataset, its dictionary or outputs
In case of multiple definitions, warnings will thrown and the following priorities will be applied:
1. Use units from outputs
2. Use units from observed dataset
3. Use units from dictionary
}
