% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifyAnyPreviousFilesRemoved}
\alias{verifyAnyPreviousFilesRemoved}
\title{verifyAnyPreviousFilesRemoved}
\usage{
verifyAnyPreviousFilesRemoved(anyPreviousPartialResultsRemoved)
}
\arguments{
\item{anyPreviousPartialResultsRemoved}{result files to be removed}
}
\description{
Check that any existing results from individual cores have been removed
}
\keyword{internal}
