% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifyPartialResultsExported}
\alias{verifyPartialResultsExported}
\title{verifyPartialResultsExported}
\usage{
verifyPartialResultsExported(partialResultsExported, numberOfCores)
}
\arguments{
\item{partialResultsExported}{logical vector indicating success of result file export}

\item{numberOfCores}{number of cores from which result files are to be exported}
}
\description{
Check that results from individual cores have been exported
}
\keyword{internal}
