% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifySensitivityAnalysisRunSuccessful}
\alias{verifySensitivityAnalysisRunSuccessful}
\title{verifySensitivityAnalysisRunSuccessful}
\usage{
verifySensitivityAnalysisRunSuccessful(sensitivityRunSuccess, logFolder)
}
\arguments{
\item{sensitivityRunSuccess}{logical vector indicating success of sensitivity analysis run on all cores}
}
\description{
Check that all cores ran sensitivity analysis successfully
}
