% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pop-pk-parameters.R
\name{vpcParameterPlot}
\alias{vpcParameterPlot}
\title{vpcParameterPlot}
\usage{
vpcParameterPlot(
  data,
  metaData = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{list of metaData about `data`}

\item{plotConfiguration}{PlotConfiguration R6 class object}

\item{plotObject}{ggplot object to which layer is added}
}
\value{
ggplot object
}
\description{
Plot vpc like plot of yParameter along xParameter
}
