% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{WorkflowCodeIdentifiers}
\alias{WorkflowCodeIdentifiers}
\title{WorkflowCodeIdentifiers}
\format{
An object of class \code{list} of length 30.
}
\usage{
WorkflowCodeIdentifiers
}
\description{
Workflow Code Identifiers
}
\keyword{internal}
