% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\docType{data}
\name{WorkflowMandatoryVariables}
\alias{WorkflowMandatoryVariables}
\title{WorkflowMandatoryVariables}
\format{
An object of class \code{list} of length 2.
}
\usage{
WorkflowMandatoryVariables
}
\description{
Mandatory Columns for standard sheets
}
\keyword{internal}
