% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{activateWorkflowTasks}
\alias{activateWorkflowTasks}
\title{activateWorkflowTasks}
\usage{
activateWorkflowTasks(workflow, tasks = workflow$getAllTasks())
}
\arguments{
\item{workflow}{`MeanModelWorklfow` or `PopulationWorklfow` object}

\item{tasks}{names of the tasks to activate
Default activates all tasks of the workflow using workflow method `workflow$getAllTasks()`}
}
\description{
activates a series of `Tasks` from a `Workflow`
}
