% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addFigureChunk}
\alias{addFigureChunk}
\title{addFigureChunk}
\usage{
addFigureChunk(fileName, figureFile, figureCaption = "", logFolder = getwd())
}
\arguments{
\item{fileName}{name of .md file}

\item{figureFile}{figure path to include}

\item{figureCaption}{caption of figure}

\item{logFolder}{folder where the logs are saved}

\item{figureWidth}{figure width within report. Default is 100 percent.}
}
\description{
Add a figure in a .md document
}
