% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addFigureChunk}
\alias{addFigureChunk}
\title{addFigureChunk}
\usage{
addFigureChunk(
  fileName,
  figureFileRelativePath,
  figureFileRootDirectory,
  figureCaption = ""
)
}
\arguments{
\item{fileName}{name of .md file}

\item{figureFileRelativePath}{path to figure relative to working directory}

\item{figureFileRootDirectory}{working directory}

\item{figureCaption}{caption of figure}
}
\description{
Add a figure in a .md document
}
\seealso{
Other reporting: 
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
