% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addFigureChunk}
\alias{addFigureChunk}
\title{addFigureChunk}
\usage{
addFigureChunk(
  fileName,
  figureFile,
  figureCaption = "",
  figureWidth = "100\%",
  logFolder = getwd()
)
}
\arguments{
\item{fileName}{name of .md file}

\item{figureFile}{figure path to include}

\item{figureCaption}{caption of figure}

\item{figureWidth}{figure width within report. Default is 100 percent.}

\item{logFolder}{folder where the logs are saved}
}
\description{
Add a figure in a .md document
}
