% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{addLLOQLegend}
\alias{addLLOQLegend}
\title{addLLOQLegend}
\usage{
addLLOQLegend(plotObject, captions, prefix = "LLOQ for")
}
\arguments{
\item{plotObject}{A ggplot object}

\item{captions}{Current observed data captions for which lloq legend is needed}

\item{prefix}{Prefix for legend}
}
\value{
A list of aesthetic values
}
\description{
Add LLOQ displayed legend to the legend of a `plotObject`
TODO: create and export that function in `{tlf}` package
Fix coloring issues
}
\keyword{internal}
