% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addMarkdownToc}
\alias{addMarkdownToc}
\title{addMarkdownToc}
\usage{
addMarkdownToc(tocContent, fileName, logFolder = getwd())
}
\arguments{
\item{tocContent}{Table of content referencing sections following a markdown format}

\item{fileName}{name of .md file to update}

\item{logFolder}{folder where the logs are saved}
}
\description{
Add table of content to a markdown file
}
\keyword{internal}
