% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addSectionNumbersToMarkdown}
\alias{addSectionNumbersToMarkdown}
\title{addSectionNumbersToMarkdown}
\usage{
addSectionNumbersToMarkdown(fileName)
}
\arguments{
\item{fileName}{name of .md file to update}
}
\description{
Update section numbers of a report
}
\keyword{internal}
