% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{addStudyParameters}
\alias{addStudyParameters}
\title{addStudyParameters}
\usage{
addStudyParameters(population, studyDesignFile = NULL)
}
\arguments{
\item{population}{`Population` object}

\item{studyDesignFile}{file name of study design table}
}
\description{
addStudyParameters
}
