% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addTableChunk}
\alias{addTableChunk}
\title{addTableChunk}
\usage{
addTableChunk(
  fileName,
  tableFileRelativePath,
  tableFileRootDirectory,
  tableCaption = "",
  logFolder = getwd()
)
}
\arguments{
\item{fileName}{name of .md file}

\item{tableFileRelativePath}{path to table relative to working directory}

\item{tableFileRootDirectory}{working directory}

\item{tableCaption}{caption of the table}

\item{logFolder}{folder where the logs are saved}
}
\description{
Add a table in a .md document
}
