% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{addTextChunk}
\alias{addTextChunk}
\title{addTextChunk}
\usage{
addTextChunk(fileName, text, logFolder = getwd())
}
\arguments{
\item{fileName}{name of .md file}

\item{text}{text to include in the document}

\item{logFolder}{folder where the logs are saved}
}
\description{
Add a text chunk to a .md document
}
\examples{
\dontrun{
resetReport("report.md")
addTextChunk(fileName = "report.md", text = "new text")
}
}
