% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{addUserDefinedTask}
\alias{addUserDefinedTask}
\title{addUserDefinedTask}
\usage{
addUserDefinedTask(
  workflow,
  taskFunction,
  taskName = "userDefinedTask",
  active = TRUE,
  settings = NULL
)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{taskFunction}{Function to be called by `Task` object.
For `MeanModelWorkflow`, input arguments of `taskFunction` should include: `simulationSet`, `logFolder` and `settings`.
For `PopulationWorkflow`, input arguments of `taskFunction` should include: `simulationSets`, `logFolder`, `xParameters`, `yParameters` and `settings`}

\item{taskName}{character name of task to initialize reporting of task}

\item{active}{logical to set if task is run by workflow}

\item{settings}{list of input arguments that can be used by `settings` input argument of `taskFunction`}
}
\value{
Updated `Workflow` object
}
\description{
Add a user-defined function as a task accessible in the `Workflow` field `$userDefinedTasks`
}
