% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{addXParametersForPkParametersPlot}
\alias{addXParametersForPkParametersPlot}
\title{addXParametersForPkParametersPlot}
\usage{
addXParametersForPkParametersPlot(workflow, parameters)
}
\description{
addXParametersForPkParametersPlot
}
