% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{addYParametersForPKParametersPlot}
\alias{addYParametersForPKParametersPlot}
\title{addYParametersForPKParametersPlot}
\usage{
addYParametersForPKParametersPlot(workflow, parameters)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}

\item{parameters}{list of parameters to be used as y-parameters}
}
\description{
Apppend y-parameters for range plots of PK parameters plot task.
The method update directly the input workflow
}
