% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\docType{data}
\name{AggregationConfiguration}
\alias{AggregationConfiguration}
\title{AggregationConfiguration}
\format{
An object of class \code{list} of length 4.
}
\usage{
AggregationConfiguration
}
\description{
To be deprecated:
Aggregation default properties  (which functions and their captions).
}
\section{Fields}{

\describe{
\item{\code{functions}}{list of `middle`, `ymin` and `ymax` functions for aggregation}

\item{\code{names}}{list of legend captions for `middle` and `range` from aggregation}

\item{\code{bins}}{default number of bins in plots}

\item{\code{binUsingQuantiles}}{logical to choose a binning based on the quantiles rather thanon a constant interval width}
}}

\keyword{datasets}
