% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{alignXTicks}
\alias{alignXTicks}
\title{alignXTicks}
\usage{
alignXTicks(plotObject)
}
\arguments{
\item{plotObject}{A `ggplot` object}
}
\value{
A `ggplot` object
}
\description{
Use the `plotConfiguration` of a `plotObject`
to check and perform vertical alignment of x-axis tick labels
}
\keyword{internal}
