% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\docType{data}
\name{AllAvailableTasks}
\alias{AllAvailableTasks}
\title{AllAvailableTasks}
\format{
An object of class \code{list} of length 9.
}
\usage{
AllAvailableTasks
}
\description{
Names of all existing tasks that can be performed by `MeanModelWorkflow` or `PopulationWorkflow` objects
}
\examples{

# Lists all available task names available in mean or population workflows
AllAvailableTasks

}
\seealso{
Other enum helpers: 
\code{\link{ApplicationRanges}},
\code{\link{DataSelectionKeys}},
\code{\link{PopulationWorkflowTypes}},
\code{\link{ResidualScales}},
\code{\link{StandardPlotTasks}},
\code{\link{StandardSimulationTasks}},
\code{\link{StatisticsTypes}},
\code{\link{tlfSettingsNames}}
}
\concept{enum helpers}
\keyword{datasets}
