% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sensitivity-analysis.R
\name{analyzeCoreSensitivity}
\alias{analyzeCoreSensitivity}
\title{analyzeCoreSensitivity}
\usage{
analyzeCoreSensitivity(
  simulation,
  parametersToPerturb = NULL,
  totalSensitivityThreshold = 1,
  resultsFilePath = paste0(getwd(), "sensitivityAnalysisResults.csv"),
  numberOfCoresToUse = NULL
)
}
\arguments{
\item{simulation}{simulation class object}

\item{totalSensitivityThreshold}{numeric value between 0 and 1.
Close to 0, only the most sensitive output paths are returned.
Close to 1, almost all the output paths are returned.}

\item{pkParameterName}{name of parameter to be analyzed}
}
\value{
sensitivityResults
}
\description{
Run a sensitivity analysis from a simulation
}
