% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sensitivity-analysis.R
\name{analyzeSensitivity}
\alias{analyzeSensitivity}
\title{analyzeSensitivity}
\usage{
analyzeSensitivity(
  simFilePath,
  parametersToPerturb = NULL,
  popFilePath = NULL,
  individualID = NULL,
  numberOfCores = 1,
  resultsFileFolder = "getwd()",
  resultsFileName = "sensitivityAnalysisResults"
)
}
\value{
SA results for individual or population
}
\description{
Determine whether to run SA for individual or population.  If for individual,  pass simulation to individualSensitivityAnalysis.
If SA is for population, loop thru population file, extract parameters for each individual, and pass them to individualSensitivityAnalysis.
}
