% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sensitivity-analysis.R
\name{analyzeSensitivity}
\alias{analyzeSensitivity}
\title{analyzeSensitivity}
\usage{
analyzeSensitivity(
  simFilePath,
  perturbationParameterNamesVector = NULL,
  totalSensitivityThreshold = 1,
  resultsFilePath = paste0(getwd(), "sensitivityAnalysisResults.csv"),
  numberOfCoresToUse = NULL
)
}
\arguments{
\item{totalSensitivityThreshold}{numeric value between 0 and 1.
Close to 0, only the most sensitive output paths are returned.
Close to 1, almost all the output paths are returned.}

\item{simulation}{simulation class object}

\item{pkParameterName}{name of parameter to be analyzed}
}
\value{
sensitivityResults
}
\description{
Run a sensitivity analysis from a simulation
}
