% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{analyzeSensitivity}
\alias{analyzeSensitivity}
\title{analyzeSensitivity}
\usage{
analyzeSensitivity(
  simFilePath,
  parametersToPerturb = NULL,
  popFilePath = NULL,
  individualID = NULL,
  numberOfCores = 1,
  resultsFileFolder = "getwd()",
  resultsFileName = "sensitivityAnalysisResults"
)
}
\arguments{
\item{simFilePath}{path to simulation file}

\item{parametersToPerturb}{paths to parameters to perturb in sensitivity analysis}

\item{popFilePath}{path to the population data file}

\item{individualID}{ID of individual in population data file for whom to perform sensitivity analysis}

\item{numberOfCores}{number of cores over which to parallelize the sensitivity analysis}

\item{resultsFileFolder}{path to population sensitivity analysis results CSV files}

\item{resultsFileName}{root name of population sensitivity analysis results CSV files}
}
\value{
SA results for individual or population
}
\description{
Determine whether to run SA for individual or population.  If for individual,  pass simulation to individualSensitivityAnalysis.
If SA is for population, loop thru population file, extract parameters for each individual, and pass them to individualSensitivityAnalysis.
}
