% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\docType{data}
\name{ApplicationRanges}
\alias{ApplicationRanges}
\title{ApplicationRanges}
\format{
An object of class \code{list} of length 3.
}
\usage{
ApplicationRanges
}
\description{
Keys of reported ranges when simulation includes multiple applications
}
\examples{
ApplicationRanges$total
ApplicationRanges$firstApplication
ApplicationRanges$lastApplication
}
\keyword{datasets}
