% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{autoAxesLimits}
\alias{autoAxesLimits}
\title{autoAxesLimits}
\usage{
autoAxesLimits(x, scale = NULL)
}
\arguments{
\item{x}{Values from which the limits are calculated}
}
\value{
A list of units for goodness of fit results
}
\description{
Defines auto axis limits
}
\keyword{internal}
