% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{buildQualificationDDIResidualsVsObserved}
\alias{buildQualificationDDIResidualsVsObserved}
\title{buildQualificationGOFResidualsVsObserved}
\usage{
buildQualificationDDIResidualsVsObserved(dataframe, metadata, pkParameter)
}
\arguments{
\item{dataframe}{data.frame}

\item{metadata}{meta data on `data`}

\item{pkParameter}{for which DDI rations are to be evaluated}
}
\value{
dataframe for plotting goodness of fit of residuals vs time type
}
\description{
Build dataframe for residuals vs observed
}
