% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{buildQualificationDDIResidualsVsObserved}
\alias{buildQualificationDDIResidualsVsObserved}
\title{#' @title buildQualificationGOFResidualsVsObserved
#' @description Build dataframe for residuals vs observed
#' @param dataframe data.frame
#' @param metadata meta data on `data`
#' @return dataframe for plotting goodness of fit of residuals vs time type}
\usage{
buildQualificationDDIResidualsVsObserved(dataframe, metadata)
}
\description{
#' @title buildQualificationGOFResidualsVsObserved
#' @description Build dataframe for residuals vs observed
#' @param dataframe data.frame
#' @param metadata meta data on `data`
#' @return dataframe for plotting goodness of fit of residuals vs time type
}
