% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{calculatePKParameters}
\alias{calculatePKParameters}
\title{calculatePKParameters}
\usage{
calculatePKParameters(
  simulationFilePath,
  simulationResultFilePaths,
  pkParametersToEvaluate = NULL,
  userDefinedPKFunctions = NULL,
  pkParameterResultsFilePath
)
}
\value{
generatedResultFileNames
}
\description{
Calculate PK parameters from simulated time profiles
}
