% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{calculatePKParameters}
\alias{calculatePKParameters}
\title{calculatePKParameters}
\usage{
calculatePKParameters(
  simulationFilePath,
  simulationResultFilePaths,
  pkParametersToEvaluate = NULL,
  userDefinedPKFunctions = NULL,
  pkParameterResultsFilePath
)
}
\arguments{
\item{simulationFilePath}{path to pkml model file}

\item{simulationResultFilePaths}{path to simulation CSV results files}

\item{pkParametersToEvaluate}{vector of PK parameters to evaluate}

\item{userDefinedPKFunctions}{vector of userDefinedPKFunction objects}

\item{pkParameterResultsFilePath}{path to PK analysis result file}
}
\value{
pkParameterResultsFilePath, paths to pk results file CSV
}
\description{
Calculate PK parameters from simulated time profiles
}
