% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{calculatePKParameters}
\alias{calculatePKParameters}
\title{calculatePKParameters}
\usage{
calculatePKParameters(structureSet, settings = NULL, logFolder = getwd())
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object contain paths of files to be used}

\item{settings}{list of options to be passed on the function}

\item{logFolder}{folder where the logs are saved}
}
\value{
pkAnalysis object
}
\description{
Calculate PK parameters from simulated time profiles
}
