% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{calculatePKParameters}
\alias{calculatePKParameters}
\title{calculatePKParameters}
\usage{
calculatePKParameters(structureSet)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}
}
\value{
pkAnalyses object
}
\description{
Calculate PK parameters from simulated time profiles
}
