% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-of-work.R,
%   R/utils-calculate-pk-parameters.R
\name{calculatePKParameters}
\alias{calculatePKParameters}
\title{calculatePKParameters}
\usage{
calculatePKParameters()

calculatePKParameters()
}
\arguments{
\item{populationSimulation}{populationSimulation object}

\item{populationSimulation}{populationSimulation object}
}
\value{
pkParametersResults

pkParametersResults
}
\description{
Calculate PK parameters from simulated time profiles

Calculate PK parameters from simulated time profiles
}
