% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-pk-parameters-task.R
\docType{class}
\name{CalculatePKParametersTask}
\alias{CalculatePKParametersTask}
\title{CalculatePKParametersTask}
\description{
Calculating PK parameters task settings for Reporting Engine
}
\section{Methods}{

\describe{
\item{new()}{Initilialize Task settings}
\item{activate()}{Set Task as active in workflow}
\item{inactivate()}{Set Task as inactive in workflow}
\item{print()}{Show task settings}
}
}

\section{Super class}{
\code{\link[ospsuite.reportingengine:Tasks]{ospsuite.reportingengine::Tasks}} -> \code{CalculatePKParametersTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resultsFileName}}{string with no file extension.  Name of CSV file to which simulation results will be saved.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{resultsFileName}}{string with no file extension.  Name of CSV file to which simulation results will be saved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CalculatePKParametersTask$new()}}
\item \href{#method-clone}{\code{CalculatePKParametersTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="activate">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-activate}{\code{ospsuite.reportingengine::Tasks$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-inactivate}{\code{ospsuite.reportingengine::Tasks$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="print">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-print}{\code{ospsuite.reportingengine::Tasks$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Tasks" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Tasks.html#method-validateInput}{\code{ospsuite.reportingengine::Tasks$validateInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$new(
  simulationFilePath,
  simulationResultFilePaths,
  pkParametersToEvaluate = NULL,
  userDefinedPKFunctions = NULL,
  pkParameterResultsFilePath = file.path(getwd(), "pkParameters.csv"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
