% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-pk-parameters-task.R
\name{CalculatePKParametersTask}
\alias{CalculatePKParametersTask}
\title{CalculatePKParametersTask}
\description{
R6 class for defining how pk parameters are calculated and save
}
\keyword{internal}
\section{Super classes}{
\code{ospsuite.reportingengine::Task} -> \code{ospsuite.reportingengine::SimulationTask} -> \code{CalculatePKParametersTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ratioComparison}}{logical defining if a ratio comparison is required}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalculatePKParametersTask-saveResults}{\code{CalculatePKParametersTask$saveResults()}}
\item \href{#method-CalculatePKParametersTask-runTask}{\code{CalculatePKParametersTask$runTask()}}
\item \href{#method-CalculatePKParametersTask-clone}{\code{CalculatePKParametersTask$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-activate'><code>ospsuite.reportingengine::Task$activate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getAbsolutePath"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getAbsolutePath'><code>ospsuite.reportingengine::Task$getAbsolutePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getInputs"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getInputs'><code>ospsuite.reportingengine::Task$getInputs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getRelativePath"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getRelativePath'><code>ospsuite.reportingengine::Task$getRelativePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-inactivate'><code>ospsuite.reportingengine::Task$inactivate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-print'><code>ospsuite.reportingengine::Task$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-validateInput'><code>ospsuite.reportingengine::Task$validateInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateStructureSetInput"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-validateStructureSetInput'><code>ospsuite.reportingengine::Task$validateStructureSetInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="SimulationTask" data-id="initialize"><a href='../../ospsuite.reportingengine/html/SimulationTask.html#method-SimulationTask-initialize'><code>ospsuite.reportingengine::SimulationTask$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalculatePKParametersTask-saveResults"></a>}}
\if{latex}{\out{\hypertarget{method-CalculatePKParametersTask-saveResults}{}}}
\subsection{Method \code{saveResults()}}{
Save the task results related to a \code{structureSet}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$saveResults(structureSet, taskResults)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{structureSet}}{A \code{SimulationStructure} object defining the properties of a simulation set}

\item{\code{taskResults}}{list of results from task run.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalculatePKParametersTask-runTask"></a>}}
\if{latex}{\out{\hypertarget{method-CalculatePKParametersTask-runTask}{}}}
\subsection{Method \code{runTask()}}{
Run task and save its output results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$runTask(structureSets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{structureSets}}{list of \code{SimulationStructure} objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalculatePKParametersTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalculatePKParametersTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalculatePKParametersTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
