% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{checkTaskInputsExist}
\alias{checkTaskInputsExist}
\title{checkTaskInputsExist}
\usage{
checkTaskInputsExist(task)
}
\arguments{
\item{task}{`Task` object or derived class}
}
\value{
Named list of logical values assessing if the files exist
}
\description{
Check if the files required by the task exist
}
