% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration-plan.R
\name{ConfigurationPlan}
\alias{ConfigurationPlan}
\title{ConfigurationPlan}
\description{
R6 class for ConfigurationPlan guiding Qualification Workflow
}
\seealso{
Other qualification workflow: 
\code{\link{QualificationVersionInfo}},
\code{\link{adjustTitlePage}()},
\code{\link{loadConfigurationPlan}()},
\code{\link{loadQualificationWorkflow}()},
\code{\link{startQualificationRunner}()}
}
\concept{qualification workflow}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plots}}{list defining \code{PlotConfiguration} settings for specific tasks}

\item{\code{inputs}}{data.frame or list mapping input files to their paths}

\item{\code{intro}}{data.frame or list mapping introduction files to their paths}

\item{\code{markdownIntro}}{name of markdown file that includes intro content}

\item{\code{referenceFolder}}{Reference path for accessing inputs}

\item{\code{workflowFolder}}{path of the output folder created or used by the Workflow.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sections}}{data.frame mapping section ids to their paths}

\item{\code{simulationMappings}}{data.frame mapping simulations to their paths}

\item{\code{observedDataSets}}{data.frame mapping observed datasets to their paths}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfigurationPlan-getSectionPath}{\code{ConfigurationPlan$getSectionPath()}}
\item \href{#method-ConfigurationPlan-getSectionTitle}{\code{ConfigurationPlan$getSectionTitle()}}
\item \href{#method-ConfigurationPlan-getSectionMarkdown}{\code{ConfigurationPlan$getSectionMarkdown()}}
\item \href{#method-ConfigurationPlan-getSectionLevel}{\code{ConfigurationPlan$getSectionLevel()}}
\item \href{#method-ConfigurationPlan-getIntroMarkdown}{\code{ConfigurationPlan$getIntroMarkdown()}}
\item \href{#method-ConfigurationPlan-copySectionContent}{\code{ConfigurationPlan$copySectionContent()}}
\item \href{#method-ConfigurationPlan-copyInput}{\code{ConfigurationPlan$copyInput()}}
\item \href{#method-ConfigurationPlan-copyIntro}{\code{ConfigurationPlan$copyIntro()}}
\item \href{#method-ConfigurationPlan-copyContentFiles}{\code{ConfigurationPlan$copyContentFiles()}}
\item \href{#method-ConfigurationPlan-getObservedDataPath}{\code{ConfigurationPlan$getObservedDataPath()}}
\item \href{#method-ConfigurationPlan-getMolWeightForObservedData}{\code{ConfigurationPlan$getMolWeightForObservedData()}}
\item \href{#method-ConfigurationPlan-getSimulationPath}{\code{ConfigurationPlan$getSimulationPath()}}
\item \href{#method-ConfigurationPlan-getPopulationPath}{\code{ConfigurationPlan$getPopulationPath()}}
\item \href{#method-ConfigurationPlan-getSimulationResultsPath}{\code{ConfigurationPlan$getSimulationResultsPath()}}
\item \href{#method-ConfigurationPlan-getPKAnalysisResultsPath}{\code{ConfigurationPlan$getPKAnalysisResultsPath()}}
\item \href{#method-ConfigurationPlan-updateTheme}{\code{ConfigurationPlan$updateTheme()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getSectionPath"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getSectionPath}{}}}
\subsection{Method \code{getSectionPath()}}{
Get location of directory corresponding to a specific section Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getSectionPath(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{section identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The section path corresponding to the id in the configuration plan field \code{sections}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getSectionTitle"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getSectionTitle}{}}}
\subsection{Method \code{getSectionTitle()}}{
Get markdown title to a specific section Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getSectionTitle(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{section identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The title associated with "#" corresponding to the subsection level
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getSectionMarkdown"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getSectionMarkdown}{}}}
\subsection{Method \code{getSectionMarkdown()}}{
Get location of .md file corresponding to a specific section Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getSectionMarkdown(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{section identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The markdown file corresponding to the id in the configuration plan field \code{sections}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getSectionLevel"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getSectionLevel}{}}}
\subsection{Method \code{getSectionLevel()}}{
Get section level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getSectionLevel(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{section identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Section level
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getIntroMarkdown"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getIntroMarkdown}{}}}
\subsection{Method \code{getIntroMarkdown()}}{
Get location of .md intro file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getIntroMarkdown()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The markdown file corresponding to introduction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-copySectionContent"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-copySectionContent}{}}}
\subsection{Method \code{copySectionContent()}}{
Copy content to section markdown
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$copySectionContent(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{section identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-copyInput"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-copyInput}{}}}
\subsection{Method \code{copyInput()}}{
Copy input to section markdown
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$copyInput(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list including SectionId and Path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-copyIntro"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-copyIntro}{}}}
\subsection{Method \code{copyIntro()}}{
Copy intro to intro markdown
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$copyIntro(intro)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{intro}}{list including Path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-copyContentFiles"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-copyContentFiles}{}}}
\subsection{Method \code{copyContentFiles()}}{
If available, copy all files and folders of \code{Content} directory into \verb{<workflowFolder>}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$copyContentFiles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getObservedDataPath"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getObservedDataPath}{}}}
\subsection{Method \code{getObservedDataPath()}}{
Get location of observed data corresponding to a specific observedDataSet Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getObservedDataPath(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{observedDataSet identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The observed data file path corresponding to the id in the configuration plan field \code{observedDataSet}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getMolWeightForObservedData"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getMolWeightForObservedData}{}}}
\subsection{Method \code{getMolWeightForObservedData()}}{
Get molecular weight of observed data corresponding to a specific observedDataSet Id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getMolWeightForObservedData(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{observedDataSet identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The observed data file path corresponding to the id in the configuration plan field \code{observedDataSet}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getSimulationPath"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getSimulationPath}{}}}
\subsection{Method \code{getSimulationPath()}}{
Get location of simulation file corresponding to a specific simulation and project names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getSimulationPath(project, simulation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{name of simulation project}

\item{\code{simulation}}{name of the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The simulation file path corresponding to the project and simulation in the configuration plan field \code{simulationMappings}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getPopulationPath"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getPopulationPath}{}}}
\subsection{Method \code{getPopulationPath()}}{
Get location of population file corresponding to a specific simulation and project names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getPopulationPath(project, simulation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{name of simulation project}

\item{\code{simulation}}{name of the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The population file path corresponding to the project and simulation in the configuration plan field \code{simulationMappings}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getSimulationResultsPath"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getSimulationResultsPath}{}}}
\subsection{Method \code{getSimulationResultsPath()}}{
Get location of simulation result file corresponding to a specific simulation and project names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getSimulationResultsPath(project, simulation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{name of simulation project}

\item{\code{simulation}}{name of the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The simulation results file path corresponding to the project and simulation in the configuration plan field \code{simulationMappings}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-getPKAnalysisResultsPath"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-getPKAnalysisResultsPath}{}}}
\subsection{Method \code{getPKAnalysisResultsPath()}}{
Get location of PK Analysis result file corresponding to a specific simulation and project names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$getPKAnalysisResultsPath(project, simulation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project}}{name of simulation project}

\item{\code{simulation}}{name of the simulation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The PKAnalysis results file path corresponding to the project and simulation in the configuration plan field \code{simulationMappings}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigurationPlan-updateTheme"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigurationPlan-updateTheme}{}}}
\subsection{Method \code{updateTheme()}}{
Update environment theme that will be used as default during workflow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigurationPlan$updateTheme()}\if{html}{\out{</div>}}
}

}
}
