% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration-plan.R
\docType{class}
\name{ConfigurationPlan}
\alias{ConfigurationPlan}
\title{ConfigurationPlan}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConfigurationPlan
}
\description{
Abstract class for ConfigurationPlan
}
\keyword{datasets}
