% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationShapes}
\alias{ConfigurationShapes}
\title{ConfigurationShapes}
\format{
An object of class \code{list} of length 22.
}
\usage{
ConfigurationShapes
}
\description{
List mapping configuration plan symbols to tlf shapes.
Note that the names only include lower case as the mapping
used in \code{tlfShape()} internally translate configuration plan field with \code{tolower()} function.
}
\keyword{internal}
