% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\docType{data}
\name{ConfigurationShapes}
\alias{ConfigurationShapes}
\title{ConfigurationShapes}
\format{
An object of class \code{list} of length 8.
}
\usage{
ConfigurationShapes
}
\description{
List defining shape following tlf nomenclature from a Symbol field of configuration plan
}
\keyword{internal}
