% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{copyReport}
\alias{copyReport}
\title{copyReport}
\usage{
copyReport(from, to, keep = FALSE)
}
\arguments{
\item{from}{path of initial .md file to copy}

\item{to}{path of destination .md file to be copied}

\item{keep}{logical defining if initial .md file and figures are kept}
}
\description{
Copy markdown report and its figures (using their paths)
}
