% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{data}
\name{DataSelectionKeys}
\alias{DataSelectionKeys}
\title{DataSelectionKeys}
\format{
An object of class \code{list} of length 2.
}
\usage{
DataSelectionKeys
}
\description{
List of available short keys for observed data selection
}
\examples{

# Lists available Data Selection Keys
DataSelectionKeys

}
\seealso{
Other enum helpers: 
\code{\link{AllAvailableTasks}},
\code{\link{ApplicationRanges}},
\code{\link{PopulationWorkflowTypes}},
\code{\link{ResidualScales}},
\code{\link{StandardPlotTasks}},
\code{\link{StandardSimulationTasks}},
\code{\link{StatisticsTypes}},
\code{\link{reSettingsNames}}
}
\concept{enum helpers}
\keyword{datasets}
