% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{defaultMassBalanceGroupings}
\alias{defaultMassBalanceGroupings}
\title{defaultMassBalanceGrouping}
\usage{
defaultMassBalanceGroupings(compoundNames)
}
\arguments{
\item{compoundNames}{Names of simulation molecules}
}
\description{
Get mass balance default inclusion/exclusion defined as groupings list
Default groups are
\itemize{
\item Plasma
\item BloodCells
\item Interstitial
\item Intracellular
\item Endosome
\item Gallbladder
\item Urine
\item Feces
\item Rest (includes everything else but Saliva compartments)
}
}
\keyword{internal}
