% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{generateDDIQualificationDDIPlot}
\alias{generateDDIQualificationDDIPlot}
\title{generateDDIQualificationDDIPlot}
\usage{
generateDDIQualificationDDIPlot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
ggplot DDI plot object for DDI qualification workflow
}
\description{
Plot observation vs prediction for qualification workflow
}
