% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateResultFileNames}
\alias{generateResultFileNames}
\title{generateResultFileNames}
\usage{
generateResultFileNames(
  numberOfCores,
  folderName,
  fileName,
  separator = "-",
  extension = ".csv"
)
}
\arguments{
\item{numberOfCores}{to be used in parallel computation}

\item{folderName}{where result files will be saved}

\item{fileName}{prefix of result file names}

\item{separator}{used between file name prefix and index}

\item{extension}{for result file type.  default is CSV}
}
\value{
A list of filenames to be output by each core
}
\description{
#Generate a listcontaining names of CSV result files that will be output by each core in parallel computation
}
