% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getAnchorName}
\alias{getAnchorName}
\title{getAnchorName}
\usage{
getAnchorName(tag)
}
\value{
A character Name/identifier of the anchor tag
}
\description{
Get the name/identifier an anchor tag
}
\examples{
getAnchorName('<a id="section-1"></a>')

# Works also for tag name instead of id 
getAnchorName('<a name="section-1"></a>')
}
