% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getAxesProperties}
\alias{getAxesProperties}
\title{getAxesProperties}
\usage{
getAxesProperties(axesSettings)
}
\arguments{
\item{axesSettings}{list of axes properties defined in `Axes` field of configuration plan}
}
\value{
A list of properties for axes identified for `x`, `y` and `y2` axes.
The identified properties are directly compatible with `tlf` package nomenclature
}
\description{
Get identified axes properties from global field `AxesSettings` of configuration plan or
from plot specific field `Axes`. The identified properties are directly compatible with `tlf` package nomenclature
}
