% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{getAxesPropertiesFromName}
\alias{getAxesPropertiesFromName}
\title{getAxesPropertiesFromName}
\usage{
getAxesPropertiesFromName(configurationPlan, plotName)
}
\arguments{
\item{configurationPlan}{A \code{ConfigurationPlan} object}

\item{plotName}{Field name of the plot in the configuration plan \code{AxesSettings}}
}
\value{
A list of properties for axes identified for \code{x}, \code{y} and \code{y2} axes.
The identified properties are directly compatible with \code{tlf} package nomenclature
}
\description{
Get axes properties from global field \code{AxesSettings} of configuration plan if defined,
otherwise use default values.
}
\keyword{internal}
