% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getBoxWhiskerPlotConfiguration}
\alias{getBoxWhiskerPlotConfiguration}
\title{getBoxWhiskerPlotConfiguration}
\usage{
getBoxWhiskerPlotConfiguration(
  plotScale = "log",
  colorGrouping = NULL,
  data,
  metaData,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{plotScale}{Scale of Y Axis}

\item{data}{A data.frame}

\item{metaData}{dimensions and units in the data.frame}

\item{dataMapping}{`DataMapping` object}

\item{plotConfiguration}{A user-defined `PlotConfiguration` object}
}
\value{
A `PlotConfiguration` object
}
\description{
Define a `PlotConfiguration` object
}
\keyword{internal}
