% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getDefaultPkParametersXParameters}
\alias{getDefaultPkParametersXParameters}
\title{getDefaultPkParametersXParameters}
\usage{
getDefaultPkParametersXParameters(workflowType)
}
\arguments{
\item{workflowType}{Name of workflow type.
Use enum `PopulationWorkflowTypes` to get a list of available workflow types.}
}
\value{
names of default parameters
}
\description{
Get names of default parameters in x axis of pk parameters plots.
}
