% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pop-pk-parameters.R
\name{getDefaultPKParametersXParameters}
\alias{getDefaultPKParametersXParameters}
\alias{getDefaultPkParametersXParameters}
\title{getDefaultPKParametersXParameters}
\usage{
getDefaultPKParametersXParameters(workflowType)

getDefaultPkParametersXParameters(workflowType)
}
\arguments{
\item{workflowType}{Name of workflow type.
Use enum \code{PopulationWorkflowTypes} to get a list of available workflow types.}
}
\value{
names of default parameters
}
\description{
Get names of default parameters in x axis of pk parameters plots.
}
\examples{

getDefaultPKParametersXParameters(PopulationWorkflowTypes$pediatric)

}
