% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getDefaultPropertiesFromTheme}
\alias{getDefaultPropertiesFromTheme}
\title{getDefaultPropertyFromTheme}
\usage{
getDefaultPropertiesFromTheme(
  plotName,
  propertyType = "points",
  propertyNames = as.character(tlf::AestheticProperties)
)
}
\arguments{
\item{plotName}{Name of the plot in Theme (e.g. \code{"plotTimeProfile"})}

\item{propertyType}{One of \code{"points"}, \verb{"lines}, \code{"ribbons"} or \code{"errorbars"}}

\item{propertyNames}{Names of the aesthetic property (e.g. \code{"color"})}
}
\value{
Property value
}
\description{
Get default property value from current reEnv theme
}
\keyword{internal}
