% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{getGOFDataForMapping}
\alias{getGOFDataForMapping}
\title{getGOFDataForMapping}
\usage{
getGOFDataForMapping(outputMapping, configurationPlan, axesUnits)
}
\arguments{
\item{outputMapping}{list of mapping elements from `OutputMappings` field in configuration plan}

\item{configurationPlan}{A `ConfigurationPlan` object}

\item{axesUnits}{list of axes properties obtained from `getGOFAxesUnits`}
}
\value{
A data.frame as obtained by `getResiduals` whose values are in base unit
}
\description{
Get data of goodness of fit from field `GOFMergedPlots` of configuration plan
}
\keyword{internal}
