% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getGOFPlotConfiguration}
\alias{getGOFPlotConfiguration}
\title{getGOFPlotConfiguration}
\usage{
getGOFPlotConfiguration(
  plotType,
  group,
  data,
  metaData,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{plotType}{Plot type for residuals}

\item{group}{A data.frame mapping properties to output groups}

\item{data}{A data.frame}

\item{metaData}{List of metaData defining dimensions and units in the data.frame}

\item{dataMapping}{List \code{DataMapping} object}

\item{plotConfiguration}{A user-defined \code{PlotConfiguration} object}
}
\value{
A \code{PlotConfiguration} object
}
\description{
Define a \code{PlotConfiguration} object
}
\keyword{internal}
