% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getLegendDimensions}
\alias{getLegendDimensions}
\title{getLegendDimensions}
\usage{
getLegendDimensions(plotObject)
}
\arguments{
\item{plotObject}{A \code{ggplot} object}
}
\value{
A list of \code{width} and \code{height} if legend is found,
otherwise \code{NULL}
}
\description{
Get legend dimensions
}
\keyword{internal}
