% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{getMassBalanceData}
\alias{getMassBalanceData}
\title{getMassBalanceData}
\usage{
getMassBalanceData(groupings, compoundNames, simulation, simulationResults)
}
\arguments{
\item{groupings}{A list of grouping lists that define naming and inclusion/exclusion criteria}

\item{compoundNames}{A vector of compound names}

\item{simulation}{A `Simulation` object}

\item{simulationResults}{A `SimulationResults` object}
}
\value{
A data.frame that includes `Time`, `Amount` and `Legend` as variables
}
\description{
Get mass balance data for a set of compounds
}
\keyword{internal}
