% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getMeanSensitivityAnalysisResultsFileNames}
\alias{getMeanSensitivityAnalysisResultsFileNames}
\title{getMeanSensitivityAnalysisResultsFileNames}
\usage{
getMeanSensitivityAnalysisResultsFileNames(workflow)
}
\arguments{
\item{workflow}{`MeanModelWorkflow` object or derived class}
}
\value{
Names of the the expected sensitivity analysis result files
}
\description{
Get the expected sensitivity analysis result files obtained from a mean model workflow
}
