% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getMetaDataFrame}
\alias{getMetaDataFrame}
\title{getMetaDataFrame}
\usage{
getMetaDataFrame(listOfMetaData)
}
\arguments{
\item{listOfMetaData}{list including \code{Path} and \code{Concentration}}
}
\value{
A data.frame with variables \code{path}, \code{dimension}, \code{unit}, \code{group}, \code{color}, \code{fill}
}
\description{
Get time profile list of metaData as a single data.frame
}
\keyword{internal}
