% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getMetaDataFrame}
\alias{getMetaDataFrame}
\title{getMetaDataFrame}
\usage{
getMetaDataFrame(listOfMetaData)
}
\arguments{
\item{listOfMetaData}{list including `Path` and `Concentration`}
}
\value{
A data.frame with variables `path`, `dimension` and `unit`
}
\description{
Get time profile metaData as a data.frame
}
\keyword{internal}
