% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{getObservedDataFromConfigurationPlan}
\alias{getObservedDataFromConfigurationPlan}
\title{getObservedDataFromConfigurationPlan}
\usage{
getObservedDataFromConfigurationPlan(
  observedDataId,
  configurationPlan,
  logFolder
)
}
\arguments{
\item{observedDataId}{Identifier of observed data}

\item{configurationPlan}{A `ConfigurationPlan` object that includes methods to find the data}

\item{logFolder}{folder where the logs are saved}
}
\value{
list of including data and metaData to perform time profile plot
}
\description{
Get selected observed data from a `ConfigurationPlan` object
}
\keyword{internal}
