% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getObservedErrorValues}
\alias{getObservedErrorValues}
\title{getObservedErrorValues}
\usage{
getObservedErrorValues(
  observedValues,
  observedResults,
  axesProperties,
  molWeight = NA
)
}
\arguments{
\item{observedValues}{Numeric values of observed data}

\item{observedResults}{A named list, including `data` and `metaData`, of observed results.}

\item{axesProperties}{list of axes properties obtained from `getAxesProperties`}

\item{molWeight}{Molecular weight if unit conversion is required}
}
\value{
A named list, with `ymin` and `ymax`, of the observed data error range
}
\description{
Get the observed data error range to display on time profile plots
}
\keyword{internal}
