% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getOutputContent}
\alias{getOutputContent}
\title{getOutputContent}
\usage{
getOutputContent(excelFile, outputInfo)
}
\arguments{
\item{excelFile}{name of the Excel file from which the R script is created}

\item{outputInfo}{list of information about `Output` object. Include sheetName, dataSelection and dataDisplayName}
}
\value{
Character vector defining the `Output` object
}
\description{
Creates a character vector to be written in a workflow .R script defining `Output` object.
}
