% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-user-helpers.R
\name{getPKParameterGroupsInOutput}
\alias{getPKParameterGroupsInOutput}
\title{getPKParameterGroupsInOutput}
\usage{
getPKParameterGroupsInOutput(output)
}
\arguments{
\item{output}{Output object}
}
\value{
Names of pkParameter groups in `output`
}
\description{
getPKParameterGroupsInOutput
}
\examples{
myOutput <- Output$new(
  path = "a",
  pkParameters = c("C_max", "AUC_tEnd")
)
getPKParametersInOutput(myOutput)

# If a group is defined
myGroupOutput <- Output$new(
  path = "a",
  pkParameters = c(
    C_max = PkParameterInfo$new("C_max", group = 1),
    AUC_tEnd = PkParameterInfo$new("AUC_tEnd", group = 2)
  )
)
getPKParametersInOutput(myGroupOutput)

}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemogrpahyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemogrpahyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemogrpahyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemogrpahyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setXParametersForDemogrpahyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemogrpahyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
