% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getPkParameterPopulationSensitivityPlot}
\alias{getPkParameterPopulationSensitivityPlot}
\title{getPkParameterPopulationSensitivityPlot}
\usage{
getPkParameterPopulationSensitivityPlot(data, settings)
}
\arguments{
\item{data}{data.frame of sensitivity analysis results}

\item{settings}{`SensitivityPlotSettings` object}
}
\value{
plt sensitivity plot based on results in input data
}
\description{
build sensitvity plot object for a population for one output and pk parameter
}
