% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getPkParametersAcrossPopulations}
\alias{getPkParametersAcrossPopulations}
\title{getPkParametersAcrossPopulations}
\usage{
getPkParametersAcrossPopulations(structureSets)
}
\arguments{
\item{structureSets}{list of `SimulationStructures` objects}
}
\value{
list of data.frame and its metaData including the values of PK parameters across Population Simulation sets
}
\description{
Get the values of PK parameters across Population Simulation sets
}
