% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPKParametersInOutput}
\alias{getPKParametersInOutput}
\title{getPKParametersInOutput}
\usage{
getPKParametersInOutput(output)
}
\arguments{
\item{output}{Output object}
}
\value{
Names of pkParameters in `output`
}
\description{
getPKParametersInOutput
}
