% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPKParametersInSimulationSet}
\alias{getPKParametersInSimulationSet}
\title{getPKParametersInSimulationSet}
\usage{
getPKParametersInSimulationSet(simulationSet)
}
\arguments{
\item{simulationSet}{SimulationSet object or derived class}
}
\value{
Data.frame with \code{path} and \code{pkParameter} in `simulationSet`
}
\description{
getPKParametersInSimulationSet
}
