% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{getPKParametersInSimulationSet}
\alias{getPKParametersInSimulationSet}
\title{getPKParametersInSimulationSet}
\usage{
getPKParametersInSimulationSet(simulationSet)
}
\arguments{
\item{simulationSet}{A `SimulationSet` object}
}
\value{
A data.frame including \code{path} and \code{pkParameter} from the `outputs` field of `simulationSet`
}
\description{
getPKParametersInSimulationSet
}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemographyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemographyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemographyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemographyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setXParametersForDemographyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemographyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
