% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getPKRatioForMapping}
\alias{getPKRatioForMapping}
\title{getPKRatioForMapping}
\usage{
getPKRatioForMapping(
  pkRatioMapping,
  pkParameterNames,
  configurationPlan,
  logFolder
)
}
\arguments{
\item{pkRatioMapping}{List providing the PK ratio mapping from field `PKRatios` of configuration plan}

\item{pkParameterNames}{Names of PK Parameters as defined by users}

\item{configurationPlan}{A `ConfigurationPlan` object}

\item{logFolder}{folder where the logs are saved}
}
\value{
list with `data` and `metaData`
}
\description{
Get data of pk ratio from field `PKRatios` of configuration plan
}
\keyword{internal}
