% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ratio-comparison.R
\name{getPKRatioSummaryStatistics}
\alias{getPKRatioSummaryStatistics}
\title{getPKRatioSummaryStatistics}
\usage{
getPKRatioSummaryStatistics(pkData, referencePKData, simulationSetName)
}
\arguments{
\item{pkData}{A data.frame of PK Parameter values for Population to compare}

\item{referencePKData}{A data.frame of PK Parameter values for reference Population}

\item{simulationSetName}{Name of simulation set}
}
\value{
A data.frame of the PK Parameter ratios summary statistics
}
\description{
Calculate and save summary statistics of ratios of PK parameters
}
\keyword{internal}
