% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getPKResultsDataFrame}
\alias{getPKResultsDataFrame}
\title{getPKResultsDataFrame}
\usage{
getPKResultsDataFrame(structureSet)
}
\arguments{
\item{structureSet}{\code{SimulationStructure} R6 class object}
}
\value{
pkResultsDataFrame, a dataframe storing the contents of the CSV file with path pkParameterResultsFilePath
}
\description{
Read PK parameter results into a dataframe and set QuantityPath,Parameter and Unit columns as factors
}
\keyword{internal}
