% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getPopSensDfForPkAndOutput}
\alias{getPopSensDfForPkAndOutput}
\title{getPopSensDfForPkAndOutput}
\usage{
getPopSensDfForPkAndOutput(
  indexDf,
  output,
  pkParameter,
  quantiles,
  rankFilter = NULL
)
}
\arguments{
\item{output}{pathID of output for which to obtain the population sensitivity results}

\item{pkParameter}{name of PK parameter for which to obtain the population sensitivity results}

\item{quantiles}{of population distribution of pkParameter, used to select individuals for sensitivity analysis}

\item{rankFilter}{results for only the 'rankFilter' most sensitive parameters will be returned}

\item{structureSet}{`SimulationStructure` R6 class object}
}
\value{
sortedFilteredIndividualsDfForPKParameter dataframe of population-wide sensitivity results for pkParameter and output
}
\description{
Retrieve dataframe of ranked and filtered population sensitivity results for a given PK parameter and model output pathID
}
