% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioData}
\alias{getQualificationPKRatioData}
\title{getQualificationPKRatioData}
\usage{
getQualificationPKRatioData(pkRatioPlan, configurationPlan, logFolder)
}
\arguments{
\item{pkRatioPlan}{List providing the PK ratio mapping from field `PKRatioPlots` of configuration plan}

\item{configurationPlan}{A `ConfigurationPlan` object}

\item{logFolder}{folder where the logs are saved}
}
\value{
list with `data` and `metaData`
}
\description{
Get data of pk ratio from field `PKRatioPlots` of configuration plan
}
