% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{getQualificationPKRatioGMFE}
\alias{getQualificationPKRatioGMFE}
\title{getQualificationPKRatioGMFE}
\usage{
getQualificationPKRatioGMFE(pkParameterNames, data)
}
\arguments{
\item{data}{data.frame with PK Ratios}

\item{pkParameterName}{Name of PK Parameter as defined by users}
}
\value{
A data.frame
}
\description{
Get Geometric Mean Fold Error for PK ratio plots
}
\keyword{internal}
