% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getQuantileIndividualIds}
\alias{getQuantileIndividualIds}
\title{getQuantileIndividualIds}
\usage{
getQuantileIndividualIds(
  pkAnalysisResultsDataframe,
  quantileVec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{pkAnalysisResultsDataframe}{Dataframe storing the PK analysis results for multiple individuals for a single PK parameter and single output path}
}
\value{
ids, IDs of individuals whose PK analysis results closest to quantiles given by vector of quantiles quantileVec
}
\description{
Find IDs of individuals whose PK analysis results closest to quantiles given by vector of quantiles quantileVec
}
