% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{getResiduals}
\alias{getResiduals}
\title{getResiduals}
\usage{
getResiduals(data, metaData = NULL, dataMapping = NULL)
}
\arguments{
\item{data}{data.frame of time profile data}

\item{metaData}{meta data on `data`}

\item{dataMapping}{XYDataMapping class object mapping what variable is to compare}
}
\value{
residualsData data.frame with Time, Observed, Simulated, Residuals
}
\description{
Get a data.frame matching simulated data to observed data and compute the residuals
}
