% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-of-work.R, R/utils-plots.R
\name{getResiduals}
\alias{getResiduals}
\title{getResiduals}
\usage{
getResiduals(simulatedData, observedData, dataMapping = NULL)

getResiduals(simulatedData, observedData, dataMapping = NULL)
}
\arguments{
\item{simulatedData}{data.frame of simulated data}

\item{observedData}{data.frame of observed data}

\item{dataMapping}{XYDataMapping class object mapping what variable is to compare}

\item{simulatedData}{data.frame of simulated data}

\item{observedData}{data.frame of observed data}

\item{dataMapping}{XYDataMapping class object mapping what variable is to compare}
}
\value{
residualsData data.frame with Time, Observed, Simulated, Residuals

residualsData data.frame with Time, Observed, Simulated, Residuals
}
\description{
Get a data.frame matching simulated data to observed data and compute the residuals

Get a data.frame matching simulated data to observed data and compute the residuals
}
