% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getSAFileIndex}
\alias{getSAFileIndex}
\title{getSAFileIndex}
\usage{
getSAFileIndex(structureSet, settings, logFolder, resultsFileName)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{settings}{list of settings for the population sensitivity analysis}

\item{logFolder}{folder where the logs are saved}

\item{resultsFileName}{root name of population sensitivity analysis results CSV files}
}
\description{
Function to build and write to CSV a dataframe that stores all
sensitivity analysis result files that will be output by a population sensitivity analysis.
}
