% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{getSAFileIndex}
\alias{getSAFileIndex}
\title{getSAFileIndex}
\usage{
getSAFileIndex(
  pkParameterResultsFilePath,
  quantileVec,
  resultsFileFolder,
  resultsFileName,
  fileNamesIndex
)
}
\arguments{
\item{pkParameterResultsFilePath}{path to pk parameter results CSV file}

\item{quantileVec}{quantiles to seek}
}
\description{
Function to build and write to CSV a dataframe that
}
