% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{getSectionTOC}
\alias{getSectionTOC}
\title{getSectionTOC}
\usage{
getSectionTOC(
  fileName,
  logFolder = getwd(),
  numberSections = TRUE,
  tocPattern = "#",
  tocLevels = 6
)
}
\arguments{
\item{fileName}{name of .md file to update}

\item{logFolder}{folder where the logs are saved}

\item{numberSections}{logical defining if sections are numbered}

\item{tocPattern}{character pattern referencing sections in first element of line}

\item{tocLevels}{levels of sections in the report}
}
\value{
Table of content referencing sections following a markdown format
}
\description{
Reference sections of a report
}
\keyword{internal}
