% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getSimulatedCurveProperties}
\alias{getSimulatedCurveProperties}
\title{getSimulatedCurveProperties}
\usage{
getSimulatedCurveProperties(
  configurationPlanCurve,
  simulation,
  simulationResults,
  axesProperties,
  configurationPlan
)
}
\arguments{
\item{configurationPlanCurve}{`Curves` fields of configuration plan}

\item{simulation}{A `Simulation` object from `ospsuite` package
that includes required information to identify and convert the data requested from `configurationPlanCurve` properties}

\item{simulationResults}{A `SimulationResults` object from `ospsuite` package
that includes the data requested from `configurationPlanCurve` properties}

\item{axesProperties}{list of axes properties obtained from `getAxesProperties`}

\item{configurationPlan}{A `ConfigurationPlan` object that includes methods to find observed data}
}
\value{
A named list data and meta data parameters
}
\description{
Get Curve Properties and values of Simulated data
}
\keyword{internal}
