% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSimulationParameterDisplayPaths}
\alias{getSimulationParameterDisplayPaths}
\title{getSimulationParameterDisplayPaths}
\usage{
getSimulationParameterDisplayPaths(parameterPaths, simulation, dictionary)
}
\arguments{
\item{parameterPaths}{Paths of a parameter in simulation}

\item{simulation}{`Simulation` object from `ospsuite`}

\item{dictionary}{parameterDisplayPaths data.frame mapping user defined display names}
}
\value{
parameterDisplayPath
}
\description{
getSimulationParameterDisplayPaths
}
