% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{getSimulationTimeRanges}
\alias{getSimulationTimeRanges}
\title{getSimulationTimeRanges}
\usage{
getSimulationTimeRanges(simulation, path, simulationSet)
}
\arguments{
\item{simulation}{A `Simulation` object}

\item{path}{Field `path` from `Output` object}

\item{simulationSet}{A `SimulationSet` or `PopulationSimulationSet` object}
}
\value{
Lists including `values` and `name` of time ranges.
Also includes logical field `keep` to define if a specific application range is kept in report.
}
\description{
Get time ranges for time profile plots according to applications and user defined settings
}
\keyword{internal}
