% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-ddi.R
\name{getSmartZoomLimits}
\alias{getSmartZoomLimits}
\title{getSmartZoomLimits}
\usage{
getSmartZoomLimits(isResidualsVsObserved, dataVector)
}
\arguments{
\item{isResidualsVsObserved}{a logical argument indicating if the data pertains to a residual vs observed (`TRUE`) or a predicted vs observed (`FALSE`) plot}

\item{dataVector}{a numeric vector containing the x or y values of datapoints to be plotted}
}
\value{
a list containing the minimum and maximum axes limits
}
\description{
Get default axis limits for DDI Ratio plot
}
\keyword{internal}
