% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingengine-env.R
\name{getStatisticsFromType}
\alias{getStatisticsFromType}
\title{getStatisticsFromType}
\usage{
getStatisticsFromType(statisticsType)
}
\arguments{
\item{statisticsType}{Statistics summarizing time profile simulated data
as defined by helper enum \code{StatisticsType}}
}
\value{
A list including \code{y}, \code{ymin} and \code{ymax} summary statistics as well as their \code{caption}
}
\description{
Get statistics
}
\examples{
\dontrun{
getStatisticsFromType(StatisticsTypes$`Arithmetic mean`)
}

}
