% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{getTaskInputs}
\alias{getTaskInputs}
\title{getTaskInputs}
\usage{
getTaskInputs(task)
}
\arguments{
\item{task}{`Task` object or derived class}
}
\value{
Names of the files required to perform the task
}
\description{
Get the names of the files required to perform the task
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- MeanModelWorkflow$new(workflowFolder, simulationSets)

# Get paths of required files for running a workflow task
getTaskInputs(myWorkflow$plotSensitivity)
}

}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemographyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemographyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemographyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemographyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setXParametersForDemographyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemographyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
