% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{getTimeProfileObservedDataFromResults}
\alias{getTimeProfileObservedDataFromResults}
\title{getTimeProfileObservedDataFromResults}
\usage{
getTimeProfileObservedDataFromResults(
  observedResults,
  molWeight,
  axesProperties,
  observedDataId
)
}
\arguments{
\item{observedResults}{List that includes
\itemize{
\item \code{data}: a data.frame with column 1 = Time, column 2 = Concentration, column 3 = Error
\item \code{metaData}: a list for each column of data that includes their unit
}}

\item{molWeight}{Molecular weight of compound}

\item{axesProperties}{list of axes properties obtained from \code{getAxesProperties}}

\item{observedDataId}{Id of the observed data for better handling error messages}
}
\value{
List with \code{time}, \code{y} and \code{error} values
}
\description{
Get the time profile observed data in appropriate unit from the results from the observed data file
}
\keyword{internal}
