% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-excel-input.R
\name{getWorkflowContent}
\alias{getWorkflowContent}
\title{getWorkflowContent}
\usage{
getWorkflowContent(workflowTable, excelFile)
}
\arguments{
\item{workflowTable}{Data.frame read from the Excel sheet "Workflow and Tasks"}

\item{excelFile}{name of the Excel file from which the R script is created}
}
\value{
Character vector defining the \code{Workflow} object
}
\description{
Creates a character vector to be written in a workflow .R script defining \code{Workflow} object
}
\keyword{internal}
