% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getXParametersForPKParametersPlot}
\alias{getXParametersForPKParametersPlot}
\title{getXParametersForPKParametersPlot}
\usage{
getXParametersForPKParametersPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of x-parameters used for PK parameters range plots
}
\description{
getXParametersForPKParametersPlot
}
