% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getXParametersForPkParametersPlot}
\alias{getXParametersForPkParametersPlot}
\title{getXParametersForPkParametersPlot}
\usage{
getXParametersForPkParametersPlot(workflow)
}
\description{
getXParametersForPkParametersPlot
}
