% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{getXParametersForPKParametersPlot}
\alias{getXParametersForPKParametersPlot}
\alias{getXParametersForPkParametersPlot}
\title{getXParametersForPKParametersPlot}
\usage{
getXParametersForPKParametersPlot(workflow)

getXParametersForPkParametersPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of x-parameters used for PK parameters range plots
}
\description{
getXParametersForPKParametersPlot
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- PopulationWorkflow$new(worflowType, workflowFolder, simulationSets)

# Get the list of parameters in x-axis for range plots
getXParametersForPKParametersPlot(workflow = myWorkflow)
}

}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemographyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemographyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemographyPlot}()},
\code{\link{getYParametersForDemographyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setXParametersForDemographyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemographyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
