% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getYParametersForPkParametersPlot}
\alias{getYParametersForPkParametersPlot}
\title{getYParametersForPkParametersPlot}
\usage{
getYParametersForPkParametersPlot(workflow)
}
\description{
getYParametersForPkParametersPlot
}
