% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{getYParametersForPkParametersPlot}
\alias{getYParametersForPkParametersPlot}
\title{getYParametersForPkParametersPlot}
\usage{
getYParametersForPkParametersPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of y-parameters used for PK parameters range plots and boxplots
}
\description{
getYParametersForPkParametersPlot
}
