% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{getYParametersForPKParametersPlot}
\alias{getYParametersForPKParametersPlot}
\alias{getYParametersForPkParametersPlot}
\title{getYParametersForPKParametersPlot}
\usage{
getYParametersForPKParametersPlot(workflow)

getYParametersForPkParametersPlot(workflow)
}
\arguments{
\item{workflow}{`PopulationWorkflow` R6 class object}
}
\value{
list of y-parameters used for PK parameters range plots and boxplots
}
\description{
getYParametersForPKParametersPlot
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- PopulationWorkflow$new(worflowType, workflowFolder, simulationSets)

# Get the list of parameters in y-axis for range plots and boxplots
getYParametersForPKParametersPlot(workflow = myWorkflow)
}

}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemogrpahyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemogrpahyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemogrpahyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemogrpahyPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setWorkflowParameterDisplayPaths}()},
\code{\link{setXParametersForDemogrpahyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemogrpahyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
