% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{inactivateWorkflowTasks}
\alias{inactivateWorkflowTasks}
\title{inactivateWorkflowTasks}
\usage{
inactivateWorkflowTasks(workflow, tasks = workflow$getAllTasks())
}
\arguments{
\item{workflow}{`MeanModelWorklfow` or `PopulationWorklfow` object}

\item{tasks}{names of the tasks to activate
Default inactivates all tasks of the workflow using workflow method `workflow$getAllTasks()`}
}
\description{
inactivates a series of `Tasks` from a `Workflow`
}
