% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{individualSensitivityAnalysis}
\alias{individualSensitivityAnalysis}
\title{individualSensitivityAnalysis}
\usage{
individualSensitivityAnalysis(structureSet, settings, individualParameters)
}
\arguments{
\item{structureSet}{\code{SimulationStructure} R6 class object}

\item{settings}{list of settings for the sensitivity analysis}

\item{individualParameters}{is an object storing an individual's parameters, obtained from a
population object's getParameterValuesForIndividual() function.}
}
\value{
SA results for an individual
}
\description{
Run SA for an individual, possibly after modifying the simulation using individualParameters.
Determine whether to run SA for on single core or in parallel.
If on single core, pass simulation to analyzeSensitivity.
If in parallel, pass simulation to runParallelSensitivityAnalysis.
}
\keyword{internal}
