% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{individualSensitivityAnalysis}
\alias{individualSensitivityAnalysis}
\title{individualSensitivityAnalysis}
\usage{
individualSensitivityAnalysis(
  simFilePath,
  variableParameterPaths = NULL,
  individualParameters,
  variationRange,
  numberOfCores = 1,
  resultsFileFolder = resultsFileFolder,
  resultsFileName = resultsFileName,
  logFolder = getwd()
)
}
\arguments{
\item{simFilePath}{path to simulation file}

\item{variableParameterPaths}{paths to parameters to vary in sensitivity analysis}

\item{individualParameters}{is an object storing an individual's parameters, obtained from a
population object's getParameterValuesForIndividual() function.}

\item{variationRange}{variation range for sensitivity analysis}

\item{numberOfCores}{is the number of cores over which to parallelize the sensitivity analysis}

\item{resultsFileFolder}{path to population sensitivity analysis results CSV files}

\item{resultsFileName}{root name of population sensitivity analysis results CSV files}
}
\value{
allResultsFileNames, the paths to CSV files containing results of sensitivity analysis
}
\description{
Run SA for an individual, possibly after modifying the simulation using individualParameters.
Determine whether to run SA for on single core or in parallel.
If on single core, pass simulation to analyzeCoreSensitivity.
If in parallel, pass simulation to runParallelSensitivityAnalysis.
}
