% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sensitivity-analysis.R
\name{individualSensitivityAnalysis}
\alias{individualSensitivityAnalysis}
\title{individualSensitivityAnalysis}
\usage{
individualSensitivityAnalysis(
  simFilePath,
  parametersToPerturb = NULL,
  individualParameters,
  numberOfCores = 1,
  resultsFileFolder = resultsFileFolder,
  resultsFileName = resultsFileName
)
}
\value{
SA results for individual or population
}
\description{
Run SA for an individual, possibly after modifying the simulation using individualParameters.  Determine whether to run SA for on single core or in parallel.
If on single core, pass simulation to analyzeCoreSensitivity.  If in parallel, pass simulation to runParallelSensitivityAnalysis.
}
