% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{initializeRmdFile}
\alias{initializeRmdFile}
\title{initializeRmdFile}
\usage{
initializeRmdFile(fileName, title = "", author = "")
}
\arguments{
\item{fileName}{name of Rmd file to initialize}

\item{title}{title of the document}

\item{author}{author of the}
}
\value{
R markdown text and save the Rmd document
}
\description{
Initialize a R markdown document by Writing its header in YAML
}
