% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isLoadedPopulation}
\alias{isLoadedPopulation}
\title{isLoadedPopulation}
\usage{
isLoadedPopulation(population)
}
\arguments{
\item{population}{A `Population` object}
}
\value{
Logical indicating if the `population` exists
}
\description{
Check that the population has been loaded
}
