% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isLoadedSimulation}
\alias{isLoadedSimulation}
\title{isLoadedSimulation}
\usage{
isLoadedSimulation(simulation)
}
\arguments{
\item{simulation}{A \code{Simulation} object}
}
\value{
Logical indicating if the \code{simulation} exists
}
\description{
Check that the simulation has been loaded
}
