% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-observed-data.R
\name{isObservedData}
\alias{isObservedData}
\title{isObservedData}
\usage{
isObservedData(path)
}
\arguments{
\item{path}{A quantitiy path from the configuration plan
For instance, "S2|Organism|PeripheralVenousBlood|Midazolam|Plasma (Peripheral Venous Blood)"
or "Midazolam 600mg SD|ObservedData|Peripheral Venous Blood|Plasma|Rifampin|Conc"}
}
\value{
A logical checking if path corresponds to observed data
}
\description{
Check if a configuration plan quantitiy path corresponds to observed data
}
\examples{
isObservedData("S2|Organism|PeripheralVenousBlood|Midazolam|Plasma (Peripheral Venous Blood)")
# > FALSE
isObservedData("Midazolam 600mg SD|ObservedData|Peripheral Venous Blood|Plasma|Rifampin|Conc")
# > TRUE
}
