% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{loadConfigurationPlan}
\alias{loadConfigurationPlan}
\title{loadConfigurationPlan}
\usage{
loadConfigurationPlan(configurationPlanFile, workflowFolder)
}
\arguments{
\item{configurationPlanFile}{path to the json file corresponding to the Configuration Plan of a Qualification workflow}

\item{workflowFolder}{path of the output folder created or used by the Workflow.}
}
\value{
A `ConfigurationPlan` object including the content of json file
}
\description{
Load a `ConfigurationPlan` object from json file
}
\seealso{
Other qualification workflow: 
\code{\link{ConfigurationPlan}},
\code{\link{adjustTitlePage}()},
\code{\link{loadQualificationWorkflow}()},
\code{\link{startQualificationRunner}()}
}
\concept{qualification workflow}
