% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{loadPKAnalysesFromCSV}
\alias{loadPKAnalysesFromCSV}
\title{loadPKAnalysesFromCSV}
\usage{
loadPKAnalysesFromCSV(filePath, simulation, to = "PKAnalyses")
}
\arguments{
\item{filePath}{Full path of the file containing the PK-Analyses to load}

\item{simulation}{A \code{Simulation} object}

\item{to}{Format of the loaded output`}
}
\value{
A \code{PKAnalyses} object if \code{to="PKAnalyses"}
A \code{data.frame} if \code{to="data.frame"}
A \code{tibble} if \code{to="tibble"}
}
\description{
Load PK analyses from a CSV file
Wrap the \code{ospsuite::importPKAnalysesFromCSV()} to provide more useful warning messages.
}
\keyword{internal}
