% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{loadPKAnalysesFromStructureSet}
\alias{loadPKAnalysesFromStructureSet}
\title{loadPKAnalysesFromStructureSet}
\usage{
loadPKAnalysesFromStructureSet(
  structureSet,
  to = "PKAnalyses",
  useCache = FALSE
)
}
\arguments{
\item{structureSet}{A `SimulationStructure` object}

\item{to}{Format of the loaded output`}
}
\value{
A `PKAnalyses` object if `to="PKAnalyses"`
A `data.frame` if `to="data.frame"`
A `tibble` if `to="tibble"`
}
\description{
loadPKAnalysesFromStructureSet
}
\keyword{internal}
