% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotAbsorptionTask}
\alias{loadPlotAbsorptionTask}
\title{loadPlotAbsorptionTask}
\usage{
loadPlotAbsorptionTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`MeanModelWorkflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `FALSE`}

\item{settings}{specific settings for `plotAbsorption` task}
}
\value{
A `PlotTask` object
}
\description{
Define `plotAbsorption` task and its settings
}
