% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotMassBalanceTask}
\alias{loadPlotMassBalanceTask}
\title{loadPlotMassBalanceTask}
\usage{
loadPlotMassBalanceTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{\code{MeanModelWorkflow} object}

\item{active}{logical defining if \code{Task} will be run by workflow.
Default value is \code{FALSE}}

\item{settings}{specific settings for \code{plotMassBalance} task}
}
\value{
A \code{PlotTask} object
}
\description{
Define \code{plotMassBalance} task and its settings
}
\examples{
\dontrun{

# A workflow object needs to be created first
myWorkflow <- MeanModelWorkflow$new(workflowFolder, simulationSets)

# (Re)load a default plotMassBalance task for workflow
myWorkflow$plotMassBalance <- loadPlotMassBalanceTask(
  workflow = myWorkflow,
  active = TRUE
)

# Load a user-defined simulation task for workflow
myWorkflow$userDefinedTasks[["plotMassBalance"]] <- loadPlotMassBalanceTask(
  workflow = myWorkflow,
  active = TRUE
)
}

}
\seealso{
Other workflow tasks: 
\code{\link{GofPlotTask}},
\code{\link{PlotTask}},
\code{\link{PopulationPlotTask}},
\code{\link{PopulationSensitivityAnalysisTask}},
\code{\link{QualificationTask}},
\code{\link{SensitivityAnalysisTask}},
\code{\link{Task}},
\code{\link{addUserDefinedTask}()},
\code{\link{loadCalculatePKParametersTask}()},
\code{\link{loadCalculateSensitivityTask}()},
\code{\link{loadGOFMergedTask}()},
\code{\link{loadPlotAbsorptionTask}()},
\code{\link{loadPlotDDIRatioTask}()},
\code{\link{loadPlotDemographyTask}()},
\code{\link{loadPlotPKParametersTask}()},
\code{\link{loadPlotPKRatioTask}()},
\code{\link{loadPlotSensitivityTask}()},
\code{\link{loadPlotTimeProfilesAndResidualsTask}()},
\code{\link{loadQualificationComparisonTimeProfileTask}()},
\code{\link{loadQualificationTimeProfilesTask}()},
\code{\link{loadSimulateTask}()}
}
\concept{workflow tasks}
