% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadPlotTimeProfilesAndResidualsTask}
\alias{loadPlotTimeProfilesAndResidualsTask}
\title{loadPlotTimeProfilesAndResidualsTask}
\usage{
loadPlotTimeProfilesAndResidualsTask(workflow, active = FALSE, settings = NULL)
}
\arguments{
\item{workflow}{`Workflow` object or derived class}

\item{active}{logical defining if `Task` will be run by workflow.
Default value is `TRUE`}

\item{settings}{specific settings for `simulate` task}
}
\value{
A `GofPlotTask` object
}
\description{
Define `plotTimeProfilesAndResiduals` task and its settings
}
