% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{loadPopulationOnCores}
\alias{loadPopulationOnCores}
\title{loadPopulationOnCores}
\usage{
loadPopulationOnCores(populationFiles, logFolder)
}
\arguments{
\item{logFolder}{folder where the logs are saved}

\item{structureSet}{containing simulationSet which contains path to simulation file and pathIDs to be loaded in simulation object as outputs}
}
\description{
Send population file names to cores, check that population is loaded on each core successfully
}
