% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadQualificationComparisonTimeProfileTask}
\alias{loadQualificationComparisonTimeProfileTask}
\title{loadQualificationComparisonTimeProfileTask}
\usage{
loadQualificationComparisonTimeProfileTask(workflow, configurationPlan)
}
\arguments{
\item{workflow}{\code{QualificationWorkflow} object}

\item{configurationPlan}{A \code{ConfigurationPlan} object}
}
\value{
A \code{QualificationTask} object
}
\description{
Define \code{plotComparisonTimeProfileTask} task and its settings
}
\seealso{
Other workflow tasks: 
\code{\link{GofPlotTask}},
\code{\link{PlotTask}},
\code{\link{PopulationPlotTask}},
\code{\link{PopulationSensitivityAnalysisTask}},
\code{\link{QualificationTask}},
\code{\link{SensitivityAnalysisTask}},
\code{\link{Task}},
\code{\link{addUserDefinedTask}()},
\code{\link{loadCalculatePKParametersTask}()},
\code{\link{loadCalculateSensitivityTask}()},
\code{\link{loadGOFMergedTask}()},
\code{\link{loadPlotAbsorptionTask}()},
\code{\link{loadPlotDDIRatioTask}()},
\code{\link{loadPlotDemographyTask}()},
\code{\link{loadPlotMassBalanceTask}()},
\code{\link{loadPlotPKParametersTask}()},
\code{\link{loadPlotPKRatioTask}()},
\code{\link{loadPlotSensitivityTask}()},
\code{\link{loadPlotTimeProfilesAndResidualsTask}()},
\code{\link{loadQualificationTimeProfilesTask}()},
\code{\link{loadSimulateTask}()}
}
\concept{workflow tasks}
