% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\name{loadQualificationTimeProfilesTask}
\alias{loadQualificationTimeProfilesTask}
\title{loadPlotTimeProfilesTask}
\usage{
loadQualificationTimeProfilesTask(workflow, configurationPlan)
}
\arguments{
\item{workflow}{`QualificationWorkflow` object}

\item{configurationPlan}{A `ConfigurationPlan` object}
}
\value{
A `QualificationTask` object
}
\description{
Define `plotTimeProfiles` task and its settings
}
