% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{loadQualificationWorkflow}
\alias{loadQualificationWorkflow}
\title{loadQualificationWorkflow}
\usage{
loadQualificationWorkflow(workflowFolder, configurationPlanFile)
}
\arguments{
\item{workflowFolder}{path of the output folder created or used by the Workflow.}

\item{configurationPlanFile}{path to the json file corresponding to the Configuration Plan of a Qualification workflow}
}
\value{
A `QualificationWorkflow` object
}
\description{
Load a `ConfigurationPlan` object from json file
}
\seealso{
Other qualification workflow: 
\code{\link{ConfigurationPlan}},
\code{\link{adjustTitlePage}()},
\code{\link{loadConfigurationPlan}()},
\code{\link{startQualificationRunner}()}
}
\concept{qualification workflow}
