% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSimulationWithUpdatedPaths}
\alias{loadSimulationWithUpdatedPaths}
\title{loadSimulationWithUpdatedPaths}
\usage{
loadSimulationWithUpdatedPaths(simulationSet)
}
\arguments{
\item{simulationSet}{simulation set containing path to simulation file and pathIDs for quantities to be loaded into simulation object}
}
\value{
simulation object with pathIDs updated from simulationSet
}
\description{
loadSimulationWithUpdatedPaths
}
