% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation-set.R
\name{loadWorkflowPopulation}
\alias{loadWorkflowPopulation}
\title{loadWorkflowPopulation}
\usage{
loadWorkflowPopulation(simulationSet)
}
\arguments{
\item{simulationSet}{A `PopulationSimulationSet` object}
}
\description{
loadWorkflowPopulation
}
