% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logDebug}
\alias{logDebug}
\title{logDebug}
\usage{
logDebug(message, printConsole = NULL)
}
\arguments{
\item{message}{message to save in log file}

\item{printConsole}{logical to print error on console}
}
\description{
Save intermediate messages into a log debug file
}
