% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logDebug}
\alias{logDebug}
\title{logDebug}
\usage{
logDebug(
  message,
  file = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logDebugFile()),
  printConsole = FALSE
)
}
\arguments{
\item{message}{message to save in log file}

\item{file}{log file name.
Default is set in function logFileName}

\item{printConsole}{logical to print error on console}
}
\description{
Save intermediate messages into a log debug file
}
