% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logError}
\alias{logError}
\title{logError}
\usage{
logError(
  message,
  file = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logErrorFile()),
  printConsole = TRUE
)
}
\arguments{
\item{message}{message to save in log file}

\item{file}{log file name.
Default is set in function logFileName}

\item{printConsole}{logical to print error on console}
}
\description{
Save error messages into a log error file
}
