% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logErrorThenStop}
\alias{logErrorThenStop}
\title{logErrorThenStop}
\usage{
logErrorThenStop(message, logFolder = NULL)
}
\arguments{
\item{message}{message to display and then log}

\item{logFolder}{\strong{Deprecated} path where logs are saved}
}
\description{
Log the error with a message and then stop, displaying same message.
}
\keyword{internal}
