% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{logErrorThenStop}
\alias{logErrorThenStop}
\title{Log the error with a message and then stop, displaying same message.}
\usage{
logErrorThenStop(message)
}
\arguments{
\item{message}{message to display and then log}
}
\description{
Log the error with a message and then stop, displaying same message.
}
