% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logInfo}
\alias{logInfo}
\title{logInfo}
\usage{
logInfo(message, printConsole = NULL)
}
\arguments{
\item{message}{message to save in log file}

\item{printConsole}{logical to print error on console}
}
\description{
Save info messages into a log info file
}
