% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\docType{data}
\name{LogTypes}
\alias{LogTypes}
\title{LogTypes}
\format{
An object of class \code{list} of length 3.
}
\usage{
LogTypes
}
\description{
List of log types for recording workflow information
}
\keyword{datasets}
