% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{mapStudyDesignSources}
\alias{mapStudyDesignSources}
\title{mapStudyDesignSources}
\usage{
mapStudyDesignSources(data)
}
\arguments{
\item{data}{data.frame read from a study design file}
}
\value{
vector of expressions assigning target values
Must be the same length as target values
}
\description{
mapStudyDesignSources
}
