% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{mapStudyDesignTargets}
\alias{mapStudyDesignTargets}
\title{mapStudyDesignTargets}
\usage{
mapStudyDesignTargets(data, population, simulation)
}
\arguments{
\item{data}{data.frame read from a study design file}

\item{population}{`Population` object}

\item{simulation}{`Simulation` object}
}
\value{
list of `StudyDesignTarget` objects
}
\description{
mapStudyDesignTargets
}
\keyword{internal}
