% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-qualification.R
\name{massMoleConversion}
\alias{massMoleConversion}
\title{massMoleConversion}
\usage{
massMoleConversion(dimension)
}
\arguments{
\item{dimension}{is string that is from among the valid OSP dimensions}
}
\value{
If `dimension` is `Concentration (mass)` or `Mass`, then return `Concentration (molar)` or `Amount` respectively, otherwise return `dimension`.
}
\description{
Function to map `Concentration (mass)` or `Mass` dimensions to `Concentration (molar)` and `Amount` respectively.
}
