% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-set.R
\name{MeanModelSet}
\alias{MeanModelSet}
\title{MeanModelSet}
\description{
R6 class representing Reporting Engine Mean Model Set
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{simulationName}}{display name of simulation}

\item{\code{pathID}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{pathName}}{display name for `pathID`}

\item{\code{pathUnit}}{display unit for `pathID`}

\item{\code{pkParameters}}{PK parameters function names to be calculated from the simulation (e.g. `C_max`).}

\item{\code{pkParametersNames}}{display names for `pkParameters`}

\item{\code{pkParametersUnits}}{display units for `pkParameters`}

\item{\code{observedDataFile}}{name of csv file to be used for observed data}

\item{\code{observedMetaDataFile}}{name of csv file to be used as dictionary of the observed data}

\item{\code{dataFilter}}{vector of characters to filter the observed data}

\item{\code{timeUnit}}{display unit for time variable}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanModelSet$new()}}
\item \href{#method-copyInputFiles}{\code{MeanModelSet$copyInputFiles()}}
\item \href{#method-clone}{\code{MeanModelSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `MeanModelSet` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelSet$new(
  simulationSetName = NULL,
  simulationFile,
  simulationName = NULL,
  pathID = NULL,
  pathName = NULL,
  pathUnit = NULL,
  pkParameters = enum(ospsuite::allPKParameterNames()),
  pkParametersNames = NULL,
  pkParametersUnits = NULL,
  observedDataFile = NULL,
  observedMetaDataFile = NULL,
  dataFilter = NULL,
  timeUnit = "h"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{simulationName}}{display name of simulation}

\item{\code{pathID}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{pathName}}{display name for `pathID`}

\item{\code{pathUnit}}{display unit for `pathID`}

\item{\code{pkParameters}}{PK parameters function names to be calculated from the simulation (e.g. `C_max`).
Default value is obtained from enum(ospsuite::allPKParameterNames()).}

\item{\code{pkParametersNames}}{display names for `pkParameters`}

\item{\code{pkParametersUnits}}{display units for `pkParameters`}

\item{\code{observedDataFile}}{name of csv file to be used for observed data}

\item{\code{observedMetaDataFile}}{name of csv file to be used as dictionary of the observed data}

\item{\code{dataFilter}}{vector of characters to filter the observed data}

\item{\code{timeUnit}}{display unit for time variable. Default is "h"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelSet` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyInputFiles"></a>}}
\subsection{Method \code{copyInputFiles()}}{
Create a copy of the input files (pkml simulation and data if available) within a folder dedicated to the simulation set
Rename the input file if `simulationName` is explicitely defined
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelSet$copyInputFiles(inputFilesFolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputFilesFolder}}{path where the files are copied}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
