% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-workflow.R
\name{MeanModelWorkflow}
\alias{MeanModelWorkflow}
\title{MeanModelWorkflow}
\description{
R6 class for Reporting Engine Mean Model Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{MeanModelWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meanModelSimulation}}{R6 class `Task` for simulation}

\item{\code{meanModelPKParameters}}{R6 class `Task` for PK parameters calculation}

\item{\code{meanModelSensitivityAnalysis}}{R6 class `Task` for sensitivity analysis}

\item{\code{plotGoF}}{R6 class `Task` for goodness of fit plots}

\item{\code{plotMassBalance}}{R6 class `Task` for mass balance plot}

\item{\code{plotAbsorption}}{R6 class `Task` for absorption plot}

\item{\code{plotPKParameters}}{R6 class `Task` for PK parameters plot}

\item{\code{plotSensitivity}}{R6 class `Task` for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanModelWorkflow$new()}}
\item \href{#method-meanModelSimulationSettings}{\code{MeanModelWorkflow$meanModelSimulationSettings()}}
\item \href{#method-calculatePKParametersSettings}{\code{MeanModelWorkflow$calculatePKParametersSettings()}}
\item \href{#method-meanModelSensitivityAnalysisSettings}{\code{MeanModelWorkflow$meanModelSensitivityAnalysisSettings()}}
\item \href{#method-plotGoFSettings}{\code{MeanModelWorkflow$plotGoFSettings()}}
\item \href{#method-plotPKParametersSettings}{\code{MeanModelWorkflow$plotPKParametersSettings()}}
\item \href{#method-plotMassBalanceSettings}{\code{MeanModelWorkflow$plotMassBalanceSettings()}}
\item \href{#method-plotAbsorptionSettings}{\code{MeanModelWorkflow$plotAbsorptionSettings()}}
\item \href{#method-plotSensitivitySettings}{\code{MeanModelWorkflow$plotSensitivitySettings()}}
\item \href{#method-runWorkflow}{\code{MeanModelWorkflow$runWorkflow()}}
\item \href{#method-print}{\code{MeanModelWorkflow$print()}}
\item \href{#method-clone}{\code{MeanModelWorkflow$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `MeanModelWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanModelSimulationSettings"></a>}}
\subsection{Method \code{meanModelSimulationSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$meanModelSimulationSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  inputFolderName = self$inputFolder,
  simulationFileName = self$simulation,
  resultsFolderName = self$simulationFolder,
  resultsFileName = "meanModelSimulation"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculatePKParametersSettings"></a>}}
\subsection{Method \code{calculatePKParametersSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$calculatePKParametersSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  simulationFilePath = file.path(self$inputFolder, paste0(self$simulation, ".pkml")),
  simulationResultFilePaths = self$meanModelSimulation$generatedResultFileNames,
  pkParametersToEvaluate = NULL,
  userDefinedPKFunctions = NULL,
  pkParameterResultsFilePath = file.path(self$pkParametersFolder,
    "meanModelPKParameters.csv")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanModelSensitivityAnalysisSettings"></a>}}
\subsection{Method \code{meanModelSensitivityAnalysisSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$meanModelSensitivityAnalysisSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  inputFolderName = self$inputFolder,
  simulationFileName = self$simulation,
  resultsFolderName = self$sensitivityFolder,
  resultsFileName = "meanModelSensitivityAnalysis",
  numberOfCores = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGoFSettings"></a>}}
\subsection{Method \code{plotGoFSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotGoFSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPKParametersSettings"></a>}}
\subsection{Method \code{plotPKParametersSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotPKParametersSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMassBalanceSettings"></a>}}
\subsection{Method \code{plotMassBalanceSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotMassBalanceSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotAbsorptionSettings"></a>}}
\subsection{Method \code{plotAbsorptionSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotAbsorptionSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSensitivitySettings"></a>}}
\subsection{Method \code{plotSensitivitySettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotSensitivitySettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\subsection{Method \code{runWorkflow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print workflow list of tasks
TO DO: add simulationSets to print() method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Task list information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
