% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-workflow.R
\docType{class}
\name{MeanModelWorkflow}
\alias{MeanModelWorkflow}
\title{MeanModelWorkflow}
\usage{
MeanModelWorkflow
}
\description{
Mean Model Workflow Task for Reporting Engine
}
\section{Fields}{

\describe{
\item{\code{meanModelSimulation}}{Population simulation task}

\item{\code{pkParametersCalculation}}{PK parameters calculation task}

\item{\code{sensitivityAnalysis}}{Sensitivity analysis task}
}}

\section{Methods}{

\describe{
\item{new()}{Initilialize reporting engine mean model workflow}
\item{setMeanModelSimulationSettings()}{Define mean model simulation task settings}
\item{runWorkflow()}{Run the active tasks of mean model worklfow}
}
}

\keyword{datasets}
