% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-workflow.R
\name{MeanModelWorkflow}
\alias{MeanModelWorkflow}
\title{MeanModelWorkflow}
\description{
R6 class for Reporting Engine Mean Model Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{MeanModelWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meanModelSimulation}}{R6 class `Task` for simulation}

\item{\code{meanModelPKParameters}}{R6 class `Task` for PK parameters calculation}

\item{\code{meanModelSensitivityAnalysis}}{R6 class `Task` for sensitivity analysis}

\item{\code{plotGoF}}{R6 class `Task` for goodness of fit plots}

\item{\code{plotMassBalance}}{R6 class `Task` for mass balance plot}

\item{\code{plotAbsorption}}{R6 class `Task` for absorption plot}

\item{\code{plotPKParameters}}{R6 class `Task` for PK parameters plot}

\item{\code{plotSensitivity}}{R6 class `Task` for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanModelWorkflow$new()}}
\item \href{#method-meanModelSimulationSettings}{\code{MeanModelWorkflow$meanModelSimulationSettings()}}
\item \href{#method-calculatePKParametersSettings}{\code{MeanModelWorkflow$calculatePKParametersSettings()}}
\item \href{#method-meanModelSensitivityAnalysisSettings}{\code{MeanModelWorkflow$meanModelSensitivityAnalysisSettings()}}
\item \href{#method-plotGoFSettings}{\code{MeanModelWorkflow$plotGoFSettings()}}
\item \href{#method-plotPKParametersSettings}{\code{MeanModelWorkflow$plotPKParametersSettings()}}
\item \href{#method-plotMassBalanceSettings}{\code{MeanModelWorkflow$plotMassBalanceSettings()}}
\item \href{#method-plotAbsorptionSettings}{\code{MeanModelWorkflow$plotAbsorptionSettings()}}
\item \href{#method-plotSensitivitySettings}{\code{MeanModelWorkflow$plotSensitivitySettings()}}
\item \href{#method-runWorkflow}{\code{MeanModelWorkflow$runWorkflow()}}
\item \href{#method-print}{\code{MeanModelWorkflow$print()}}
\item \href{#method-clone}{\code{MeanModelWorkflow$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getActiveTasks}{\code{ospsuite.reportingengine::Workflow$getActiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllTasks}{\code{ospsuite.reportingengine::Workflow$getAllTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getInactiveTasks}{\code{ospsuite.reportingengine::Workflow$getInactiveTasks()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `MeanModelWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanModelSimulationSettings"></a>}}
\subsection{Method \code{meanModelSimulationSettings()}}{
Define simulate `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$meanModelSimulationSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{inputFolderName}}{folder of input}

\item{\code{simulationFileName}}{name of simulation file}

\item{\code{resultsFolderName}}{folder where output is saved}

\item{\code{resultsFileName}}{file where output is saved}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculatePKParametersSettings"></a>}}
\subsection{Method \code{calculatePKParametersSettings()}}{
Define calculate PK Parameters `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$calculatePKParametersSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{simulationFilePath}}{TO DO}

\item{\code{simulationResultFilePaths}}{TO DO}

\item{\code{pkParametersToEvaluate}}{TO DO}

\item{\code{userDefinedPKFunctions}}{TO DO}

\item{\code{pkParameterResultsFilePath}}{TO DO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanModelSensitivityAnalysisSettings"></a>}}
\subsection{Method \code{meanModelSensitivityAnalysisSettings()}}{
Define mean Model Sensitivity Analysis `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$meanModelSensitivityAnalysisSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = FALSE,
  message = NULL,
  variationRange = 0.1,
  numberOfCores = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{variationRange}}{TO DO}

\item{\code{numberOfCores}}{TO DO}

\item{\code{inputFolderName}}{TO DO}

\item{\code{simulationFileName}}{TO DO}

\item{\code{resultsFolderName}}{TO DO}

\item{\code{resultsFileName}}{TO DO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGoFSettings"></a>}}
\subsection{Method \code{plotGoFSettings()}}{
Define Goodness of fit `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotGoFSettings(
  input = NULL,
  output = file.path(self$resultsFolder, "TimeProfile"),
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPKParametersSettings"></a>}}
\subsection{Method \code{plotPKParametersSettings()}}{
Define plot PK parameters `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotPKParametersSettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMassBalanceSettings"></a>}}
\subsection{Method \code{plotMassBalanceSettings()}}{
Define plot mass balance `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotMassBalanceSettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotAbsorptionSettings"></a>}}
\subsection{Method \code{plotAbsorptionSettings()}}{
Define plot absorption `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotAbsorptionSettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSensitivitySettings"></a>}}
\subsection{Method \code{plotSensitivitySettings()}}{
Define plot sensisitivity `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$plotSensitivitySettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\subsection{Method \code{runWorkflow()}}{
Run population workflow tasks for a single simulation set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in a folder specific to simulation set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print workflow list of tasks
TO DO: add simulationSets to print() method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Task list information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
