% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-model-workflow.R
\name{MeanModelWorkflow}
\alias{MeanModelWorkflow}
\title{MeanModelWorkflow}
\description{
R6 class for Reporting Engine Mean Model Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{MeanModelWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulate}}{`SimulationTask` object for time profile simulations}

\item{\code{calculatePKParameters}}{`CalculatePKParametersTask` object for PK parameters calculation}

\item{\code{calculateSensitivity}}{`SensitivityAnalysisTask` object for sensitivity analysis}

\item{\code{plotTimeProfilesAndResiduals}}{`PlotTask` object for goodness of fit plots}

\item{\code{plotMassBalance}}{`PlotTask` object for mass balance plot}

\item{\code{plotAbsorption}}{`PlotTask` object for absorption plot}

\item{\code{plotPKParameters}}{`PlotTask` object for PK parameters plot}

\item{\code{plotSensitivity}}{`PlotTask` object for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanModelWorkflow$new()}}
\item \href{#method-runWorkflow}{\code{MeanModelWorkflow$runWorkflow()}}
\item \href{#method-clone}{\code{MeanModelWorkflow$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-activateTasks}{\code{ospsuite.reportingengine::Workflow$activateTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getActiveTasks}{\code{ospsuite.reportingengine::Workflow$getActiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllPlotTasks}{\code{ospsuite.reportingengine::Workflow$getAllPlotTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllTasks}{\code{ospsuite.reportingengine::Workflow$getAllTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getInactiveTasks}{\code{ospsuite.reportingengine::Workflow$getInactiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getParameterDisplayPaths">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getParameterDisplayPaths}{\code{ospsuite.reportingengine::Workflow$getParameterDisplayPaths()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getWatermark">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getWatermark}{\code{ospsuite.reportingengine::Workflow$getWatermark()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-inactivateTasks}{\code{ospsuite.reportingengine::Workflow$inactivateTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="print">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-print}{\code{ospsuite.reportingengine::Workflow$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="printReportingEngineInfo">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-printReportingEngineInfo}{\code{ospsuite.reportingengine::Workflow$printReportingEngineInfo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setParameterDisplayPaths">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-setParameterDisplayPaths}{\code{ospsuite.reportingengine::Workflow$setParameterDisplayPaths()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setWatermark">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-setWatermark}{\code{ospsuite.reportingengine::Workflow$setWatermark()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `MeanModelWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\if{latex}{\out{\hypertarget{method-runWorkflow}{}}}
\subsection{Method \code{runWorkflow()}}{
Run mean model workflow tasks for all simulation sets if tasks are activated
The order of tasks is as follows:
# 1) Run simulations
# 2) Perform PK and sensitivity analyses
# 3) Perform plot tasks
## 3.a) time profiles and residual plots
## 3.b) absorption plots
## 3.c) mass balance plots
## 3.d) PK and sensitivity analyses tables and plots
# 4) Render report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in the workflow folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanModelWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
