% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{measureValuesBetween}
\alias{measureValuesBetween}
\title{measureValuesBetween}
\usage{
measureValuesBetween(x, left, right, method = "count", strict = FALSE)
}
\arguments{
\item{x}{Numeric values to assess}

\item{left}{Numeric value(s) used as lower bound}

\item{right}{Numeric value(s) used as upper bound}

\item{method}{One of the following methods \code{"count"}, \code{"ratio"}, and \code{"percent"}.}

\item{strict}{Logical value defining if \code{x} is strictly between \code{left} and \code{right}.
Default value is \code{FALSE}.}
}
\value{
Measure of \code{x} values between \code{left} and \code{right} bounds
}
\description{
Measure the values of \code{x} between \code{left} and \code{right} bounds according to \code{method}.
}
\examples{
measureValuesBetween(1:12, 7, 9)
measureValuesBetween(1:12, 7, 9, method = "percent")

x <- rnorm(1e2)
measureValuesBetween(x, -1, 1)
measureValuesBetween(x, -1, 1, method = "ratio")

measureValuesBetween(x, cos(x) + 1, cos(x) - 1)
}
