% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{mergeMarkdowndFiles}
\alias{mergeMarkdowndFiles}
\title{mergeMarkdowndFiles}
\usage{
mergeMarkdowndFiles(
  inputFiles,
  outputFile,
  logFolder = getwd(),
  keepInputFiles = FALSE
)
}
\arguments{
\item{inputFiles}{names of .md files to merge}

\item{outputFile}{name of merged .md file}

\item{logFolder}{folder where the logs are saved}

\item{keepInputFiles}{logical option to prevent the input files to be deleted after merging them}
}
\description{
Merge markdown files into one unique file
}
