% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{mergeMarkdownFiles}
\alias{mergeMarkdownFiles}
\title{mergeMarkdownFiles}
\usage{
mergeMarkdownFiles(inputFiles, outputFile, keepInputFiles = FALSE)
}
\arguments{
\item{inputFiles}{names of .md files to merge}

\item{outputFile}{name of merged .md file}

\item{keepInputFiles}{logical option to prevent the input files to be deleted after merging them}
}
\description{
Merge markdown files into one unique file
}
\examples{
\dontrun{
resetReport("chapter-1.md")
addTextChunk(fileName = "chapter-1.md", text = "Chapter 1")
resetReport("chapter-2.md")
addTextChunk(fileName = "chapter-2.md", text = "Chapter 2")
mergeMarkdownFiles(
  inputFiles = c("chapter-1.md", "chapter-2.md"),
  outputFile = "chapters-1and2.md"
)
}

}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{hasAnchor}()},
\code{\link{renderReport}()},
\code{\link{renderWordReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
