% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{Output}
\alias{Output}
\title{Output}
\description{
R6 class representing workflow output
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{displayName}}{display name for `path`}

\item{\code{displayUnit}}{display unit for `path`}

\item{\code{dataFilter}}{character or expression used to filter the observed data}

\item{\code{dataDisplayName}}{display name of the observed data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Output$new()}}
\item \href{#method-clone}{\code{Output$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `Output` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Output$new(
  path,
  displayName = NULL,
  displayUnit = NULL,
  dataFilter = NULL,
  dataDisplayName = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{displayName}}{display name for `path`}

\item{\code{displayUnit}}{display unit for `path`}

\item{\code{dataFilter}}{characters or expression to filter the observed data}

\item{\code{dataDisplayName}}{display name of the observed data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Output` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Output$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
