% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{Output}
\alias{Output}
\title{Output}
\description{
R6 class representing workflow output
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{displayName}}{display name for `path`}

\item{\code{displayUnit}}{display unit for `path`}

\item{\code{dataSelection}}{character or expression used to filter the observed data}

\item{\code{dataDisplayName}}{display name of the observed data}

\item{\code{pkParameters}}{R6 class `PkParameterInfo` objects}

\item{\code{residualScale}}{Scale for calculation of residuals as included in enum `ResidualScales`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Output$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Output` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Output$new(
  path,
  displayName = NULL,
  displayUnit = NULL,
  dataSelection = DataSelectionKeys$NONE,
  dataDisplayName = NULL,
  pkParameters = NULL,
  residualScale = ResidualScales$Logarithmic
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{displayName}}{display name for `path`}

\item{\code{displayUnit}}{display unit for `path`}

\item{\code{dataSelection}}{characters or expression to filter the observed data}

\item{\code{dataDisplayName}}{display name of the observed data}

\item{\code{pkParameters}}{R6 class `PkParameterInfo` objects or their names}

\item{\code{residualScale}}{Scale for calculation of residuals as included in enum `ResidualScales`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Output` object
}
}
}
