% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{Output}
\alias{Output}
\title{Output}
\description{
R6 class representing workflow output
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{displayName}}{display name for `path`}

\item{\code{displayUnit}}{display unit for `path`}

\item{\code{groupID}}{Grouping Identifier.
Outputs with same identifier and unit are plotted together}

\item{\code{color}}{Displayed line/point color of the Output in plots}

\item{\code{fill}}{Displayed range color of the Output in plots}

\item{\code{dataSelection}}{character or expression used to select a subset of observed data}

\item{\code{dataUnit}}{Unit of the observed data}

\item{\code{dataDisplayName}}{display name of the observed data}

\item{\code{pkParameters}}{R6 class `PkParameterInfo` objects}

\item{\code{residualScale}}{Scale for calculation of residuals as included in enum `ResidualScales`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Output-new}{\code{Output$new()}}
\item \href{#method-Output-clone}{\code{Output$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Output-new"></a>}}
\if{latex}{\out{\hypertarget{method-Output-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Output` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Output$new(
  path,
  displayName = NULL,
  displayUnit = NULL,
  groupID = NULL,
  color = NULL,
  fill = NULL,
  dataSelection = DataSelectionKeys$NONE,
  dataUnit = NULL,
  dataDisplayName = NULL,
  pkParameters = NULL,
  residualScale = ResidualScales$Logarithmic
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path name for the simulation (e.g. `Organism|PeripheralVenousBlood|Raltegravir|Plasma (Peripheral Venous Blood)`)}

\item{\code{displayName}}{display name for `path`}

\item{\code{displayUnit}}{display unit for `path`}

\item{\code{groupID}}{Grouping Identifier.
Outputs with same identifier and unit are plotted together}

\item{\code{color}}{Displayed line/point color of the `Output` in plots}

\item{\code{fill}}{Displayed range color of the `Output` in plots}

\item{\code{dataSelection}}{characters or expression to select subset the observed data
By default, no data is selected.
When using a character array, selections are concatenated with the `&` sign}

\item{\code{dataUnit}}{Unit of the observed data}

\item{\code{dataDisplayName}}{display name of the observed data}

\item{\code{pkParameters}}{R6 class `PkParameterInfo` objects or their names}

\item{\code{residualScale}}{Scale for calculation of residuals as included in enum `ResidualScales`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Output` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Output-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Output-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Output$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
