% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{plotDemographyHistogram}
\alias{plotDemographyHistogram}
\title{plotDemographyHistogram}
\usage{
plotDemographyHistogram(
  data,
  metaData,
  dataMapping = NULL,
  plotConfiguration = NULL,
  bins = AggregationConfiguration$bins,
  dodge = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{list of metaData about `data`}

\item{dataMapping}{`HistogramDataMapping` class object}

\item{plotConfiguration}{`PlotConfiguration` class object}

\item{bins}{Number of bins for continuous demography parameters}

\item{dodge}{For continuous demography parameters,
Logical defining if histogram bars should dodge for continuous parameters}
}
\value{
ggplot object
}
\description{
Plot histograms for demography parameters
}
