% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{plotDemographyRange}
\alias{plotDemographyRange}
\title{plotDemographyRange}
\usage{
plotDemographyRange(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  parameterClass = "numeric"
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{list of metaData about `data`}

\item{dataMapping}{A `TimeProfileDataMapping` or `BoxWhiskerDataMapping` object}

\item{plotConfiguration}{`PlotConfiguration`  object}

\item{parameterClass}{Class of the parameter, either "numeric" or "character"}
}
\value{
ggplot object
}
\description{
Plot demography range plot using visual predictive check style if data is numeric,
or using box whisker plot if data is categorical
}
