% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-demography.R
\name{plotDemographyRange}
\alias{plotDemographyRange}
\title{plotDemographyRange}
\usage{
plotDemographyRange(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{list of metaData about `data`}

\item{plotConfiguration}{`PlotConfiguration`  object}

\item{daraMapping}{A `TimeProfileDataMapping` object}
}
\value{
ggplot object
}
\description{
Plot demography range plot using visual predictive check style
}
