% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-absorption.R
\name{plotMeanAbsorption}
\alias{plotMeanAbsorption}
\title{plotMeanAbsorption}
\usage{
plotMeanAbsorption(structureSet, plotConfigurations = NULL)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{plotConfigurations}{list of `PlotConfiguration` R6 class objects}
}
\value{
`ggplot` object
}
\description{
Plot absorption diagnostics time profile
}
