% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotMeanGoodnessOfFit}
\alias{plotMeanGoodnessOfFit}
\title{plotMeanGoodnessOfFit}
\usage{
plotMeanGoodnessOfFit(structureSet, settings = NULL)
}
\arguments{
\item{structureSet}{\code{SimulationStructure} R6 class object}

\item{settings}{List of settings such as \code{PlotConfiguration} R6 class objects for each goodness of fit plot}
}
\value{
list with \code{plots}, \code{tables} and \code{residuals} objects to be saved
}
\description{
Plot goodness of fit diagnostics including time profiles,
observations vs predictions, residuals plots (residuals vs time, vs predictions, qq-plots and histogram)
}
\keyword{internal}
