% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mass-balance.R
\name{plotMeanMassBalance}
\alias{plotMeanMassBalance}
\title{plotMeanMassBalance}
\usage{
plotMeanMassBalance(structureSet, logFolder = getwd(), settings = NULL)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{Options to be defined
(e.g. how to group the plot results, or which molecule to exclude)}
}
\value{
list of `ggplot` objects
}
\description{
Plot mass balance diagnostics time profiles, cumulative time profiles,
normalized time profiles and cumulative normalized time profiles
}
\keyword{internal}
