% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-goodness-of-fit.R
\name{plotMeanTimeProfile}
\alias{plotMeanTimeProfile}
\title{plotMeanTimeProfile}
\usage{
plotMeanTimeProfile(
  simulatedData,
  observedData = NULL,
  lloqData = NULL,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL
)
}
\arguments{
\item{simulatedData}{data.frame of simulated data}

\item{observedData}{data.frame of observed data}

\item{lloqData}{data.frame of lloq data}

\item{metaData}{meta data on `data`}

\item{dataMapping}{`XYGDataMapping` R6 class object from `tlf` library}

\item{plotConfiguration}{`TimeProfilePlotConfiguration` R6 class object from `tlf` library}
}
\value{
ggplot object of time profile for mean model workflow
}
\description{
Plot time profile for mean model workflow
}
