% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plotPKParameters}
\alias{plotPKParameters}
\title{plotPKParameters}
\usage{
plotPKParameters()
}
\arguments{
\item{populationSimulation}{class object from runSimulation(simulation, population)}

\item{population}{population class object}

\item{observedData}{list of datasets corresponding to observations}

\item{parameterNames}{names of PK parameters to plot}

\item{plotConfiguration}{List of PlotConfiguration class objects for each plot}
}
\value{
pkParametersPlot ggplot object
}
\description{
Plot boxplot of the PK parameters
}
