% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pop-pk-parameters.R
\name{plotPopulationPKParameters}
\alias{plotPopulationPKParameters}
\title{plotPopulationPKParameters}
\usage{
plotPopulationPKParameters(
  structureSets,
  settings = NULL,
  workflowType = PopulationWorkflowTypes$parallelComparison,
  xParameters = getDefaultPkParametersXParameters(workflowType),
  yParameters = NULL
)
}
\arguments{
\item{structureSets}{`SimulationStructure` R6 class object}

\item{settings}{list of settings for the output table/plot}

\item{workflowType}{workflowType Type of population workflow.
Use enum `PopulationWorkflowTypes` to get list of workflow types.}

\item{xParameters}{list of parameters to be plotted along x axis}

\item{yParameters}{list of parameters to be plotted along y axis}
}
\value{
list of plots and tables with summary of PK parameters
}
\description{
Plot PK parameters box plots and tables
}
\keyword{internal}
