% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{plotPopulationSensitivity}
\alias{plotPopulationSensitivity}
\title{plotPopulationSensitivity}
\usage{
plotPopulationSensitivity(structureSet, logFolder = getwd(), settings)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{list of settings for the population sensitivity plot}
}
\value{
a structured list of plots for each possible combination of pathID output-pkParameter that is found in sensitivity results index file
}
\description{
Retrieve dataframe of ranked and filtered population sensitivity results for a given PK parameter and model output pathID
}
