% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{plotPopulationSensitivity}
\alias{plotPopulationSensitivity}
\title{plotPopulationSensitivity}
\usage{
plotPopulationSensitivity(
  structureSets,
  logFolder = NULL,
  settings,
  workflowType = NULL,
  xParameters = NULL,
  yParameters = NULL
)
}
\arguments{
\item{logFolder}{folder where the logs are saved}

\item{settings}{list of settings for the population sensitivity plot}

\item{structureSet}{`SimulationStructure` R6 class object}
}
\value{
a structured list of plots for each possible combination of pathID output-pkParameter that is found in sensitivity results index file
}
\description{
Retrieve list of plots of population sensitivity analyses across all populations
}
