% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-comparison-time-profile.R
\name{plotQualificationComparisonTimeProfile}
\alias{plotQualificationComparisonTimeProfile}
\title{plotQualificationComparisonTimeProfile}
\usage{
plotQualificationComparisonTimeProfile(configurationPlan, settings)
}
\arguments{
\item{configurationPlan}{A \code{ConfigurationPlan} object}

\item{settings}{\code{ConfigurationPlan} object}
}
\value{
list with \code{plots} and \code{tables}
}
\description{
Plot comparison time profile for qualification workflow
}
\keyword{internal}
