% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-gof.R
\name{plotQualificationGOFs}
\alias{plotQualificationGOFs}
\title{plotQualificationGOFs}
\usage{
plotQualificationGOFs(configurationPlan, settings)
}
\arguments{
\item{configurationPlan}{A `ConfigurationPlan` object}

\item{settings}{settings for the task}
}
\value{
list of qualification GOF ggplot objects
}
\description{
Plot observation vs prediction for qualification workflow
}
\keyword{internal}
