% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{plotQualificationMeanTimeProfile}
\alias{plotQualificationMeanTimeProfile}
\title{plotQualificationMeanTimeProfile}
\usage{
plotQualificationMeanTimeProfile(
  configurationPlanCurves,
  simulation,
  simulationResults,
  axesProperties,
  configurationPlan,
  plotConfiguration
)
}
\arguments{
\item{configurationPlanCurves}{`Curves` fields of configuration plan}

\item{simulation}{A `Simulation` object from `ospsuite` package
that includes required information to identify and convert the data requested from `configurationPlanCurve` properties}

\item{simulationResults}{A `SimulationResults` object from `ospsuite` package
that includes the data requested from `configurationPlanCurve` properties}

\item{axesProperties}{list of axes properties obtained from `getAxesProperties`}

\item{configurationPlan}{A `ConfigurationPlan` object that includes methods to find observed data}

\item{plotConfiguration}{A `TimeProfilePlotConfiguration` object}
}
\value{
Mean time profile plot as a `ggplot` object
}
\description{
Plot mean time profile for qualification workflow
}
\keyword{internal}
