% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-pk-ratio.R
\name{plotQualificationPKRatio}
\alias{plotQualificationPKRatio}
\title{plotQualificationPKRatio}
\usage{
plotQualificationPKRatio(configurationPlan, logFolder = getwd(), settings)
}
\arguments{
\item{configurationPlan}{A `ConfigurationPlan` object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{settings for the task}
}
\value{
list with `plots` and `tables`
}
\description{
Plot PK Ratio plots for qualification workflow
}
