% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{plotQualificationPopulationTimeProfile}
\alias{plotQualificationPopulationTimeProfile}
\title{plotQualificationPopulationTimeProfile}
\usage{
plotQualificationPopulationTimeProfile(
  simulationAnalysis,
  observedDataCollection,
  simulation,
  simulationResults,
  axesProperties,
  configurationPlan,
  plotObject
)
}
\arguments{
\item{simulationAnalysis}{Field `Analysis` from `ConfigurationPlan` population time profile plot}

\item{observedDataCollection}{Field `ObservedDataCollection` from `ConfigurationPlan` population time profile plot}

\item{simulation}{A `Simulation` object from `ospsuite` package
that includes required information to identify and convert the data requested from `configurationPlanCurve` properties}

\item{simulationResults}{A `SimulationResults` object from `ospsuite` package
that includes the data requested from `configurationPlanCurve` properties}

\item{axesProperties}{list of axes properties obtained from `getAxesProperties`}

\item{configurationPlan}{A `ConfigurationPlan` object that includes methods to find observed data}

\item{plotObject}{A `ggplot` object}
}
\value{
Population time profile plot as a `ggplot` object
}
\description{
Plot population time profile for qualification workflow
}
\keyword{internal}
