% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-time-profile.R
\name{plotQualificationTimeProfiles}
\alias{plotQualificationTimeProfiles}
\title{plotQualificationTimeProfiles}
\usage{
plotQualificationTimeProfiles(configurationPlan, logFolder = getwd(), settings)
}
\arguments{
\item{configurationPlan}{A `ConfigurationPlan` object}

\item{logFolder}{folder where the logs are saved}

\item{settings}{`ConfigurationPlan` object}
}
\value{
list with `plots` and `tables`
}
\description{
Plot time profile for qualification workflow
}
