% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{plotQualificationTimeProfiles}
\alias{plotQualificationTimeProfiles}
\title{plotQualificationTimeProfiles}
\usage{
plotQualificationTimeProfiles(configurationPlan, settings)
}
\arguments{
\item{configurationPlan}{A `ConfigurationPlan` object}

\item{settings}{`ConfigurationPlan` object}
}
\value{
list with `plots` and `tables`
}
\description{
Plot time profile for qualification workflow
}
\keyword{internal}
