% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-time-profile.R
\name{plotStatisticsFromPlan}
\alias{plotStatisticsFromPlan}
\title{plotStatisticsFromPlan}
\usage{
plotStatisticsFromPlan(
  time,
  outputValues,
  statisticId,
  outputName,
  color,
  linetype,
  plotObject
)
}
\arguments{
\item{time}{Time values on which output values are aggregated}

\item{outputValues}{Output values to be aggregated}

\item{statisticId}{Statistic Id as defined in `ConfigurationPlan` used for data aggregation}

\item{outputName}{Display name of output}

\item{color}{Color of the line or ribbon}

\item{linetype}{Linetype of the line}

\item{plotObject}{A `ggplot` object with previous statistics displayed}
}
\value{
A `ggplot` object updated with new displayed statistic
}
\description{
Add summary statistics to a time profile plot from configuration plan
}
\keyword{internal}
