% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-task.R
\name{PlotTask}
\alias{PlotTask}
\title{PlotTask}
\description{
R6 class for PlotTask settings
}
\seealso{
Other workflow tasks: 
\code{\link{GofPlotTask}},
\code{\link{PopulationPlotTask}},
\code{\link{PopulationSensitivityAnalysisTask}},
\code{\link{QualificationTask}},
\code{\link{SensitivityAnalysisTask}},
\code{\link{Task}},
\code{\link{addUserDefinedTask}()},
\code{\link{loadCalculatePKParametersTask}()},
\code{\link{loadCalculateSensitivityTask}()},
\code{\link{loadGOFMergedTask}()},
\code{\link{loadPlotAbsorptionTask}()},
\code{\link{loadPlotDDIRatioTask}()},
\code{\link{loadPlotDemographyTask}()},
\code{\link{loadPlotMassBalanceTask}()},
\code{\link{loadPlotPKParametersTask}()},
\code{\link{loadPlotPKRatioTask}()},
\code{\link{loadPlotSensitivityTask}()},
\code{\link{loadPlotTimeProfilesAndResidualsTask}()},
\code{\link{loadQualificationComparisonTimeProfileTask}()},
\code{\link{loadQualificationTimeProfilesTask}()},
\code{\link{loadSimulateTask}()}
}
\concept{workflow tasks}
\section{Super class}{
\code{ospsuite.reportingengine::Task} -> \code{PlotTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{section title in the report corresponding to the task}

\item{\code{reference}}{id of anchor tag referencing title}

\item{\code{fileName}}{name of report appendix file associated to task}

\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{nameTaskResults}}{name of the function that returns task results,}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotTask-new}{\code{PlotTask$new()}}
\item \href{#method-PlotTask-saveResults}{\code{PlotTask$saveResults()}}
\item \href{#method-PlotTask-runTask}{\code{PlotTask$runTask()}}
\item \href{#method-PlotTask-clone}{\code{PlotTask$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-activate'><code>ospsuite.reportingengine::Task$activate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getAbsolutePath"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getAbsolutePath'><code>ospsuite.reportingengine::Task$getAbsolutePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getInputs"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getInputs'><code>ospsuite.reportingengine::Task$getInputs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getRelativePath"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-getRelativePath'><code>ospsuite.reportingengine::Task$getRelativePath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-inactivate'><code>ospsuite.reportingengine::Task$inactivate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-print'><code>ospsuite.reportingengine::Task$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-validateInput'><code>ospsuite.reportingengine::Task$validateInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateStructureSetInput"><a href='../../ospsuite.reportingengine/html/Task.html#method-Task-validateStructureSetInput'><code>ospsuite.reportingengine::Task$validateStructureSetInput()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlotTask-new}{}}}
\subsection{Method \code{new()}}{
Create a `PlotTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$new(
  reportTitle = NULL,
  reportReference = NULL,
  fileName = NULL,
  getTaskResults = NULL,
  nameTaskResults = "none",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reportTitle}}{title to be printed in the report}

\item{\code{reportReference}}{id of anchor tag referencing title}

\item{\code{fileName}}{name of report appendix file associated to task}

\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{nameTaskResults}}{name of the function that returns task results,}

\item{\code{...}}{input parameters inherited from `Task` R6 class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PlotTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotTask-saveResults"></a>}}
\if{latex}{\out{\hypertarget{method-PlotTask-saveResults}{}}}
\subsection{Method \code{saveResults()}}{
Save the task results related to a `structureSet`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$saveResults(structureSet, taskResults)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{structureSet}}{A `SimulationStructure` object defining the properties of a simulation set}

\item{\code{taskResults}}{list of `TaskResults` objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotTask-runTask"></a>}}
\if{latex}{\out{\hypertarget{method-PlotTask-runTask}{}}}
\subsection{Method \code{runTask()}}{
Run task and save its output results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$runTask(structureSets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{structureSets}}{list of `SimulationStructure` objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
