% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-sensitivity-analysis-task.R
\name{PopulationSensitivityAnalysisTask}
\alias{PopulationSensitivityAnalysisTask}
\title{PopulationSensitivityAnalysisTask}
\description{
R6 class for PopulationSensitivityAnalysisTask settings
}
\section{Super classes}{
\code{\link[ospsuite.reportingengine:Task]{ospsuite.reportingengine::Task}} -> \code{\link[ospsuite.reportingengine:SensitivityAnalysisTask]{ospsuite.reportingengine::SensitivityAnalysisTask}} -> \code{PopulationSensitivityAnalysisTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{settings}}{instance of SensitivityAnalysisSettings class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationSensitivityAnalysisTask$new()}}
\item \href{#method-saveResults}{\code{PopulationSensitivityAnalysisTask$saveResults()}}
\item \href{#method-clone}{\code{PopulationSensitivityAnalysisTask$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate">}\href{../../ospsuite.reportingengine/html/Task.html#method-activate}{\code{ospsuite.reportingengine::Task$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="getInputs">}\href{../../ospsuite.reportingengine/html/Task.html#method-getInputs}{\code{ospsuite.reportingengine::Task$getInputs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Task.html#method-inactivate}{\code{ospsuite.reportingengine::Task$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print">}\href{../../ospsuite.reportingengine/html/Task.html#method-print}{\code{ospsuite.reportingengine::Task$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Task.html#method-validateInput}{\code{ospsuite.reportingengine::Task$validateInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateStructureSetInput">}\href{../../ospsuite.reportingengine/html/Task.html#method-validateStructureSetInput}{\code{ospsuite.reportingengine::Task$validateStructureSetInput()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="SensitivityAnalysisTask" data-id="runTask">}\href{../../ospsuite.reportingengine/html/SensitivityAnalysisTask.html#method-runTask}{\code{ospsuite.reportingengine::SensitivityAnalysisTask$runTask()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `PopulationSensitivityAnalysisTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisTask$new(
  getTaskResults = NULL,
  settings = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{getTaskResults}}{function called by task that computes and format figure results}

\item{\code{settings}}{instance of SensitivityAnalysisSettings class}

\item{\code{...}}{parameters inherited from R6 class `Task` object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationSensitivityAnalysisTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveResults"></a>}}
\subsection{Method \code{saveResults()}}{
Save results from task run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisTask$saveResults(set, taskResults)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set}}{R6 class `SimulationStructure`}

\item{\code{taskResults}}{list of results from task run.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSensitivityAnalysisTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
