% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-simulation-set.R
\name{PopulationSimulationSet}
\alias{PopulationSimulationSet}
\title{PopulationSimulationSet}
\description{
R6 class representing Reporting Engine Population Model Set
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:SimulationSet]{ospsuite.reportingengine::SimulationSet}} -> \code{PopulationSimulationSet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{referencePopulation}}{logical for reference population used in Pediatric and Ratio Comparison workflows}

\item{\code{populationFile}}{name of csv file to be used for the population}

\item{\code{populationName}}{display name of population}

\item{\code{studyDesignFile}}{name of study design csv file}

\item{\code{plotReferenceObsData}}{logical for plotting reference observed data in Pediatric and Ratio Comparison workflows}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PopulationSimulationSet-new}{\code{PopulationSimulationSet$new()}}
\item \href{#method-PopulationSimulationSet-copyInputFiles}{\code{PopulationSimulationSet$copyInputFiles()}}
\item \href{#method-PopulationSimulationSet-clone}{\code{PopulationSimulationSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationSimulationSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationSimulationSet-new}{}}}
\subsection{Method \code{new()}}{
Create a new `PopulationSimulationSet` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSet$new(
  referencePopulation = FALSE,
  simulationSetName,
  simulationFile,
  populationFile,
  populationName = NULL,
  studyDesignFile = NULL,
  plotReferenceObsData = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{referencePopulation}}{logical for reference population used in Pediatric and Ratio Comparison workflows}

\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{populationFile}}{name of csv file to be used for the population}

\item{\code{populationName}}{display name of population}

\item{\code{studyDesignFile}}{name of study design csv file}

\item{\code{plotReferenceObsData}}{logical for plotting reference observed data in Pediatric and Ratio Comparison workflows}

\item{\code{...}}{inputs inherited from `SimulationSet`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationSimulationSet` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationSimulationSet-copyInputFiles"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationSimulationSet-copyInputFiles}{}}}
\subsection{Method \code{copyInputFiles()}}{
Copy input files into a simulation set specific folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSet$copyInputFiles(inputFilesFolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputFilesFolder}}{where input are located}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationSimulationSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationSimulationSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulationSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
