% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R, R/workflow.R
\docType{class}
\name{PopulationWorkflow}
\alias{PopulationWorkflow}
\title{PopulationWorkflow}
\usage{
PopulationWorkflow

PopulationWorkflow
}
\description{
Population Workflow Task for Reporting Engine

Population Workflow Task for Reporting Engine
}
\section{Fields}{

\describe{
\item{\code{populationSimulation}}{Population simulation task}

\item{\code{pkParametersCalculation}}{PK parameters calculation task}

\item{\code{sensitivityAnalysis}}{Sensitivity analysis task}

\item{\code{demographyPlot}}{Plot demography task}

\item{\code{timeProfilePlot}}{Plot time profile task}

\item{\code{pkParametersPlot}}{Plot PK parameters task}

\item{\code{sensitivityPlot}}{Plot sensitiviy task}

\item{\code{populationSimulation}}{Population simulation task}

\item{\code{pkParametersCalculation}}{PK parameters calculation task}

\item{\code{sensitivityAnalysis}}{Sensitivity analysis task}

\item{\code{demographyPlot}}{Plot demography task}

\item{\code{timeProfilePlot}}{Plot time profile task}

\item{\code{pkParametersPlot}}{Plot PK parameters task}

\item{\code{sensitivityPlot}}{Plot sensitiviy task}
}}

\section{Methods}{

\describe{
\item{new()}{Initilialize reporting engine population workflow}
\item{setPopulationSimulationSettings()}{Define population simulation task settings}
\item{setPlotDemographySettings()}{Define demography plot settings}
\item{setPlotTimeProfileSettings()}{Define time profile plot settings}
\item{runWorkflow()}{Run the active tasks of population worklfow}
}


\describe{
\item{new()}{Initilialize reporting engine population workflow}
\item{setPopulationSimulationSettings()}{Define population simulation task settings}
\item{setPKParametersCalculationSettings()}{Define PK parameters calculation task settings}
\item{setSensitivityAnalysisSettings()}{Define sensitivity analysis task settings}
\item{setPlotDemographySettings()}{Define demography plots settings}
\item{setPlotTimeProfileSettings()}{Define goodness of fit plots settings}
\item{setPlotPKParametersSettings()}{Define PK parameters plot settings}
\item{setPlotSensitivitySettings()}{Define sensitivity plot settings}
\item{runWorkflow()}{Run the active tasks of population worklfow}
}
}

\keyword{datasets}
