% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R
\name{PopulationWorkflow}
\alias{PopulationWorkflow}
\title{PopulationWorkflow}
\description{
R6 class for Reporting Engine Population Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{PopulationWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{populationSimulation}}{R6 class `Task` for population simulation}

\item{\code{populationPKParameters}}{R6 class `Task` for PK parameters calculation}

\item{\code{calculateSensitivity}}{R6 class `Task` for sensitivity analysis}

\item{\code{plotDemography}}{R6 class `Task` for demography plots}

\item{\code{plotGoF}}{R6 class `Task` for goodness of fit plots}

\item{\code{plotPKParameters}}{R6 class `Task` for PK parameters plot}

\item{\code{plotSensitivity}}{R6 class `Task` for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationWorkflow$new()}}
\item \href{#method-populationSimulationSettings}{\code{PopulationWorkflow$populationSimulationSettings()}}
\item \href{#method-populationPKParameterSettings}{\code{PopulationWorkflow$populationPKParameterSettings()}}
\item \href{#method-plotDemographySettings}{\code{PopulationWorkflow$plotDemographySettings()}}
\item \href{#method-plotGoFSettings}{\code{PopulationWorkflow$plotGoFSettings()}}
\item \href{#method-plotPKParametersSettings}{\code{PopulationWorkflow$plotPKParametersSettings()}}
\item \href{#method-plotSensitivitySettings}{\code{PopulationWorkflow$plotSensitivitySettings()}}
\item \href{#method-runWorkflow}{\code{PopulationWorkflow$runWorkflow()}}
\item \href{#method-print}{\code{PopulationWorkflow$print()}}
\item \href{#method-clone}{\code{PopulationWorkflow$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `PopulationWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationSimulationSettings"></a>}}
\subsection{Method \code{populationSimulationSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationSimulationSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  inputFolderName = self$inputFolder,
  simulationFileName = self$simulation,
  populationFileName = self$population,
  resultsFolderName = self$simulationFolder,
  resultsFileName = "populationSimulation",
  numberOfCores = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationPKParameterSettings"></a>}}
\subsection{Method \code{populationPKParameterSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationPKParameterSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  simulationFilePath = file.path(self$inputFolder, paste0(self$simulation, ".pkml")),
  simulationResultFilePaths = self$populationSimulation$generatedResultFileNames,
  pkParametersToEvaluate = NULL,
  userDefinedPKFunctions = NULL,
  pkParameterResultsFilePath = file.path(self$pkParametersFolder,
    "populationPKParameters.csv")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDemographySettings"></a>}}
\subsection{Method \code{plotDemographySettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotDemographySettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGoFSettings"></a>}}
\subsection{Method \code{plotGoFSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotGoFSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPKParametersSettings"></a>}}
\subsection{Method \code{plotPKParametersSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotPKParametersSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSensitivitySettings"></a>}}
\subsection{Method \code{plotSensitivitySettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotSensitivitySettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\subsection{Method \code{runWorkflow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print workflow list of tasks
TO DO: add simulationSets to print() method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Task list information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
