% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R
\name{PopulationWorkflow}
\alias{PopulationWorkflow}
\title{PopulationWorkflow}
\description{
R6 class for Reporting Engine Population Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{PopulationWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulatePopulation}}{R6 class `Task` for population simulation}

\item{\code{populationPKParameters}}{R6 class `Task` for PK parameters calculation}

\item{\code{populationSensitivityAnalysis}}{R6 class `Task` for sensitivity analysis}

\item{\code{plotDemography}}{R6 class `Task` for demography plots}

\item{\code{plotGoF}}{R6 class `Task` for goodness of fit plots}

\item{\code{plotPKParameters}}{R6 class `Task` for PK parameters plot}

\item{\code{plotSensitivity}}{R6 class `Task` for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationWorkflow$new()}}
\item \href{#method-simulatePopulationSettings}{\code{PopulationWorkflow$simulatePopulationSettings()}}
\item \href{#method-populationPKParameterSettings}{\code{PopulationWorkflow$populationPKParameterSettings()}}
\item \href{#method-populationSensitivityAnalysisSettings}{\code{PopulationWorkflow$populationSensitivityAnalysisSettings()}}
\item \href{#method-plotDemographySettings}{\code{PopulationWorkflow$plotDemographySettings()}}
\item \href{#method-plotGoFSettings}{\code{PopulationWorkflow$plotGoFSettings()}}
\item \href{#method-plotPKParametersSettings}{\code{PopulationWorkflow$plotPKParametersSettings()}}
\item \href{#method-plotSensitivitySettings}{\code{PopulationWorkflow$plotSensitivitySettings()}}
\item \href{#method-runWorkflow}{\code{PopulationWorkflow$runWorkflow()}}
\item \href{#method-clone}{\code{PopulationWorkflow$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-activateTasks}{\code{ospsuite.reportingengine::Workflow$activateTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getActiveTasks}{\code{ospsuite.reportingengine::Workflow$getActiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllPlotTasks}{\code{ospsuite.reportingengine::Workflow$getAllPlotTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllTasks}{\code{ospsuite.reportingengine::Workflow$getAllTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getInactiveTasks}{\code{ospsuite.reportingengine::Workflow$getInactiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-inactivateTasks}{\code{ospsuite.reportingengine::Workflow$inactivateTasks()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `PopulationWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulatePopulationSettings"></a>}}
\subsection{Method \code{simulatePopulationSettings()}}{
Define simulate `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$simulatePopulationSettings(
  taskFunction = simulateModelForPopulation,
  outputFolder = defaultTaskOutputFolders$simulate,
  settings = NULL,
  active = TRUE,
  message = defaultWorkflowMessages$simulate
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{numberOfCores}}{number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationPKParameterSettings"></a>}}
\subsection{Method \code{populationPKParameterSettings()}}{
Define calculate PK parameters `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationPKParameterSettings(
  taskFunction,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{input}}{file or folder of input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationSensitivityAnalysisSettings"></a>}}
\subsection{Method \code{populationSensitivityAnalysisSettings()}}{
Define population Sensitivity Analysis `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationSensitivityAnalysisSettings(
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  variationRange = NULL,
  numberOfCores = NULL,
  quantileVec = NULL,
  variableParameterPaths = NULL,
  pkParameterSelection = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{variationRange}}{variation range for sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{quantileVec}}{vector of quantiles to be calculated}

\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}

\item{\code{pkParameterSelection}}{list of selected PK parameters for sensitivity analysis}

\item{\code{input}}{file or folder of input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SensitivityAnalysisTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDemographySettings"></a>}}
\subsection{Method \code{plotDemographySettings()}}{
Define plot demography `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotDemographySettings(
  taskFunction,
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGoFSettings"></a>}}
\subsection{Method \code{plotGoFSettings()}}{
Define plot goodness of fit `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotGoFSettings(
  taskFunction,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{input}}{file or folder of input}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPKParametersSettings"></a>}}
\subsection{Method \code{plotPKParametersSettings()}}{
Define plot PK parameters `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotPKParametersSettings(
  taskFunction,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{input}}{file or folder of input}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSensitivitySettings"></a>}}
\subsection{Method \code{plotSensitivitySettings()}}{
Define plot sensisitivity `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotSensitivitySettings(
  taskFunction,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{input}}{file or folder of input}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\subsection{Method \code{runWorkflow()}}{
Loop through all simulation sets and run active population model workflow tasks for each.
# POPULATION WORKFLOW
# CORE STAGE 0:  DEMOGRAPHY SUMMARY
# 0a - Demography Plots/Tables
# CORE STAGE 1:  SIMULATION
# 1a - Time profile Plot
# CORE STAGE 2:  CALCULATE PK PARAMETER
# 2a - PK parameter Plot
# CORE STAGE 3:  CALCULATE SENSITIVITY
# 3a - Plots and Tables based on sensitivity results
#' @description
#' Run population workflow tasks for a single simulation set
#' @param set R6 class `SimulationStructure` object
#' @return All results and plots as a structured output in a folder specific to simulation set
runSingleSetWorkflow = function(set) {
  logInfo(message = "Start of population workflow")

  if (self$simulatePopulation$active) {
    if (self$simulatePopulation$validateInput()) {
      # if (self$simulatePopulation$numberOfCores == 1) {
      #   logInfo(message = "Starting population simulation")
      #   createFolder(set$simulationResultsFolder)
      #   resultsFilePath <- file.path(set$simulationResultsFolder, defaultFileNames$simulationResultsFile(set$simulationSet$simulationSetName))
      #   simulateModel(
      #     simFilePath = file.path(set$inputFilesFolder, paste0(set$simulationSet$simulationName, ".pkml")),
      #     popDataFilePath = file.path(set$inputFilesFolder, paste0(set$simulationSet$populationName, ".csv")),
      #     resultsFilePath = resultsFilePath
      #   )
      #   set$simulationResultFileNames <- resultsFilePath
      #   logInfo(message = "Population simulation completed.")
      # }
      # else if (self$simulatePopulation$numberOfCores > 1) {
      #   logInfo(message = "Starting parallel population simulation")
      #   createFolder(set$simulationResultsFolder)
      #   set$simulationResultFileNames <- runParallelsimulatePopulation(
      #     numberOfCores = self$simulatePopulation$numberOfCores,
      #     inputFolderName = set$inputFilesFolder,
      #     simulationFileName = set$simulationSet$simulationName,
      #     populationFileName = set$simulationSet$populationName,
      #     resultsFolderName = set$simulationResultsFolder,
      #     resultsFileName = trimFileName(defaultFileNames$simulationResultsFile(set$simulationSet$simulationSetName), extension = "csv")
      #   )
      #   logInfo(message = "Parallel population simulation completed.")
      # }
    }
  }


  if (self$populationPKParameters$active) {
    if (self$populationPKParameters$validateInput()) {
      logInfo("Starting PK parameter calculation")
      createFolder(set$pkAnalysisResultsFolder)
      set$pkAnalysisResultsFileNames <- calculatePKParameters(
        simulationFilePath = file.path(set$inputFilesFolder, paste0(set$simulationSet$simulationName, ".pkml")),
        simulationResultFilePaths = set$simulationResultFileNames,
        pkParameterResultsFilePath = file.path(set$pkAnalysisResultsFolder, defaultFileNames$pkAnalysisResultsFile(set$simulationSet$simulationSetName))
      )
      logInfo("PK parameter calculation completed.")
    }
  }


  if (self$populationSensitivityAnalysis$active) {
    if (self$populationSensitivityAnalysis$validateInput()) {
      logInfo("Starting population sensitivity analysis")
      createFolder(set$sensitivityAnalysisResultsFolder)
      set$sensitivityAnalysisResultsFileNames <- runPopulationSensitivityAnalysis(
        simFilePath = file.path(set$inputFilesFolder, paste0(set$simulationSet$simulationName, ".pkml")),
        variableParameterPaths = self$populationSensitivityAnalysis$variableParameterPaths,
        popDataFilePath = file.path(set$inputFilesFolder, paste0(set$simulationSet$populationName, ".csv")),
        pkParameterResultsFilePath = file.path(set$pkAnalysisResultsFolder, defaultFileNames$pkAnalysisResultsFile(set$simulationSet$simulationSetName)),
        pkParameterSelection = self$populationSensitivityAnalysis$pkParameterSelection,
        resultsFileFolder = set$sensitivityAnalysisResultsFolder,
        resultsFileName = trimFileName(defaultFileNames$sensitivityAnalysisResultsFile(set$simulationSet$simulationSetName), extension = "csv"),
        variationRange = self$populationSensitivityAnalysis$variationRange,
        quantileVec = self$populationSensitivityAnalysis$quantileVec,
        numberOfCores = self$populationSensitivityAnalysis$numberOfCores
      )
      logInfo("Population sensitivity analysis completed.")
    }
  }

  # TO DO: plug plot tasks to actual results
  if (self$plotDemography$active) {
    logInfo(message = self$plotDemography$message)
  }
  if (self$plotGoF$active) {
    logInfo(message = self$plotGoF$message)
  }
  if (self$plotPKParameters$active) {
    logInfo(message = self$plotPKParameters$message)
  }
  if (self$plotSensitivity$active) {
    logInfo(message = self$plotSensitivity$message)
  }
},


#' @description
#' Print workflow list of tasks
#' TO DO: add simulationSets to print() method
#' @return Task list information
print = function() {
  taskOrder <- list(
    "Task 1" = self$simulatePopulation$print(),
    "Task 2" = self$pkParametersCalculation$print(),
    "Task 3" = self$sensitivityAnalysis$print(),
    "Task 4" = self$plotDemography$print(),
    "Task 5" = self$plotGoF$print(),
    "Task 6" = self$plotPKParameters$print(),
    "Task 7" = self$plotSensitivity$print()
  )
  invisible(self)

  return(taskOrder)
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
