% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R
\name{PopulationWorkflow}
\alias{PopulationWorkflow}
\title{PopulationWorkflow}
\description{
R6 class for Reporting Engine Population Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{PopulationWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{populationSimulation}}{R6 class `Task` for population simulation}

\item{\code{populationPKParameters}}{R6 class `Task` for PK parameters calculation}

\item{\code{calculateSensitivity}}{R6 class `Task` for sensitivity analysis}

\item{\code{plotDemography}}{R6 class `Task` for demography plots}

\item{\code{plotGoF}}{R6 class `Task` for goodness of fit plots}

\item{\code{plotPKParameters}}{R6 class `Task` for PK parameters plot}

\item{\code{plotSensitivity}}{R6 class `Task` for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{calculateSensitivity}}{R6 class `Task` for sensitivity analysis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationWorkflow$new()}}
\item \href{#method-populationSimulationSettings}{\code{PopulationWorkflow$populationSimulationSettings()}}
\item \href{#method-populationPKParameterSettings}{\code{PopulationWorkflow$populationPKParameterSettings()}}
\item \href{#method-populationSensitivityAnalysisSettings}{\code{PopulationWorkflow$populationSensitivityAnalysisSettings()}}
\item \href{#method-plotDemographySettings}{\code{PopulationWorkflow$plotDemographySettings()}}
\item \href{#method-plotGoFSettings}{\code{PopulationWorkflow$plotGoFSettings()}}
\item \href{#method-plotPKParametersSettings}{\code{PopulationWorkflow$plotPKParametersSettings()}}
\item \href{#method-plotSensitivitySettings}{\code{PopulationWorkflow$plotSensitivitySettings()}}
\item \href{#method-runWorkflow}{\code{PopulationWorkflow$runWorkflow()}}
\item \href{#method-runSingleSetWorkflow}{\code{PopulationWorkflow$runSingleSetWorkflow()}}
\item \href{#method-print}{\code{PopulationWorkflow$print()}}
\item \href{#method-clone}{\code{PopulationWorkflow$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-activateTasks}{\code{ospsuite.reportingengine::Workflow$activateTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getActiveTasks}{\code{ospsuite.reportingengine::Workflow$getActiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllPlotTasks}{\code{ospsuite.reportingengine::Workflow$getAllPlotTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllTasks}{\code{ospsuite.reportingengine::Workflow$getAllTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getInactiveTasks}{\code{ospsuite.reportingengine::Workflow$getInactiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-inactivateTasks}{\code{ospsuite.reportingengine::Workflow$inactivateTasks()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `PopulationWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PopulationWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationSimulationSettings"></a>}}
\subsection{Method \code{populationSimulationSettings()}}{
Define simulate `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationSimulationSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  numberOfCores = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{numberOfCores}}{number of cores for parallelization}

\item{\code{inputFolderName}}{folder of input}

\item{\code{simulationFileName}}{name of simulation file}

\item{\code{resultsFolderName}}{folder where output is saved}

\item{\code{resultsFileName}}{file where output is saved}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationPKParameterSettings"></a>}}
\subsection{Method \code{populationPKParameterSettings()}}{
Define calculate PK parameters `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationPKParameterSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{simulationFilePath}}{TO DO}

\item{\code{simulationResultFilePaths}}{TO DO}

\item{\code{pkParametersToEvaluate}}{TO DO}

\item{\code{userDefinedPKFunctions}}{TO DO}

\item{\code{pkParameterResultsFilePath}}{TO DO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populationSensitivityAnalysisSettings"></a>}}
\subsection{Method \code{populationSensitivityAnalysisSettings()}}{
Define population Sensitivity Analysis `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$populationSensitivityAnalysisSettings(
  input = NULL,
  output = NULL,
  settings = NULL,
  active = TRUE,
  message = NULL,
  variationRange = 0.1,
  numberOfCores = 1,
  quantileVec = c(0.05, 0.5, 0.95)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{settings}}{specific settings for task}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{variationRange}}{TO DO}

\item{\code{numberOfCores}}{TO DO}

\item{\code{inputFolderName}}{TO DO}

\item{\code{simulationFileName}}{TO DO}

\item{\code{resultsFolderName}}{TO DO}

\item{\code{resultsFileName}}{TO DO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDemographySettings"></a>}}
\subsection{Method \code{plotDemographySettings()}}{
Define plot demography `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotDemographySettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGoFSettings"></a>}}
\subsection{Method \code{plotGoFSettings()}}{
Define plot goodness of fit `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotGoFSettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPKParametersSettings"></a>}}
\subsection{Method \code{plotPKParametersSettings()}}{
Define plot PK parameters `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotPKParametersSettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSensitivitySettings"></a>}}
\subsection{Method \code{plotSensitivitySettings()}}{
Define plot sensisitivity `task` settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$plotSensitivitySettings(
  input = NULL,
  output = NULL,
  active = FALSE,
  message = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{file or folder of input}

\item{\code{output}}{file or folder of output}

\item{\code{active}}{logical indicating if `task` is performed in worklfow.
Default value is `TRUE`}

\item{\code{message}}{title of the `task`.
Default value indicates `task` name.}

\item{\code{settings}}{specific settings for task}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Task` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\subsection{Method \code{runWorkflow()}}{
Loop through all simulation sets and run active population model workflow tasks for each
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runSingleSetWorkflow"></a>}}
\subsection{Method \code{runSingleSetWorkflow()}}{
Run population workflow tasks for a single simulation set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$runSingleSetWorkflow(set)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in a folder specific to simulation set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print workflow list of tasks
TO DO: add simulationSets to print() method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Task list information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
