% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R
\name{PopulationWorkflow}
\alias{PopulationWorkflow}
\title{PopulationWorkflow}
\description{
R6 class for Reporting Engine Population Workflow
}
\seealso{
Other workflows: 
\code{\link{MeanModelWorkflow}},
\code{\link{QualificationWorkflow}}
}
\concept{workflows}
\section{Super class}{
\code{ospsuite.reportingengine::Workflow} -> \code{PopulationWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{workflowType}}{Type of population workflow}

\item{\code{simulate}}{\code{SimulationTask} object for time profile simulations}

\item{\code{calculatePKParameters}}{\code{CalculatePKParametersTask} object for PK parameters calculation}

\item{\code{calculateSensitivity}}{\code{SensitivityAnalysisTask} object for sensitivity analysis}

\item{\code{plotDemography}}{R6 class \code{Task} for demography plots}

\item{\code{plotTimeProfilesAndResiduals}}{\code{PlotTask} object for goodness of fit plots}

\item{\code{plotPKParameters}}{R6 class \code{Task} for PK parameters plot}

\item{\code{plotSensitivity}}{R6 class \code{Task} for sensitivity plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PopulationWorkflow-new}{\code{PopulationWorkflow$new()}}
\item \href{#method-PopulationWorkflow-runWorkflow}{\code{PopulationWorkflow$runWorkflow()}}
\item \href{#method-PopulationWorkflow-clone}{\code{PopulationWorkflow$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-activateTasks'><code>ospsuite.reportingengine::Workflow$activateTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getActiveTasks'><code>ospsuite.reportingengine::Workflow$getActiveTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getAllPlotTasks'><code>ospsuite.reportingengine::Workflow$getAllPlotTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getAllTasks'><code>ospsuite.reportingengine::Workflow$getAllTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getInactiveTasks'><code>ospsuite.reportingengine::Workflow$getInactiveTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getParameterDisplayPaths"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getParameterDisplayPaths'><code>ospsuite.reportingengine::Workflow$getParameterDisplayPaths()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getSimulationDescriptor"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getSimulationDescriptor'><code>ospsuite.reportingengine::Workflow$getSimulationDescriptor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getWatermark"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-getWatermark'><code>ospsuite.reportingengine::Workflow$getWatermark()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-inactivateTasks'><code>ospsuite.reportingengine::Workflow$inactivateTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="print"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-print'><code>ospsuite.reportingengine::Workflow$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="printReportingEngineInfo"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-printReportingEngineInfo'><code>ospsuite.reportingengine::Workflow$printReportingEngineInfo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setParameterDisplayPaths"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setParameterDisplayPaths'><code>ospsuite.reportingengine::Workflow$setParameterDisplayPaths()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setSimulationDescriptor"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setSimulationDescriptor'><code>ospsuite.reportingengine::Workflow$setSimulationDescriptor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setWatermark"><a href='../../ospsuite.reportingengine/html/Workflow.html#method-Workflow-setWatermark'><code>ospsuite.reportingengine::Workflow$setWatermark()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationWorkflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationWorkflow-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PopulationWorkflow} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$new(
  workflowType,
  simulationSets,
  workflowFolder,
  createWordReport = TRUE,
  watermark = NULL,
  simulationSetDescriptor = NULL,
  numberSections = TRUE,
  reportTitle = NULL,
  theme = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workflowType}}{Type of population workflow. Use enum \code{PopulationWorkflowTypes} to get list of workflow types.}

\item{\code{simulationSets}}{list of \code{SimulationSet} R6 class objects}

\item{\code{workflowFolder}}{path of the output folder created or used by the Workflow.}

\item{\code{createWordReport}}{logical of option for creating Markdown-Report only but not a Word-Report.}

\item{\code{watermark}}{displayed watermark in every plot background}

\item{\code{simulationSetDescriptor}}{character Descriptor of simulation sets indicated in reports}

\item{\code{numberSections}}{logical defining if the report sections should be numbered}

\item{\code{reportTitle}}{report title internally added as a cover page
If \code{reportTitle} is an existing file, it will be merged to the report as cover page.}

\item{\code{theme}}{A \code{Theme} object from \code{{tlf}} package}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PopulationWorkflow} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationWorkflow-runWorkflow"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationWorkflow-runWorkflow}{}}}
\subsection{Method \code{runWorkflow()}}{
Run population workflow tasks for all simulation sets if tasks are activated
The order of tasks is as follows:
\enumerate{
\item Run simulations
\item Perform PK and sensitivity analyses
\item Perform plot tasks\if{html}{\out{<br>}}
a. time profiles and residual plots\if{html}{\out{<br>}}
b. demography plots\if{html}{\out{<br>}}
c. PK and sensitivity analyses tables and plots\if{html}{\out{<br>}}
\item Render report
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in the workflow folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationWorkflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationWorkflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
