% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-workflow.R
\docType{data}
\name{PopulationWorkflowTypes}
\alias{PopulationWorkflowTypes}
\title{PopulationWorkflowTypes}
\format{
An object of class \code{list} of length 3.
}
\usage{
PopulationWorkflowTypes
}
\description{
List of population workflow available types
}
\examples{

# Lists available Population Workflow types
PopulationWorkflowTypes

}
\seealso{
Other enum helpers: 
\code{\link{AllAvailableTasks}},
\code{\link{ApplicationRanges}},
\code{\link{DataSelectionKeys}},
\code{\link{ResidualScales}},
\code{\link{StandardPlotTasks}},
\code{\link{StandardSimulationTasks}},
\code{\link{StatisticsTypes}},
\code{\link{tlfSettingsNames}}
}
\concept{enum helpers}
\keyword{datasets}
