% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{prettyCaption}
\alias{prettyCaption}
\title{prettyCaption}
\usage{
prettyCaption(captions, plotObject, element = "legend")
}
\arguments{
\item{captions}{Array of character strings to render}

\item{plotObject}{A \code{ggplot} object}

\item{element}{The name of element to which the line break should be added.
If applied to the legend, use \code{"legend"}.
If applied to a plot label use e.g. \code{"ylabel"} or \code{"title"}.
If applied to tick labels use \code{"yticks"} or \code{"yticklabels"}.}
}
\value{
A character vector of wrapped strings with line breaks at sensible places.
}
\description{
Get prettied captions with line breaks to prevent cropping of long captions
}
