% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-version-info.R
\name{QualificationVersionInfo}
\alias{QualificationVersionInfo}
\title{QualificationVersionInfo}
\description{
Object defining Qualification Version Information to be displayed on title page
}
\concept{qualification worfklow}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{qualificationPlanRelease}}{Qualification Plan Release Version (string, e.g. `"1.1"`)}

\item{\code{osp}}{OSP version (string, e.g. `"10.0"`)}

\item{\code{qualificationFramework}}{Qualification Framework version (string, e.g. `"3.0"`)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QualificationVersionInfo-new}{\code{QualificationVersionInfo$new()}}
\item \href{#method-QualificationVersionInfo-print}{\code{QualificationVersionInfo$print()}}
\item \href{#method-QualificationVersionInfo-updateText}{\code{QualificationVersionInfo$updateText()}}
\item \href{#method-QualificationVersionInfo-getQualificationPlanReleasePattern}{\code{QualificationVersionInfo$getQualificationPlanReleasePattern()}}
\item \href{#method-QualificationVersionInfo-getOSPPattern}{\code{QualificationVersionInfo$getOSPPattern()}}
\item \href{#method-QualificationVersionInfo-getQualificationFrameworkPattern}{\code{QualificationVersionInfo$getQualificationFrameworkPattern()}}
\item \href{#method-QualificationVersionInfo-setQualificationPlanReleasePattern}{\code{QualificationVersionInfo$setQualificationPlanReleasePattern()}}
\item \href{#method-QualificationVersionInfo-setOSPPattern}{\code{QualificationVersionInfo$setOSPPattern()}}
\item \href{#method-QualificationVersionInfo-setQualificationFrameworkPattern}{\code{QualificationVersionInfo$setQualificationFrameworkPattern()}}
\item \href{#method-QualificationVersionInfo-clone}{\code{QualificationVersionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-new}{}}}
\subsection{Method \code{new()}}{
Create a new `QualificationVersionInfo` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$new(
  qualificationPlanRelease = NULL,
  osp = NULL,
  qualificationFramework = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{qualificationPlanRelease}}{Qualification Plan Release Version (string, e.g. `"1.1"`)}

\item{\code{osp}}{OSP version (string, e.g. `"10.0"`)}

\item{\code{qualificationFramework}}{Qualification Framework version (string, e.g. `"3.0"`)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `QualificationVersionInfo` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-print}{}}}
\subsection{Method \code{print()}}{
Print version information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A text with system information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-updateText"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-updateText}{}}}
\subsection{Method \code{updateText()}}{
Update `text` if patterns of version info are found
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$updateText(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{character string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-getQualificationPlanReleasePattern"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-getQualificationPlanReleasePattern}{}}}
\subsection{Method \code{getQualificationPlanReleasePattern()}}{
Get the Qualification Plan Release Version pattern to be replaced in title page
**Caution, the function `gsub` is used to replace the pattern**
**Escape characters may be included in the pattern**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$getQualificationPlanReleasePattern()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-getOSPPattern"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-getOSPPattern}{}}}
\subsection{Method \code{getOSPPattern()}}{
Get the OSP version pattern to be replaced in title page
**Caution, the function `gsub` is used to replace the pattern**
**Escape characters may be included in the pattern**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$getOSPPattern()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-getQualificationFrameworkPattern"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-getQualificationFrameworkPattern}{}}}
\subsection{Method \code{getQualificationFrameworkPattern()}}{
Get the Qualification Framework version pattern to be replaced in title page
**Caution, the function `gsub` is used to replace the pattern**
**Escape characters may be included in the pattern**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$getQualificationFrameworkPattern()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-setQualificationPlanReleasePattern"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-setQualificationPlanReleasePattern}{}}}
\subsection{Method \code{setQualificationPlanReleasePattern()}}{
Set the Qualification Plan Release Version pattern to be replaced in title page
**Caution, the function `gsub` is used to replace the pattern.**
**You may need to include escape characters in `pattern`**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$setQualificationPlanReleasePattern(pattern)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{characters to be replaced in title page}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-setOSPPattern"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-setOSPPattern}{}}}
\subsection{Method \code{setOSPPattern()}}{
Set the OSP version pattern to be replaced in title page
**Caution, the function `gsub` is used to replace the pattern.**
**You may need to include escape characters in `pattern`**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$setOSPPattern(pattern)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{characters to be replaced in title page}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-setQualificationFrameworkPattern"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-setQualificationFrameworkPattern}{}}}
\subsection{Method \code{setQualificationFrameworkPattern()}}{
Set the Qualification Framework version pattern to be replaced in title page
**Caution, the function `gsub` is used to replace the pattern.**
**You may need to include escape characters in `pattern`**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$setQualificationFrameworkPattern(pattern)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{characters to be replaced in title page}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QualificationVersionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QualificationVersionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationVersionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
