% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification-workflow.R
\name{QualificationWorkflow}
\alias{QualificationWorkflow}
\title{QualificationWorkflow}
\description{
R6 class for Reporting Engine Qualification Workflow
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Workflow]{ospsuite.reportingengine::Workflow}} -> \code{QualificationWorkflow}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configurationPlan}}{`ConfigurationPlan` object}

\item{\code{simulate}}{`SimulationTask` object for time profile simulations}

\item{\code{calculatePKParameters}}{`CalculatePKParametersTask` object for PK parameters calculation}

\item{\code{plotTimeProfiles}}{`PlotTask` object for time profile plots}

\item{\code{plotComparisonTimeProfiles}}{`PlotTask` object for comparison of time profiles plots}

\item{\code{plotObsVsPred}}{`PlotTask` object for goodness of fit plots}

\item{\code{plotPKRatio}}{`PlotTask` object for PK ratio plot}

\item{\code{plotDDIRatio}}{`PlotTask` object for DDI ratio plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QualificationWorkflow$new()}}
\item \href{#method-runWorkflow}{\code{QualificationWorkflow$runWorkflow()}}
\item \href{#method-clone}{\code{QualificationWorkflow$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="activateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-activateTasks}{\code{ospsuite.reportingengine::Workflow$activateTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getActiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getActiveTasks}{\code{ospsuite.reportingengine::Workflow$getActiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllPlotTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllPlotTasks}{\code{ospsuite.reportingengine::Workflow$getAllPlotTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getAllTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getAllTasks}{\code{ospsuite.reportingengine::Workflow$getAllTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getInactiveTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getInactiveTasks}{\code{ospsuite.reportingengine::Workflow$getInactiveTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getParameterDisplayPaths">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getParameterDisplayPaths}{\code{ospsuite.reportingengine::Workflow$getParameterDisplayPaths()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getSimulationDescriptor">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getSimulationDescriptor}{\code{ospsuite.reportingengine::Workflow$getSimulationDescriptor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="getWatermark">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-getWatermark}{\code{ospsuite.reportingengine::Workflow$getWatermark()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="inactivateTasks">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-inactivateTasks}{\code{ospsuite.reportingengine::Workflow$inactivateTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="print">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-print}{\code{ospsuite.reportingengine::Workflow$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="printReportingEngineInfo">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-printReportingEngineInfo}{\code{ospsuite.reportingengine::Workflow$printReportingEngineInfo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setParameterDisplayPaths">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-setParameterDisplayPaths}{\code{ospsuite.reportingengine::Workflow$setParameterDisplayPaths()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setSimulationDescriptor">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-setSimulationDescriptor}{\code{ospsuite.reportingengine::Workflow$setSimulationDescriptor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Workflow" data-id="setWatermark">}\href{../../ospsuite.reportingengine/html/Workflow.html#method-setWatermark}{\code{ospsuite.reportingengine::Workflow$setWatermark()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `QualificationWorkflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input parameters inherited from R6 class object `Workflow`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MeanModelWorkflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runWorkflow"></a>}}
\if{latex}{\out{\hypertarget{method-runWorkflow}{}}}
\subsection{Method \code{runWorkflow()}}{
Run qualification workflow tasks for all simulation sets if tasks are activated
The order of tasks is as follows:
# 1) Run simulations
# 2) Perform PK analyses
# 3) Perform plot tasks
## 3.a) time profiles and residual plots
## 3.b) comparison time profiles plots
## 3.c) PK ratio tables and plots
## 3.d) DDI ratio tables and plots
# 4) Render report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$runWorkflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
All results and plots as a structured output in the workflow folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QualificationWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
