% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{renderReport}
\alias{renderReport}
\title{renderReport}
\usage{
renderReport(
  fileName,
  logFolder = getwd(),
  createWordReport = FALSE,
  numberSections = TRUE,
  intro = NULL
)
}
\arguments{
\item{fileName}{name of .md file to render}

\item{logFolder}{folder where the logs are saved}

\item{createWordReport}{option for creating Markdwon-Report only but not a Word-Report}

\item{numberSections}{logical defining if sections are numbered}

\item{intro}{name of .md file that include introduction (before toc)}
}
\description{
Render report with number sections and table of content
}
