% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{renderWordReport}
\alias{renderWordReport}
\title{renderWordReport}
\usage{
renderWordReport(
  fileName,
  intro = NULL,
  logFolder = getwd(),
  createWordReport = FALSE,
  wordConversionTemplate = NULL
)
}
\arguments{
\item{fileName}{name of .md file to render}

\item{intro}{name of .md file that include introduction (before toc)}

\item{logFolder}{folder where the logs are saved}

\item{createWordReport}{option for creating Markdwon-Report only but not a Word-Report}

\item{wordConversionTemplate}{optional docx template for rendering a tuned Word-Report document}
}
\description{
Render docx report with number sections and table of content
}
