% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{renderWordReport}
\alias{renderWordReport}
\title{renderWordReport}
\usage{
renderWordReport(
  fileName,
  intro = NULL,
  createWordReport = FALSE,
  wordConversionTemplate = NULL
)
}
\arguments{
\item{fileName}{name of .md file to render}

\item{intro}{name of .md file that include introduction (before toc)}

\item{createWordReport}{option for creating Markdown-Report only but not a Word-Report}

\item{wordConversionTemplate}{optional docx template for rendering a tuned Word-Report document}
}
\description{
Render docx report with number sections and table of content
}
\seealso{
Other reporting: 
\code{\link{addFigureChunk}()},
\code{\link{addTableChunk}()},
\code{\link{addTextChunk}()},
\code{\link{anchor}()},
\code{\link{copyReport}()},
\code{\link{getAnchorName}()},
\code{\link{getFigurePathsFromReport}()},
\code{\link{hasAnchor}()},
\code{\link{mergeMarkdownFiles}()},
\code{\link{renderReport}()},
\code{\link{resetReport}()}
}
\concept{reporting}
