% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-writing-report.R
\name{renderWordReport}
\alias{renderWordReport}
\title{renderWordReport}
\usage{
renderWordReport(fileName, logFolder = getwd())
}
\arguments{
\item{fileName}{name of .md file to render}

\item{logFolder}{folder where the logs are saved}
}
\description{
Render docx report with number sections and table of content
}
