% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replaceInfWithNA}
\alias{replaceInfWithNA}
\title{replaceInfWithNA}
\usage{
replaceInfWithNA(data, logFolder = getwd())
}
\arguments{
\item{data}{numeric vector}

\item{logFolder}{folder where the logs are saved}
}
\value{
numeric vector
}
\description{
replaceInfWithNA
}
\keyword{internal}
