% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-info.R
\name{ReportingEngineInfo}
\alias{ReportingEngineInfo}
\title{ReportingEngineInfo}
\description{
R6 class representing system information for Reporting Engine
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Date}}{Date at which the class is initialized}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportingEngineInfo$new()}}
\item \href{#method-print}{\code{ReportingEngineInfo$print()}}
\item \href{#method-clone}{\code{ReportingEngineInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `ReportingEngineInfo` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportingEngineInfo$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `ReportingEngineInfo` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print system information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportingEngineInfo$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A text with system information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportingEngineInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
