% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{data}
\name{ResidualScales}
\alias{ResidualScales}
\title{ResidualScales}
\format{
An object of class \code{list} of length 2.
}
\usage{
ResidualScales
}
\description{
List of available scales for calculation of residuals
}
\keyword{datasets}
