% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulation.R
\name{runParallelPopulationSimulation}
\alias{runParallelPopulationSimulation}
\title{runParallelPopulationSimulation}
\usage{
runParallelPopulationSimulation(numberOfCores, workingDirectory,
  inputFolder, outputFolder, simFileName, popFileName, resultsFileName)
}
\value{
Simulation results for population
}
\description{
Spawn cores, divide population among cores, run population simulation on cores, save results as CSV.
}
