% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{runParallelPopulationSimulation}
\alias{runParallelPopulationSimulation}
\title{runParallelPopulationSimulation}
\usage{
runParallelPopulationSimulation(
  structureSet,
  numberOfCores,
  settings,
  logFolder
)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object contain paths of files to be used}

\item{numberOfCores}{number of cores do be used by the parallel simulation}

\item{settings}{list of options to be passed on the function}

\item{logFolder}{folder where the logs are saved}
}
\value{
Simulation results for population
}
\description{
Spawn cores, divide population among cores, run population simulation on cores, save results as CSV.
}
