% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sensitivity-analysis.R
\name{runParallelSensitivityAnalysis}
\alias{runParallelSensitivityAnalysis}
\title{runParallelSensitivityAnalysis}
\usage{
runParallelSensitivityAnalysis(simFilePath, parametersToPerturb,
  individualParameters, numberOfCores, resultsFileFolder, resultsFileName)
}
\value{
Simulation results for population
}
\description{
Spawn cores, divide parameters among cores, run sensitivity analysis on cores, save results as CSV.
}
