% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runParallelSensitivityAnalysis}
\alias{runParallelSensitivityAnalysis}
\title{runParallelSensitivityAnalysis}
\usage{
runParallelSensitivityAnalysis(
  structureSet,
  settings = settings,
  individualParameters,
  logFolder = getwd()
)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{settings}{list of settings for the sensitivity analysis}

\item{individualParameters}{is an object storing an individual's parameters, obtained
from a population object's `getParameterValuesForIndividual()`` function.}

\item{logFolder}{folder where the logs are saved}
}
\value{
SA results for population
}
\description{
Spawn cores, divide parameters among cores, run sensitivity analysis on cores
for a single individual, save results as CSV.
}
\keyword{internal}
