% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runPopulationSensitivityAnalysis}
\alias{runPopulationSensitivityAnalysis}
\title{runPopulationSensitivityAnalysis}
\usage{
runPopulationSensitivityAnalysis(structureSet, settings, logFolder = getwd())
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object}

\item{settings}{list of settings for the population sensitivity analysis}

\item{logFolder}{folder where the logs are saved}
}
\description{
runPopulationSensitivityAnalysis
}
