% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity-analysis.R
\name{runPopulationSensitivityAnalysis}
\alias{runPopulationSensitivityAnalysis}
\title{runPopulationSensitivityAnalysis}
\usage{
runPopulationSensitivityAnalysis(
  simFilePath,
  variableParameterPaths = NULL,
  popDataFilePath,
  pkParameterResultsFilePath,
  pkParameterSelection = NULL,
  resultsFileFolder,
  resultsFileName,
  popSAResultsIndexFile = "sensitivityAnalysesResultsIndexFile",
  variationRange,
  quantileVec,
  numberOfCores,
  logFolder = getwd()
)
}
\arguments{
\item{simFilePath}{path to simulation file}

\item{popDataFilePath}{path to population file}

\item{pkParameterResultsFilePath}{path to pk parameter results CSV}

\item{resultsFileFolder}{folder where results of sensitivity analysis will be saved}

\item{resultsFileName}{root name of sensitivity analysis results CSV files}

\item{popSAResultsIndexFile}{Names of CSV file containing index of sensitivity analysis CSV files}

\item{variationRange}{variation range for sensitivity analysis}

\item{quantileVec}{vector of quantiles in (0,1).  For each output and pk parameter,
there will be a distribution of pk parameter values for the population.
The individuals yielding pk parameters closest to these quantiles will be selected for sensitivity analysis.}

\item{numberOfCores}{the number of cores to be used for parallelization of the sensitivity analysis.
Default is 1 core (no parallelization).}
}
\description{
runPopulationSensitivityAnalysis
}
