% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{saveRESettings}
\alias{saveRESettings}
\title{saveRESettings
Save the current reporting engine global settings in a \code{.RData} file}
\usage{
saveRESettings(file)
}
\arguments{
\item{file}{\code{.RData} file containing the settings}
}
\description{
saveRESettings
Save the current reporting engine global settings in a \code{.RData} file
}
