% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-analysis-task.R
\name{SensitivityAnalysisTask}
\alias{SensitivityAnalysisTask}
\title{SensitivityAnalysisTask}
\description{
R6 class for SensitivityAnalysisTask settings
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Task]{ospsuite.reportingengine::Task}} -> \code{SensitivityAnalysisTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{variationRange}}{variation range for sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{generatedResultFileNames}}{name of files where PK parameters are saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensitivityAnalysisTask$new()}}
\item \href{#method-clone}{\code{SensitivityAnalysisTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate">}\href{../../ospsuite.reportingengine/html/Task.html#method-activate}{\code{ospsuite.reportingengine::Task$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Task.html#method-inactivate}{\code{ospsuite.reportingengine::Task$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print">}\href{../../ospsuite.reportingengine/html/Task.html#method-print}{\code{ospsuite.reportingengine::Task$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Task.html#method-validateInput}{\code{ospsuite.reportingengine::Task$validateInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `SensitivityAnalysisTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityAnalysisTask$new(
  variationRange = NULL,
  numberOfCores = NULL,
  quantileVec = NULL,
  variableParameterPaths = NULL,
  pkParameterSelection = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variationRange}}{variation range for sensitivity analysis}

\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{variableParameterPaths}}{vector of paths of parameters to vary when performing sensitivity analysis}

\item{\code{...}}{parameters inherited from R6 class `Task` object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SensitivityAnalysisTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityAnalysisTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
