% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity-plot-settings.R
\name{SensitivityPlotSettings}
\alias{SensitivityPlotSettings}
\title{SensitivityPlotSettings}
\description{
R6 class for sensitivity analysis plot settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{totalSensitivityThreshold}}{cut-off used for plots of the most sensitive parameters}

\item{\code{maximalParametersPerSensitivityPlot}}{is the maximal number of parameters to display in a sensitivity plot}

\item{\code{plotConfiguration}}{`PlotConfiguration` R6 class object from `tlf` library}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensitivityPlotSettings$new()}}
\item \href{#method-clone}{\code{SensitivityPlotSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `SensitivityPlotSettings` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityPlotSettings$new(
  totalSensitivityThreshold = NULL,
  variableParameterPaths = NULL,
  maximalParametersPerSensitivityPlot = 50,
  plotConfiguration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{totalSensitivityThreshold}}{cut-off used for plots of the most sensitive parameters}

\item{\code{variableParameterPaths}}{paths that were varied in the sensitivity analysis.  If supplied totalSensitivityThreshold = 1, else 0.9.}

\item{\code{maximalParametersPerSensitivityPlot}}{maximalParametersPerSensitivityPlot is the maximal number of parameters to display in a sensitivity plot}

\item{\code{plotConfiguration}}{`PlotConfiguration` R6 class object from `tlf` library}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SensitivityPlotSettings` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensitivityPlotSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
