% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reportingengine-env.R
\name{setPlotFormat}
\alias{setPlotFormat}
\title{setPlotFormat}
\usage{
setPlotFormat(
  format = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  dpi = NULL
)
}
\arguments{
\item{format}{file format of the exported plots. E.g. "png" or "pdf"}

\item{width}{plot width in `unit`}

\item{height}{plot height in `unit`}

\item{units}{units of `width` and `height` included in "in", "cm", "mm", or "px"}

\item{dpi}{Plot resolution in dots per inch. Caution, font sizes depend on resolution.}
}
\description{
Set default plot format to be exported by reporting engine
}
