% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{setQuadraticDimension}
\alias{setQuadraticDimension}
\title{setQuadraticDimension}
\usage{
setQuadraticDimension(plotObject, plotConfiguration = NULL)
}
\arguments{
\item{plotObject}{A `ggplot` object}

\item{plotConfiguration}{`PlotConfiguration` object defined in task settings}
}
\value{
A `ggplot` object
}
\description{
Set quadratic dimensions if plot configuration is not user-defined
}
\keyword{internal}
