% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{setWorkflowParameterDisplayPaths}
\alias{setWorkflowParameterDisplayPaths}
\title{setWorkflowParameterDisplayPaths}
\usage{
setWorkflowParameterDisplayPaths(parameterDisplayPaths, workflow)
}
\arguments{
\item{parameterDisplayPaths}{data.frame mapping Parameters with their display paths
Variables of the data.frame should include `parameter` and `displayPath`.}

\item{workflow}{Object of class `MeanModelWorkflow` or `PopulationWorkflow`}
}
\description{
Set mapping between parameters and their display paths in a workflow
to replace standard display of parameter paths.
}
\seealso{
Other workflow helpers: 
\code{\link{activateWorkflowTasks}()},
\code{\link{addStudyParameters}()},
\code{\link{addXParametersForDemogrpahyPlot}()},
\code{\link{addXParametersForPKParametersPlot}()},
\code{\link{addYParametersForDemogrpahyPlot}()},
\code{\link{addYParametersForPKParametersPlot}()},
\code{\link{checkTaskInputsExist}()},
\code{\link{createWorkflowFromExcelInput}()},
\code{\link{getMeanSensitivityAnalysisResultsFileNames}()},
\code{\link{getOutputPathsInSimulationSet}()},
\code{\link{getPKAnalysisResultsFileNames}()},
\code{\link{getPKParameterGroupsInOutput}()},
\code{\link{getPKParametersInOutput}()},
\code{\link{getPKParametersInSimulationSet}()},
\code{\link{getPopulationPKData}()},
\code{\link{getPopulationPKMetaData}()},
\code{\link{getPopulationSensitivityAnalysisResultsFileNames}()},
\code{\link{getSimulationDescriptor}()},
\code{\link{getSimulationParameterDisplayPaths}()},
\code{\link{getSimulationResultFileNames}()},
\code{\link{getTaskInputs}()},
\code{\link{getWorkflowParameterDisplayPaths}()},
\code{\link{getWorkflowReferencePopulationName}()},
\code{\link{getXParametersForDemogrpahyPlot}()},
\code{\link{getXParametersForPKParametersPlot}()},
\code{\link{getYParametersForDemogrpahyPlot}()},
\code{\link{getYParametersForPKParametersPlot}()},
\code{\link{inactivateWorkflowTasks}()},
\code{\link{setSimulationDescriptor}()},
\code{\link{setWorkflowParameterDisplayPathsFromFile}()},
\code{\link{setXParametersForDemogrpahyPlot}()},
\code{\link{setXParametersForPKParametersPlot}()},
\code{\link{setYParametersForDemogrpahyPlot}()},
\code{\link{setYParametersForPKParametersPlot}()}
}
\concept{workflow helpers}
