% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-calculate-pk-parameters.R
\name{setXParametersForPkParametersPlot}
\alias{setXParametersForPkParametersPlot}
\title{setXParametersForPkParametersPlot}
\usage{
setXParametersForPkParametersPlot(workflow, parameters)
}
\description{
setXParametersForPkParametersPlot
}
