% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulation.R
\name{simulateModel}
\alias{simulateModel}
\title{simulateModel}
\usage{
simulateModel(simFilePath, popDataFilePath = NULL, resultsFilePath)
}
\value{
Simulation results for individual or population
}
\description{
Simulate model, either for an individual or for a given population.  Calculate and save PK parameters as an option.
}
