% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModel}
\alias{simulateModel}
\title{simulateModel}
\usage{
simulateModel(
  structureSet,
  logFolder = getwd(),
  nodeName = NULL,
  showProgress = FALSE
)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object contain paths of files to be used}

\item{logFolder}{folder where the logs are saved}

\item{nodeName}{node name for parallel simulations}
}
\value{
Simulation results for individual or population
}
\description{
Simulate model, either for an individual or for a given population.  Calculate and save PK parameters as an option.
}
