% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModel}
\alias{simulateModel}
\title{simulateModel}
\usage{
simulateModel(structureSet, settings = NULL, logFolder = getwd())
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object contain paths of files to be used}

\item{settings}{list of options to be passed on the function}

\item{logFolder}{folder where the logs are saved}
}
\value{
Simulation results for individual or population
}
\description{
Simulate model, either for an individual or for a given population.
}
\keyword{internal}
