% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModelForPopulation}
\alias{simulateModelForPopulation}
\title{simulateModelForPopulation}
\usage{
simulateModelForPopulation(structureSet, settings = NULL, logFolder = getwd())
}
\arguments{
\item{structureSet}{`SimulationStructure` object contain paths of files to be used}

\item{settings}{list of simulation settings}

\item{logFolder}{folder where the logs are saved}
}
\value{
Simulation results for individual or population
}
\description{
Simulate model for a population.  Run parallel workflow if numberOfCores > 1 AND population size is >1.
}
