% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{simulateModelOnCore}
\alias{simulateModelOnCore}
\title{simulateModelOnCore}
\usage{
simulateModelOnCore(
  structureSet,
  populationFilePath = NULL,
  resultsFilePath,
  debugLogFileName = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logDebugFile()),
  errorLogFileName = file.path(defaultFileNames$workflowFolderPath(),
    defaultFileNames$logErrorFile()),
  nodeName = NULL
)
}
\arguments{
\item{structureSet}{`SimulationStructure` R6 class object contain paths of files to be used}

\item{populationFilePath}{path to population file to be used for simulation}

\item{resultsFilePath}{path to CSV files to which results will be saved,}

\item{debugLogFileName}{path to file where core debug logs are saved}

\item{errorLogFileName}{path to file where core error logs are saved}

\item{nodeName}{node name for parallel simulations}

\item{showProgress}{option to print progress of simulation to console}
}
\description{
Simulate model, either for an individual or for a given population.
}
