% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-set.R
\name{SimulationSet}
\alias{SimulationSet}
\title{SimulationSet}
\description{
R6 class representing Reporting Engine Mean Model Set
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{outputs}}{list of `Output` R6 class objects}

\item{\code{observedDataFile}}{name of csv file to be used for observed data}

\item{\code{observedMetaDataFile}}{name of csv file to be used as dictionary of the observed data}

\item{\code{timeUnit}}{display unit for time variable}

\item{\code{applicationRanges}}{named list of logicals defining which Application ranges are included in
time profiles and residual plots when applicable}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimulationSet$new()}}
\item \href{#method-clone}{\code{SimulationSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SimulationSet` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationSet$new(
  simulationSetName,
  simulationFile,
  outputs = NULL,
  observedDataFile = NULL,
  observedMetaDataFile = NULL,
  timeUnit = "h",
  applicationRanges = list(total = TRUE, firstApplication = TRUE, lastApplication =
    TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationSetName}}{display name of simulation set}

\item{\code{simulationFile}}{names of pkml file to be used for the simulation}

\item{\code{outputs}}{list of `Output` R6 class objects}

\item{\code{observedDataFile}}{name of csv file to be used for observed data}

\item{\code{observedMetaDataFile}}{name of csv file to be used as dictionary of the observed data}

\item{\code{timeUnit}}{display unit for time variable. Default is "h"}

\item{\code{applicationRanges}}{named list of logicals defining which application ranges are included in
time profiles and residual plots when applicable. Names are available in enum `ApplicationRanges`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SimulationSet` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
