% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-task.R
\name{SimulationTask}
\alias{SimulationTask}
\title{SimulationTask}
\description{
R6 class for SimulationTask settings
}
\section{Super class}{
\code{\link[ospsuite.reportingengine:Task]{ospsuite.reportingengine::Task}} -> \code{SimulationTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numberOfCores}}{number of cores for parallel computation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimulationTask$new()}}
\item \href{#method-updateNumberOfCores}{\code{SimulationTask$updateNumberOfCores()}}
\item \href{#method-clone}{\code{SimulationTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="activate">}\href{../../ospsuite.reportingengine/html/Task.html#method-activate}{\code{ospsuite.reportingengine::Task$activate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="inactivate">}\href{../../ospsuite.reportingengine/html/Task.html#method-inactivate}{\code{ospsuite.reportingengine::Task$inactivate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="print">}\href{../../ospsuite.reportingengine/html/Task.html#method-print}{\code{ospsuite.reportingengine::Task$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ospsuite.reportingengine" data-topic="Task" data-id="validateInput">}\href{../../ospsuite.reportingengine/html/Task.html#method-validateInput}{\code{ospsuite.reportingengine::Task$validateInput()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a `SimulationTask` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationTask$new(numberOfCores = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{number of cores for parallel computation}

\item{\code{...}}{parameters inherited from R6 class `Task` object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SimulationTask` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateNumberOfCores"></a>}}
\subsection{Method \code{updateNumberOfCores()}}{
Update the `numberOfCores`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationTask$updateNumberOfCores(numberOfCores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numberOfCores}}{is the number of cores to use for simulation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulationTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
