% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\docType{data}
\name{StandardPlotTasks}
\alias{StandardPlotTasks}
\title{StandardPlotTasks}
\format{
An object of class \code{list} of length 3.
}
\usage{
StandardPlotTasks
}
\description{
Names of plot tasks performed by both `MeanModelWorkflow` and `PopulationWorkflow` objects
}
\keyword{datasets}
