% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-task.R
\docType{data}
\name{StandardSimulationTasks}
\alias{StandardSimulationTasks}
\title{StandardSimulationTasks}
\format{
An object of class \code{list} of length 3.
}
\usage{
StandardSimulationTasks
}
\description{
Names of simulation tasks performed by both `MeanModelWorkflow` and `PopulationWorkflow` objects
}
\examples{

# Lists all available standard simulation task names available in both mean and population workflows
StandardSimulationTasks

}
\seealso{
Other enum helpers: 
\code{\link{AllAvailableTasks}},
\code{\link{ApplicationRanges}},
\code{\link{DataSelectionKeys}},
\code{\link{PopulationWorkflowTypes}},
\code{\link{ResidualScales}},
\code{\link{StandardPlotTasks}},
\code{\link{StatisticsTypes}},
\code{\link{tlfSettingsNames}}
}
\concept{enum helpers}
\keyword{datasets}
