% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-study-design.R
\name{StudyDesignTarget}
\alias{StudyDesignTarget}
\title{StudyDesignTarget}
\description{
StudyDesignTarget
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{path name of study design target}

\item{\code{values}}{values assigned to study design target}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudyDesignTarget$new()}}
\item \href{#method-print}{\code{StudyDesignTarget$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `StudyDesign` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDesignTarget$new(name, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{path name of study design target}

\item{\code{values}}{values assigned to study design target.
`values` must be the same length as source condition expressions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`StudyDesignTarget` class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print study design target features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDesignTarget$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
}
