% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\docType{class}
\name{Task}
\alias{Task}
\title{Task}
\usage{
Task
}
\description{
Task settings for Reporting Engine
}
\section{Fields}{

\describe{
\item{\code{active}}{logical}

\item{\code{input}}{character}

\item{\code{output}}{character}

\item{\code{settings}}{class}
}}

\section{Methods}{

\describe{
\item{new()}{Initilialize Task settings}
\item{activate()}{Set Task as active in workflow}
\item{inactivate()}{Set Task as inactive in workflow}
\item{print()}{Show task settings}
}
}

\keyword{datasets}
