% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-results.R
\name{TaskResults}
\alias{TaskResults}
\title{TaskResults}
\description{
R6 class for TaskResults settings
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{unique identifer of results, helps in deriving final plot name}

\item{\code{sectionId}}{unique identifer of section results, used only by qualification workflows}

\item{\code{plot}}{`ggplot` object corresponding to the figure to be saved}

\item{\code{plotCaption}}{text included into the report explaining the figure}

\item{\code{includePlot}}{logical indicating if the plot should be included in final report}

\item{\code{table}}{data.frame corresponding to the table to be saved}

\item{\code{tableCaption}}{text included into the report explaining the table}

\item{\code{includeTable}}{logical indicating if the table should be included in final report}

\item{\code{textChunk}}{text included into the report explaining the table}

\item{\code{includeTextChunk}}{logical indicating if the text chunk should be included in final report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-saveFigure}{\code{TaskResults$saveFigure()}}
\item \href{#method-saveTable}{\code{TaskResults$saveTable()}}
\item \href{#method-addFigureToReport}{\code{TaskResults$addFigureToReport()}}
\item \href{#method-addTableToReport}{\code{TaskResults$addTableToReport()}}
\item \href{#method-addTextChunkToReport}{\code{TaskResults$addTextChunkToReport()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveFigure"></a>}}
\if{latex}{\out{\hypertarget{method-saveFigure}{}}}
\subsection{Method \code{saveFigure()}}{
Save ggplot figure, available as plot task result, into a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskResults$saveFigure(fileName, logFolder = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{path of file corresponding to the figure to save}

\item{\code{logFolder}}{folder were logs are saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveTable"></a>}}
\if{latex}{\out{\hypertarget{method-saveTable}{}}}
\subsection{Method \code{saveTable()}}{
Save data.frame, available as table task result, into a csv file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskResults$saveTable(fileName, logFolder = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{path of csv file corresponding to the table to save}

\item{\code{logFolder}}{folder were logs are saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addFigureToReport"></a>}}
\if{latex}{\out{\hypertarget{method-addFigureToReport}{}}}
\subsection{Method \code{addFigureToReport()}}{
Write markdown content that adds a figure, available as plot task result, into a markdown report
To be displayed, figure path must be relative to report location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskResults$addFigureToReport(
  reportFile,
  fileRelativePath,
  fileRootDirectory,
  logFolder = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reportFile}}{markdown file in which the figure and its caption should be added}

\item{\code{fileRelativePath}}{figure path relative to `reportFile` location}

\item{\code{fileRootDirectory}}{root/working directory needed by `tracelib` package}

\item{\code{logFolder}}{folder were logs are saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTableToReport"></a>}}
\if{latex}{\out{\hypertarget{method-addTableToReport}{}}}
\subsection{Method \code{addTableToReport()}}{
Write markdown content that adds a data.frame, available as table task result, into a markdown report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskResults$addTableToReport(
  reportFile,
  fileRelativePath,
  fileRootDirectory,
  digits = NULL,
  scientific = NULL,
  logFolder = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reportFile}}{markdown file in which the table and its caption should be added}

\item{\code{fileRelativePath}}{table path relative to `reportFile` location}

\item{\code{fileRootDirectory}}{root/working directory needed by `tracelib` package}

\item{\code{digits}}{number of decimal digits in displayed numbers}

\item{\code{scientific}}{logical defining if displayed numbers use scientific writing}

\item{\code{logFolder}}{folder were logs are saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTextChunkToReport"></a>}}
\if{latex}{\out{\hypertarget{method-addTextChunkToReport}{}}}
\subsection{Method \code{addTextChunkToReport()}}{
Write markdown content that adds text, available as textChunk task result, into a markdown report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskResults$addTextChunkToReport(reportFile, logFolder = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reportFile}}{markdown file in which the text should be added}

\item{\code{logFolder}}{folder were logs are saved}
}
\if{html}{\out{</div>}}
}
}
}
