% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{tlfLinetype}
\alias{tlfLinetype}
\title{tlfLinetype}
\usage{
tlfLinetype(configurationLinetype)
}
\arguments{
\item{configurationLinetype}{Value of the configuration plan plot property}
}
\value{
`linetype` values that follows `tlf` package nomenclature
}
\description{
Translate value of `LineStyle` property from configuration plan into tlf `linetype`
}
