% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{tlfScale}
\alias{tlfScale}
\title{tlfScale}
\usage{
tlfScale(configurationScale)
}
\arguments{
\item{configurationScale}{Value of the configuration plan plot property}
}
\value{
\code{scale} values that follows \code{tlf} package nomenclature
}
\description{
Translate value of \code{Scaling} property from configuration plan into tlf \code{scale}
}
\keyword{internal}
