% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configuration-plan.R
\name{tlfShape}
\alias{tlfShape}
\title{tlfShape}
\usage{
tlfShape(configurationShape)
}
\arguments{
\item{configurationShape}{Value of the configuration plan plot property}
}
\value{
\code{shape} values that follows \code{tlf} package nomenclature
}
\description{
Translate value of \code{Symbol} property from configuration plan into tlf \code{shape}
}
\keyword{internal}
