% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{updatePlotConfigurationTimeTicks}
\alias{updatePlotConfigurationTimeTicks}
\title{updatePlotConfigurationTimeTicks}
\usage{
updatePlotConfigurationTimeTicks(
  data,
  metaData,
  dataMapping,
  plotConfiguration
)
}
\arguments{
\item{data}{data.frame}

\item{metaData}{meta data on \code{data}}

\item{dataMapping}{\code{XYGDataMapping} R6 class object from \code{tlf} library}

\item{plotConfiguration}{\code{PlotConfiguration} R6 class object from \code{tlf} library}
}
\value{
A \code{PlotConfiguration} object
}
\description{
Update time ticks based on selected time unit in \code{PlotConfiguration} objects
}
\keyword{internal}
