% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulation.R
\name{updateSimulationIndividualParameters}
\alias{updateSimulationIndividualParameters}
\title{updateSimulationIndividualParameters}
\usage{
updateSimulationIndividualParameters(simulation, individualParameters = NULL)
}
\description{
Update individual parameters with parameters in individualParameters,
individualParameters is obtained from a population object's getParameterValuesForIndividual() function.
}
