% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{updateSimulationIndividualParameters}
\alias{updateSimulationIndividualParameters}
\title{updateSimulationIndividualParameters}
\usage{
updateSimulationIndividualParameters(simulation, individualParameters = NULL)
}
\arguments{
\item{simulation}{is the simulation object to be updated}

\item{individualParameters}{is an object storing an individual's parameters, obtained from a population object's getParameterValuesForIndividual() function.}
}
\description{
Update simulation with parameters from individualParameters
}
