% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plots.R
\name{updateWatermarkDimensions}
\alias{updateWatermarkDimensions}
\title{updateWatermarkDimensions}
\usage{
updateWatermarkDimensions(plotObject)
}
\arguments{
\item{plotObject}{A `ggplot` object}
}
\value{
A `ggplot` object
}
\description{
Update Watermark dimensions
}
\keyword{internal}
