% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifyAnyPreviousFilesRemoved}
\alias{verifyAnyPreviousFilesRemoved}
\title{verifyAnyPreviousFilesRemoved}
\usage{
verifyAnyPreviousFilesRemoved(anyPreviousPartialResultsRemoved, logFolder)
}
\arguments{
\item{sensitivityRunSuccess}{logical vector indicating success of result file removal}
}
\description{
Check that any existing partial sensitivity results from individual cores have been removed
}
