% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifyPartialResultsExported}
\alias{verifyPartialResultsExported}
\title{verifyPartialResultsExported}
\usage{
verifyPartialResultsExported(partialResultsExported, logFolder)
}
\arguments{
\item{logFolder}{folder where the logs are saved}

\item{sensitivityRunSuccess}{logical vector indicating success of result file export}
}
\description{
Check that results from individual cores have been exported
}
