% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifySimulationRunSuccessful}
\alias{verifySimulationRunSuccessful}
\title{verifySimulationRunSuccessful}
\usage{
verifySimulationRunSuccessful(
  simulationRunSuccess,
  tempPopDataFiles,
  logFolder
)
}
\arguments{
\item{simulationRunSuccess}{logical vector indicating success of simulation run on all cores}

\item{tempPopDataFiles}{name of all temporary population files sent to cores}

\item{logFolder}{folder where the logs are saved}
}
\description{
Check that all cores ran simulation successfully
}
