% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mpi-error-handling.R
\name{verifySimulationRunSuccessful}
\alias{verifySimulationRunSuccessful}
\title{verifySimulationRunSuccessful}
\usage{
verifySimulationRunSuccessful(simulationRunSuccess, logFolder = getwd())
}
\arguments{
\item{simulationRunSuccess}{logical vector indicating success of simulation run on all cores}
}
\description{
Check that all cores ran simulation successfully
}
