% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\description{
R6 class representing Reporting Engine generic Workflow
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reportingEngineInfo}}{R6 class object with relevant information about reporting engine}

\item{\code{simulationSets}}{list of `MeanModelSet` R6 class objects}

\item{\code{observedData}}{list of observed `data` and `metaData`}

\item{\code{reportFolder}}{path where report and logs are saved}

\item{\code{resultsFolder}}{path where results are saved}

\item{\code{figuresFolder}}{path where figure are saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Workflow$new()}}
\item \href{#method-clone}{\code{Workflow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new `Workflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$new(
  simulationSets,
  observedDataFile = NULL,
  observedMetaDataFile = NULL,
  reportFolder = file.path(getwd(), defaultFileNames$reportFolder()),
  resultsFolder = file.path(getwd(), defaultFileNames$resultsFolder()),
  figuresFolder = file.path(getwd(), defaultFileNames$figuresFolder())
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationSets}}{names of pkml files to be used for simulations}

\item{\code{observedDataFile}}{name of csv file to be used for observations}

\item{\code{observedMetaDataFile}}{name of csv file to be used as dictionary for observed data}

\item{\code{reportFolder}}{name of folder where reports and logs are saved}

\item{\code{resultsFolder}}{name of folder where results are saved}

\item{\code{figuresFolder}}{name of folder where figures are saved}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Workflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
