% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{class}
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Workflow
}
\description{
Workflow Task for Reporting Engine
}
\section{Fields}{

\describe{
\item{\code{settings}}{setting object}

\item{\code{models}}{simulation class object}

\item{\code{population}}{population class object}

\item{\code{observedData}}{List of observed data (use Nonmem format ?)}

\item{\code{outputFolder}}{path where output is saved}
}}

\keyword{datasets}
