% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{class}
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\description{
Workflow Task for Reporting Engine
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reportingEngineInfo}}{R6 class object with relevant information about reporting engine}

\item{\code{settings}}{setting object}

\item{\code{models}}{simulation class object}

\item{\code{population}}{population class object}

\item{\code{observedData}}{List of observed data (use Nonmem format ?)}

\item{\code{outputFolder}}{path where output is saved
#' @section Methods:
#' \describe{
\item{setPKParametersCalculationSettings()}{Define PK parameters calculation task settings}
\item{setSensitivityAnalysisSettings()}{Define sensitivity analysis task settings}
\item{setpkParametersPlotSettings()}{Define PK parameters plot settings}
\item{setSensitivityPlotSettings()}{Define sensitivity plot settings}
}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{models}}{simulation class object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Workflow$new()}}
\item \href{#method-setPKParametersCalculationSettings}{\code{Workflow$setPKParametersCalculationSettings()}}
\item \href{#method-setSensitivityAnalysisSettings}{\code{Workflow$setSensitivityAnalysisSettings()}}
\item \href{#method-setpkParametersPlotSettings}{\code{Workflow$setpkParametersPlotSettings()}}
\item \href{#method-setSensitivityPlotSettings}{\code{Workflow$setSensitivityPlotSettings()}}
\item \href{#method-clone}{\code{Workflow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$new(
  simulationFile,
  populationFile,
  observedDataFile = NULL,
  workflowFolder = paste0("Workflow", "_", format(Sys.Date(), "\%Y\%m\%d"), "_",
    format(Sys.time(), "\%H\%M\%S")),
  inputFolder = "Inputs",
  simulationFolder = "Simulations",
  sensitivityFolder = "Sensitivities",
  outputFolder = "Outputs",
  settings = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPKParametersCalculationSettings"></a>}}
\subsection{Method \code{setPKParametersCalculationSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setPKParametersCalculationSettings(message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSensitivityAnalysisSettings"></a>}}
\subsection{Method \code{setSensitivityAnalysisSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setSensitivityAnalysisSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setpkParametersPlotSettings"></a>}}
\subsection{Method \code{setpkParametersPlotSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setpkParametersPlotSettings(message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSensitivityPlotSettings"></a>}}
\subsection{Method \code{setSensitivityPlotSettings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setSensitivityPlotSettings(
  input = NULL,
  output = NULL,
  active = TRUE,
  message = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
