% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\description{
R6 class representing Reporting Engine generic Workflow
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulationStructures}}{`SimulationStructure` R6 class object managing the structure of the workflow output}

\item{\code{workflowFolder}}{path of the folder create by the Workflow}

\item{\code{taskNames}}{Enum of task names}

\item{\code{reportFileName}}{name of the Rmd report file}

\item{\code{createWordReport}}{logical of option for creating Markdwon-Report only but not a Word-Report.}

\item{\code{userDefinedTasks}}{List of user-defined tasks (to update with loadUserDefinedTask)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Workflow$new()}}
\item \href{#method-getAllTasks}{\code{Workflow$getAllTasks()}}
\item \href{#method-getAllPlotTasks}{\code{Workflow$getAllPlotTasks()}}
\item \href{#method-getActiveTasks}{\code{Workflow$getActiveTasks()}}
\item \href{#method-getInactiveTasks}{\code{Workflow$getInactiveTasks()}}
\item \href{#method-activateTasks}{\code{Workflow$activateTasks()}}
\item \href{#method-inactivateTasks}{\code{Workflow$inactivateTasks()}}
\item \href{#method-printReportingEngineInfo}{\code{Workflow$printReportingEngineInfo()}}
\item \href{#method-getWatermark}{\code{Workflow$getWatermark()}}
\item \href{#method-setWatermark}{\code{Workflow$setWatermark()}}
\item \href{#method-setParameterDisplayPaths}{\code{Workflow$setParameterDisplayPaths()}}
\item \href{#method-getParameterDisplayPaths}{\code{Workflow$getParameterDisplayPaths()}}
\item \href{#method-setSimulationDescriptor}{\code{Workflow$setSimulationDescriptor()}}
\item \href{#method-getSimulationDescriptor}{\code{Workflow$getSimulationDescriptor()}}
\item \href{#method-print}{\code{Workflow$print()}}
\item \href{#method-clone}{\code{Workflow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Workflow` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$new(
  simulationSets,
  workflowFolder,
  createWordReport = TRUE,
  watermark = NULL,
  simulationSetDescriptor = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulationSets}}{list of `SimulationSet` R6 class objects}

\item{\code{workflowFolder}}{path of the output folder created or used by the Workflow.}

\item{\code{createWordReport}}{logical of option for creating Markdown-Report only but not a Word-Report.}

\item{\code{watermark}}{displayed watermark in figures background}

\item{\code{simulationSetDescriptor}}{character Descriptor of simulation sets indicated in reports}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Workflow` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllTasks"></a>}}
\if{latex}{\out{\hypertarget{method-getAllTasks}{}}}
\subsection{Method \code{getAllTasks()}}{
Get a vector with all the names of the tasks within the `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getAllTasks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of `Task` names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAllPlotTasks"></a>}}
\if{latex}{\out{\hypertarget{method-getAllPlotTasks}{}}}
\subsection{Method \code{getAllPlotTasks()}}{
Get a vector with all the names of the plot tasks within the `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getAllPlotTasks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of `PlotTask` names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getActiveTasks"></a>}}
\if{latex}{\out{\hypertarget{method-getActiveTasks}{}}}
\subsection{Method \code{getActiveTasks()}}{
Get a vector with all the names of active tasks within the `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getActiveTasks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of active `Task` names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInactiveTasks"></a>}}
\if{latex}{\out{\hypertarget{method-getInactiveTasks}{}}}
\subsection{Method \code{getInactiveTasks()}}{
Get a vector with all the names of inactive tasks within the `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getInactiveTasks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of inactive `Task` names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-activateTasks"></a>}}
\if{latex}{\out{\hypertarget{method-activateTasks}{}}}
\subsection{Method \code{activateTasks()}}{
Activates a series of `Tasks` from current `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$activateTasks(tasks = self$getAllTasks())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tasks}}{names of the worklfow tasks to activate.
Default activates all tasks of the workflow using workflow method `workflow$getAllTasks()`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of inactive `Task` names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inactivateTasks"></a>}}
\if{latex}{\out{\hypertarget{method-inactivateTasks}{}}}
\subsection{Method \code{inactivateTasks()}}{
Inactivates a series of `Tasks` from current `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$inactivateTasks(tasks = self$getAllTasks())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tasks}}{names of the worklfow tasks to inactivate.
Default inactivates all tasks of the workflow using workflow method `workflow$getAllTasks()`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of inactive `Task` names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-printReportingEngineInfo"></a>}}
\if{latex}{\out{\hypertarget{method-printReportingEngineInfo}{}}}
\subsection{Method \code{printReportingEngineInfo()}}{
Print reporting engine information obtained from initiliazing a `Workflow`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$printReportingEngineInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWatermark"></a>}}
\if{latex}{\out{\hypertarget{method-getWatermark}{}}}
\subsection{Method \code{getWatermark()}}{
Get the current watermark to be reprted on figures background
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getWatermark()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWatermark"></a>}}
\if{latex}{\out{\hypertarget{method-setWatermark}{}}}
\subsection{Method \code{setWatermark()}}{
Set the watermark to be reported on figure background.
The default value `NULL` leads to check if the computer has a validated environment.
If the environment is validated, no watermark is reported on the background.
If the environment is NOT validated, \code{workflowWatermarkMessage} is reported on the background.
Any user-defined text will overwrite this default feature and be reported on the figure background.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setWatermark(watermark)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{watermark}}{text to be reported on figures background.
Default value is `NULL`, which leads to check if the computer has a validated environment.
If the environment is validated, no watermark is reported on the background.
If the environment is NOT validated, \code{workflowWatermarkMessage} is reported on the background.
Any user-defined text will overwrite this default feature and be reported on the figure background.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setParameterDisplayPaths"></a>}}
\if{latex}{\out{\hypertarget{method-setParameterDisplayPaths}{}}}
\subsection{Method \code{setParameterDisplayPaths()}}{
Set mapping between parameters and their display paths in workflow
to replace standard display of parameter paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setParameterDisplayPaths(parameterDisplayPaths)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameterDisplayPaths}}{data.frame mapping Parameters with their display paths
Variables of the data.frame should include `parameter` and `displayPath`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParameterDisplayPaths"></a>}}
\if{latex}{\out{\hypertarget{method-getParameterDisplayPaths}{}}}
\subsection{Method \code{getParameterDisplayPaths()}}{
Get mapping between parameters and their display paths in workflow
to replace standard display of parameter paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getParameterDisplayPaths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with `parameter` and `displayPath` variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSimulationDescriptor"></a>}}
\if{latex}{\out{\hypertarget{method-setSimulationDescriptor}{}}}
\subsection{Method \code{setSimulationDescriptor()}}{
Set descriptor of simulation sets indicated in reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$setSimulationDescriptor(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{character describing what simulation sets refer to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSimulationDescriptor"></a>}}
\if{latex}{\out{\hypertarget{method-getSimulationDescriptor}{}}}
\subsection{Method \code{getSimulationDescriptor()}}{
Get descriptor of simulation sets indicated in reports
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$getSimulationDescriptor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character Descriptor of simulation sets
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print workflow list of tasks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Task list information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Workflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
